% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{AddAzimuthResults}
\alias{AddAzimuthResults}
\title{Add Azimuth Results}
\usage{
AddAzimuthResults(object = NULL, filename)
}
\arguments{
\item{object}{A \code{\link[SeuratObject]{Seurat}} object}

\item{filename}{Path to Azimuth mapping scores file}
}
\value{
\code{object} with Azimuth results added
}
\description{
Add mapping and prediction scores, UMAP embeddings, and imputed assay (if
available)
from Azimuth to an existing or new \code{\link[SeuratObject]{Seurat}} object
}
\examples{
\dontrun{
object <- AddAzimuthResults(object, filename = "azimuth_results.Rds")
}

}
