% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{AddAzimuthScores}
\alias{AddAzimuthScores}
\title{Add Azimuth Scores}
\usage{
AddAzimuthScores(object, filename)
}
\arguments{
\item{object}{A \code{\link[SeuratObject]{Seurat}} object}

\item{filename}{Path to Azimuth mapping scores file}
}
\value{
\code{object} with the mapping scores added
}
\description{
Add mapping and prediction scores from Azimuth to a
\code{\link[SeuratObject]{Seurat}} object
}
\examples{
\dontrun{
object <- AddAzimuthScores(object, filename = "azimuth_pred.tsv")
}

}
