% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seurat.R
\name{addImputedScore}
\alias{addImputedScore}
\title{Impute gene expression and fill in Seurat object.}
\usage{
addImputedScore(object, genes.use = NULL, genes.fit = NULL,
  scheme = "lasso", s.use = 20, do.print = FALSE, gram = TRUE)
}
\arguments{
\item{object}{Seurat object}

\item{genes.use}{A vector of genes (predictors) that can be used for
building the imputation models.}

\item{genes.fit}{A vector of response genes to be imputed.}

\item{scheme}{Imputation strategies: "lasso", "plsr", "tlasso" (tilling lasso).}

\item{s.use}{Maximum number of steps taken by the algorithm (lower values
indicate a greater degree of smoothing). Only used for "lasso" scheme.}

\item{do.print}{Print progress (output the name of each gene after it has
been imputed).}

\item{gram}{The use.gram argument passed to \link{lars} of lars package. See
\link{lars} for parameter details. Only used for "lasso" scheme.}
}
\value{
Returns a Seurat object where the imputed values have been added to
\code{object@data}. If \code{genes.fit} have already existed in object,
their values are imputed again via the selected \code{scheme} and refilled;
otherwise they are imputed and added to \code{object@data}.
}
\description{
Impute the expression values of import genes, e.g. landmark,
genes with other genes as predictors via L1-constrained linear models, i.e.
lasso (default), or partial least squares regression (plsr), or tilling LASSO #' (tlasso).
}

