% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{AddSamples}
\alias{AddSamples}
\title{Add samples into existing Seurat object.}
\usage{
AddSamples(object, new.data, min.genes = 2500, names.field = 1,
  names.delim = "_")
}
\arguments{
\item{object}{Seurat object}

\item{new.data}{Data to be added to the object}

\item{min.genes}{Gene threshold for new data. Discard all cells with fewer genes detected}

\item{names.field}{For the initial identity class for each cell, choose this
field from the cell's column name}

\item{names.delim}{For the initial identity class for each cell, choose this
delimiter from the cell's column name}
}
\description{
Add samples into existing Seurat object.
}

