% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seurat.R
\name{addSmoothedScore}
\alias{addSmoothedScore}
\title{Calculate smoothed expression values}
\usage{
addSmoothedScore(object, genes.fit = NULL, dim.1 = 1, dim.2 = 2,
  reduction.use = "tSNE", k = 30, do.log = FALSE, do.print = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{genes.fit}{Genes to calculate smoothed values for}

\item{dim.1}{Dimension for x-axis (default 1)}

\item{dim.2}{Dimension for y-axis (default 2)}

\item{reduction.use}{Which dimensionality reduction to use. Default is
"tsne", can also be "pca", or "ica", assuming these are precomputed.}

\item{do.print}{Print progress (output the name of each gene after it has
been imputed).}

\item{inheritParams}{Seurat object}
}
\description{
Smooths expression values across the k-nearest neighbors based on dimensional reduction
}

