% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{AddTCC}
\alias{AddTCC}
\title{Add in raw TCC data to Seurat object}
\usage{
AddTCC(object, tcc.counts, cell.names, ec.map, gene.map,
  dataset = "hsapiens_gene_ensembl", tcc.assay.name = "tcc",
  intersect.cells = FALSE, min.ec.filter = 0, verbose = TRUE)
}
\arguments{
\item{object}{Seurat object}

\item{tcc.counts}{Path to TCC matrix -- should be provided as a tsv file in
triplet sparse matrix format (i, j, x) (matrix.tsv from Kallisto)}

\item{cell.names}{File with cell names (matrix.cells from Kallisto)}

\item{ec.map}{File mapping ECs to transcript ids (matrix.ec from Kallisto)}

\item{gene.map}{File mapping ENST to ENSG names}

\item{tcc.assay.name}{Name for tcc Assay object}

\item{intersect.cells}{Return Seurat object with only cells with that also have TCC
quantifications.}

\item{min.ec.filter}{Only keep ECs with at least this many counts across all
cells}

\item{verbose}{prints output/progress bars}
}
\description{
Add in raw TCC data to Seurat object
}
