% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{AnnotateAnchors}
\alias{AnnotateAnchors}
\title{Add info to anchor matrix}
\usage{
AnnotateAnchors(object, annotation = NULL)
}
\arguments{
\item{object}{An \code{\link{AnchorSet}} object}

\item{annotation}{Name in metadata to annotate anchors with}

\item{object.list}{List of objects using in FindIntegrationAnchors call}
}
\value{
Returns the anchor dataframe with additional columns for annotation
metadata
}
\description{
Add info to anchor matrix
}
