% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{AnnoyNN}
\alias{AnnoyNN}
\title{Run Annoy approximate nearest neighbors algorithm}
\usage{
AnnoyNN(
  data,
  query = data,
  metric = "euclidean",
  n.trees = 50,
  k,
  search.k = -1,
  include.distance = TRUE,
  nn.idx = NULL,
  return.annoy_index = FALSE
)
}
\arguments{
\item{data}{Data to build the index with}

\item{query}{A set of data to be queried against data}

\item{metric}{Distance metric; can be one of "euclidean", "cosine", "manhattan",
"hamming"}

\item{n.trees}{More trees gives higher precision when querying}

\item{k}{Number of neighbors}

\item{search.k}{During the query it will inspect up to search_k nodes which
gives you a run-time tradeoff between better accuracy and speed.}

\item{include.distance}{Include the corresponding distances}

\item{nn.idx}{the precomputed AnnoyBuildIndex index}

\item{return.annoy_index}{Return the index built by AnnoyBuildIndex?}
}
\description{
Run Annoy approximate nearest neighbors algorithm
}
\details{
if `return.annoy_index = TRUE`, the returned object from AnnoySearch also contains
a `metric` field (corresponding to the `metric` argument to this function), a `ndim` 
field (the number of dimensions), and an `annoy_index` field which 
is the index object.
}
