% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{as.CellDataSet}
\alias{as.CellDataSet}
\alias{as.CellDataSet.Seurat}
\title{Convert objects to CellDataSet objects}
\usage{
as.CellDataSet(x, ...)

\method{as.CellDataSet}{Seurat}(x, assay = NULL, dim.res = NULL, ...)
}
\arguments{
\item{x}{An object to convert to class \code{CellDataSet}}

\item{assay}{Assay to convert}
}
\description{
Convert objects to CellDataSet objects
}
