% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{as.Graph}
\alias{as.Graph}
\alias{as.Graph.Matrix}
\alias{as.Graph.matrix}
\title{Convert a matrix (or Matrix) to the Graph class.}
\usage{
as.Graph(x, ...)

\method{as.Graph}{Matrix}(x, ...)

\method{as.Graph}{matrix}(x, ...)
}
\arguments{
\item{x}{The matrix to convert}

\item{...}{Arguments passed to other methods (ignored for now)}
}
\description{
Convert a matrix (or Matrix) to the Graph class.
}
