% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{as.list.SeuratCommand}
\alias{as.list.SeuratCommand}
\title{Coerce a SeuratCommand to a list}
\usage{
\method{as.list}{SeuratCommand}(x, complete = FALSE, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{complete}{Include slots besides just parameters (eg. call string, name, timestamp)}

\item{...}{objects, possibly named.}
}
\value{
A list with the parameters and, if \code{complete = TRUE}, the call string, name, and timestamp
}
\description{
Coerce a SeuratCommand to a list
}
