% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{as.Neighbor}
\alias{as.Neighbor}
\alias{as.Neighbor.Graph}
\title{Convert objects to Neighbor ojbects}
\usage{
as.Neighbor(x, ...)

\method{as.Neighbor}{Graph}(x, ...)
}
\arguments{
\item{x}{An object to convert to \code{Neighbor}}

\item{...}{Arguments passed to other methods}
}
\description{
Convert objects to Neighbor ojbects
}
