% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{as.SingleCellExperiment}
\alias{as.SingleCellExperiment}
\alias{as.SingleCellExperiment.Seurat}
\title{Convert objects to SingleCellExperiment objects}
\usage{
as.SingleCellExperiment(x, ...)

\method{as.SingleCellExperiment}{Seurat}(x, assay = NULL, ...)
}
\arguments{
\item{x}{An object to convert to class \code{SingleCellExperiment}}

\item{...}{Arguments passed to other methods}

\item{assay}{Assays to convert}
}
\description{
Convert objects to SingleCellExperiment objects
}
\concept{objects}
