% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{Assay-class}
\alias{Assay-class}
\alias{Assay}
\title{The Assay Class}
\description{
The Assay object is the basic unit of Seurat; each Assay stores raw, normalized, and scaled data
as well as cluster information, variable features, and any other assay-specific metadata.
Assays should contain single cell expression data such as RNA-seq, protein, or imputed expression
data.
}
\section{Slots}{

\describe{
\item{\code{counts}}{Unnormalized data such as raw counts or TPMs}

\item{\code{data}}{Normalized expression data}

\item{\code{scale.data}}{Scaled expression data}

\item{\code{key}}{Key for the Assay}

\item{\code{assay.orig}}{Original assay that this assay is based off of. Used to track
assay provenence}

\item{\code{var.features}}{Vector of features exhibiting high variance across single cells}

\item{\code{meta.features}}{Feature-level metadata}

\item{\code{misc}}{Utility slot for storing additional data associated with the assay}
}}

