% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{Assay}
\alias{Assay}
\title{The Assay Class}
\description{
The Assay object is the basic unit of Seurat; each Assay stores raw, normalized, and scaled data
as well as cluster information, variable features, and any other assay-specific metadata.
Assays should contain single cell expression data such as RNA-seq, protein, or imputed expression data.
}
\section{Slots}{

\describe{
\item{\code{raw.data}}{Raw expression data}

\item{\code{data}}{Normalized expression data}

\item{\code{scale.data}}{Scaled expression data}

\item{\code{ident}}{...}

\item{\code{key}}{...}

\item{\code{hvf.info}}{Output of mean/variability analysis for all features}

\item{\code{var.features}}{Vector of features exhibiting high variance across single cells}

\item{\code{meta.data}}{Contains meta-information about each cell, starting with number of genes detected (nGene)
and the original identity class (orig.ident); more information is added using \code{AddMetaData}}

\item{\code{meta.features}}{Feature-level meteadata}

\item{\code{cluster.tree}}{...}

\item{\code{kmeans}}{...}
}}

