% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_validation.R
\name{AssessSplit}
\alias{AssessSplit}
\title{Assess Cluster Split}
\usage{
AssessSplit(object, node)
}
\arguments{
\item{object}{Seurat object}

\item{node}{Node in the cluster tree in question}
}
\value{
Returns the Out of Bag error for a random forest classifier 
trained on the split from the given node
}
\description{
Method for determining confidence in specific bifurcations in
the cluster tree. Use the Out of Bag (OOB) error of a random
forest classifier to judge confidence.
}

