% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{AutoPointSize}
\alias{AutoPointSize}
\title{Automagically calculate a point size for ggplot2-based scatter plots}
\usage{
AutoPointSize(data, raster = NULL)
}
\arguments{
\item{data}{A data frame being passed to ggplot2}

\item{raster}{If TRUE, point size is set to 1}
}
\value{
The "optimal" point size for visualizing these data
}
\description{
It happens to look good
}
\examples{
df <- data.frame(x = rnorm(n = 10000), y = runif(n = 10000))
AutoPointSize(data = df)

}
\concept{visualization}
