% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{AverageExpression}
\alias{AverageExpression}
\title{Averaged gene expression by identity class}
\usage{
AverageExpression(object, features = NULL, assays = NULL,
  group.by = NULL, slot = "data", verbose = TRUE, ...)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Features to analyze; default is all features}

\item{assays}{A vector of assays to analyze, default is the default assay}

\item{group.by}{Meta data to group by, default is identity classes}

\item{slot}{Slot to pull data from}

\item{verbose}{Show progress updates}

\item{...}{Ignored for now}
}
\value{
Returns a list of matrices with features as rows and identity classes as columns
}
\description{
Returns gene expression for an 'average' single cell in each identity class
}
\details{
Output is in log-space when \code{return.seurat = TRUE}, otherwise it's in non-log space.
Averaging is done in non-log space.
}
\examples{
head(AverageExpression(object = pbmc_small))

}
