% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{AveragePCA}
\alias{AveragePCA}
\title{Average PCA scores by identity class}
\usage{
AveragePCA(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{
Returns a matrix with genes as rows, identity classes as columns
}
\description{
Returns the PCA scores for an 'average' single cell in each identity class
}
\examples{

head(AveragePCA(object = pbmc_small))

}
