% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_used_yet.R
\name{BatchGene}
\alias{BatchGene}
\title{Identify potential genes associated with batch effects}
\usage{
BatchGene(object, idents.use, genes.use = NULL, auc.cutoff = 0.6,
  thresh.use = 0)
}
\arguments{
\item{object}{Seurat object}

\item{idents.use}{Batch names to test}

\item{genes.use}{Gene list to test}

\item{auc.cutoff}{Minimum AUC needed to qualify as a 'batch gene'}

\item{thresh.use}{Limit testing to genes which show, on average, at least
X-fold difference (log-scale) in any one batch}
}
\value{
Returns a list of genes that are strongly correlated with batch.
}
\description{
Test for genes whose expression value is strongly predictive of batch (based
on ROC classification). Important note: Assumes that the 'batch' of each
cell is assigned to the cell's identity class (will be improved in a future
release)
}
