% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utilities.R
\name{BlackAndWhite}
\alias{BlackAndWhite}
\title{A black and white color palette}
\usage{
BlackAndWhite(...)
}
\arguments{
\item{...}{Extra parameters to CustomPalette}
}
\value{
A color palette
}
\description{
A black and white color palette
}
\seealso{
\code{CustomPalette}
}
