% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{BridgeReferenceSet-class}
\alias{BridgeReferenceSet-class}
\alias{BridgeReferenceSet}
\title{The BridgeReferenceSet Class
The BridgeReferenceSet is an output from PrepareBridgeReference}
\description{
The BridgeReferenceSet Class
The BridgeReferenceSet is an output from PrepareBridgeReference
}
\section{Slots}{

\describe{
\item{\code{bridge}}{The multi-omic object}

\item{\code{reference}}{The Reference object only containing bridge representation assay}

\item{\code{params}}{A list of parameters used in the PrepareBridgeReference}

\item{\code{command}}{Store log of parameters that were used}
}}

\concept{objects}
