% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seurat.R
\name{build.SNN}
\alias{build.SNN}
\title{SNN Graph Construction}
\usage{
build.SNN(object, genes.use = NULL, pc.use = NULL, k_param = 10,
  k_scale = 10, plot.SNN = FALSE, prune.SNN = FALSE, do_sparse = FALSE,
  update = 0.25)
}
\arguments{
\item{object}{Seurat object}

\item{genes.use}{Gene expression data}

\item{pc.use}{Which PCs to use for construction of the SNN graph}

\item{k_param}{Defines k for the k-nearest neighbor algorithm}

\item{k_scale}{granularity option for k_param}

\item{plot.SNN}{Plot the SNN graph}

\item{prune.SNN}{Prune the SNN graph}

\item{do_sparse}{Whether to compute and return the SNN graph as a sparse matrix or not}

\item{update}{Adjust how verbose the output is}
}
\value{
Returns the object with object@snn.k and object@snn.sparse or object@snn.dense filled
}
\description{
Construct a Shared Nearest Neighbor (SNN) Graph for a given
dataset.
}

