% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{BuildRFClassifier}
\alias{BuildRFClassifier}
\title{Build Random Forest Classifier}
\usage{
BuildRFClassifier(object, training.genes = NULL, training.classes = NULL,
  ...)
}
\arguments{
\item{object}{Seurat object on which to train the classifier}

\item{training.genes}{Vector of genes to build the classifier on}

\item{training.classes}{Vector of classes to build the classifier on}

\item{...}{additional parameters passed to ranger}
}
\value{
Returns a list with the random forest classifier object as the first element,
the original classes as the second, and the classes used by the classifier as the third
}
\description{
Train the random forest classifier
}

