% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seuratFxns.R
\name{calc_cell_centroid}
\alias{calc_cell_centroid}
\title{Calculate spatial centroid of given cell.}
\usage{
calc_cell_centroid(probs, bins = 64)
}
\arguments{
\item{probs}{A vector of probability of cell originating from all bins. Length
is equal to \code{bins}.}

\item{bins}{A scalar. Number of bins (default: 64).}
}
\description{
For given one cell, calculate spatial centroid, i.e. center of mass, of the
spatial probability map. [helper function]
}

