% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seuratFxns.R
\name{calc_dist}
\alias{calc_dist}
\title{Distance Calculation}
\usage{
calc_dist(x, v, struct = "fn", method = c("euclidean"))
}
\arguments{
\item{x}{A matrix or data.frame.}

\item{v}{A query vector.}

\item{struct}{A character indicating structure of the \code{x}. String
"\code{fn}" means feature by sample; "\code{nf}" means sample by feature.
Default: '\code{fn}'.}

\item{method}{Distance metric (Default: euclidean).}
}
\value{
A vector of distance or pairwise distance matrix.
}
\description{
Compute distance between query vector and all rows (or columns)
of data matrix in 1 v.s. N fashion. Alternatively pairwise compute distance
among rows (or columns).
}
\note{
Default structure of matrix is feature by sample, which is opposite to
conventions. But this is a good compromise so that friendly to vectorized
computation and smaller memory usage.
}

