% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing5.R
\name{CalcDispersion}
\alias{CalcDispersion}
\title{Calculate dispersion of features}
\usage{
CalcDispersion(
  object,
  mean.function = FastExpMean,
  dispersion.function = FastLogVMR,
  num.bin = 20,
  binning.method = "equal_width",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Data matrix}

\item{mean.function}{Function to calculate mean}

\item{dispersion.function}{Function to calculate dispersion}

\item{num.bin}{Number of bins to use}

\item{binning.method}{Method to use for binning. Options are 'equal_width' or 'equal_frequency'}

\item{verbose}{Display progress}
}
\description{
Calculate dispersion of features
}
\keyword{internal}
