% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{CalcProjectedVar}
\alias{CalcProjectedVar}
\title{Calculate percent variance explained}
\usage{
CalcProjectedVar(object, reduction.type = "pca", dims.use, genes.use)
}
\arguments{
\item{object}{Seurat object}

\item{reduction.type}{Name of the reduction to use for the projection}

\item{dims.use}{Vector of dimensions to project onto (default is the 1:number stored for given technique)}

\item{genes.use}{vector of genes to use in calculation}
}
\description{
Projects dataset onto the orthonormal space defined by some dimensional reduction technique 
(e.g. PCA, CCA) and calculates the percent of the variance in gene expression explained by each cell
in that lower dimensional space.
}
