% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{CalculateSaturation}
\alias{CalculateSaturation}
\title{Good-Turing QC}
\usage{
CalculateSaturation(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{
A density plot with the expected unobserved expression for all cells in the dataset
}
\description{
Good-Turing QC
}
