% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_generics.R, R/preprocessing.R
\name{CalcUMI}
\alias{CalcUMI}
\alias{CalcUMI.loom}
\title{Calculate nUMI and nGene}
\usage{
CalcUMI(object, ...)

\method{CalcUMI}{loom}(object, cells.use = NULL, is.expr = 0,
  chunk.size = 1000, display.progress = TRUE, overwrite = FALSE)
}
\arguments{
\item{object}{An object}

\item{cells.use}{Optional numeric vector of cells to include}

\item{is.expr}{Expression threshold for 'detected' gene. For most datasets, particularly UMI
datasets, will be set to 0 (default). If not, when initializing, this should be set to a level
based on pre-normalized counts (i.e. require at least 5 counts to be treated as expresesd) All
values less than this will be set to 0 (though maintained in object@raw.data).}

\item{chunk.size}{Number of cells to chunk over}

\item{overwrite}{Overwrite existing datasets?}

\item{display.progres}{Display a progress bar?}
}
\value{
Stores nUMI in \code{col_attrs/nUMI}; stores nGene in \code{col_attrs/nGene}
}
\description{
Calculate nUMI and nGene
}
\section{Methods (by class)}{
\itemize{
\item \code{loom}: Calculate nUMI and nGene in loom objects
}}

