% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CaseMatch}
\alias{CaseMatch}
\title{Match the case of character vectors}
\usage{
CaseMatch(search, match)
}
\arguments{
\item{search}{A vector of search terms}

\item{match}{A vector of characters whose case should be matched}
}
\value{
Values from search present in match with the case of match
}
\description{
Match the case of character vectors
}
\examples{
data("pbmc_small")
cd_genes <- c('Cd79b', 'Cd19', 'Cd200')
CaseMatch(search = cd_genes, match = rownames(x = pbmc_small))

}
\concept{utilities}
