% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc.genes.updated.2019}
\alias{cc.genes.updated.2019}
\title{Cell cycle genes: 2019 update}
\format{A list of two vectors
\describe{
  \item{s.genes}{Genes associated with S-phase}
  \item{g2m.genes}{Genes associated with G2M-phase}
}}
\source{
\url{http://science.sciencemag.org/content/352/6282/189}
}
\usage{
cc.genes.updated.2019
}
\description{
A list of genes used in cell-cycle regression, updated with 2019 symbols
}
\section{Updated symbols}{

The following symbols were updated from \code{\link{cc.genes}}
\describe{
  \item{s.genes}{
    \itemize{
      \item \emph{MLF1IP}: \emph{CENPU}
    }
  }
  \item{g2m.genes}{
    \itemize{
      \item \emph{FAM64A}: \emph{PICALM}
      \item \emph{HN1}: \emph{JPT}
    }
  }
}
}

\seealso{
\code{\link{cc.genes}}
}
\keyword{datasets}
