% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr.R
\name{CellCycleEnrichment}
\alias{CellCycleEnrichment}
\title{Enrich for cell cycle phases}
\usage{
CellCycleEnrichment(object, g2m.genes, s.genes, set.ident = FALSE)
}
\arguments{
\item{object}{A Seurat object}

\item{g2m.genes}{A vector of genes associated with G2M phase}

\item{s.genes}{A vector of genes associated with S phases}

\item{set.ident}{If true, sets identity to phase assignments
Stashes old identities in 'old.ident'}
}
\value{
A Seurat object with the following columns added to object@meta.data

\list{\code{S.Score}} S phase score

\list{\code{G2M.Score}} G2M phase score

\list{\code{Phase}} Cell-cycle phase assignment
}
\description{
Enrich for cell cycle phases
}
\seealso{
\link{\code{AddEnrichScore}}
}
