% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{CellPlot}
\alias{CellPlot}
\title{Cell-cell scatter plot}
\usage{
CellPlot(object, cell1, cell2, features.use = NULL, cols.use = NULL,
  pt.size = 1, ...)
}
\arguments{
\item{object}{Seurat object}

\item{cell1}{Cell 1 name}

\item{cell2}{Cell 2 name}

\item{features.use}{Genes to plot (default, all genes)}

\item{cols.use}{Colors to use for identity class plotting.}

\item{pt.size}{Size of the points on the plot}

\item{\dots}{Extra parameters passed to kde2d}
}
\value{
A ggplot object
}
\description{
Creates a plot of scatter plot of genes across two single cells. Pearson
correlation between the two cells is displayed above the plot.
}
\examples{
CellPlot(object = pbmc_small, cell1 = 'ATAGGAGAAACAGA', cell2 = 'CATCAGGATGCACA')

}
\seealso{
\code{\link{MASS::kde2d}}
}
