% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Cells}
\alias{Cells}
\alias{Cells.default}
\title{Get cells present in an object}
\usage{
Cells(object, ...)

\method{Cells}{default}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
A vector of cell names
}
\description{
Get cells present in an object
}
\examples{
Cells(object = pbmc_small)

}
