% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CellsByIdentities}
\alias{CellsByIdentities}
\title{Get cell names grouped by identity class}
\usage{
CellsByIdentities(object, idents = NULL, cells = NULL)
}
\arguments{
\item{object}{A Seurat object}

\item{idents}{A vector of identity class levels to limit resulting list to;
defaults to all identity class levels}

\item{cells}{A vector of cells to grouping to}
}
\value{
A named list where names are identity classes and values are vectors
of cells beloning to that class
}
\description{
Get cell names grouped by identity class
}
\examples{
CellsByIdentities(object = pbmc_small)

}
