% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CellsByImage}
\alias{CellsByImage}
\title{Get a vector of cell names associated with an image (or set of images)}
\usage{
CellsByImage(object, images = NULL, unlist = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{images}{Vector of image names}

\item{unlist}{Return as a single vector of cell names as opposed to a list,
named by image name.}
}
\value{
A vector of cell names
}
\description{
Get a vector of cell names associated with an image (or set of images)
}
\examples{
\dontrun{
CellsByImage(object = object, images = "slice1")
}

}
