% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CheckWorkflow}
\alias{CheckWorkflow}
\title{Checks if a workflow is defined for a Seurat object}
\usage{
CheckWorkflow(object, workflow.name)
}
\arguments{
\item{object}{Seurat object}

\item{workflow.name}{Workflow name, should already be initialized using InitializeWorkflow#'}
}
\value{
TRUE if workflow is defined. STOP otherwise
}
\description{
Checks if a workflow is defined for a Seurat object
}
\examples{
\dontrun{
CheckWorkflow(pbmc_small, "cluster")
}

}
