% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{CheckWorkflowUpdate}
\alias{CheckWorkflowUpdate}
\title{Check if workflow command needs update}
\usage{
CheckWorkflowUpdate(object, workflow.name, command.name)
}
\arguments{
\item{object}{Seurat object}

\item{workflow.name}{Workflow name, should already be initialized using InitializeWorkflow}

\item{command.name}{Name of the command to check}
}
\value{
Returns TRUE if the dependency has changed (or has not been run), and an update is needed. FALSE otherwise
}
\description{
Compares the stored timestamp with the most recently recorded timestamp to see if a dependency has been updated
}
\examples{
CheckWorkflowUpdate(object = pbmc_small,workflow.name = "cluster", command.name = "ScaleData")
#'
}
