% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{classifyCells}
\alias{classifyCells}
\title{Sample classification from MULTI-seq}
\usage{
classifyCells(data.n, q)
}
\arguments{
\item{q}{Scale the data. Default is 1e4}

\item{data}{Data frame with the raw count data (cell x tags)}
}
\value{
Returns a named vector with demultiplexed identities
}
\description{
Identify singlets, doublets and negative cells from multiplexing experiments.
Written by Chris McGinnis from Gartner Lab, UCSF
}
\examples{
demux_result <- classifyCells(data = counts_data,q = 0.7)

}
