% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seurat.R
\name{cluster.alpha}
\alias{cluster.alpha}
\title{Probability of detection by identity class}
\usage{
cluster.alpha(object, thresh.min = 0)
}
\arguments{
\item{object}{Seurat object}

\item{thresh.min}{Minimum threshold to define 'detected' (log-scale)}
}
\value{
Returns a matrix with genes as rows, identity classes as columns.
}
\description{
For each gene, calculates the probability of detection for each identity
class.
}

