% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{ClusterAlpha}
\alias{ClusterAlpha}
\title{Probability of detection by identity class}
\usage{
ClusterAlpha(object, thresh.min = 0)
}
\arguments{
\item{object}{Seurat object}

\item{thresh.min}{Minimum threshold to define 'detected' (log-scale)}
}
\value{
Returns a matrix with genes as rows, identity classes as columns.
}
\description{
For each gene, calculates the probability of detection for each identity
class.
}
