% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Command}
\alias{Command}
\alias{Command.Seurat}
\title{Get SeuratCommands}
\usage{
Command(object, command, ..., value)

\method{Command}{Seurat}(object, command, value = NULL)
}
\arguments{
\item{object}{An object}

\item{command}{Name of the command to pull}

\item{value}{Name of the parameter to pull the value for}
}
\value{
Either the SeuratCommand object or the paramter value
}
\description{
Pull information on previously run commands in the Seurat object.
}
\section{Methods (by class)}{
\itemize{
\item \code{Seurat}: Get the SeuratCommands
}}

