% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{ComputeRMetric}
\alias{ComputeRMetric}
\title{Computes the metric at a given r (radius) value and stores in meta.features}
\usage{
ComputeRMetric(object, assay = NULL, r.metric = 20)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Assay on which the mark variogram computations were made}

\item{r.metric}{r value at which to report the "trans" value of the mark 
variogram}
}
\value{
Returns a Seurat object with the "r.metric" column in the Assay's 
meta.features data.frame updated for relevant features
}
\description{
Computes the metric at a given r (radius) value and stores in meta.features
}
