% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{ComputeSNNwidth}
\alias{ComputeSNNwidth}
\title{Calculate mean distance of the farthest neighbors from SNN graph}
\usage{
ComputeSNNwidth(
  snn.graph,
  embeddings,
  k.nn,
  l2.norm = TRUE,
  nearest.dist = NULL
)
}
\description{
This function will compute the average distance of the farthest k.nn 
neighbors with the lowest nonzero SNN edge weight. First, for each cell it 
finds the k.nn neighbors with the smallest edge weight. If there are multiple
cells with the same edge weight at the k.nn-th index, consider all of those 
cells in the next step. Next, it computes the euclidean distance to all k.nn
cells in the space defined by the embeddings matrix and returns the average 
distance to the farthest k.nn cells.
}
