% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{contrast-theory}
\alias{contrast-theory}
\alias{Intensity}
\alias{Luminance}
\title{Get the intensity and/or luminance of a color}
\source{
\url{https://stackoverflow.com/questions/3942878/how-to-decide-font-color-in-white-or-black-depending-on-background-color}
}
\usage{
Intensity(color)

Luminance(color)
}
\arguments{
\item{color}{A vector of colors}
}
\value{
A vector of intensities/luminances for each color
}
\description{
Get the intensity and/or luminance of a color
}
\examples{
Intensity(color = c('black', 'white', '#E76BF3'))

Luminance(color = c('black', 'white', '#E76BF3'))

}
\concept{visualization}
