% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{ConvertSeurat}
\alias{ConvertSeurat}
\title{Convert old Seurat object to accomodate new features}
\usage{
ConvertSeurat(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{
Returns a Seurat object compatible with latest changes
}
\description{
Adds the object@dr slot to older objects and moves the stored PCA/ICA analyses to new slot
}

