% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketching.R
\name{CountSketch}
\alias{CountSketch}
\title{Generate CountSketch random matrix}
\usage{
CountSketch(nsketch, ncells, seed = NA_integer_, ...)
}
\arguments{
\item{nsketch}{Number of sketching random cells}

\item{ncells}{Number of cells in the original data}

\item{seed}{a single value, interpreted as an integer, or \code{NULL}
    (see \sQuote{Details}).}

\item{...}{Ignored}
}
\value{
...
}
\description{
Generate CountSketch random matrix
}
\references{
Clarkson, KL. & Woodruff, DP.
Low-rank approximation and regression in input sparsity time.
Journal of the ACM (JACM). 2017 Jan 30;63(6):1-45.
\url{https://dl.acm.org/doi/abs/10.1145/3019134};
}
\keyword{internal}
