% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{CreateAssayObject}
\alias{CreateAssayObject}
\title{Create an Assay object}
\usage{
CreateAssayObject(raw.data, min.cells = 0, min.features = 0)
}
\arguments{
\item{raw.data}{Raw input data}

\item{min.cells}{Include features detected in at least this many cells. Will subset the raw.data 
matrix as well. To reintroduce excluded features, create a new object with a lower cutoff.}

\item{min.features}{Include cells where at least this many features are detected.}
}
\description{
Create an Assay object from a feature (e.g. gene) expression matrix. The expected format of the 
input matrix is features x cells.
}
\details{
Non-unique cell or feature names are not allowed. Please make unique before calling this function.
}
\examples{
pbmc_raw <- read.table(
  file = system.file('extdata', 'pbmc_raw.txt', package = 'Seurat'),
  as.is = TRUE
)
pbmc_rna <- CreateAssayObject(raw.data = pbmc_raw)
pbmc_rna

}
