% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CreateCategoryMatrix}
\alias{CreateCategoryMatrix}
\title{Create one hot matrix for a given label}
\usage{
CreateCategoryMatrix(
  labels,
  method = c("aggregate", "average"),
  cells.name = NULL
)
}
\arguments{
\item{labels}{A vector of labels}

\item{method}{Method to aggregate cells with the same label. Either 'aggregate' or 'average'}

\item{cells.name}{A vector of cell names}
}
\description{
Create one hot matrix for a given label
}
