% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{CreateCoo}
\alias{CreateCoo}
\title{Create a graph connectivity object in coo format}
\usage{
CreateCoo(data = NULL)
}
\arguments{
\item{data}{A square matrix with pairwise cell similarities}
}
\value{
Return a simplicial set
}
\description{
This function takes an adjacency matrix and transforms it to a simplicial set in RunkUMAP()
}
