% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreateDimReducObject}
\alias{CreateDimReducObject}
\title{Create a DimReduc object}
\usage{
CreateDimReducObject(embeddings = matrix(), loadings = matrix(),
  projected = matrix(), assay = NULL, stdev = numeric(),
  key = NULL, jackstraw = NULL, misc = list(), ...)
}
\arguments{
\item{embeddings}{...}

\item{loadings}{...}

\item{projected}{...}

\item{assay}{...}

\item{stdev}{...}

\item{key}{...}

\item{jackstraw}{...}

\item{misc}{...}

\item{...}{Ignored for now}
}
\description{
Create a DimReduc object
}
