% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreateSCTAssayObject}
\alias{CreateSCTAssayObject}
\title{Create a SCT Assay object}
\usage{
CreateSCTAssayObject(
  counts,
  data,
  scale.data = NULL,
  umi.assay = "RNA",
  min.cells = 0,
  min.features = 0,
  SCTModel.list = NULL
)
}
\arguments{
\item{scale.data}{a residual matrix}

\item{umi.assay}{The UMI assay name. Default is RNA}

\item{SCTModel.list}{list of SCTModels}
}
\description{
Create a SCT object from a feature (e.g. gene) expression matrix and a list of SCTModels.
The expected format of the input matrix is features x cells.
}
\details{
Non-unique cell or feature names are not allowed. Please make unique before
calling this function.
}
\concept{objects}
