% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreateWorkflowObject}
\alias{CreateWorkflowObject}
\title{Create a workflow object}
\usage{
CreateWorkflowObject(file)
}
\arguments{
\item{file}{Path to workflow file}
}
\value{
A SeuratWorkflow object
}
\description{
Create a workflow object
}
