% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{CustomNormalize}
\alias{CustomNormalize}
\title{Normalize raw data}
\usage{
CustomNormalize(data, custom_function, across)
}
\arguments{
\item{data}{Matrix with the raw count data}
}
\value{
Returns a matrix with the custom normalization
}
\description{
Normalize count data per cell and transform to centered log ratio
}
