% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utilities.R
\name{DarkTheme}
\alias{DarkTheme}
\title{Dark Theme}
\usage{
DarkTheme(...)
}
\arguments{
\item{...}{Extra parameters to be passed to theme()}
}
\value{
A ggplot2 theme object
}
\description{
Add a dark theme to ggplot objects
}
\examples{
df <- data.frame(x = rnorm(n = 100, mean = 20, sd = 2), y = rbinom(n = 100, size = 100, prob = 0.2))
p <- ggplot(data = df, mapping = aes(x = x, y = y)) + geom_point(mapping = aes(color = 'red'))
p + DarkTheme(legend.position = 'none')

}
\seealso{
\code{theme}
}
