% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_generics.R, R/seurat.R, R/dimreduc.R
\name{DefaultAssay}
\alias{DefaultAssay}
\alias{DefaultAssay<-}
\alias{DefaultAssay.Seurat}
\alias{DefaultAssay<-.Seurat}
\alias{DefaultAssay.DimReduc}
\title{Get the default assay}
\usage{
DefaultAssay(object, ...)

DefaultAssay(object, ...) <- value

\method{DefaultAssay}{Seurat}(object, ...)

\method{DefaultAssay}{Seurat}(object, ...) <- value

\method{DefaultAssay}{DimReduc}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{Name of assay to set as default}
}
\value{
The name of the default assay

An object with the new default assay
}
\description{
Get the default assay
}
\section{Methods (by class)}{
\itemize{
\item \code{Seurat}: Get the default assay of a Seurat object

\item \code{Seurat}: Set the default assay of a Seurat object

\item \code{DimReduc}: Get the name of the assay used to calculate this DimReduc
}}

