% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{DietSeurat}
\alias{DietSeurat}
\title{Slim down a Seurat object}
\usage{
DietSeurat(
  object,
  counts = TRUE,
  data = TRUE,
  scale.data = FALSE,
  features = NULL,
  assays = NULL,
  dimreducs = NULL,
  graphs = NULL,
  misc = TRUE
)
}
\arguments{
\item{object}{Seurat object}

\item{counts}{Preserve the count matrices for the assays specified}

\item{data}{Preserve the data slot for the assays specified}

\item{scale.data}{Preserve the scale.data slot for the assays specified}

\item{features}{Only keep a subset of features, defaults to all features}

\item{assays}{Only keep a subset of assays specified here}

\item{dimreducs}{Only keep a subset of DimReducs specified here (if NULL,
remove all DimReducs)}

\item{graphs}{Only keep a subset of Graphs specified here (if NULL, remove
all Graphs)}

\item{misc}{Preserve the \code{misc} slot; default is \code{TRUE}}
}
\description{
Keep only certain aspects of the Seurat object. Can be useful in functions that utilize merge as
it reduces the amount of data in the merge.
}
\concept{objects}
