% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seurat.R
\name{diff.t.test}
\alias{diff.t.test}
\title{Differential expression testing using Student's t-test}
\usage{
diff.t.test(object, cells.1, cells.2, genes.use = NULL, thresh.use = log(2))
}
\arguments{
\item{object}{Seurat object}

\item{cells.1}{Group 1 cells}

\item{cells.2}{Group 2 cells}

\item{genes.use}{Genes to test. Default is to use all genes.}

\item{thresh.use}{Limit testing to genes which show, on average, at least
X-fold difference (log-scale) between the two groups of cells.

Increasing thresh.use speeds up the function, but can miss weaker signals.}
}
\value{
Returns a p-value ranked matrix of putative differentially expressed
genes.
}
\description{
Identify differentially expressed genes between two groups of cells using
the Student's t-test
}

