% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seurat.R
\name{dim.plot}
\alias{dim.plot}
\title{Dimensional reduction plot}
\usage{
dim.plot(object, reduction.use = "pca", dim.1 = 1, dim.2 = 2,
  cells.use = NULL, pt.size = 3, do.return = FALSE, do.bare = FALSE,
  cols.use = NULL, group.by = "ident", pt.shape = NULL, pt.alpha = 1)
}
\arguments{
\item{object}{Seurat object}

\item{reduction.use}{Which dimensionality reduction to use. Default is
"pca", can also be "tsne", or "ica", assuming these are precomputed.}

\item{dim.1}{Dimension for x-axis (default 1)}

\item{dim.2}{Dimension for y-axis (default 2)}

\item{cells.use}{Vector of cells to plot (default is all cells)}

\item{pt.size}{Adjust point size for plotting}

\item{do.return}{Return a ggplot2 object (default : FALSE)}

\item{do.bare}{Do only minimal formatting (default : FALSE)}

\item{cols.use}{Vector of colors, each color corresponds to an identity
class. By default, ggplot assigns colors.}

\item{group.by}{Group (color) cells in different ways (for example, orig.ident)}

\item{pt.shape}{If NULL, all points are circles (default). You can specify any
cell attribute (that can be pulled with fetch.data) allowing for both different colors and different shapes on cells.}

\item{pt.alpha}{A value between 0 and 1 to set points transparent. (default: 1).}
}
\value{
If do.return==TRUE, returns a ggplot2 object. Otherwise, only
graphical output.
}
\description{
Graphs the output of a dimensional reduction technique (PCA by default).
Cells are colored by their identity class.
}

