% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{DimPlot}
\alias{DimPlot}
\title{Dimensional reduction plot}
\usage{
DimPlot(object, dims = c(1, 2), cells = NULL, cols = NULL,
  pt.size = NULL, reduction = "pca", group.by = NULL,
  shape.by = NULL, order = NULL, label = FALSE, label.size = 4,
  cells.highlight = NULL, cols.highlight = "red",
  sizes.highlight = 1, na.value = "grey50", ...)
}
\arguments{
\item{object}{Seurat object}

\item{dims}{Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions}

\item{cells}{Vector of cells to plot (default is all cells)}

\item{cols}{Vector of colors, each color corresponds to an identity class. By default, ggplot2 assigns colors}

\item{pt.size}{Adjust point size for plotting}

\item{reduction}{Which dimensionality reduction to use}

\item{group.by}{Group (color) cells in different ways (for example, orig.ident)}

\item{shape.by}{If NULL, all points are circles (default). You can specify any
cell attribute (that can be pulled with FetchData) allowing for both
different colors and different shapes on cells}

\item{order}{Specify the order of plotting for the idents. This can be
useful for crowded plots if points of interest are being buried. Provide
either a full list of valid idents or a subset to be plotted last (on top)}

\item{label}{Whether to label the clusters}

\item{label.size}{Sets size of labels}

\item{cells.highlight}{A list of character or numeric vectors of cells to
highlight. If only one group of cells desired, can simply
pass a vector instead of a list. If set, colors selected cells to the color(s)
in \code{cols.highlight} and other cells black (white if dark.theme = TRUE);
will also resize to the size(s) passed to \code{sizes.highlight}}

\item{cols.highlight}{A vector of colors to highlight the cells as; will
repeat to the length groups in cells.highlight}

\item{sizes.highlight}{Size of highlighted cells; will repeat to the length
groups in cells.highlight}

\item{na.value}{Color value for NA points when using custom scale.}

\item{...}{Ignored for now}
}
\value{
A ggplot object
}
\description{
Graphs the output of a dimensional reduction technique (PCA by default).
Cells are colored by their identity class.
}
\examples{
DimPlot(object = pbmc_small)

}
\seealso{
\code{\link{FeaturePlot}} \code{\link{HoverLocator}}
\code{\link{FeatureLocator}}
}
