% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{DimReduc}
\alias{DimReduc}
\title{The Dimmensional Reduction Class}
\description{
The DimReduc object is ...
}
\section{Slots}{

\describe{
\item{\code{cell.embeddings}}{...}

\item{\code{gene.loadings}}{...}

\item{\code{gene.loadings.projected}}{...}

\item{\code{stdev}}{...}

\item{\code{key}}{...}

\item{\code{jackstraw}}{...}

\item{\code{misc}}{...}
}}

