% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{DimReduction}
\alias{DimReduction}
\title{Dimensional Reduction}
\usage{
DimReduction(object, reduction.type = NULL, genes.use = NULL,
  dims.store = 40, dims.compute = 40, use.imputed = FALSE,
  rev.reduction = FALSE, print.results = TRUE, dims.print = 1:5,
  genes.print = 30, ica.fxn = icafast, seed.use = 1, ...)
}
\arguments{
\item{object}{Seurat object}

\item{reduction.type}{Type of dimensional reduction to run. Options include "pca", "pcafast", 
"ica", "icafast"}

\item{genes.use}{Genes to use as input for the dimensional reduction technique. Default is 
object@var.genes.}

\item{dims.store}{Number of dimensions to store}

\item{dims.compute}{Number of dimensions to compute (for fast approximations)}

\item{use.imputed}{Whether to run the dimensional reduction on imputed values.}

\item{rev.reduction}{By default, computes the dimensional reduction on the cell x gene matrix.
Setting to true will compute it on the transpose (gene x cell matrix).}

\item{print.results}{Print the top genes associated with each dimension}

\item{dims.print}{Number of dimensions to print genes for}

\item{genes.print}{Number of genes to print for each dimension}

\item{ica.fxn}{Function to use if calculating ica}

\item{seed.use}{Random seed}

\item{...}{Additional arguments to be passed to specific reduction technique}
}
\value{
Returns a Seurat object with the dimensional reduction information stored
}
\description{
Various methods for dimensional reductions
}

