% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{DimRot}
\alias{DimRot}
\title{Dimensional Reduction Rotation Accessor Function}
\usage{
DimRot(object, reduction.type = "pca", dims.use = NULL, cells.use = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{reduction.type}{Dimensional reduction to use (default is PCA)}

\item{dims.use, }{Dimensions to include (default is all stored dims)}

\item{cells.use, }{Cells to include (default is all)}
}
\value{
Rotation matrix for given reduction, cells, and dimensions
}
\description{
Pull rotoation matrix for specified stored dimensional reduction analysis
}
