% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{DimX}
\alias{DimX}
\title{Dimensional Reduction Gene Loadings Accessor Function}
\usage{
DimX(object, reduction.type = "pca", dims.use = NULL, genes.use = NULL,
  use.full = F)
}
\arguments{
\item{object}{Seurat object}

\item{reduction.type}{Dimensional reduction to use (default is PCA)}

\item{dims.use, }{Dimensions to include (default is all stored dims)}

\item{genes.use, }{Genes to include (default is all)}

\item{use.full}{return projected gene loadings (default is F)}
}
\value{
Rotation matrix for given reduction, cells, and genes
}
\description{
Pull rotoation matrix for specified stored dimensional reduction analysis
}

