% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing5.R
\name{DISP}
\alias{DISP}
\title{Find variable features based on dispersion}
\usage{
DISP(data, nselect = 2000L, verbose = TRUE, ...)
}
\arguments{
\item{data}{Data matrix}

\item{nselect}{Number of top features to select based on dispersion values}

\item{verbose}{Display progress}
}
\description{
Find variable features based on dispersion
}
\keyword{internal}
