% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Distances}
\alias{Distances}
\alias{Distances.Neighbor}
\title{Get the Neighbor nearest neighbors distance matrix}
\usage{
Distances(object, ...)

\method{Distances}{Neighbor}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\description{
Get the Neighbor nearest neighbors distance matrix
}
\concept{objects}
