% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction_utilities.R
\name{DMEmbed}
\alias{DMEmbed}
\title{Diffusion Maps Cell Embeddings Accessor Function}
\usage{
DMEmbed(object, dims.use = NULL, cells.use = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{dims.use}{Dimensions to include (default is all stored dims)}

\item{cells.use}{Cells to include (default is all cells)}
}
\value{
Diffusion maps embedding matrix for given cells and DMs
}
\description{
Pull Diffusion maps cell embedding matrix
}
\examples{
\dontrun{
pbmc_small
pbmc_small <- RunDiffusion(pbmc_small, genes.use = pbmc_small@var.genes)
head(DMEmbed(object = pbmc_small))
}

}
