% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction_utilities.R
\name{DMLoad}
\alias{DMLoad}
\title{Diffusion Maps Gene Loading Accessor Function}
\usage{
DMLoad(object, dims.use = NULL, genes.use = NULL, use.full = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{dims.use}{Dimensions to include (default is all stored dims)}

\item{genes.use}{Genes to include (default is all)}

\item{use.full}{Return projected gene loadings (default is FALSE)#'}
}
\value{
Diffusion maps gene loading matrix for given genes and DMs
}
\description{
Pull the diffusion maps gene loadings matrix
}
