% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{DMPlot}
\alias{DMPlot}
\title{Plot Diffusion map}
\usage{
DMPlot(object, ...)
}
\arguments{
\item{object}{Seurat object}

\item{\dots}{Additional parameters to DimPlot, for example, which dimensions to plot.}
}
\description{
Graphs the output of a Diffusion map analysis
Cells are colored by their identity class.
}
\details{
This function is a wrapper for DimPlot. See ?DimPlot for a full list of possible
arguments which can be passed in here.
}
\examples{
pbmc_small <- RunDiffusion(object = pbmc_small)
DMPlot(object = pbmc_small)

}
