% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{DMRot}
\alias{DMRot}
\title{Diffusion Maps Rotation Accessor Function}
\usage{
DMRot(object, dims.use = NULL, cells.use = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{dims.use}{Dimensions to include (default is all stored dims)}

\item{cells.use}{Cells to include (default is all cells)}
}
\value{
Diffusion maps rotation matrix for given cells and DMs
}
\description{
Pull Diffusion maps rotation matrix
}

