% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{DoHeatmap}
\alias{DoHeatmap}
\title{Feature expression heatmap}
\usage{
DoHeatmap(object, cells.use = NULL, features.use = NULL, disp.min = -2.5,
  disp.max = 2.5, group.by = "ident", slot.use = "data",
  assay.use = NULL, check.plot = FALSE, ...)
}
\arguments{
\item{object}{Seurat object}

\item{cells.use}{A vector of cells to plot}

\item{features.use}{A vector of features to plot, defaults to \code{VariableFeatures(object = object)}}

\item{disp.min}{Minimum display value (all values below are clipped)}

\item{disp.max}{Maximum display value (all values above are clipped)}

\item{group.by}{Name of variable to group cells by}

\item{slot.use}{Data slot to use, choose from 'raw.data', 'data', or 'scale.data'}

\item{assay.use}{Assay to pull from}

\item{check.plot}{Check that plotting will finish in a reasonable amount of time}

\item{...}{Extra parameters passed to superheat}
}
\value{
Invisbly returns the final grob
}
\description{
Draws a heatmap of single cell feature expression using superheat
}
\examples{
DoHeatmap(object = pbmc_small)

}
\seealso{
\code{\link{superheat::superheat}}
}
