% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_generics.R, R/interaction.R
\docType{methods}
\name{DownsampleSeurat}
\alias{DownsampleSeurat}
\alias{DownsampleSeurat,seurat-method}
\alias{DownsampleSeurat,loom-method}
\title{Density Dependent Downsampling}
\usage{
DownsampleSeurat(object, ...)

\S4method{DownsampleSeurat}{seurat}(object, ...)

\S4method{DownsampleSeurat}{loom}(object, ...)
}
\arguments{
\item{object}{An object to downsample}

\item{...}{Parameters to methods}
}
\description{
Density Dependent Downsampling
}
