% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreduc_generics.R
\name{Embeddings}
\alias{Embeddings}
\title{Get cell embeddings}
\usage{
Embeddings(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\description{
Get cell embeddings
}
