% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expMean}
\alias{expMean}
\title{Calculate the mean of logged values}
\usage{
expMean(x)
}
\arguments{
\item{x}{value or vector of values}
}
\value{
Returns the mean in log-space
}
\description{
Calculate mean of logged values in non-log space (return answer in log-space)
}
