% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expSD}
\alias{expSD}
\title{Calculate the standard deviation of logged values}
\usage{
expSD(x)
}
\arguments{
\item{x}{value or vector of values}
}
\value{
Returns the standard deviation in log-space
}
\description{
Calculate SD of logged values in non-log space (return answer in log-space)
}
