% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expVar}
\alias{expVar}
\title{Calculate the variance of logged values}
\usage{
expVar(x)
}
\arguments{
\item{x}{value or vector of values}
}
\value{
Returns the variance in log-space
}
\description{
Calculate variance of logged values in non-log space (return answer in
log-space)
}
