% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{FastRowScale}
\alias{FastRowScale}
\title{Scale and/or center matrix rowwise}
\usage{
FastRowScale(mat, center = TRUE, scale = TRUE, scale_max = 10)
}
\arguments{
\item{mat}{A matrix}

\item{center}{a logical value indicating whether to center the rows}

\item{scale}{a logical value indicating whether to scale the rows}

\item{scale_max}{clip all values greater than scale_max to scale_max. Don't
clip if Inf.}
}
\value{
Returns the center/scaled matrix
}
\description{
Performs row scaling and/or centering. Equivalent to using t(scale(t(mat)))
in R except in the case of NA values.
}
\concept{utilities}
