% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{FastScaleData}
\alias{FastScaleData}
\title{Scale and center the data using C++}
\usage{
FastScaleData(object, genes.use = NULL, data.use = NULL, do.scale = TRUE,
  do.center = TRUE, scale.max = 10, display.progress = TRUE,
  assay.type = "RNA")
}
\arguments{
\item{object}{Seurat object}

\item{genes.use}{Vector of gene names to scale/center. Default is all genes in object@data.}

\item{data.use}{Can optionally pass a matrix of data to scale, default is object@data[genes.use,]}

\item{do.scale}{Whether to scale the data.}

\item{do.center}{Whether to center the data.}

\item{scale.max}{Max value to return for scaled data. The default is 10. Setting this can help
reduce the effects of genes that are only expressed in a very small number of cells.}

\item{assay.type}{Assay to scale data for. Default is RNA. Can be changed for multimodal analyses.}
}
\value{
Returns a seurat object with object@scale.data updated with scaled and/or centered data.
}
\description{
Note: will give slightly different results than R due to differences in numerical precision
between R and C++. This could cause the results of some stochastic processes like tSNE to change.
}
