% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{FastWhichCells}
\alias{FastWhichCells}
\title{FastWhichCells
Identify cells matching certain criteria (limited to character values)}
\usage{
FastWhichCells(object, group.by, subset.value, invert = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Group cells in different ways (for example, orig.ident). Should be a column name in object@meta.data}

\item{subset.value}{Return cells matching this value}

\item{invert}{invert cells to return.FALSE by default}
}
\description{
FastWhichCells
Identify cells matching certain criteria (limited to character values)
}
\examples{
FastWhichCells(object = pbmc_small, group.by = 'res.1', subset.value = 1)

}
