% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{FeatureLocator}
\alias{FeatureLocator}
\title{Feature Locator}
\usage{
FeatureLocator(plot, data.plot, ...)
}
\arguments{
\item{plot}{A ggplot2 plot}

\item{data.plot}{The oridinal data that went into the ggplot2 plot}

\item{...}{Extra parameters, such as dark.theme, recolor, or smooth for using a dark theme,
recoloring based on selected cells, or using a smooth scatterplot, respectively}
}
\value{
The names of the points selected
}
\description{
Select points on a scatterplot and get information about them
}
\seealso{
\code{\link{locator}}

\code{\link{ggplot2::ggplot_build}}
}
