% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{FeatureMap}
\alias{FeatureMap}
\alias{FeatureHeatmap}
\title{Vizualization of multiple features}
\usage{
FeatureMap(object, features.plot, min.cutoff = NA, max.cutoff = NA,
  reduction.use = "tsne", dims.use = c(1, 2), group.by = NULL,
  cols.use = c("grey", "red"), pt.size = 2, group.display = NULL,
  slot.use = "data", scale.group = FALSE, horizontal = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{features.plot}{Vector of features to plot}

\item{min.cutoff}{Vector of minimum cutoff values for each feature, may specify quantile in the form of 'q##' where '##' is the quantile (eg, 1, 10)}

\item{max.cutoff}{Vector of maximum cutoff values for each feature, may specify quantile in the form of 'q##' where '##' is the quantile (eg, 1, 10)}

\item{reduction.use}{Which dimensionality reduction to use. Default is
"tsne", can also be "pca", or "ica", assuming these are precomputed.}

\item{dims.use}{Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions}

\item{group.by}{Group (color) cells in different ways (for example, orig.ident)}

\item{cols.use}{The two colors to form the gradient over. Provide as string vector with
the first color corresponding to low values, the second to high. Also accepts a Brewer
color scale or vector of colors. Note: this will bin the data into number of colors provided.}

\item{pt.size}{Adjust point size for plotting}

\item{group.display}{Which identity classes to display (default is all identity classes)}

\item{slot.use}{Dataset to use for plotting, choose from 'raw.data', 'data', or 'scale.data'}

\item{scale.group}{Scale each group separately. Default is FALSE.}

\item{horizontal}{rotate the plot such that the features are columns, groups are the rows}
}
\value{
A ggplot object
}
\description{
Similar to FeaturePlot, however, also splits the plot by visualizing each
identity class separately.
}
\details{
Particularly useful for seeing if the same groups of cells co-exhibit a
common feature (i.e. co-express a gene), even within an identity class. Best
understood by example.
}
\examples{
pbmc_small
FeatureMap(object = pbmc_small, features.plot = "PC1")

}
\seealso{
\code{\link{FeaturePlot}} \code{\link{DimPlot}}
}
