% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{fetch.data}
\alias{fetch.data}
\title{Access cellular data}
\usage{
fetch.data(object, vars.all = NULL, cells.use = NULL, use.imputed = FALSE,
  use.scaled = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{vars.all}{List of all variables to fetch}

\item{cells.use}{Cells to collect data for (default is all cells)}

\item{use.imputed}{For gene expression, use imputed values}

\item{use.scaled}{For gene expression, use scaled values}
}
\value{
A data frame with cells as rows and cellular data as columns
}
\description{
Retreives data (gene expression, PCA scores, etc, metrics, etc.) for a set
of cells in a Seurat object
}

