% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seuratFxns.R
\name{fetch_closest}
\alias{fetch_closest}
\title{For given bin, fetch cloeset cell-centroids.}
\usage{
fetch_closest(b, centroids, cells.num, bins = 64, compatible = TRUE)
}
\arguments{
\item{b}{A scalar indicating the bin index.}

\item{centroids}{A data.frame recording all the cell-centroids candidates
with cell names as rows and coordinates as columns.}

\item{cells.num}{A scalar indicating the number of cell-centroids to be
fetched.}

\item{bins}{A scalar indicating total bins number. Default: 64.}

\item{compatible}{A logic indicating whether running function compatible with
original \code{seura} package function. If \code{TRUE}, this function will
actually yieid \code{cells.num + 1} closest centroids. If \code{FALSE},
exactly \code{cells.num} centroids will be reported. Default: TRUE.}
}
\value{
A character vector of cells names.
}
\description{
For given bin, find closest cell-centroids in the space.
Distance metric is L2-distance.
}
\note{
Please notice the original \code{fetch.closest} function in
\code{seurat} package actually
report \code{2n+1} cells where \code{n} is the number of landmarked genes,
while in paper it claimed \code{2n}.
}

