% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing5.R
\name{FetchResiduals_reference}
\alias{FetchResiduals_reference}
\title{Calculate pearson residuals of features not in the scale.data}
\usage{
FetchResiduals_reference(
  object,
  reference.SCT.model = NULL,
  features = NULL,
  nCount_UMI = NULL,
  verbose = FALSE
)
}
\arguments{
\item{object}{A seurat object}

\item{reference.SCT.model}{SCTModel object to use for calculating residuals}

\item{features}{Name of features to add into the scale.data}

\item{nCount_UMI}{UMI Counts per cell provided to vst}

\item{verbose}{Whether to print messages and progress bars}
}
\value{
Returns a Seurat object containing Pearson residuals of added
features in its scale.data

temporal function to get residuals from reference
}
\description{
This function calls sctransform::get_residuals.
}
