% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{FilterCells}
\alias{FilterCells}
\title{Filter out cells from a Seurat object}
\usage{
FilterCells(...)
}
\arguments{
\item{...}{Ignored}
}
\description{
Defunct, replaced with \code{subset.Seurat}
}
\seealso{
\code{\link{subset.Seurat}}
}
