% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs.R
\name{FindGeneMarkerPairs}
\alias{FindGeneMarkerPairs}
\title{Find marker pairs for cell-cycle phase prediction}
\usage{
FindGeneMarkerPairs(object, g1.cells, g2m.cells, s.cells, genes.list,
  threshold.fraction = 0.5)
}
\arguments{
\item{object}{A Seurat object}

\item{g1.cells}{A vector of cells known to be in G1 phase}

\item{g2m.cells}{A vector of cells known to be in G2M phase}

\item{s.cells}{A vector of cells known to be in S phase}

\item{genes.list}{A vector of genes to train the model on}

\item{threshold.fraction}{Fraction of differences needed to be significant}
}
\value{
A list with the following fields

\item{g1.pairs}{A data frame of gene pairs marking G1 phase}

\item{g2m.pairs}{A data frame of gene pairs marking G2M phase}

\item{s.pairs}{A data frame of gene pairs marking S phase}
}
\description{
Find marker pairs for cell-cycle phase prediction
}
\references{
Scialdone A, Natarajan KN, Saraiva LR, Proserpio V, Teichmann SA, Stegle O, Marioni JC, Buettner F. (2015) Computational assignment of cell-cycle stage from single-cell transcriptome data. Methods
}
