% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_used_yet.R
\name{FindGeneTerms}
\alias{FindGeneTerms}
\title{Find gene terms from Enrichr}
\usage{
FindGeneTerms(QueryGene = NULL)
}
\arguments{
\item{QueryGene}{A gene to query on Enrichr}
}
\value{
An XML document with information on the queried gene
}
\description{
Fing gene terms from Enrichr and return the XML information
}
