% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{FindMappingAnchors}
\alias{FindMappingAnchors}
\title{Find mapping anchors}
\usage{
FindMappingAnchors(
  reference,
  query,
  normalization.method = c("LogNormalize", "SCT"),
  reference.assay = NULL,
  reference.nn = NULL,
  reference.nnidx = NULL,
  query.assay = NULL,
  reference.reduction = "pca",
  reduction = "pcaproject",
  features = NULL,
  npcs = NULL,
  l2.norm = TRUE,
  dims = 1:30,
  k.anchor = 5,
  k.filter = NA,
  k.score = 30,
  max.features = 200,
  nn.method = "annoy",
  eps = 0,
  approx.pca = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{reference}{\code{\link{Seurat}} object to use as the reference}

\item{query}{\code{\link{Seurat}} object to use as the query}

\item{normalization.method}{Name of normalization method used on the reference: 
LogNormalize or SCT;  the same method should be applied to the query}

\item{reference.assay}{Name of the Assay to use from reference}

\item{reference.nn}{Precomputed nearest neighbors for the reference, returned by AnnoyNN}

\item{reference.nnidx}{Precomputed Annoy NN index object returned by AnnoyNN}

\item{query.assay}{Name of the Assay to use from query}

\item{reference.reduction}{Dimensional reduction to use from the reference}

\item{reduction}{Name for the projected reduction in the query object}

\item{features}{Features to use for dimensional reduction. If not specified,
set as variable features of the reference object which are also present in 
the query.}

\item{npcs}{Number of PCs to compute on reference. If null, then use an 
existing PCA structure in the reference object}

\item{l2.norm}{Perform L2 normalization on the cell embeddings after 
dimensional reduction}

\item{dims}{Which dimensions to use from the reduction to specify the 
neighbor search space}

\item{k.anchor}{How many neighbors (k) to use when finding anchors}

\item{k.filter}{How many neighbors (k) to use when filtering anchors}

\item{k.score}{How many neighbors (k) to use when scoring anchors}

\item{max.features}{The maximum number of features to use when specifying the 
neighborhood search space in the anchor filtering}

\item{nn.method}{Method for nearest neighbor finding. Options include: rann,
annoy}

\item{eps}{Error bound on the neighbor finding algorithm (from 
\code{\link{RANN}})}

\item{approx.pca}{Use truncated singular value decomposition to approximate 
PCA}

\item{verbose}{Print progress bars and output}
}
\description{
Find a set of anchors between a reference and query object. These 
anchors are used in the \code{\link{MapQueryData}} function
}
