% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{FindMultiModalNeighbors}
\alias{FindMultiModalNeighbors}
\title{Construct multimodal neighbors, KNN and SNN Graph}
\usage{
FindMultiModalNeighbors(
  object,
  reduction.list,
  dims.list,
  modality.weight.name = NULL,
  k.nn = 20,
  l2.norm = TRUE,
  sd.scale = 1,
  cross.contant.list = list(1e-04, 1e-04),
  smooth = FALSE,
  knn.range = 200,
  knn.graph.name = "wknn",
  snn.graph.name = "wsnn",
  weighted.nn.name = "weighted.nn",
  modality.weight = NULL,
  prune.SNN = 1/15,
  weighted.graph = FALSE,
  return.intermediate = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{A Seurat object}

\item{reduction.list}{A list of name of dimension reduction}

\item{dims.list}{A list of number of dimensions to use}

\item{modality.weight.name}{The variable name of first modality weights
stored in the meta.data.}

\item{k.nn}{the number of multi-modal neighbors computed}

\item{l2.norm}{Perform L2 normalization on the cell embeddings after
dimensional reduction}

\item{sd.scale}{The scaling factor for kernel width, and the default is 1.}

\item{cross.contant.list}{the minimal cross-modality prediction similarity
used in the modality score calculation.}

\item{smooth}{Smoothing modality score across each individual modality
neighbors.}

\item{knn.range}{The number of approximate neighbors to compute}

\item{knn.graph.name}{The name of multimodal knn graph}

\item{snn.graph.name}{The name of multimodal snn graph}

\item{weighted.nn.name}{The name of multimodal neighbors}

\item{modality.weight}{A \code{\link{ModalityWeights}} object generated by
\code{FindModalityWeights}}

\item{prune.SNN}{Sets the cutoff for acceptable Jaccard index when computing
the neighborhood overlap for the SNN construction}

\item{weighted.graph}{Add consider neighbor distance as the edges to
construct KNN graph}

\item{return.intermediate}{Return intermediate results in the FindModalityWeights.
It will be stored in the misc as a ModalityClass.}

\item{verbose}{Print progress bars and output}
}
\value{
return an object containing multimodal KNN, SNN and neighbors
}
\description{
This function will construct multimodal neighbors, Kth Nearest Neighbors
(KNN) and Shared Nearest Neighbor (SNN) Graphs. According to the input
\code{modality.weight}, it constructs a cell-specific weighted joint kernel.
Then, for each cell, it will find \code{knn.range}s individual modal
neighbors, and get the union of those neighbors. Next, it will find
\code{k.nn} multimodal neighbors by the weighted joint kernel. Given the set
of multimodal neighbors, we construct its KNN and SNN Graph.
}
