% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/clustering.R
\name{FindSpatialNeighbors}
\alias{FindSpatialNeighbors}
\alias{FindSpatialNeighbors.Seurat}
\title{SNN Graph Construction}
\usage{
FindSpatialNeighbors(object, ...)

\method{FindSpatialNeighbors}{Seurat}(
  object,
  spots.coordinates,
  profileDistance = 2,
  spotDistance = 2,
  spotDistanceTransformationWeight = 1,
  dims = 1:10
)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{spots.coordinates}{nSpots x 2 matrix or dataframe with the coordinates of each spot}

\item{profileDistance}{integer. Param of the minkowski distance between transcriptional profiles}

\item{spotDistance}{integer. Param of the minkowski distance between spot positions}

\item{spotDistanceTransformationWeight}{double in [0,1]. Weight for the linear transormation of spot distance.
1 means spot distance weight as much as profile distance, 0 means spot distance doesn't contribute at all in the overall 
distance measure.}

\item{dims}{Dimensions of reduction to use as input}

\item{do.plot}{Plot SNN graph on tSNE coordinates}
}
\value{
When running on a \code{\link{Seurat}} object, returns fills the
\code{graphs} slot; names of graphs can be found with
\code{Filter(function(x) inherits(object[[x]], "Graph"), names(object))}
}
\description{
Constructs a Shared Nearest Neighbor (SNN) Graph for a given dataset. We
first determine the k-nearest neighbors of each spot (pca + phisical distance between spots). 
We use this knn graph to construct the SNN graph by calculating the neighborhood overlap
(Jaccard index) between every spot and its k.param nearest neighbors.
}
\examples{
pbmc_small

# We build the SNN on the pca of the data such (ex. 10 principle components).

pbmc_small <- FindSpatialNeighbors(pbmc_small, dims = 1:10, spots.coordinates = coord)

}
