% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{FindSubCluster}
\alias{FindSubCluster}
\title{Find subclusters under one cluster}
\usage{
FindSubCluster(
  object,
  cluster,
  graph.name,
  subcluster.name = "sub.cluster",
  resolution = 0.5,
  algorithm = 1
)
}
\arguments{
\item{object}{An object}

\item{cluster}{the cluster to be sub-clustered}

\item{graph.name}{Name of graph to use for the clustering algorithm}

\item{subcluster.name}{the name of sub cluster added in the meta.data}

\item{resolution}{Value of the resolution parameter, use a value above
(below) 1.0 if you want to obtain a larger (smaller) number of communities.}

\item{algorithm}{Algorithm for modularity optimization (1 = original Louvain
algorithm; 2 = Louvain algorithm with multilevel refinement; 3 = SLM
algorithm; 4 = Leiden algorithm). Leiden requires the leidenalg python.}
}
\value{
return a object with sub cluster labels in the sub-cluster.name variable
}
\description{
Find subclusters under one cluster
}
\concept{clustering}
