% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{findThresh}
\alias{findThresh}
\title{Inter-maxima quantile sweep to find ideal barcode thresholds}
\usage{
findThresh(call.list, id)
}
\arguments{
\item{call.list}{A list of sample classification result from different quantiles using classifyCells}

\item{id}{The id as the suffix for output dataframe (named as res_id)}
}
\value{
A data.frame named res_id documenting the quantile used, subset, number of cells and proportion
}
\description{
Finding ideal thresholds for positive-negative signal classification per multiplex barcode
Written by Chris McGinnis from Gartner Lab, UCSF
}
\examples{
findThresh(call.list=bar.table_sweep.list, id="round") #returns res_round in the environment

}
