% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R, R/preprocessing_generics.R
\name{FindVariableFeaturesNew.Assay}
\alias{FindVariableFeaturesNew.Assay}
\alias{FindVariableFeaturesNew.Seurat}
\alias{FindVariableFeaturesNew}
\title{Find variable features}
\usage{
\method{FindVariableFeaturesNew}{Assay}(object, num.features = 1000,
  loess.span = 0.3, clip.max = "auto", verbose = TRUE)

\method{FindVariableFeaturesNew}{Seurat}(object, assay.use = NULL,
  num.features = 1000, loess.span = 0.3, clip.max = "auto",
  verbose = TRUE, workflow.name = NULL, ...)

FindVariableFeaturesNew(object, num.features, loess.span, clip.max, verbose)
}
\arguments{
\item{object}{An object}

\item{num.features}{Number of features to select as top variable features}

\item{loess.span}{Loess span parameter used when fitting the variance-mean relationship}

\item{clip.max}{After standardization values larger than clip.max will be set to clip.max;
default is 'auto' which sets this value to the square root of the number of cells}

\item{verbose}{Show progress bar for calculations}

\item{assay.use}{Which assay to use}

\item{workflow.name}{Name of workflow}

\item{num.features}{Number of features to select as top variable features}

\item{loess.span}{Loess span parameter used when fitting the variance-mean relationship}

\item{clip.max}{After standardization values larger than clip.max will be set to clip.max;
default is 'auto' which sets this value to the square root of the number of cells}

\item{verbose}{Show progress bar for calculations}
}
\description{
Identifies features that are outliers on a 'mean variability plot'. First, fits a
line to the relationship of log(variance) and log(mean) using local polynomial 
regression (loess). Then standardizes the feature values using the observed mean and 
expected variance (given by the fitted line). Feature variance is the calculated on 
the standardized values after clipping to a maximum (default is 50).
}
\section{Methods (by class)}{
\itemize{
\item \code{Assay}: Find variable features in an Assay object

\item \code{Seurat}: Find variable features in a Seurat object
}}

