% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit.gene.k.2.R
\name{fit.gene.k.fast}
\alias{fit.gene.k.fast}
\title{Build mixture models of gene expression}
\usage{
fit.gene.k.fast(object, gene, do.k = 2, num.iter = 1, do.plot = FALSE,
  genes.use = NULL, start.pct = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{gene}{Gene to fit}

\item{do.k}{Number of modes for the mixture model (default is 2)}

\item{num.iter}{Number of 'greedy k-means' iterations (default is 1)}

\item{do.plot}{Plot mixture model results}

\item{genes.use}{Genes to use in the greedy k-means step (See manuscript for details)}

\item{start.pct}{Initial estimates of the percentage of cells in the 'on'
state (usually estimated from the in situ map)}
}
\value{
A Seurat object, where the posterior of each cell being in the 'on'
or 'off' state for each gene is stored in object@mix.probs
}
\description{
Proof-of-concept to accelarate runing speed of fit.gene.k function
}

