% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/differential_expression.R
\name{FoldChange}
\alias{FoldChange}
\alias{FoldChange.default}
\alias{FoldChange.Assay}
\alias{FoldChange.DimReduc}
\alias{FoldChange.Seurat}
\title{FoldChange}
\usage{
FoldChange(object, ...)

\method{FoldChange}{default}(object, cells.1, cells.2, features, mean.fxn, fc.name = "avg_logFC")

\method{FoldChange}{Assay}(object, cells.1, cells.2, features, slot, pseudocount.use = 1)

\method{FoldChange}{DimReduc}(
  object,
  cells.1,
  cells.2,
  features,
  slot = NULL,
  pseudocount.use = NULL,
  fc.name = "avg_diff"
)

\method{FoldChange}{Seurat}(
  object,
  ident.1 = NULL,
  ident.2 = NULL,
  group.by = NULL,
  subset.ident = NULL,
  assay = NULL,
  slot = "data",
  reduction = NULL,
  features = NULL,
  pseudocount.use = 1
)
}
\arguments{
\item{object}{An object}

\item{cells.1}{Vector of cell names belonging to group 1}

\item{cells.2}{Vector of cell names belonging to group 2}

\item{features}{Genes to calculate fold change for; default is to use all genes}

\item{mean.fxn}{Function to use for fold change or average difference calculation}

\item{fc.name}{Name of fold change column in output data.frame}

\item{slot}{Slot to pull data from}

\item{pseudocount.use}{Pseudocount to add to averaged expression values when
calculating logFC. 1 by default.}

\item{ident.1}{Identity class to calculate fold change for; pass an object of class
\code{phylo} or 'clustertree' to calculate fold change for a node in a cluster tree;
passing 'clustertree' requires \code{\link{BuildClusterTree}} to have been run}

\item{ident.2}{A second identity class for comparison; if \code{NULL},
use all other cells for comparison; if an object of class \code{phylo} or
'clustertree' is passed to \code{ident.1}, must pass a node to calculate fold change for}

\item{group.by}{Regroup cells into a different identity class prior to calculating fold change (see example)}

\item{subset.ident}{Subset a particular identity class prior to regrouping. Only relevant if group.by is set (see example)}

\item{assay}{Assay to use in fold change calculation}

\item{reduction}{Reduction to use - will calculate average difference on cell embeddings}
}
\description{
Calculate log fold change and percent expressed of features for identity classes
}
\details{
If the slot is scale.data or a reduction is specified, average difference
is returned instead of log fold change
}
