% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{Format10X_GeoJson_CellID}
\alias{Format10X_GeoJson_CellID}
\title{Format 10X Genomics GeoJson cell IDs}
\usage{
Format10X_GeoJson_CellID(ids, prefix = "cellid_", suffix = "-1", digits = 9)
}
\arguments{
\item{ids}{Vector of cell IDs to format}

\item{prefix}{Optional prefix string}

\item{suffix}{Optional suffix string}

\item{digits}{Number of digits to zero-pad

A helper function to format cell IDs from the segmentation GeoJson to the same type as in the matrix.h5
The GeoJson has cell IDs as integers (eg 1). They need to be in the format cellid_000000001-1}
}
\value{
Vector of formatted cell IDs
}
\description{
Format 10X Genomics GeoJson cell IDs
}
