% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{fortify-Spatial}
\alias{fortify-Spatial}
\alias{fortify.Centroids}
\alias{fortify}
\alias{fortify.Molecules}
\alias{fortify.Segmentation}
\title{Prepare Coordinates for Spatial Plots}
\usage{
\method{fortify}{Centroids}(model, data, ...)

\method{fortify}{Molecules}(model, data, nmols = NULL, seed = NA_integer_, ...)

\method{fortify}{Segmentation}(model, data, ...)
}
\arguments{
\item{model}{A \code{\linkS4class{Segmentation}},
\code{\linkS4class{Centroids}},
or \code{\linkS4class{Molecules}} object}

\item{data}{Extra data to be used for annotating the cell segmentations; the
easiest way to pass data is a one-column
\code{\link[base:data.frame]{data frame}} with the values to color by and
the cell names are rownames}

\item{...}{Arguments passed to other methods}
}
\description{
Prepare Coordinates for Spatial Plots
}
\keyword{internal}
