% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{GadgetPlot}
\alias{GadgetPlot}
\alias{GadgetDimPlot}
\alias{GadgetFeaturePlot}
\title{Visualize spatial and clustering (dimensional reduction) data in a linked,
interactive framework}
\usage{
GadgetDimPlot(
  object,
  dims = 1:2,
  reduction = NULL,
  image = NULL,
  group.by = NULL,
  alpha = c(0.3, 1)
)

GadgetFeaturePlot(
  object,
  feature,
  dims = 1:2,
  reduction = NULL,
  image = NULL,
  slot = "data",
  alpha = c(0.3, 1)
)
}
\arguments{
\item{object}{Seurat object}

\item{dims}{Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions}

\item{reduction}{Which dimensionality reduction to use. If not specified, first searches for umap, then tsne, then pca}

\item{image}{Name of the image to use in the plot}

\item{group.by}{Name of one or more metadata columns to group (color) cells by
(for example, orig.ident); pass 'ident' to group by identity class}

\item{alpha}{Controls opacity of spots. Provide as a vector specifying the
min and max}

\item{feature}{Feature to visualize}

\item{slot}{Which slot to pull expression data from?}
}
\value{
Runs linked plot in a Shiny gadget session; does not return any value
}
\description{
Visualize spatial and clustering (dimensional reduction) data in a linked,
interactive framework
}
