% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketching.R
\name{GaussianSketch}
\alias{GaussianSketch}
\title{Gaussian sketching}
\usage{
GaussianSketch(nsketch, ncells, seed = NA_integer_, ...)
}
\arguments{
\item{nsketch}{Number of sketching random cells}

\item{ncells}{Number of cells in the original data}

\item{seed}{a single value, interpreted as an integer, or \code{NULL}
    (see \sQuote{Details}).}

\item{...}{Ignored}
}
\value{
...
}
\description{
Gaussian sketching
}
\keyword{internal}
