% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_generics.R, R/utilities_internal.R
\docType{methods}
\name{GetAllCalcParam}
\alias{GetAllCalcParam}
\alias{GetAllCalcParam,seurat-method}
\alias{GetAllCalcParam,loom-method}
\title{Get All CalcParam information for given calculation}
\usage{
GetAllCalcParam(object, ...)

\S4method{GetAllCalcParam}{seurat}(object, calculation)

\S4method{GetAllCalcParam}{loom}(object, dataset.use)
}
\arguments{
\item{object}{A Seurat object}

\item{calculation}{The name of the calculation that was done}
}
\value{
list of parameter values for given calculation
}
\description{
Get All CalcParam information for given calculation
}
