% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{GetAssay}
\alias{GetAssay}
\alias{GetAssay.Seurat}
\title{Get an assay from an object}
\usage{
GetAssay(object, assay, ...)

\method{GetAssay}{Seurat}(object, assay = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{assay}{Assay to get}

\item{...}{Arguments passed to other methods}
}
\value{
Returns an Assay object
}
\description{
Get an assay from an object
}
\section{Methods (by class)}{
\itemize{
\item \code{Seurat}: Get an assay from a Seurat object
}}

