% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{GetAssayData}
\alias{GetAssayData}
\alias{GetAssayData.Assay}
\alias{GetAssayData.Seurat}
\title{Accessor function for multimodal data}
\usage{
GetAssayData(object, slot, ...)

\method{GetAssayData}{Assay}(object, slot = "data")

\method{GetAssayData}{Seurat}(object, slot = "data", assay.use = NULL,
  ...)
}
\arguments{
\item{object}{An object}

\item{slot}{Specific information to pull (i.e. raw.data, data, scale.data,...). Default is data}

\item{...}{Arguments passed to other methods}

\item{assay.use}{Name of assay to pull data from}
}
\value{
Returns assay data
}
\description{
Pull information for specified stored dimensional reduction analysis
}
\section{Methods (by class)}{
\itemize{
\item \code{Assay}: Get assay data for an Assay object

\item \code{Seurat}: Get assay data from a Seurat object
}}

