% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_generics.R, R/utilities_internal.R
\docType{methods}
\name{GetCalcParam}
\alias{GetCalcParam}
\alias{GetCalcParam,seurat-method}
\alias{GetCalcParam,loom-method}
\title{Get CalcParam information}
\usage{
GetCalcParam(object, ...)

\S4method{GetCalcParam}{seurat}(object, calculation, parameter)

\S4method{GetCalcParam}{loom}(object, dataset.use, parameter)
}
\arguments{
\item{object}{A Seurat object}

\item{calculation}{The name of the calculation that was done}

\item{parameter}{Parameter for the calculation to pull}

\item{dataset.use}{dataset from which to pull the attribute (e.g. "layers/norm_data")}
}
\value{
parameter value for given calculation
}
\description{
Get CalcParam information
}
