% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_determination.R
\name{GetClusters}
\alias{GetClusters}
\title{Get Cluster Assignments}
\usage{
GetClusters(object)
}
\arguments{
\item{object}{Seurat object with cluster assignments}
}
\value{
Returns a dataframe with cell names and cluster assignments
}
\description{
Retrieve cluster IDs as a dataframe. First column will be the cell name,
second column will be the current cluster identity (pulled from object@ident).
}
