% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction_generics.R,
%   R/dimensional_reduction_utilities.R
\name{GetDimReduction}
\alias{GetDimReduction}
\alias{GetDimReduction.seurat}
\alias{GetDimReduction.loom}
\title{Dimensional Reduction Accessor Function}
\usage{
GetDimReduction(object, ...)

\method{GetDimReduction}{seurat}(object, reduction.type = "pca",
  slot = "gene.loadings")

\method{GetDimReduction}{loom}(object, reduction.type = "pca",
  slot = "gene_loadings", dataset.use = NULL,
  gene.names = "row_attrs/gene_names", cell.names = "col_attrs/cell_names")
}
\arguments{
\item{object}{An object}

\item{reduction.type}{Type of dimensional reduction to fetch (default is PCA)}

\item{slot}{Specific information to pull (must be one of the following:
"cell.embeddings", "gene.loadings", "gene.loadings.full", "sdev", "key", "misc")
(replace the '.' with a '_' for loom objects)}

\item{dataset.use}{Use this specific dataset rather than try to find the correct one automatically}

\item{gene.names}{Dataset with gene names, can pass \code{NULL}}

\item{cell.names}{Dataset will cell names, can pass \code{NULL}}
}
\value{
Returns specified slot results from given reduction technique
}
\description{
General accessor function for dimensional reduction objects. Pulls slot
contents for specified stored dimensional reduction analysis.
}
\section{Methods (by class)}{
\itemize{
\item \code{seurat}: Get dimensional reduction information for Seurat objects

\item \code{loom}: Get dimensional reduction information for loom objects
}}

\examples{
pbmc_small
# Get the PCA cell embeddings and print the top left corner
GetDimReduction(object = pbmc_small, reduction.type = "pca",
                slot = "cell.embeddings")[1:5, 1:5]
# Get the standard deviation of each PC
GetDimReduction(object = pbmc_small, reduction.type = "pca", slot = "sdev")

}
