% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_generics.R, R/seurat.R, R/assay.R
\name{GetHVFInfo}
\alias{GetHVFInfo}
\alias{GetHVFInfo.Seurat}
\alias{GetHVFInfo.Assay}
\title{Get highly variable feature information}
\usage{
GetHVFInfo(object, ...)

\method{GetHVFInfo}{Seurat}(object, assay.use = NULL, ...)

\method{GetHVFInfo}{Assay}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{assay.use}{Name of assay to pull highly variable feature information for}
}
\value{
A dataframe with feature means, dispersion, and scaled dispersion
}
\description{
Get highly variable feature information
}
\section{Methods (by class)}{
\itemize{
\item \code{Seurat}: Get highly variable feature information from a Seurat object

\item \code{Assay}: Get highly variable feature information from an Assay
}}

