% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_generics.R, R/assay.R, R/interaction.R
\name{GetIdent}
\alias{GetIdent}
\alias{GetIdent.Assay}
\alias{GetIdent}
\title{Get the idents of multimodal data}
\usage{
GetIdent(object, uniq = TRUE, cells.use = NULL)

\method{GetIdent}{Assay}(object, ...)

GetIdent(object, uniq = TRUE, cells.use = NULL)
}
\arguments{
\item{object}{An object}

\item{uniq}{logic, indicating whether to return unique ident values or ident
of all cells}

\item{cells.use}{A vector of cell names. If specified, only the identity of
these cells will be returned.}

\item{...}{Arguments passed to other methods}

\item{object}{Seurat object}
}
\value{
A factor with the cell identities

Return the cell identites of this Seurat object
}
\description{
Get the idents of multimodal data

Get identity of cells
}
\section{Methods (by generic)}{
\itemize{
\item \code{GetIdent}: Get cell identities from an Assay object
}}

\examples{
GetIdent(pbmc_small)

}
