% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/spatial.R
\name{GetImage}
\alias{GetImage}
\alias{GetImage.Seurat}
\alias{GetImage.SliceImage}
\alias{GetImage.SpatialImage}
\title{Get image data}
\usage{
GetImage(object, grob = TRUE, ...)

\method{GetImage}{Seurat}(object, grob = TRUE, image = NULL, ...)

\method{GetImage}{SliceImage}(object, grob = TRUE, ...)

\method{GetImage}{SpatialImage}(object, grob = TRUE, ...)
}
\arguments{
\item{object}{An object}

\item{grob}{Return a grob representation of the image}

\item{...}{Arguments passed to other methods}

\item{image}{Name of \code{SpatialImage} object to pull image data for; if
\code{NULL}, will attempt to select an image automatically}
}
\value{
If \code{grob}, a grob representation of the image; otherwise the default
representation of the image (typically an array)
}
\description{
Get image data
}
