% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{GetIntegrationData}
\alias{GetIntegrationData}
\title{Get integration data}
\usage{
GetIntegrationData(object, integration.name, slot)
}
\arguments{
\item{object}{Seurat object}

\item{integration.name}{Name of integration object}

\item{slot}{Which slot in integration object to get}
}
\value{
Returns data from the requested slot within the integrated object
}
\description{
Get integration data
}
\concept{objects}
