% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackstraw.R, R/jackstraw_generics.R
\name{GetJS.JackStrawData}
\alias{GetJS.JackStrawData}
\alias{GetJS}
\title{Get JackStrawData slot}
\usage{
\method{GetJS}{JackStrawData}(object, slot)

GetJS(object, slot, ...)
}
\arguments{
\item{object}{JackStrawData object}

\item{slot}{Name of slot to get}
}
\value{
Returns contents of specified slot
}
\description{
Accessor method for JackStrawData
}
\section{Methods (by class)}{
\itemize{
\item \code{JackStrawData}: Get a slot for a given JackStrawData
}}

