% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{GetOntologyTable}
\alias{GetOntologyTable}
\title{Get Ontology Table from Ontology Analysis}
\usage{
GetOntologyTable(object, pval.cutoff = 1)
}
\arguments{
\item{pval.cutoff}{p-value above which GO terms will be excluted. Default is 1}

\item{ontology.results}{A list with two topGO objects: ontology and test as returned by OntologyAnalysis}
}
\description{
Creates a data frame with GO terms and enrichment p-values as given
by Fisher's exact test. It takes as an input the result of OntologyAnalysos
}
