% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{GetRMetric}
\alias{GetRMetric}
\title{Quick accessor function for getting the Mark Variogram values}
\usage{
GetRMetric(object, assay = NULL, sorted = TRUE)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Assay on which the mark variogram computations were made}

\item{sorted}{Return the values in sorted decreasing order}
}
\value{
Returns a vector of r metric values as a name vector
}
\description{
Quick accessor function for getting the Mark Variogram values
}
