% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetSfPlotData}
\alias{GetSfPlotData}
\title{Extract data from a plot-ready \code{sf} object in a format suitable
for use with \code{ggplot2::geom_polygon()}.}
\usage{
GetSfPlotData(sf_data)
}
\arguments{
\item{sf_data}{A \code{sf} object}
}
\value{
A data.frame with columns x.vertex, y.vertex, cell (barcode), and columns corresponding to plotting variables, such that
each row represents plotting information for a single vertex of the segmentation corresponding to that barcode.
}
\description{
Extract data from a plot-ready \code{sf} object in a format suitable
for use with \code{ggplot2::geom_polygon()}.
}
\keyword{internal}
