% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_generics.R
\name{GetVariableFeatures}
\alias{GetVariableFeatures}
\title{Get the top variable Features}
\usage{
GetVariableFeatures(object, num.features, mean.cutoff, dispersion.cutoff,
  selection.method, ...)
}
\arguments{
\item{object}{An object}

\item{num.features}{Number of top variable features to get}

\item{mean.cutoff}{Cutoff for feature means, must be a two-length numeric for low and high values}

\item{dispersion.cutoff}{Cutoff for feature dispersion values, must be a two-length numeric for low and high values}

\item{selection.method}{Specifies how to select the features to store in @var.genes.
\itemize{
  \item{mean.var.plot: }{Default method, placing cutoffs on the mean variablility plot}
  \item{dispersion: }{Choose the top.features with the highest dispersion}
}}

\item{...}{Arguments passed to other methods}
}
\description{
Get the top variable Features
}
