% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_generics.R, R/preprocessing.R
\name{GetVariableGenes}
\alias{GetVariableGenes}
\alias{GetVariableGenes.loom}
\title{Get the variable genes dataframe}
\usage{
GetVariableGenes(object, ...)

\method{GetVariableGenes}{loom}(object)
}
\arguments{
\item{object}{An object}
}
\value{
A dataframe with four columns
\item{gene.means}{Gene means}
\item{gene.dispersion}{Gene dispersion}
\item{gene.dispersion.scaled}{Gene dispersion scaled}
\item{index}{Index for given gene in the dataset}
}
\description{
Get the variable genes dataframe
}
\section{Methods (by class)}{
\itemize{
\item \code{loom}: Get variable genes dataframe from a loom file
}}

