% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{Graph-class}
\alias{Graph-class}
\alias{Graph}
\title{The Graph Class}
\description{
The Graph class simply inherits from dgCMatrix. We do this to enable future expandability of graphs.
}
\seealso{
\code{\link[Matrix]{dgCMatrix-class}}
}
