% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Graphs}
\alias{Graphs}
\title{Pull Graph or Graph names}
\usage{
Graphs(object, slot = NULL)
}
\arguments{
\item{object}{A Seurat object}

\item{slot}{Name of Graph object}
}
\value{
If \code{slot} is \code{NULL}, the names of all \code{Graph} objects
in this Seurat object. Otherwise, the \code{Graph} object requested
}
\description{
Lists the names of \code{\link{Graph}} objects present in
a Seurat object. If slot is provided, pulls specified Graph object.
}
