% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{GroupCorrelationPlot}
\alias{GroupCorrelationPlot}
\title{Boxplot of correlation of a variable (e.g. number of UMIs) with expression
data}
\usage{
GroupCorrelationPlot(
  object,
  assay = NULL,
  feature.group = "feature.grp",
  cor = "nCount_RNA_cor"
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Assay where the feature grouping info and correlations are
stored}

\item{feature.group}{Name of the column in meta.features where the feature
grouping info is stored}

\item{cor}{Name of the column in meta.features where correlation info is
stored}
}
\value{
Returns a ggplot boxplot of correlations split by group
}
\description{
Boxplot of correlation of a variable (e.g. number of UMIs) with expression
data
}
\concept{visualization}
