% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{GroupedMultiCCA}
\alias{GroupedMultiCCA}
\title{Perform Canonical Correlation Analysis with more than two groups, without memory-intensive merging operations.}
\usage{
GroupedMultiCCA(object, group.by, genes.use = object@var.genes, num.ccs = 1,
  standardize = TRUE, ...)
}
\arguments{
\item{genes.use}{Genes to use in mCCA.}

\item{num.ccs}{Number of canonical vectors to calculate}

\item{standardize}{standardize scale.data matrices to be centered (mean zero)
and scaled to have a standard deviation of 1.}

\item{...}{Additional parameters for InternalMultiCCA, such as niter.}

\item{object.list}{List of Seurat objects}
}
\value{
Returns the Seurat object with CCA stored in the @dr$cca slot.
}
\description{
Runs a canonical correlation analysis
}
\examples{
pbmc_small

pbmc_small@meta.data$group <- sample( paste0( "group", 1:3 ) , replace = T )
pbmc_cca <- LiteMultiCCA(object = pbmc_small, genes.use = pbmc_small@var.genes, num.ccs = 4)
# Print results
PrintDim(pbmc_cca,reduction.type = 'cca')

}
