% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{HVFInfo}
\alias{HVFInfo}
\alias{HVFInfo.Assay}
\alias{HVFInfo.Seurat}
\title{Get highly variable feature information}
\usage{
HVFInfo(object, ...)

\method{HVFInfo}{Assay}(object, ...)

\method{HVFInfo}{Seurat}(object, assay = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{assay}{Name of assay to pull highly variable feature information for}
}
\value{
A dataframe with feature means, dispersion, and scaled dispersion
}
\description{
Get highly variable feature information
}
