% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{HVFInfo.SCTAssay}
\alias{HVFInfo.SCTAssay}
\title{Get Variable Feature Information}
\usage{
\method{HVFInfo}{SCTAssay}(object, selection.method, status = FALSE, ...)
}
\description{
Get variable feature information from \code{\link{SCTAssay}} objects
}
\examples{
# Get the HVF info directly from an SCTAssay object
pbmc_small <- SCTransform(pbmc_small)
HVFInfo(pbmc_small[["SCT"]], selection.method = 'sct')[1:5, ]

}
\seealso{
\code{\link[SeuratObject]{HVFInfo}}
}
