% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seurat.R
\name{ica.plot}
\alias{ica.plot}
\title{Plot ICA map}
\usage{
ica.plot(object, ...)
}
\arguments{
\item{object}{Seurat object}

\item{\dots}{Additional parameters to dim.plot, for example, which dimensions to plot.}
}
\description{
Graphs the output of a ICA analysis
Cells are colored by their identity class.
}
\details{
This function is a wrapper for dim.plot. See ?dim.plot for a full list of possible
arguments which can be passed in here.
}

