% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ICAPlot}
\alias{ICAPlot}
\title{Plot ICA map}
\usage{
ICAPlot(object, ...)
}
\arguments{
\item{object}{Seurat object}

\item{\dots}{Additional parameters to DimPlot, for example, which dimensions to plot.}
}
\description{
Graphs the output of a ICA analysis
Cells are colored by their identity class.
}
\details{
This function is a wrapper for DimPlot. See ?DimPlot for a full list of possible
arguments which can be passed in here.
}
\examples{
pbmc_small <- RunICA(object = pbmc_small, ics.compute = 25, print.results = FALSE)
ICAPlot(object = pbmc_small)

}
