% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{ICARot}
\alias{ICARot}
\title{ICA Rotation Accessor Function}
\usage{
ICARot(object, dims.use = NULL, cells.use = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{dims.use, }{Dimensions to include (default is all stored dims)}

\item{cells.use, }{Cells to include (default is all cells)}
}
\value{
ICA rotation matrix for given cells and ICs
}
\description{
Pull ICA rotation matrix
}
