% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{ICAX}
\alias{ICAX}
\title{ICA Gene Loadings Accessor Function}
\usage{
ICAX(object, dims.use = NULL, genes.use = NULL, use.full = F)
}
\arguments{
\item{object}{Seurat object}

\item{dims.use, }{Dimensions to include (default is all stored dims)}

\item{genes.use, }{Genes to include (default is all)}
}
\value{
ICA loading matrix for given genes and ICs
}
\description{
Pull ICA Gene Loadings
}

