% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{icHeatmap}
\alias{icHeatmap}
\title{Independent component heatmap}
\usage{
icHeatmap(object, ic.use = 1, cells.use = NULL, num.genes = 30,
  disp.min = -2.5, disp.max = 2.5, do.return = FALSE, col.use = pyCols,
  use.scale = TRUE, do.balanced = FALSE, ...)
}
\arguments{
\item{object}{Seurat object}

\item{ic.use}{Independent components to use}

\item{cells.use}{Cells to include in the heatmap (default is all cells)}

\item{num.genes}{Number of genes to return}

\item{disp.min}{Minimum display value (all values below are clipped)}

\item{disp.max}{Maximum display value (all values above are clipped)}

\item{do.return}{Default is FALSE. If TRUE, return a matrix of scaled values
which would be passed to heatmap.2}

\item{col.use}{Color palette to use}

\item{use.scale}{Default is TRUE: plot scaled data. If FALSE, plot raw data on the heatmap.}

\item{do.balanced}{Return an equal number of genes with both + and - IC scores.}

\item{...}{Additional parameters to heatmap.2. Common examples are cexRow
and cexCol, which set row and column text sizes}
}
\value{
If do.return==TRUE, a matrix of scaled values which would be passed
to heatmap.2. Otherwise, no return value, only a graphical output
}
\description{
Draws a heatmap focusing on a principal component. Both cells and genes are sorted by their principal component scores. Allows for nice visualization of sources of heterogeneity in the dataset.
}

