% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Idents}
\alias{Idents}
\alias{Idents<-}
\alias{StashIdent}
\alias{Idents.Seurat}
\alias{Idents<-.Seurat}
\alias{StashIdent.Seurat}
\title{Get an object's cell identities}
\usage{
Idents(object, ...)

Idents(object, ...) <- value

StashIdent(object, save.name, ...)

\method{Idents}{Seurat}(object, ...)

\method{Idents}{Seurat}(object, cells.use = NULL, ...) <- value

\method{StashIdent}{Seurat}(object, save.name = "orig.ident", ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{The name of the identites to pull or the identities themselves}

\item{save.name}{Store current identity information under this name}

\item{cells.use}{Set cell identities for specific cells}
}
\value{
The cell identies

An object with the cell identites changed

An object with the identities stashed
}
\description{
Get an object's cell identities

Stash an object's identity information
}
\section{Methods (by class)}{
\itemize{
\item \code{Seurat}: Get the active identities of a Seurat object

\item \code{Seurat}: Set the active identities of a Seurat object

\item \code{Seurat}: Stash cell identities of a Seurat object
}}

\examples{
head(x = pbmc_small[])
pbmc_small <- StashIdent(object = pbmc_small, save.name = 'cluster.ident')
head(x = pbmc_small[])

}
