% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{IFeaturePlot}
\alias{IFeaturePlot}
\title{Visualize features in dimensional reduction space interactively}
\usage{
IFeaturePlot(object, feature, dims = c(1, 2), reduction = NULL, slot = "data")
}
\arguments{
\item{object}{Seurat object}

\item{feature}{Feature to plot}

\item{dims}{Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions}

\item{reduction}{Which dimensionality reduction to use. If not specified, first searches for umap, then tsne, then pca}

\item{slot}{Which slot to pull expression data from?}
}
\value{
Returns the final plot as a ggplot object
}
\description{
Visualize features in dimensional reduction space interactively
}
\concept{visualization}
