% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/spatial.R
\name{ImageData}
\alias{ImageData}
\alias{GetSlice}
\alias{SetSlice}
\alias{SetTissueCoordinates}
\alias{GetSlice.Seurat}
\alias{GetSlice.SpatialAssay}
\alias{SetSlice.Assay}
\alias{SetSlice.Seurat}
\alias{SetSlice.SpatialAssay}
\title{Get and set slice image data}
\usage{
GetSlice(object, slice = NULL, ...)

SetSlice(object, image, slice = NULL, ...)

SetTissueCoordinates(object, coordinates, ...)

\method{GetSlice}{Seurat}(object, slice = NULL, assay = NULL, ...)

\method{GetSlice}{SpatialAssay}(object, slice = NULL, ...)

\method{SetSlice}{Assay}(object, image, slice = NULL, ...)

\method{SetSlice}{Seurat}(object, image, slice = NULL, assay = NULL,
  ...)

\method{SetSlice}{SpatialAssay}(object, image, slice = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{slice}{Name of slice to get/set}

\item{...}{Arguments passed to other methods}

\item{image}{An object of class \code{SliceImage}}

\item{assay}{Name of assay to get \code{SliceImage} from or set in. Note: will
attempt to convert non-\code{SpatialAssay} objects into a \code{SpatialAssay}}
}
\description{
Get and set slice image data
}
