% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{ImageDimPlot}
\alias{ImageDimPlot}
\title{Spatial Plots}
\usage{
ImageDimPlot(
  object,
  images = NULL,
  layers = NULL,
  group.by = NULL,
  split.by = NULL,
  cols = NULL,
  molecules = NULL,
  mols.size = 0.1,
  mols.cols = NULL,
  nmols = 1000,
  alpha = molecules \%iff\% 0.3 \%||\% 0.6,
  border.color = "black",
  border.size = 0.3,
  na.value = "grey50",
  crop = NA,
  cells = NULL,
  overlap = FALSE,
  combine = TRUE
)
}
\arguments{
\item{object}{A \code{\link[SeuratObject]{Seurat}} object}

\item{images}{A vector if images to plot}

\item{layers}{A vector of segmentation layers per image to plot; can be a
character vector, a named character vector, or a named list. Names should
be the names of images and values should be the names of segmentation layers}

\item{group.by}{Name of one or more metadata columns to group (color) cells by
(for example, orig.ident); pass 'ident' to group by identity class}

\item{split.by}{Name of a metadata column to split plot by;
see \code{\link{FetchData}} for more details}

\item{cols}{Vector of colors, each color corresponds to an identity class. This may also be a single character
or numeric value corresponding to a palette as specified by \code{\link[RColorBrewer]{brewer.pal.info}}.
By default, ggplot2 assigns colors. We also include a number of palettes from the pals package.
See \code{\link{DiscretePalette}} for details.}

\item{na.value}{Color value for NA points when using custom scale}

\item{crop}{Crop the plots to area with cells only; choose from:
\itemize{
 \item \code{TRUE} to crop all plots to the area constrained by the layer
 \item \code{NA} to crop all plots to the area constrained by the image
 \item \code{FALSE} to skip cropping
}}

\item{cells}{Vector of cells to plot (default is all cells)}

\item{overlap}{Overlay layers from a single image to create a single plot;
if \code{TRUE}, then layers are stacked in the order they're given
(first is lowest)}

\item{combine}{Combine plots into a single
\code{\link[patchwork]{patchworked}} ggplot object.If \code{FALSE},
return a list of ggplot objects}
}
\value{
If \code{combine = TRUE}, a \code{\link[patchwork]{patchwork}ed}
ggplot object; otherwise, a list of ggplot objects
}
\description{
Spatial Plots
}
