% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Images}
\alias{Images}
\title{Pull spatial image names}
\usage{
Images(object, assay = NULL)
}
\arguments{
\item{object}{A \code{Seurat} object}

\item{assay}{Name of assay to limit search to}
}
\value{
A list of image names
}
\description{
List the names of \code{SpatialImage} objects present in a \code{Seurat} object.
If \code{assay} is provided, limits search to images associated with that assay
}
\examples{
\dontrun{
Images(object)
}

}
