% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Index}
\alias{Index}
\alias{Index<-}
\alias{Index.Neighbor}
\alias{Index<-.Neighbor}
\title{Get Neighbor algorithm index}
\usage{
Index(object, ...)

Index(object, ...) <- value

\method{Index}{Neighbor}(object)

\method{Index}{Neighbor}(object, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods;}

\item{value}{The index to store}
}
\value{
Returns the value in the alg.idx slot of the Neighbor object

\code{Idents<-}: A Neighbor bject with the index stored
}
\description{
Get Neighbor algorithm index
}
