% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{InitializeWorkflow}
\alias{InitializeWorkflow}
\title{Initialize a Seurat workflow}
\usage{
InitializeWorkflow(object, file)
}
\arguments{
\item{object}{Seurat object}

\item{file}{Ini configuration file. See cluster.workflow.ini for an example.}
}
\value{
Object with modified workflows
}
\description{
Reads dependencies from a file and initializes a Seurat workflow
}
\examples{
\dontrun{
pbmc_small <- InitializeWorkflow(object = pbmc_small, file = 'workflows/cluster.workflow.txt')
}

}
