% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{IntegrateData}
\alias{IntegrateData}
\title{Integrate data}
\usage{
IntegrateData(anchorset, new.assay.name = "integrated",
  features = NULL, features.to.integrate = NULL, dims = 1:30,
  k.weight = 100, sd.weight = 1, sample.tree = NULL,
  preserve.order = FALSE, do.cpp = TRUE, eps = 0, verbose = TRUE)
}
\arguments{
\item{anchorset}{Results from FindIntegrationAnchors}

\item{new.assay.name}{Name for the new assay containing the integrated data}

\item{features}{Vector of features to use when computing the PCA to determine the weights. Only set
if you want a different set from those used in the anchor finding process}

\item{features.to.integrate}{Vector of features to integrate. By default, will use the features
used in anchor finding.}

\item{dims}{Number of PCs to use in the weighting procedure}

\item{k.weight}{Number of neighbors to consider when weighting}

\item{sd.weight}{Controls the bandwidth of the Gaussian kernel for weighting}

\item{sample.tree}{Specify the order of integration. If null, will compute automatically.}

\item{preserve.order}{Integrate the objects in the order they were provided to FindAnchors}

\item{do.cpp}{Run cpp code where applicable}

\item{eps}{Error bound on the neighbor finding algorithm (from RANN)}

\item{verbose}{Print progress bars and output}
}
\value{
Returns a Seurat object with a new integrated Assay
}
\description{
Integrates the data
}
