% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{IntegrationAnchorSet-class}
\alias{IntegrationAnchorSet-class}
\alias{IntegrationAnchorSet}
\title{The IntegrationAnchorSet Class}
\description{
Inherits from the Anchorset class. Implemented mainly for method dispatch
purposes.  See \code{\link{AnchorSet}} for slot details.
}
\concept{objects}
