% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{IntegrationData-class}
\alias{IntegrationData-class}
\alias{IntegrationData}
\title{The IntegrationData Class}
\description{
The IntegrationData object is an intermediate storage container used internally throughout the
integration procedure to hold bits of data that are useful downstream.
}
\section{Slots}{

\describe{
\item{\code{neighbors}}{List of neighborhood information for cells (outputs of \code{RANN::nn2})}

\item{\code{weights}}{Anchor weight matrix}

\item{\code{integration.matrix}}{Integration matrix}

\item{\code{anchors}}{Anchor matrix}

\item{\code{offsets}}{The offsets used to enable cell look up in downstream functions}

\item{\code{objects.ncell}}{Number of cells in each object in the object.list}

\item{\code{sample.tree}}{Sample tree used for ordering multi-dataset integration}
}}

\concept{objects}
