% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction_internal.R
\name{InternalMultiCCA}
\alias{InternalMultiCCA}
\title{MultiCCA helper function - performs CCA}
\usage{
InternalMultiCCA(mat.list, num.ccs, num.sets = length(mat.list), niter = 25)
}
\arguments{
\item{mat.list}{list of input matrices}

\item{num.ccs}{latent dimension to use}

\item{num.sets}{How many input matrices}

\item{niter}{Number of iterations to perform. Set by default to 25.}
}
\description{
MultiCCA helper function - performs CCA
}
