% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{IsGlobal}
\alias{IsGlobal}
\alias{IsGlobal.DimReduc}
\alias{as.logical.DimReduc}
\title{Is an object global or persistent?}
\usage{
IsGlobal(object, ...)

\method{IsGlobal}{DimReduc}(object, ...)

\method{as.logical}{DimReduc}(x, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
\code{TRUE} if the object is global or persistent otherwise \code{FALSE}
}
\description{
Is an object global or persistent?
}
\examples{
IsGlobal(pbmc_small[['pca']])

}
