% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{ISpatialDimPlot}
\alias{ISpatialDimPlot}
\title{Visualize clusters spatially and interactively}
\usage{
ISpatialDimPlot(object, image = NULL, group.by = NULL, alpha = c(0.3, 1))
}
\arguments{
\item{object}{Seurat object}

\item{image}{Name of the image to use in the plot}

\item{group.by}{Name of one or more metadata columns to group (color) cells by
(for example, orig.ident); pass 'ident' to group by identity class}

\item{alpha}{Alpha value for plotting (default is 1)}
}
\value{
Returns final plot as a ggplot object
}
\description{
Visualize clusters spatially and interactively
}
\concept{spatial}
\concept{visualization}
