% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{ISpatialFeaturePlot}
\alias{ISpatialFeaturePlot}
\title{Visualize features spatially and interactively}
\usage{
ISpatialFeaturePlot(
  object,
  feature,
  image = NULL,
  slot = "data",
  alpha = c(0.1, 1)
)
}
\arguments{
\item{object}{Seurat object}

\item{image}{Name of the image to use in the plot}

\item{slot}{Which slot to pull expression data from?}

\item{alpha}{Controls opacity of spots. Provide as a vector specifying the
min and max}
}
\value{
Runs an interactive spatial feature plot in a Shiny gadget session;
does not return any value
}
\description{
Visualize features spatially and interactively
}
