% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{JackStrawData-class}
\alias{JackStrawData-class}
\alias{JackStrawData}
\title{The JackStrawData Class}
\description{
The JackStrawData is used to store the results of a JackStraw computation.
}
\section{Slots}{

\describe{
\item{\code{empirical.p.values}}{Empirical p-values}

\item{\code{fake.reduction.scores}}{Fake reduction scores}

\item{\code{empirical.p.values.full}}{Empirical p-values on full}

\item{\code{overall.p.values}}{Overall p-values from ScoreJackStraw}
}}

