% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration5.R
\name{JointPCAIntegration}
\alias{JointPCAIntegration}
\title{Seurat-Joint PCA Integration}
\usage{
JointPCAIntegration(
  object = NULL,
  assay = NULL,
  layers = NULL,
  orig = NULL,
  new.reduction = "integrated.dr",
  reference = NULL,
  features = NULL,
  normalization.method = c("LogNormalize", "SCT"),
  dims = 1:30,
  k.anchor = 20,
  scale.layer = "scale.data",
  dims.to.integrate = NULL,
  k.weight = 100,
  weight.reduction = NULL,
  sd.weight = 1,
  sample.tree = NULL,
  preserve.order = FALSE,
  groups = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{Seurat} object}

\item{assay}{Name of \code{Assay} in the \code{Seurat} object}

\item{layers}{Names of layers in \code{assay}}

\item{orig}{A \link[SeuratObject:DimReduc]{dimensional reduction} to correct}

\item{new.reduction}{Name of new integrated dimensional reduction}

\item{reference}{A reference \code{Seurat} object}

\item{features}{A vector of features to use for integration}

\item{normalization.method}{Name of normalization method used: LogNormalize
or SCT}

\item{dims}{Dimensions of dimensional reduction to use for integration}

\item{k.anchor}{How many neighbors (k) to use when picking anchors}

\item{scale.layer}{Name of scaled layer in \code{Assay}}

\item{verbose}{Print progress}

\item{groups}{A one-column data frame with grouping information}

\item{...}{Additional arguments passed to \code{FindIntegrationAnchors}}
}
\description{
Seurat-Joint PCA Integration
}
