% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{Kclust_dimension}
\alias{Kclust_dimension}
\title{Perform spectral k-means clustering on single cells}
\usage{
Kclust_dimension(object, dim.1 = 1, dim.2 = 2, cells.use = NULL,
  pt.size = 4, reduction.use = "tsne", k.use = 5, set.ident = FALSE,
  seed.use = 1, ...)
}
\description{
Find point clounds single cells in a low-dimensional space using k-means clustering.
}
\details{
CAn be useful for smaller datasets, not documented here yet
}

