% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_determination.R
\name{KClustDimension}
\alias{KClustDimension}
\title{Perform spectral k-means clustering on single cells}
\usage{
KClustDimension(object, dims.use = c(1, 2), cells.use = NULL, pt.size = 4,
  reduction.use = "tsne", k.use = 5, set.ident = T, seed.use = 1, ...)
}
\description{
Find point clounds single cells in a low-dimensional space using k-means clustering.
Can be useful for smaller datasets, where graph-based clustering can perform poorly
TODO : add documentation here
}
