% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreduc_generics.R, R/assay.R, R/dimreduc.R
\name{Key}
\alias{Key}
\alias{Key<-}
\alias{Key.Assay}
\alias{Key<-.Assay}
\alias{Key.DimReduc}
\title{Get a key}
\usage{
Key(object, ...)

Key(object, ...) <- value

\method{Key}{Assay}(object, ...)

\method{Key}{Assay}(object, ...) <- value

\method{Key}{DimReduc}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{Key value}
}
\description{
Get a key

Set a key
}
\section{Methods (by class)}{
\itemize{
\item \code{Assay}: Get the key for an Assay object

\item \code{Assay}: Set the key for an Assay object

\item \code{DimReduc}: Get the key for a DimReduc object
}}

