% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{L2CCA}
\alias{L2CCA}
\title{L2-Normalize CCA}
\usage{
L2CCA(object, ...)
}
\arguments{
\item{object}{Seurat object}

\item{\dots}{Additional parameters to L2Dim.}
}
\description{
Perform l2 normalization on CCs
}
\concept{dimensional_reduction}
