% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seurat.R
\name{lasso_preds_expr}
\alias{lasso_preds_expr}
\title{Impute expression of landmarked genes based on other genes via LASSO
linear modal}
\usage{
lasso_preds_expr(x, y, s.use = 20, y.name = NULL, do.print = FALSE,
  use.gram = TRUE)
}
\arguments{
\item{x}{Matrix of predictors(columns), observed genes expression.}

\item{y}{Vector of response, observed expression of given landmarked gene.}

\item{s.use}{Numeric. See \link{predict.lars}'s parameter \code{s}.}

\item{y.name}{Chractor. Name of the given landmarked gene.}

\item{do.print}{Logic. Whether print \code{y.name} on screen.}

\item{use.gram}{Logic. See \link{lars}'s parameter \code{use.Gram}}
}
\value{
Returns the imputed landmarked gene expession via LASSO modal.
}
\description{
Impute expression of landmarked genes based on other genes via LASSO
linear modal
}

