% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{LeverageScoreSampling}
\alias{LeverageScoreSampling}
\title{Sampling cells from objects based on Leverage score}
\usage{
LeverageScoreSampling(
  object,
  num.cells = 5000,
  assay = NULL,
  features = NULL,
  var.name = "leverage.score",
  over.write = FALSE,
  seed = 123,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{num.cells}{Number of sampled cells (default is 5000)}

\item{assay}{Assay used to calculate leverage score}

\item{features}{Features used to calculate leverage score}

\item{var.name}{Variable name stored leverage score in the meta.data (default is 'leverage.score')}

\item{over.write}{Whether to over write the variable with leverage score (default is FALSE)}

\item{seed}{Set a random seed (default is 123)}

\item{...}{Arguments passed to LeverageScore}
}
\value{
Returns a subset Seurat object with sampled cells
}
\description{
Sampling cells from objects based on Leverage score
}
