% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{LinkedPlots}
\alias{LinkedPlots}
\alias{LinkedFeaturePlot}
\alias{LinkedPlot}
\alias{LinkedDimPlot}
\title{Visualize spatial and clustering (dimensional reduction) data in a linked,
interactive framework}
\usage{
LinkedFeaturePlot(
  object,
  feature,
  dims = 1:2,
  reduction = NULL,
  pt.size = 6,
  image = NULL,
  slot = "data",
  min.cutoff = NA,
  max.cutoff = NA
)

LinkedDimPlot(
  object,
  dims = 1:2,
  reduction = NULL,
  pt.size = 6,
  image = NULL,
  group.by = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{feature}{Feature to visualize}

\item{dims}{Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions}

\item{reduction}{Which dimensionality reduction to use. If not specified, first searches for umap, then tsne, then pca}

\item{pt.size}{Adjust point size for plotting}

\item{image}{Which image to use}

\item{slot}{Which slot to pull expression data from?}

\item{min.cutoff}{Vector of minimum and maximum cutoff values for each feature,
may specify quantile in the form of 'q##' where '##' is the quantile (eg, 'q1', 'q10')}

\item{max.cutoff}{Vector of minimum and maximum cutoff values for each feature,
may specify quantile in the form of 'q##' where '##' is the quantile (eg, 'q1', 'q10')}

\item{group.by}{Name of meta.data column to group the data by}
}
\value{
A \code{\link[htmltools]{browsable}} HTML element with the spatial plot
on the left and the dimensional-reduction plot on the right
}
\description{
Visualize spatial and clustering (dimensional reduction) data in a linked,
interactive framework
}
\examples{
\dontrun{
LinkedFeaturePlot(seurat.object, feature = 'Hpca')
LinkedDimPlot(seurat.object)
}

}
