% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{LinkPlots}
\alias{LinkPlots}
\title{Link two ggplot plots together}
\usage{
LinkPlots(
  plot1,
  plot2,
  plot1.labels = TRUE,
  plot2.labels = TRUE,
  information = NULL,
  pt.size = 6,
  plot1.cols = eval(formals(fun = GGpointToBase)$cols),
  plot2.cols = eval(formals(fun = GGpointToBase)$cols),
  plot1.layout = list(),
  plot2.layout = list()
)
}
\arguments{
\item{plot1, plot2}{\code{ggplot} objects}

\item{plot1.labels, plot2.labels}{Logical values to inherit X/Y labels from
component \code{ggplot} objects}

\item{information}{An optional dataframe or matrix of extra information to be displayed on hover}

\item{pt.size}{Point size in px}

\item{plot1.cols, plot2.cols}{A named vector of column names to pull. Vector
names must be 'x', 'y', 'colour', 'shape', and/or 'size'; vector values must
be the names of columns in plot data that correspond to these values. May
pass only values that differ from the default
(eg. \code{cols = c('size' = 'point.size.factor')})}

\item{plot1.layout, plot2.layout}{Extra information for \code{\link[plotly]{layout}}}
}
\value{
A \code{\link[htmltools]{browsable}} HTML element with the two plots
linked and side-by-side
}
\description{
Link two ggplot plots together
}
\seealso{
\code{\link[plotly]{layout}}
}
