% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{Load10X_Spatial}
\alias{Load10X_Spatial}
\title{Load a 10X Genomics Visium Spatial Experiment into a \code{Seurat} object}
\usage{
Load10X_Spatial(data.dir, assay = "Spatial", slice = NULL,
  filter.matrix = TRUE, ...)
}
\arguments{
\item{data.dir}{Directory containing the matrix.mtx, genes.tsv (or features.tsv), and barcodes.tsv
files provided by 10X. A vector or named vector can be given in order to load
several data directories. If a named vector is given, the cell barcode names
will be prefixed with the name.}

\item{...}{Arguments passed to \code{\link{Read10X_h5}}}
}
\value{
A \code{Seurat} object
}
\description{
Load a 10X Genomics Visium Spatial Experiment into a \code{Seurat} object
}
