% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{LoadAkoya}
\alias{LoadAkoya}
\title{Load Akoya CODEX data}
\usage{
LoadAkoya(mtx, type = "inForm", fov = "fov")
}
\arguments{
\item{mtx}{Path to CSV generated by CODEX Processor version 1.7.0.6}

\item{type}{One of the following values depending on how data was preprocessed
\itemize{
 \item \dQuote{\code{processor}}: matrix generated by CODEX Processor
 \item \dQuote{\code{inForm}}: matrix generated by inForm
 \item \dQuote{\code{QuPath}}: matrix generated by QuPath
}}

\item{fov}{Name of fov to store in fovs slot of Seurat object}
}
\value{
Seurat object
}
\description{
Load Akoya CODEX data
}
\concept{preprocessing}
