% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{LoadAnnoyIndex}
\alias{LoadAnnoyIndex}
\title{Load the Annoy index file}
\usage{
LoadAnnoyIndex(object, file)
}
\arguments{
\item{object}{Neighbor object}

\item{file}{Path to file with annoy index}
}
\value{
Returns the Neighbor object with the index stored
}
\description{
Load the Annoy index file
}
\concept{utilities}
