% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{LoadCurioSeeker}
\alias{LoadCurioSeeker}
\title{Load Curio Seeker data}
\usage{
LoadCurioSeeker(data.dir, assay = "Spatial")
}
\arguments{
\item{data.dir}{location of data directory that contains the counts matrix,
gene names, barcodes/beads, and barcodes/bead location files.}

\item{assay}{Name of assay to associate spatial data to}
}
\value{
A \code{\link{Seurat}} object
}
\description{
Load Curio Seeker data
}
\concept{preprocessing}
