% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{LoadHuBMAPCODEX}
\alias{LoadHuBMAPCODEX}
\title{Load HuBMAP CODEX data}
\usage{
LoadHuBMAPCODEX(data.dir, fov = "fov")
}
\arguments{
\item{data.dir}{Path to a directory containing Vitessce cells and clusters JSONs}

\item{fov}{Name of fov to store in fovs slot of Seurat object}
}
\value{
Seurat object
}
\description{
Load HuBMAP CODEX data
}
\concept{preprocessing}
