% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreduc_generics.R, R/dimreduc.R
\name{Loadings}
\alias{Loadings}
\alias{Loadings.DimReduc}
\title{Get feature loadings}
\usage{
Loadings(object, projected, ...)

\method{Loadings}{DimReduc}(object, projected = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{projected}{Pull the projected feature loadings?}

\item{...}{Arguments passed to other methods}
}
\description{
Get feature loadings
}
\section{Methods (by class)}{
\itemize{
\item \code{DimReduc}: Get the feature loadings from a DimReduc object
}}

