% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{LoadNanostring}
\alias{LoadNanostring}
\title{Load Nanostring SMI data}
\usage{
LoadNanostring(data.dir, fov = "fov")
}
\arguments{
\item{data.dir}{Path to folder containing Nanostring SMI outputs}

\item{fov}{Name of fov to store in fovs slot of Seurat object}
}
\value{
Seurat object
}
\description{
Load Nanostring SMI data
}
\concept{preprocessing}
