% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{LoadVizgen}
\alias{LoadVizgen}
\title{Load Vizgen data}
\usage{
LoadVizgen(data.dir, fov = "fov", z.slice = 3L)
}
\arguments{
\item{data.dir}{Path to folder containing Vizgen MERFISH files}

\item{fov}{Name of fov to store in fovs slot of Seurat object}

\item{z.slice}{Z slice to load; must be between 0 and 6, inclusive}
}
\value{
Seurat object
}
\description{
Load Vizgen data
}
\concept{preprocessing}
