% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{localMaxima}
\alias{localMaxima}
\title{Local maxima estimator}
\usage{
localMaxima(x)
}
\arguments{
\item{x}{A continuous vector}
}
\value{
Returns a (named) vector showing positions of local maximas
}
\description{
Finding local maxima given a numeric vector
Written by Chris McGinnis from Gartner Lab, UCSF
From Tommy @ https://stackoverflow.com/questions/6836409/finding-local-maxima-and-minima
}
\examples{
extrema <- localMaxima(x)

}
