% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{LogSeuratCommand}
\alias{LogSeuratCommand}
\title{Log a command}
\usage{
LogSeuratCommand(object, return.command = FALSE)
}
\arguments{
\item{object}{Name of Seurat object}

\item{return.command}{Return a \link{SeuratCommand} object instead}
}
\value{
If \code{return.command}, returns a SeuratCommand object. Otherwise,
returns the Seurat object with command stored
}
\description{
Logs command run, storing the name, timestamp, and argument list. Stores in
the Seurat object
}
\seealso{
\code{\link{Command}}
}
