% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{logVarDivMean}
\alias{logVarDivMean}
\title{Calculate the dispersion of logged values}
\usage{
logVarDivMean(x)
}
\arguments{
\item{x}{value or vector of values}
}
\value{
Returns the dispersion in log-space
}
\description{
Calculate the dispersion (variance divided by mean) in non-log space
(return answer in log-space)
}
