% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{LRDETest}
\alias{LRDETest}
\title{Differential expression testing using logistic regression}
\usage{
LRDETest(object, cells.1 = NULL, cells.2 = NULL, genes.to.test = NULL,
  ecs.to.test = NULL, txs.to.test = NULL, use.tcc = FALSE,
  use.tx = FALSE, transcript.threshold = 0.9, display.progress = TRUE,
  assay.type = "RNA", ambig = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{cells.1}{Group 1 cells}

\item{cells.2}{Group 2 cells}

\item{genes.to.test}{Vector of gene to test}

\item{ecs.to.test}{Vector of equivalence classes to test}

\item{txs.to.test}{Vector of transcripts to test}

\item{use.tcc}{When testing genes, build LR model using TCCs mapping to gene
as predictors}

\item{use.tx}{When testing genes, build LR model using TXs mapping to gene as
predictors}

\item{transcript.threshold}{When testing genes using transcripts, filter any
trascript not expressed in at least this fraction of cells.}

\item{display.progress}{Print messages and progress bar}

\item{assay.type}{Type of assay to perform DE for (default is RNA)}

\item{ambig}{Allow ambiguously mapping TCCs/Txs?  (default is FALSE)}
}
\value{
Returns a matrix with the p-value results
}
\description{
Identifies differentially expressed genes between two groups of cells using a
logistic regression framework.
}
