% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreduc.R
\name{MakeDimReducObject}
\alias{MakeDimReducObject}
\title{Make a DimReduc object}
\usage{
MakeDimReducObject(cell.embeddings = matrix(), feature.loadings = matrix(),
  feature.loadings.projected = matrix(), assay.used = NULL,
  stdev = numeric(), key = NULL, jackstraw = NULL, misc = list(), ...)
}
\arguments{
\item{cell.embeddings}{...}

\item{feature.loadings}{...}

\item{feature.loadings.projected}{...}

\item{assay.used}{...}

\item{stdev}{...}

\item{key}{...}

\item{jackstraw}{...}

\item{misc}{...}

\item{...}{Ignored for now}
}
\description{
Make a DimReduc object
}
