% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{MakeSparse}
\alias{MakeSparse}
\title{Make object sparse
Converts stored data matrices to sparse matrices to save space}
\usage{
MakeSparse(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{
Returns a seurat object with object@raw.data and object@data converted to sparse matrices
}
\description{
Make object sparse
Converts stored data matrices to sparse matrices to save space
}

