% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seurat.R
\name{map.cell}
\alias{map.cell}
\title{Helper function of \code{initial_mapping}.}
\usage{
map.cell(object, cell.name, do.plot = FALSE, safe.use = TRUE,
  text.val = NULL, do.rev = FALSE)
}
\description{
Map one cell to bins.
}

