% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{MapQueryData}
\alias{MapQueryData}
\title{Map query cells to a reference}
\usage{
MapQueryData(
  reference,
  query,
  proj.reduction = "pcaproject",
  anchor.reduction = NULL,
  dims,
  anchorset = NULL,
  transfer.labels = NULL,
  transfer.expression = NULL,
  verbose = TRUE,
  k.nn = 20,
  k.weight = 50,
  sd.weight = 1,
  eps = 0,
  nn.method = "annoy",
  reference.neighbors = NULL,
  return.intermediate = FALSE
)
}
\arguments{
\item{reference}{\code{\link{Seurat}} object to use as the reference}

\item{query}{\code{\link{Seurat}} object to use as the query}

\item{proj.reduction}{Name of the projected dimensional reduction in the query
from \code{\link{FindTransferAnchors}} with \code{mapping = TRUE}}

\item{anchor.reduction}{Name of the dimensional reduction used to calculate
cell-anchor weights matrix}

\item{dims}{Number of dimensions to use in the anchor weighting procedure}

\item{anchorset}{An \code{\link{AnchorSet}} object generated by
\code{\link{FindTransferAnchors}} with \code{mapping = TRUE}}

\item{transfer.labels}{Identities to transfer to the query.
Can be a list of multiple identities to transfer.
The list must be named if it has >1 element, default name for 1 element is "id".
Element names are used to name metadata fields and prediction score assay.}

\item{transfer.expression}{Expression values to transfer to the query.
Can be a list of multiple expression matrices to transfer.
The list must be named if it has >1 element, default name for 1 element is
"transfer". Element names are used to name the transferred assays in the query.}

\item{verbose}{Print progress bars and output}

\item{k.nn}{How many neighbors (k) to use}

\item{k.weight}{Number of neighbors to consider when weighting anchors}

\item{sd.weight}{Controls the bandwidth of the Gaussian kernel for weighting}

\item{eps}{Error bound on the neighbor finding algorithm (from
\code{\link{RANN}})}

\item{nn.method}{Method for nearest neighbor finding. Options include: rann,
annoy}

\item{reference.neighbors}{Optional; name of the Neighbor to use from the reference
(generated by \code{\link{FindNeighbors}} with cache.index = TRUE)}

\item{return.intermediate}{Return intermediate results in the query object
(projected PCA as DimReduc, corrected PCA as DimReduc,
integration weights as IntegrationData)}
}
\value{
Returns a modified query object, containing: a Neighbor object (query_ref.nn)
to use with RunUMAP and a reduction model; predicted identities and prediction
scores in metadata and per-class prediction scores in an Assay for each identity
in transfer.labels; Assay with imputed values for each matrix in transfer.expression.
}
\description{
Correct batch effects of query projected dimension reduction,transfer
categorical or continuous data from reference datasets.
}
\details{
\code{refdata = GetAssayData(reference[['RNA']])})
}
