% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{marker.auc.test}
\alias{marker.auc.test}
\title{internal function to calculate AUC values}
\usage{
marker.auc.test(data1, data2, mygenes, print.bar = TRUE)
}
\description{
internal function to calculate AUC values
}
