% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{MeanVarPlot}
\alias{MeanVarPlot}
\title{Identify variable genes}
\usage{
MeanVarPlot(object, fxn.x = expMean, fxn.y = logVarDivMean,
  do.plot = TRUE, set.var.genes = TRUE, do.text = TRUE,
  x.low.cutoff = 0.1, x.high.cutoff = 8, y.cutoff = 1,
  y.high.cutoff = Inf, cex.use = 0.5, cex.text.use = 0.5,
  do.spike = FALSE, pch.use = 16, col.use = "black",
  spike.col.use = "red", plot.both = FALSE, do.contour = TRUE,
  contour.lwd = 3, contour.col = "white", contour.lty = 2, num.bin = 20,
  do.recalc = TRUE, sort.results = TRUE)
}
\arguments{
\item{object}{Seurat object}

\item{fxn.x}{Function to compute x-axis value (average expression). Default
is to take the mean of the detected (i.e. non-zero) values}

\item{fxn.y}{Function to compute y-axis value (dispersion). Default is to
take the standard deviation of all values/}

\item{do.plot}{Plot the average/dispersion relationship}

\item{set.var.genes}{Set object@var.genes to the identified variable genes
(default is TRUE)}

\item{do.text}{Add text names of variable genes to plot (default is TRUE)}

\item{x.low.cutoff}{Bottom cutoff on x-axis for identifying variable genes}

\item{x.high.cutoff}{Top cutoff on x-axis for identifying variable genes}

\item{y.cutoff}{Bottom cutoff on y-axis for identifying variable genes}

\item{y.high.cutoff}{Top cutoff on y-axis for identifying variable genes}

\item{cex.use}{Point size}

\item{cex.text.use}{Text size}

\item{do.spike}{FALSE by default. If TRUE, color all genes starting with ^ERCC a different color}

\item{pch.use}{Pch value for points}

\item{col.use}{Color to use}

\item{spike.col.use}{if do.spike, color for spike-in genes}

\item{plot.both}{Plot both the scaled and non-scaled graphs.}

\item{do.contour}{Draw contour lines calculated based on all genes}

\item{contour.lwd}{Contour line width}

\item{contour.col}{Contour line color}

\item{contour.lty}{Contour line type}

\item{num.bin}{Total number of bins to use in the scaled analysis (default
is 20)}

\item{do.recalc}{TRUE by default. If FALSE, plots and selects variable genes without recalculating statistics for each gene.}

\item{sort.results}{If TRUE (by default), sort results in object@mean.var in decreasing order of dispersion}
}
\value{
Returns a Seurat object, placing variable genes in object@var.genes.
The result of all analysis is stored in object@mean.var
}
\description{
Identifies genes that are outliers on a 'mean variability plot'. First, uses
a function to calculate average expression (fxn.x) and dispersion (fxn.y)
for each gene. Next, divides genes into num.bin (deafult 20) bins based on
their average expression, and calculates z-scores for dispersion within each
bin. The purpose of this is to identify variable genes while controlling for
the strong relationship between variability and average expression.
}
\details{
Exact parameter settings may vary empirically from dataset to dataset, and
based on visual inspection of the plot.
Setting the y.cutoff parameter to 2 identifies genes that are more than two standard
deviations away from the average dispersion within a bin. The default X-axis function
is the mean expression level, and for Y-axis it is the log(Variance/mean). All mean/variance
calculations are not performed in log-space, but the results are reported in log-space -
see relevant functions for exact details.
}
