% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_generics.R
\name{merge}
\alias{merge}
\title{Merge Seurat Objects}
\usage{
merge(x, y, ...)
}
\arguments{
\item{x}{Seurat object}

\item{y}{Seurat object (or multiple Seurat objects)}

\item{...}{additional arguments}
}
\value{
Seurat object
}
\description{
Merge Seurat Objects
}
