% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{merge.SCTAssay}
\alias{merge.SCTAssay}
\title{Merge SCTAssay objects}
\usage{
\method{merge}{SCTAssay}(
  x = NULL,
  y = NULL,
  add.cell.ids = NULL,
  merge.data = TRUE,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{\link[SeuratObject]{Seurat}} object}

\item{na.rm}{If na.rm = TRUE, this will only preserve residuals that are
present in all SCTAssays being merged. Otherwise, missing residuals will be
populated with NAs.}
}
\description{
Merge SCTAssay objects
}
\concept{objects}
