% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{MergeSeurat}
\alias{MergeSeurat}
\title{Merge Seurat Objects}
\usage{
MergeSeurat(object1, object2)
}
\arguments{
\item{object1}{First Seurat object to merge}

\item{object2}{Second Seurat object to merge}
}
\value{
Merged Seurat object
}
\description{
Merge two Seurat objects
}

