% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Misc}
\alias{Misc}
\alias{Misc<-}
\alias{Misc.Seurat}
\alias{Misc<-.Seurat}
\title{Access miscellaneous data}
\usage{
Misc(object, slot, ...)

Misc(object, slot, ...) <- value

\method{Misc}{Seurat}(object, slot = NULL, ...)

\method{Misc}{Seurat}(object, slot, ...) <- value
}
\arguments{
\item{object}{An object}

\item{slot}{Name of specific bit of meta data to pull}

\item{...}{Arguments passed to other methods}

\item{value}{Data to add}
}
\value{
Miscellaneous data

An object with miscellaneous data added
}
\description{
Access miscellaneous data

Set miscellaneous data
}
\section{Methods (by class)}{
\itemize{
\item \code{Seurat}: Get miscellaneous data from a Seurat object

\item \code{Seurat}: Set miscellaneous data for a Seurat object
}}

