% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{ModalityWeights-class}
\alias{ModalityWeights-class}
\alias{ModalityWeights}
\title{The ModalityWeights Class}
\description{
The ModalityWeights class is an intermediate data storage class that stores the modality weight and other
related information needed for performing downstream analyses - namely data integration
(\code{FindModalityWeights}) and data transfer (\code{\link{FindMultiModalNeighbors}}).
}
\section{Slots}{

\describe{
\item{\code{first.modality.weight}}{A vector of value representing for the modality weights of
the first modality}

\item{\code{modality.assay}}{Names of assays for the list of dimensional reductions}

\item{\code{params}}{A list of parameters used in the FindModalityWeights}

\item{\code{score.matrix}}{a score matrix representing cross and within-modality prediction
score, and kernel value}

\item{\code{command}}{Store log of parameters that were used}
}}

