% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{MultiModal_CCA}
\alias{MultiModal_CCA}
\title{Canonical Correlation Analysis between modalities}
\usage{
MultiModal_CCA(object, assay.1 = "RNA", assay.2 = "ADT",
  features.1 = NULL, features.2 = NULL, num.cc = 20,
  normalize.variance = TRUE)
}
\arguments{
\item{object}{Seurat object. Assumes that scale.data exist for both assays.}

\item{assay.1}{The first assay for CCA.}

\item{assay.2}{The second assay for CCA.}

\item{features.1}{Features to use for the first assay, default is all the features (use object@var.genes if this is RNA).}

\item{features.2}{Features to use for the second assay, default is all the features.}

\item{num.cc}{Minimal number of CCs to return.}

\item{normalize.variance}{Whether to scale the embeddings. Default is TRUE.}
}
\value{
Returns Seurat object with two CCA results stored (for example, object@dr$RNACCA and object@dr$ADTCCA).
}
\description{
Runs a canonical correlation analysis between two assays (for example, RNA and ADT from CITE-seq)
}
\examples{
\dontrun{
object <- MultiModal_CCA(object,assay.1 = "RNA",assay.2 = "ADT")
}

}
