% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFxns.R
\name{myPalette}
\alias{myPalette}
\title{Create a custom color palette}
\usage{
myPalette(low = "white", high = "red", mid = NULL, k = 50)
}
\arguments{
\item{low}{low color}

\item{high}{high color}

\item{mid}{middle color. Optional.}

\item{k}{number of steps (colors levels) to include between low and high values}
}
\description{
Creates a custom color palette based on low, middle, and high color values
}
