% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{NegBinomRegDETest}
\alias{NegBinomRegDETest}
\title{Negative binomial test for UMI-count based data (regularized version)}
\usage{
NegBinomRegDETest(object, cells.1, cells.2, genes.use = NULL,
  latent.vars = NULL, print.bar = TRUE, min.cells = 3)
}
\arguments{
\item{object}{Seurat object}

\item{cells.1}{Group 1 cells}

\item{cells.2}{Group 2 cells}

\item{genes.use}{Genes to test. Default is to use all genes}

\item{print.bar}{Print a progress bar once expression testing begins (uses pbapply to do this)}

\item{min.cells}{Minimum number of cells expressing the gene in at least one of the two groups}
}
\value{
Returns a p-value ranked data frame of test results.
}
\description{
Identifies differentially expressed genes between two groups of cells using
a likelihood ratio test of negative binomial generalized linear models where
the overdispersion parameter theta is determined by pooling information
across genes.
}
