% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Neighbors}
\alias{Neighbors}
\title{Pull Neighbor or Neighbor names}
\usage{
Neighbors(object, slot = NULL)
}
\arguments{
\item{object}{A Seurat object}

\item{slot}{Name of Neighbor object}
}
\value{
If \code{slot} is \code{NULL}, the names of all \code{Neighbor} objects
in this Seurat object. Otherwise, the \code{Neighbor} object requested
}
\description{
Lists the names of \code{\link{Neighbor}} objects present in
a Seurat object. If slot is provided, pulls specified Neighbors object.
}
