% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{NNPlot}
\alias{NNPlot}
\title{Highlight Neighbors in DimPlot}
\usage{
NNPlot(
  object,
  reduction,
  nn.idx,
  cells,
  dims = 1:2,
  label = FALSE,
  label.size = 4,
  repel = TRUE,
  highlight.size = 2,
  pt.size = 1,
  highlight.col = c("#377eb8", "#e41a1c"),
  other.col = "#bdbdbd",
  order = c("self", "neighbors", "other"),
  show.all.cells = TRUE,
  highlight = TRUE,
  group.by = NULL,
  cols = NULL,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Which dimensionality reduction to use. If not specified, first searches for umap, then tsne, then pca}

\item{nn.idx}{the neighbor index of all cells}

\item{cells}{Vector of cells to plot (default is all cells)}

\item{dims}{Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions}

\item{label}{Whether to label the clusters}

\item{label.size}{Sets size of labels}

\item{repel}{Repel labels}

\item{pt.size}{Adjust point size for plotting}

\item{order}{Specify the order of plotting for the idents. This can be
useful for crowded plots if points of interest are being buried. Provide
either a full list of valid idents or a subset to be plotted last (on top)}

\item{group.by}{Name of one or more metadata columns to group (color) cells by
(for example, orig.ident); pass 'ident' to group by identity class}

\item{cols}{Vector of colors, each color corresponds to an identity class. This may also be a single character
or numeric value corresponding to a palette as specified by \code{\link[RColorBrewer]{brewer.pal.info}}.
By default, ggplot2 assigns colors. We also include a number of palettes from the pals package.
See \code{\link{DiscretePalette}} for details.}

\item{...}{Extra parameters passed to \code{DimPlot}}
}
\description{
It will color the query cells and the neighbors of the query cells in the
DimPlot
}
