% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{NNtoGraph}
\alias{NNtoGraph}
\title{Convert Neighbor class to an asymmetrical Graph class}
\usage{
NNtoGraph(nn.object, col.cells = NULL, weighted = FALSE)
}
\arguments{
\item{nn.object}{A neighbor class object}

\item{col.cells}{Cells names of the neighbors, cell names in nn.object is used by default}

\item{weighted}{Determine if use distance in the Graph}
}
\value{
Returns a Graph object
}
\description{
Convert Neighbor class to an asymmetrical Graph class
}
