% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{NodeHeatmap}
\alias{NodeHeatmap}
\title{Node Heatmap}
\usage{
NodeHeatmap(object, marker.list, node = NULL, max.genes = 10, ...)
}
\arguments{
\item{object}{Seurat object. Must have the cluster.tree slot filled (use BuildClusterTree)}

\item{marker.list}{List of marker genes given from the FindAllMarkersNode function}

\item{node}{Node in the cluster tree from which to start the plot, defaults to highest node in marker list}

\item{max.genes}{Maximum number of genes to keep for each division}

\item{...}{Additional parameters to pass to DoHeatmap}
}
\value{
Plots heatmap. No return value.
}
\description{
Takes an object, a marker list (output of FindAllMarkers), and a node
and plots a heatmap where genes are ordered vertically by the splits present
in the object@cluster.tree slot.
}
