% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_determination.R
\name{NumberClusters}
\alias{NumberClusters}
\title{Convert the cluster labels to a numeric representation}
\usage{
NumberClusters(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{
Returns a Seurat object with the identities relabeled numerically
starting from 1.
}
\description{
Convert the cluster labels to a numeric representation
}
