% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction_utilities.R
\name{PCAEmbed}
\alias{PCAEmbed}
\title{PCA Cell Embeddings Accessor Function}
\usage{
PCAEmbed(object, dims.use = NULL, cells.use = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{dims.use}{Dimensions to include (default is all stored dims)}

\item{cells.use}{Cells to include (default is all cells)}
}
\value{
PCA cell embedding matrix for given cells and PCs
}
\description{
Pull PCA cell embedding matrix
}
