% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{PCAFast}
\alias{PCAFast}
\title{Run Principal Component Analysis on gene expression using IRLBA}
\usage{
PCAFast(object, pc.genes = NULL, do.print = TRUE, pcs.print = 1:5,
  pcs.store = 40, pcs.compute = 20, genes.print = 30, ...)
}
\arguments{
\item{object}{Seurat object}

\item{pc.genes}{Genes to use as input for PCA. Default is object@var.genes}

\item{do.print}{Print the top genes associated with high/low loadings for
the PCs}

\item{pcs.print}{PCs to print genes for}

\item{pcs.store}{Number of PCs to store}

\item{pcs.compute}{Total Number of PCs to compute and store}

\item{genes.print}{Number of genes to print for each PC}

\item{\dots}{Additional arguments to be passed to prcomp}
}
\value{
Returns Seurat object with an PCA embedding (object@pca.rot) and
gene projection matrix (object@pca.x). The PCA object itself is stored in
object@pca.obj[[1]]
}
\description{
Run Fast PCA dimensionality reduction
}

