% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{PCARot}
\alias{PCARot}
\title{PCA Rotation Accessor Function}
\usage{
PCARot(object, dims.use = NULL, cells.use = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{dims.use, }{Dimensions to include (default is all stored dims)}

\item{cells.use, }{Cells to include (default is all cells)}
}
\value{
PCA rotation matrix for given cells and PCs
}
\description{
Pull PCA rotation matrix
}

