% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seurat.R
\name{pcHeatmap}
\alias{pcHeatmap}
\title{Principal component heatmap}
\usage{
pcHeatmap(object, pc.use = 1, cells.use = NULL, num.genes = 30,
  use.full = FALSE, disp.min = -2.5, disp.max = 2.5, do.return = FALSE,
  col.use = pyCols, use.scale = TRUE, do.balanced = FALSE,
  remove.key = FALSE, ...)
}
\arguments{
\item{object}{Seurat object}

\item{pc.use}{Principal components to use}

\item{cells.use}{A list of cells to plot. If numeric, just plots the top cells.}

\item{num.genes}{Number of genes to return}

\item{use.full}{Use the full PCA (projected PCA). Default i s FALSE}

\item{disp.min}{Minimum display value (all values below are clipped)}

\item{disp.max}{Maximum display value (all values above are clipped)}

\item{do.return}{Default is FALSE. If TRUE, return a matrix of scaled values
which would be passed to heatmap.2}

\item{col.use}{Color palette to use}

\item{use.scale}{Default is TRUE: plot scaled data. If FALSE, plot raw data on the heatmap.}

\item{do.balanced}{Return an equal number of genes with both + and - PC scores.}

\item{remove.key}{Removes the color key from the plot.}

\item{...}{Additional parameters to heatmap.2. Common examples are cexRow
and cexCol, which set row and column text sizes}
}
\value{
If do.return==TRUE, a matrix of scaled values which would be passed
to heatmap.2. Otherwise, no return value, only a graphical output
}
\description{
Draws a heatmap focusing on a principal component. Both cells and genes are sorted by their principal component scores. Allows for nice visualization of sources of heterogeneity in the dataset.
}

