% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{pcTopCells}
\alias{pcTopCells}
\title{Find cells with highest PCA scores}
\usage{
pcTopCells(object, pc.use = 1, num.cells = NULL, do.balanced = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{pc.use}{Principal component to use}

\item{do.balanced}{Return an equal number of cells with both + and - PC scores.}

\item{num.genes}{Number of cells to return}
}
\value{
Returns a vector of cells
}
\description{
Return a list of genes with the strongest contribution to a set of principal components
}

