% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{PercentAbove}
\alias{PercentAbove}
\title{Calculate the percentage of a vector above some threshold}
\usage{
PercentAbove(x, threshold)
}
\arguments{
\item{x}{Vector of values}

\item{threshold}{Threshold to use when calculating percentage}
}
\value{
Returns the percentage of \code{x} values above the given threshold
}
\description{
Calculate the percentage of a vector above some threshold
}
\examples{
set.seed(42)
PercentAbove(sample(1:100, 10), 75)

}
\concept{utilities}
