% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{PlotClusterTree}
\alias{PlotClusterTree}
\title{Plot phylogenetic tree}
\usage{
PlotClusterTree(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{
Plots dendogram (must be precomputed using BuildClusterTree), returns no value
}
\description{
Plots previously computed phylogenetic tree (from BuildClusterTree)
}

