% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{PlotNoiseModel}
\alias{PlotNoiseModel}
\title{Visualize expression/dropout curve}
\usage{
PlotNoiseModel(object, cell.ids = c(1, 2), col.use = "black", lwd.use = 2,
  do.new = TRUE, x.lim = 10, ...)
}
\arguments{
\item{object}{Seurat object}

\item{cell.ids}{Cells to use}

\item{col.use}{Color code or name}

\item{lwd.use}{Line width for curve}

\item{do.new}{Create a new plot (default) or add to existing}

\item{x.lim}{Maximum value for X axis}

\item{\dots}{Additional arguments to pass to lines function}
}
\value{
Returns no value, displays a plot
}
\description{
Plot the probability of detection vs average expression of a gene.
}
\details{
Assumes that this 'noise' model has been precomputed with CalcNoiseModels
}
