% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{PlotSaturation}
\alias{PlotSaturation}
\title{Saturation Plot}
\usage{
PlotSaturation(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{
A density plot with the expected unobserved expression for all cells in the dataset
}
\description{
Saturation Plot
}
