% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snn.R
\name{PlotSNN}
\alias{PlotSNN}
\title{SNN Graph Plotting}
\usage{
PlotSNN(object, color.by.cluster = F, tsne.layout = F)
}
\arguments{
\item{object}{Seurat object}

\item{color.by.cluster}{Whether or not nodes should be colored by cluster}

\item{tsne.layout}{Whether or not to use TSNE as graph layout}
}
\value{
The plot.igraph object
}
\description{
Plots the Shared Nearest Neighbor (SNN) Graph built
with the BuiltSNN function
}
