% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seurat.R
\name{plsr_preds_expr}
\alias{plsr_preds_expr}
\title{PLSR scheme to perform imputation}
\usage{
plsr_preds_expr(x, y, y.name, do.print = FALSE, validation = "CV")
}
\arguments{
\item{x}{data.frame of predictors(columns): observed gene expression.}

\item{y}{Vector of response, observed expression of given landmarked gene.}

\item{y.name}{Chractor. Name of given landmarked gene.}

\item{do.print}{Logic. Whether print gene name to screen.}

\item{validation}{Charactor. Scheme for validation. Default: CV.}
}
\description{
PLSR scheme to perform imputation
}

