% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{PoissonDETest}
\alias{PoissonDETest}
\title{Poisson test for UMI-count based data}
\usage{
PoissonDETest(object, cells.1, cells.2, genes.use = NULL,
  latent.vars = NULL, print.bar = TRUE, min.cells = 3)
}
\arguments{
\item{object}{Seurat object}

\item{cells.1}{Group 1 cells}

\item{cells.2}{Group 2 cells}

\item{genes.use}{Optional vector of genes to use in test. Default is all genes}

\item{latent.vars}{effects to remove from comparison}

\item{print.bar}{Prints progress bar}

\item{min.cells}{Minimum number of cells in each group that express each gene}
}
\value{
Returns a p-value ranked matrix of putative differentially expressed
genes.
}
\description{
Identifies differentially expressed genes between two groups of cells using
a poisson generalized linear model
}
