% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PolyDimPlot}
\alias{PolyDimPlot}
\title{Polygon DimPlot}
\usage{
PolyDimPlot(
  object,
  group.by = NULL,
  cells = NULL,
  poly.data = "spatial",
  flip.coords = FALSE
)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{A grouping variable present in the metadata. Default is to use the groupings present
in the current cell identities (\code{Idents(object = object)})}

\item{cells}{Vector of cells to plot (default is all cells)}

\item{poly.data}{Name of the polygon dataframe in the misc slot}

\item{flip.coords}{Flip x and y coordinates}
}
\value{
Returns a ggplot object
}
\description{
Plot cells as polygons, rather than single points. Color cells by identity, or a categorical variable
in metadata
}
\concept{visualization}
