% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PosteriorPlot}
\alias{PosteriorPlot}
\title{Posterior Plot}
\usage{
PosteriorPlot(object, name)
}
\arguments{
\item{object}{A Seurat object}

\item{name}{Spatial code}
}
\description{
Posterior Plot
}
\seealso{
\code{SubsetColumn}

\code{VlnPlot}
}
