% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs.R
\name{PredictCellCyclePhases}
\alias{PredictCellCyclePhases}
\title{Predict cell cycle phases}
\usage{
PredictCellCyclePhases(object, g1.pairs, g2m.pairs, s.pairs,
  genes.training = NULL, null = "G1", score.threshold = 0.5,
  num.replicates = 1000, random.threshold = 100, couples.threshold = 10,
  num.cores = NULL)
}
\arguments{
\item{object}{A Seurat object}

\item{g1.pairs}{A data frame of gene pairs marking G1 phase}

\item{g2m.pairs}{A data frame of gene pairs marking G2M phase}

\item{s.pairs}{A data frame of gene pairs marking S phase}

\item{genes.training}{A list of genes that the model was trained on,
will assemble from pairs data frames if not provided}

\item{null}{Which phase is considered the 'null' phase when allocating scores
ie. which phase is not compared for score allocation. See Scialdone et al (2015) for more details}

\item{score.threshold}{Threshold for comparing scores. If neither comapred score is above this
threshold, then the 'null' phase is assigned.}

\item{num.replicates}{The number of times we should repeat our testing}

\item{random.threshold}{The minimum number of times a randomly selected
pair of genes provides a significant hit}

\item{couples.threshold}{The minimum number of gene pairs that provide a significant hit}

\item{num.cores}{The number of cores to run on}
}
\value{
A Seurat object with cell-cycle scores and allocations added to object@data.info
}
\description{
Predict cell cycle phases
}
\references{
Scialdone A, Natarajan KN, Saraiva LR, Proserpio V, Teichmann SA, Stegle O, Marioni JC, Buettner F. (2015) Computational assignment of cell-cycle stage from single-cell transcriptome data. Methods
}
