% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixscape.R
\name{PrepLDA}
\alias{PrepLDA}
\title{Function to prepare data for Linear Discriminant Analysis.}
\usage{
PrepLDA(
  object,
  de.assay = "RNA",
  pc.assay = "PRTB",
  labels = "gene",
  nt.label = "NT",
  npcs = 10,
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of class Seurat.}

\item{de.assay}{Assay to use for selection of DE genes.}

\item{pc.assay}{Assay to use for running Principle components analysis.}

\item{labels}{Meta data column with target gene class labels.}

\item{nt.label}{Name of non-targeting cell class.}

\item{npcs}{Number of principle components to use.}

\item{verbose}{Print progress bar.}
}
\description{
Function to prepare data for Linear Discriminant Analysis.
}
