% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{SubsetByBarcodeDistribution}
\alias{SubsetByBarcodeDistribution}
\title{Subsets a Seurat Object By Distribution of UMIs}
\usage{
SubsetByBarcodeDistribution(object, umi.column = "nCount_RNA",
  threshold.low = NULL, threshold.high = NULL, show.plot = FALSE,
  return.plot = FALSE, return.cells = FALSE,
  return.inflections = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{umi.column}{Column to use as proxy for barcodes ("nCount_RNA" by default)}

\item{threshold.low}{Ignore barcodes of rank below this threshold in inflection calculation}

\item{threshold.high}{Ignore barcodes of rank above this threshold in inflection calculation}

\item{show.plot}{Whether to produce diagnostic plot showing inflection point and barcode distribution}

\item{return.plot}{Returns the ggplot2 object in lieu of the subsetted Seurat object (useful for visually testing thresholds)}

\item{return.cells}{Returns the vector of cells which pass the defined barcode thresholds}

\item{return.inflections}{Returns a tibble with the grouping variable, data list-column with cell, umi.column, and rank, the calculated inflection point and associated umi_threshold, and a list-column of the cells which passed the calculated threshold per group}
}
\value{
Return Seurat object subsetted. Appends to the workflow slot 'nCount_RNA_threshold'
with a `data.frame` showing the calculated thresholds per sample group that determine the passing cells.
}
\description{
This function calculates an adaptive inflection point ("knee") of the barcode distribution
for each sample group (as identified by `object@meta.data$orig.ident`).
Returns the subsetted `object` with cells that passed the calculated threshold.
}
\details{
The assumed assay is "RNA", and thus uses the "nCount_RNA" distribution for
calculating the appropriate cutoff and returning a subsetted object.
}
\examples{
\dontrun{
SubsetByBarcodeDistribution(pbmc_small)
}
}
