% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{PrepSCTFindMarkers}
\alias{PrepSCTFindMarkers}
\title{Prepare object to run differential expression on SCT assay with multiple models}
\usage{
PrepSCTFindMarkers(object, assay = "SCT", verbose = TRUE)
}
\arguments{
\item{object}{Seurat object with SCT assays}

\item{assay}{Assay name where for SCT objects are stored; Default is 'SCT'}

\item{verbose}{Print messages and progress}
}
\value{
object Seurat object with recorrected counts and data in the SCT assay.
}
\description{
Given a merged object with multiple SCT models, this function uses minimum
of the median UMI (calculated using the raw UMI counts) of individual objects
to reverse the individual SCT regression model using minimum of median UMI
as the sequencing depth covariate.
The counts slot of the SCT assay is replaced with recorrected counts and
the data slot is replaced with log1p of recorrected counts.
}
