% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{print.DimReduc}
\alias{print.DimReduc}
\alias{print}
\title{Print the results of a dimensional reduction analysis}
\usage{
\method{print}{DimReduc}(x, dims = 1:5, nfeatures = 20, projected = FALSE, ...)
}
\arguments{
\item{x}{An object}

\item{dims}{Number of dimensions to display}

\item{nfeatures}{Number of genes to display}

\item{projected}{Use projected slot}

\item{...}{Arguments passed to other methods}
}
\value{
Set of features defining the components
}
\description{
Prints a set of features that most strongly define a set of components
}
\seealso{
\code{\link[base]{cat}}
}
