% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreduc_generics.R, R/dimreduc.R
\name{Print}
\alias{Print}
\alias{Print.DimReduc}
\title{Print the results of a dimensional reduction analysis}
\usage{
Print(object, ...)

\method{Print}{DimReduc}(object, dims = 1:5, num.features = 20,
  projected = FALSE)
}
\arguments{
\item{object}{DimReduc object}

\item{dims}{Number of dimensions to display}

\item{num.features}{Number of genes to display}

\item{projected}{Use projected slot}
}
\value{
Set of features defining the components
}
\description{
Prints a set of genes that most strongly define a set of components
}
\section{Methods (by class)}{
\itemize{
\item \code{DimReduc}: Print top features for DimReduc
}}

