% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing_utilities.R
\name{PrintCCAParams}
\alias{PrintCCAParams}
\title{Print CCA Calculation Parameters}
\usage{
PrintCCAParams(object, raw = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{raw}{Print the entire contents of the calculation settings slot
(calc.params) for the RunCCA calculation. Default (FALSE) will print a nicely
formatted summary.}
}
\value{
No return value. Only prints to console.
}
\description{
Print the parameters chosen for the latest stored CCA calculation.
}
