% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing_utilities.R
\name{PrintDMParams}
\alias{PrintDMParams}
\title{Print Diffusion Map Calculation Parameters}
\usage{
PrintDMParams(object, raw = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{raw}{Print the entire contents of the calculation settings slot
(calc.params) for the RunDiffusion calculation. Default (FALSE) will print a
nicely formatted summary.}
}
\value{
No return value. Only prints to console.
}
\description{
Print the parameters chosen for the latest stored diffusion map calculation.
}
\examples{
# Run Diffusion on variable genes
pbmc_small <- RunDiffusion(pbmc_small,genes.use = pbmc_small@var.genes)
PrintDMParams(object = pbmc_small)

}
