% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing_utilities.R
\name{PrintFindClustersParams}
\alias{PrintFindClustersParams}
\title{Print FindClusters Calculation Parameters}
\usage{
PrintFindClustersParams(object, resolution, raw = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{resolution}{Optionally specify only a subset of resolutions to print
parameters for.}

\item{raw}{Print the entire contents of the calculation settings slot
(calc.params) for the FindClusters calculation. Default (FALSE) will print a
nicely formatted summary.}
}
\value{
No return value. Only prints to console.
}
\description{
Print the parameters chosen for the latest FindClusters calculation for each
stored resolution.
}
