% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{PrintICA}
\alias{PrintICA}
\title{Print the results of a PCA analysis}
\usage{
PrintICA(object, ics.print = 1:5, genes.print = 30, use.full = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{genes.print}{Number of genes to print for each PC}

\item{use.full}{Use full PCA (i.e. the projected PCA, by default FALSE)}

\item{pcs.print}{Set of PCs to print genes for}
}
\value{
Only text output
}
\description{
Prints a set of genes that most strongly define a set of principal components
}
