% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing_utilities.R
\name{PrintICAParams}
\alias{PrintICAParams}
\title{Print ICA Calculation Parameters}
\usage{
PrintICAParams(object, raw = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{raw}{Print the entire contents of the calculation settings slot (calc.params) for the ICA
calculation. Default (FALSE) will print a nicely formatted summary.}
}
\value{
No return value. Only prints to console.
}
\description{
Print the parameters chosen for the latest stored ICA calculation.
}
\examples{
pbmc_small <- RunICA(object = pbmc_small, ics.compute = 5)
PrintICAParams(object = pbmc_small, raw = TRUE)

}
