% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{PrintOntologyGraph}
\alias{PrintOntologyGraph}
\title{Plot Ontology Graph from OntologyAnalysis}
\usage{
PrintOntologyGraph(ontology.result)
}
\arguments{
\item{ontology.result}{A list with two topGO objects: ontology and test as returned by OntologyAnalysis}
}
\description{
A function that gets the output of a TopGO analysis as outputted by OntologyAnalysis
and prints the enrichment DAG.
}
