% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing_utilities.R
\name{PrintSNNParams}
\alias{PrintSNNParams}
\title{Print SNN Construction Calculation Parameters}
\usage{
PrintSNNParams(object, raw = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{raw}{Print the entire contents of the calculation settings slot (calc.params) for the
BuildSNN calculation. Default (FALSE) will print a nicely formatted summary.}
}
\value{
No return value. Only prints to console.
}
\description{
Print the parameters chosen for the latest stored SNN calculation (via BuildSNN or FindClusters).
}
\examples{
pbmc_small <- BuildSNN(object = pbmc_small)
PrintSNNParams(object = pbmc_small)

}
