% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing_utilities.R
\name{PrintTSNEParams}
\alias{PrintTSNEParams}
\title{Print TSNE Calculation Parameters}
\usage{
PrintTSNEParams(object, raw = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{raw}{Print the entire contents of the calculation settings slot (calc.params) for the
RunTSNE calculation. Default (FALSE) will print a nicely formatted summary.}
}
\value{
No return value. Only prints to console.
}
\description{
Print the parameters chosen for the latest stored TSNE calculation.
}
\examples{
PrintTSNEParams(object = pbmc_small)

}
