% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{PropFeatureSet}
\alias{PropFeatureSet}
\title{Calculate the proportion of all counts that belong to a given set of features}
\usage{
PropFeatureSet(object, pattern = NULL, features = NULL, assay = NULL,
  metadata.name = NULL)
}
\arguments{
\item{object}{A Seurat object}

\item{pattern}{A regex pattern to match features against}

\item{features}{A defined feature set. If features provided, will ignore the pattern matching}

\item{assay}{Assay to use}

\item{metadata.name}{Name of metadata column to store info in}
}
\value{
Returns a Seurat object with the proportion of the feature set stored in metadata.
}
\description{
This function enables you to easily calculate the proportion of all the counts belonging to a 
subset of the possible features for each cell. This is useful when trying to compute the fraction 
of transcripts that map to mitochondrial genes for example. The calculation here is simply the 
column sum of the matrix present in the counts slot for features belonging to the set divided by
the column sum for all features.
}
\examples{
# Calculate the proportion of transcripts mapping to mitochondrial genes 
# NOTE: The pattern provided works for human gene names. You may need to adjust depending on your
# system of interest
pbmc_small <- PropFeatureSet(object = pbmc_small, pattern = "^MT-", metadata.name = "prop.mito")

}
