% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{PseudobulkExpression.Assay}
\alias{PseudobulkExpression.Assay}
\title{Pseudobulk feature expression by identity class}
\usage{
\method{PseudobulkExpression}{Assay}(
  object,
  assay,
  category.matrix,
  features = NULL,
  layer = "data",
  slot = deprecated(),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Features to analyze. Default is all features in the assay}

\item{slot}{Slot(s) to use; if multiple slots are given, assumed to follow
the order of 'assays' (if specified) or object's assays}

\item{verbose}{Print messages and show progress bar}

\item{...}{Arguments to be passed to methods such as \code{\link{CreateSeuratObject}}}

\item{method}{Whether to 'average' (default) or 'aggregate' expression levels}

\item{assays}{Which assays to use. Default is all assays}

\item{return.seurat}{Whether to return the data as a Seurat object. Default is FALSE}

\item{group.by}{Categories for grouping (e.g, ident, replicate, celltype); 'ident' by default}

\item{add.ident}{(Deprecated) Place an additional label on each cell prior to pseudobulking
(very useful if you want to observe cluster pseudobulk values, separated by replicate, for example)}
}
\value{
Returns a matrix with genes as rows, identity classes as columns.
If return.seurat is TRUE, returns an object of class \code{\link{Seurat}}.
}
\description{
Returns a representative expression value for each identity class
}
