% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utilities.R
\name{PurpleAndYellow}
\alias{PurpleAndYellow}
\title{A purple and yellow color palette}
\usage{
PurpleAndYellow(...)
}
\arguments{
\item{...}{Extra parameters to CustomPalette}
}
\value{
A color palette
}
\description{
A purple and yellow color palette
}
\seealso{
\code{CustomPalette}
}
