% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Radius.SlideSeq}
\alias{Radius.SlideSeq}
\alias{Radius.STARmap}
\alias{Radius.VisiumV1}
\alias{Radius.VisiumV2}
\title{Get Spot Radius}
\usage{
\method{Radius}{SlideSeq}(object, ...)

\method{Radius}{STARmap}(object, ...)

\method{Radius}{VisiumV1}(object, scale = "lowres", ...)

\method{Radius}{VisiumV2}(object, scale = "lowres", ...)
}
\arguments{
\item{object}{An image object}

\item{...}{Arguments passed to other methods}

\item{scale}{A factor to scale the radius by; one of: "hires",
"lowres", or \code{NULL} for the unscaled value.}
}
\description{
Get Spot Radius
}
\seealso{
\code{\link[SeuratObject:Radius]{SeuratObject::Radius}}
}
\concept{objects}
\concept{spatial}
