% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{Read10X}
\alias{Read10X}
\title{Load in data from 10X}
\usage{
Read10X(data.dir = NULL)
}
\arguments{
\item{data.dir}{Directory containing the matrix.mtx, genes.tsv, and barcodes.tsv
files provided by 10X.}
}
\value{
Returns a sparse matrix with rows and columns labeled
}
\description{
Enables easy loading of sparse data matrices provided by 10X genomics.
}

