% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{Read10X_Coordinates}
\alias{Read10X_Coordinates}
\title{Load 10X Genomics Visium Tissue Positions}
\usage{
Read10X_Coordinates(filename, filter.matrix)
}
\arguments{
\item{filename}{Path to a \code{tissue_positions_list.csv} file}

\item{filter.matrix}{Filter spot/feature matrix to only include spots that
have been determined to be over tissue}
}
\value{
A data.frame
}
\description{
Load 10X Genomics Visium Tissue Positions
}
\concept{preprocessing}
