% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{Read10X_h5}
\alias{Read10X_h5}
\title{Read 10X hdf5 file}
\usage{
Read10X_h5(filename, ensg.names = FALSE)
}
\arguments{
\item{filename}{Path to h5 file}

\item{ensg.names}{Label row names with ENSG names rather than unique gene names}
}
\value{
Returns a sparse matrix with rows and columns labeled. If multiple genomes are present,
returns a list of sparse matrices (one per genome).
}
\description{
Read gene expression matrix from 10X CellRanger hdf5 file
}
