% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{Read10X_probe_metadata}
\alias{Read10X_probe_metadata}
\title{Read10x Probe Metadata}
\usage{
Read10X_probe_metadata(data.dir, filename = "raw_probe_bc_matrix.h5")
}
\arguments{
\item{data.dir}{The directory where the file is located.}

\item{filename}{The name of the file containing the raw probe barcode matrix in HDF5 format. The default filename is 'raw_probe_bc_matrix.h5'.}
}
\value{
Returns a data.frame containing the probe metadata.
}
\description{
This function reads the probe metadata from a 10x Genomics probe barcode matrix file in HDF5 format.
}
