% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{Read10X_ScaleFactors}
\alias{Read10X_ScaleFactors}
\title{Load 10X Genomics Visium Scale Factors}
\usage{
Read10X_ScaleFactors(filename)
}
\arguments{
\item{filename}{Path to a \code{scalefactors_json.json} file}
}
\value{
A scalefactors object
}
\description{
Load 10X Genomics Visium Scale Factors
}
\concept{preprocessing}
