% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ReadAkoya}
\alias{ReadAkoya}
\title{Read Akoya CODEX data}
\usage{
ReadAkoya(matrix, file.type = "inForm")
}
\arguments{
\item{matrix}{Path to matrix generated by upstream processing.}

\item{file.type}{Specify which type matrix is being provided.
\itemize{
 \item \dQuote{\code{processor}}: matrix generated by CODEX Processor
 \item \dQuote{\code{inForm}}: matrix generated by inForm
 \item \dQuote{\code{QuPath}}: matrix generated by QuPath
}}
}
\value{
A list with some combination of the following values
\itemize{
 \item \dQuote{\code{matrix}}: a
 \link[Matrix:dgCMatrix-class]{sparse matrix} with expression data; cells
  are columns and features are rows
 \item \dQuote{\code{centroids}}: a data frame with cell centroid
  coordinates in three columns: \dQuote{x}, \dQuote{y}, and \dQuote{cell}
}
}
\description{
Read Akoya CODEX data
}
\concept{preprocessing}
