% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{ReadAnnoyNN}
\alias{ReadAnnoyNN}
\title{Load the Annoy neighbors and C++ index from RDS}
\usage{
ReadAnnoyNN(file)
}
\arguments{
\item{file}{the name of the file where the nn object is saved by SaveAnnoyNN}
}
\description{
Load the Annoy neighbors and C++ index from RDS
}
