% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ReadMtx}
\alias{ReadMtx}
\title{Load in data from remote or local mtx files}
\usage{
ReadMtx(
  mtx,
  cells,
  features,
  cell.column = 1,
  feature.column = 2,
  unique.features = TRUE,
  strip.suffix = FALSE
)
}
\arguments{
\item{mtx}{Name or remote URL of the mtx file}

\item{cells}{Name or remote URL of the cells/barcodes file}

\item{features}{Name or remote URL of the features/genes file}

\item{cell.column}{Specify which column of cells file to use for cell names; default is 1}

\item{unique.features}{Make feature names unique (default TRUE)}

\item{strip.suffix}{Remove trailing "-1" if present in all cell barcodes.}

\item{gene.column}{Specify which column of features files to use for gene names; default is 2}
}
\value{
A sparse matrix containing the expression data.
}
\description{
Enables easy loading of sparse data matrices
}
\examples{
\dontrun{
# For local files:

expression_matrix <- ReadMtx(genes="count_matrix.mtx.gz", features="features.tsv.gz", cells="barcodes.tsv.gz")
seurat_object <- CreateSeuratObject(counts = expression_matrix)

# For remote files:

expression_matrix <- ReadMtx(mtx = "https://www.ncbi.nlm.nih.gov/geo/download/?acc=GSE132044&format=file&file=GSE132044\%5Fmixture\%5Fhg19\%5Fmm10\%5Fcount\%5Fmatrix\%2Emtx\%2Egz",
cells ="https://www.ncbi.nlm.nih.gov/geo/download/?acc=GSE132044&format=file&file=GSE132044\%5Fmixture\%5Fhg19\%5Fmm10\%5Fcell\%2Etsv\%2Egz",
features = "https://www.ncbi.nlm.nih.gov/geo/download/?acc=GSE132044&format=file&file=GSE132044\%5Fmixture\%5Fhg19\%5Fmm10\%5Fgene\%2Etsv\%2Egz",
feature.columns = 1)
seurat_object <- CreateSeuratObject(counts = data)
}

}
\concept{preprocessing}
