% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{ReadParseBio}
\alias{ReadParseBio}
\title{Read output from Parse Biosciences}
\usage{
ReadParseBio(data.dir, ...)
}
\arguments{
\item{data.dir}{Directory containing the data files}

\item{...}{Extra parameters passed to \code{\link{ReadMtx}}}
}
\description{
Read output from Parse Biosciences
}
\concept{convenience}
