% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{ReadSlideSeq}
\alias{ReadSlideSeq}
\title{Load a Slide-seq image}
\usage{
ReadSlideSeq(puck.path, assay = "Spatial")
}
\arguments{
\item{puck.path}{Path to puck directory}

\item{assay}{Name of assay to associate image to}
}
\value{
A \code{\link{SlideSeq}} object
}
\description{
Load a Slide-seq image
}
\seealso{
\code{\link{SlideSeq}}
}
