% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{RecreateWorkflows}
\alias{RecreateWorkflows}
\title{Output individual function calls to recreate workflow}
\usage{
RecreateWorkflows(object, workflow.name, command.name, depth = 1)
}
\arguments{
\item{object}{Seurat object}

\item{workflow.name}{Workflow name, should already be initialized using InitializeWorkflow}

\item{command.name}{Name of the command at the end of the workflow}

\item{depth}{depth of the recursive call. Only depth 1 outputs the parameters}
}
\value{
Seurat object with updated workflow
}
\description{
Output all commands to reproduce your analysis without shortcuts. Should enhance reproducibility, but can be confused by custom modifcations, usage of SubsetData, etc.
We hope this will be very useful, but use with care and verify that it does indeed reproduce your work.
}
\examples{
RecreateWorkflow(object = pbmc_small,workflow.name = "cluster", command.name = "FindClusters")
#
}
