% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{components}
\alias{components}
\alias{\%||\%}
\alias{\%iff\%}
\alias{reexports}
\alias{AddMetaData}
\alias{as.Graph}
\alias{as.Neighbor}
\alias{as.Seurat}
\alias{as.sparse}
\alias{Assays}
\alias{Cells}
\alias{CellsByIdentities}
\alias{Command}
\alias{CreateAssayObject}
\alias{CreateDimReducObject}
\alias{CreateSeuratObject}
\alias{DefaultAssay}
\alias{DefaultAssay<-}
\alias{Distances}
\alias{Embeddings}
\alias{FetchData}
\alias{GetAssayData}
\alias{GetImage}
\alias{GetTissueCoordinates}
\alias{HVFInfo}
\alias{Idents}
\alias{Idents<-}
\alias{Images}
\alias{Index}
\alias{Index<-}
\alias{Indices}
\alias{IsGlobal}
\alias{JS}
\alias{JS<-}
\alias{Key}
\alias{Key<-}
\alias{Loadings}
\alias{Loadings<-}
\alias{LogSeuratCommand}
\alias{Misc}
\alias{Misc<-}
\alias{Neighbors}
\alias{Project}
\alias{Project<-}
\alias{Radius}
\alias{Reductions}
\alias{RenameCells}
\alias{RenameIdents}
\alias{ReorderIdent}
\alias{RowMergeSparseMatrices}
\alias{SetAssayData}
\alias{SetIdent}
\alias{SpatiallyVariableFeatures}
\alias{StashIdent}
\alias{Stdev}
\alias{SVFInfo}
\alias{Tool}
\alias{Tool<-}
\alias{UpdateSeuratObject}
\alias{VariableFeatures}
\alias{VariableFeatures<-}
\alias{WhichCells}
\title{Objects exported from other packages}
\usage{
components(object, ...)

x \%||\% y

x \%iff\% y
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{SeuratObject}{\code{\link[SeuratObject:AddMetaData]{AddMetaData()}}, \code{\link[SeuratObject:as.Graph]{as.Graph()}}, \code{\link[SeuratObject:as.Neighbor]{as.Neighbor()}}, \code{\link[SeuratObject:as.Seurat]{as.Seurat()}}, \code{\link[SeuratObject:as.sparse]{as.sparse()}}, \code{\link[SeuratObject:Assays]{Assays()}}, \code{\link[SeuratObject:Cells]{Cells()}}, \code{\link[SeuratObject:CellsByIdentities]{CellsByIdentities()}}, \code{\link[SeuratObject:Command]{Command()}}, \code{\link[SeuratObject:CreateAssayObject]{CreateAssayObject()}}, \code{\link[SeuratObject:CreateDimReducObject]{CreateDimReducObject()}}, \code{\link[SeuratObject:CreateSeuratObject]{CreateSeuratObject()}}, \code{\link[SeuratObject:DefaultAssay]{DefaultAssay()}}, \code{\link[SeuratObject:DefaultAssay<-]{DefaultAssay<-()}}, \code{\link[SeuratObject:Distances]{Distances()}}, \code{\link[SeuratObject:Embeddings]{Embeddings()}}, \code{\link[SeuratObject:FetchData]{FetchData()}}, \code{\link[SeuratObject:GetAssayData]{GetAssayData()}}, \code{\link[SeuratObject:GetImage]{GetImage()}}, \code{\link[SeuratObject:GetTissueCoordinates]{GetTissueCoordinates()}}, \code{\link[SeuratObject:HVFInfo]{HVFInfo()}}, \code{\link[SeuratObject:Idents]{Idents()}}, \code{\link[SeuratObject:Idents<-]{Idents<-()}}, \code{\link[SeuratObject:Images]{Images()}}, \code{\link[SeuratObject:Index]{Index()}}, \code{\link[SeuratObject:Index<-]{Index<-()}}, \code{\link[SeuratObject:Indices]{Indices()}}, \code{\link[SeuratObject:IsGlobal]{IsGlobal()}}, \code{\link[SeuratObject:JS]{JS()}}, \code{\link[SeuratObject:JS<-]{JS<-()}}, \code{\link[SeuratObject:Key]{Key()}}, \code{\link[SeuratObject:Key<-]{Key<-()}}, \code{\link[SeuratObject:Loadings]{Loadings()}}, \code{\link[SeuratObject:Loadings<-]{Loadings<-()}}, \code{\link[SeuratObject:LogSeuratCommand]{LogSeuratCommand()}}, \code{\link[SeuratObject:Misc]{Misc()}}, \code{\link[SeuratObject:Misc<-]{Misc<-()}}, \code{\link[SeuratObject:Neighbors]{Neighbors()}}, \code{\link[SeuratObject:Project]{Project()}}, \code{\link[SeuratObject:Project<-]{Project<-()}}, \code{\link[SeuratObject:Radius]{Radius()}}, \code{\link[SeuratObject:Reductions]{Reductions()}}, \code{\link[SeuratObject:RenameCells]{RenameCells()}}, \code{\link[SeuratObject:RenameIdents]{RenameIdents()}}, \code{\link[SeuratObject:ReorderIdent]{ReorderIdent()}}, \code{\link[SeuratObject:RowMergeSparseMatrices]{RowMergeSparseMatrices()}}, \code{\link[SeuratObject:SetAssayData]{SetAssayData()}}, \code{\link[SeuratObject:SetIdent]{SetIdent()}}, \code{\link[SeuratObject:SpatiallyVariableFeatures]{SpatiallyVariableFeatures()}}, \code{\link[SeuratObject:StashIdent]{StashIdent()}}, \code{\link[SeuratObject:Stdev]{Stdev()}}, \code{\link[SeuratObject:SVFInfo]{SVFInfo()}}, \code{\link[SeuratObject:Tool]{Tool()}}, \code{\link[SeuratObject:Tool<-]{Tool<-()}}, \code{\link[SeuratObject:UpdateSeuratObject]{UpdateSeuratObject()}}, \code{\link[SeuratObject:VariableFeatures]{VariableFeatures()}}, \code{\link[SeuratObject:VariableFeatures<-]{VariableFeatures<-()}}, \code{\link[SeuratObject:WhichCells]{WhichCells()}}}
}}

