% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{RegressOut}
\alias{RegressOut}
\title{Regress out technical effects and cell cycle}
\usage{
RegressOut(object, latent.vars, genes.regress = NULL, do.scale = TRUE, ...)
}
\arguments{
\item{object}{Seurat object}

\item{latent.vars}{effects to regress out}

\item{genes.regress}{gene to run regression for (default is all genes)}

\item{do.scale}{Z-normalize the residual values (default is TRUE)}
}
\value{
Returns Seurat object with the scale.data (object@scale.data) genes returning the residuals from the regression model
}
\description{
Remove unwanted effects from scale.data
}

