% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{RegressOutNB}
\alias{RegressOutNB}
\title{Regress out technical effects and cell cycle using regularized Negative Binomial regression}
\usage{
RegressOutNB(object, latent.vars, genes.regress = NULL,
  pr.clip.range = c(-30, 30), min.theta = 0.01)
}
\arguments{
\item{object}{Seurat object}

\item{latent.vars}{effects to regress out}

\item{genes.regress}{gene to run regression for (default is all genes)}

\item{pr.clip.range}{numeric of length two specifying the min and max values the results will be clipped to}
}
\value{
Returns Seurat object with the scale.data (object@scale.data) genes returning the residuals fromthe regression model
}
\description{
Remove unwanted effects from umi data and set scale.data to Pearson residuals
Uses mclapply; you can set the number of cores it will use to n with command options(mc.cores = n)
}
