% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_internal.R
\name{RegressOutResid}
\alias{RegressOutResid}
\title{Regress out technical effects and cell cycle}
\usage{
RegressOutResid(object, vars.to.regress, genes.regress = NULL,
  model.use = "linear", use.umi = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{vars.to.regress}{effects to regress out}

\item{genes.regress}{gene to run regression for (default is all genes)}

\item{model.use}{Use a linear model or generalized linear model (poisson, negative binomial) for the regression. Options are 'linear' (default), 'poisson', and 'negbinom'}

\item{use.umi}{Regress on UMI count data. Default is FALSE for linear modeling, but automatically set to TRUE if model.use is 'negbinom' or 'poisson'}
}
\value{
Returns the residuals from the regression model
}
\description{
Remove unwanted effects from scale.data
}
\keyword{internal}
