% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{RegroupIdents}
\alias{RegroupIdents}
\title{Regroup idents based on meta.data info}
\usage{
RegroupIdents(object, metadata)
}
\arguments{
\item{object}{Seurat object}

\item{metadata}{Name of metadata column}
}
\value{
A Seurat object with the active idents regrouped
}
\description{
For cells in each ident, set a new identity based on the most common value
of a specified metadata column.
}
\examples{
data("pbmc_small")
pbmc_small <- RegroupIdents(pbmc_small, metadata = "groups")

}
\concept{utilities}
