% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seurat.R
\name{rename.ident}
\alias{rename.ident}
\title{Rename one identity class to another}
\usage{
rename.ident(object, old.ident.name = NULL, new.ident.name = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{old.ident.name}{The old identity class (to be renamed)}

\item{new.ident.name}{The new name to apply}
}
\value{
A Seurat object where object@ident has been appropriately modified
}
\description{
Can also be used to join identity classes together (for example, to merge clusters).
}

