% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{RenameAssays}
\alias{RenameAssays}
\title{Rename assays in a \code{Seurat} object}
\usage{
RenameAssays(object, ...)
}
\arguments{
\item{object}{A \code{Seurat} object}

\item{...}{Named arguments as \code{old.assay = new.assay}}
}
\value{
\code{object} with assays renamed
}
\description{
Rename assays in a \code{Seurat} object
}
\examples{
RenameAssays(object = pbmc_small, RNA = 'rna')

}
