% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R, R/interaction_generics.R
\name{RenameCells.Assay}
\alias{RenameCells.Assay}
\alias{RenameCells.DimReduc}
\alias{RenameCells.Seurat}
\alias{RenameCells}
\title{Rename cells}
\usage{
\method{RenameCells}{Assay}(object, new.names = NULL)

\method{RenameCells}{DimReduc}(object, new.names = NULL)

\method{RenameCells}{Seurat}(object, add.cell.id = NULL, new.names = NULL,
  for.merge = FALSE)

RenameCells(object, slot, ...)
}
\arguments{
\item{object}{Seurat object}

\item{new.names}{vector of new cell names}

\item{add.cell.id}{prefix to add cell names}

\item{for.merge}{Only rename slots needed for merging Seurat objects.
Currently only renames the raw.data and meta.data slots.}
}
\value{
Seurat object with new cell names
}
\description{
Change the cell names in all the different parts of a Seurat object. Can
be useful before combining multiple objects.
}
\details{
If \code{add.cell.id} is set a prefix is added to existing cell names. If
\code{new.names} is set these will be used to replace existing names.
}
\section{Methods (by class)}{
\itemize{
\item \code{Assay}: Rename cells in an Assay object

\item \code{DimReduc}: Rename cells in a DimReduc object

\item \code{Seurat}: Rename cells in a Seurat object
}}

\examples{
head(pbmc_small@cell.names)
pbmc_small <- RenameCells(pbmc_small, add.cell.id = "Test")
head(pbmc_small@cell.names)

}
