% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{RowMergeSparseMatrices}
\alias{RowMergeSparseMatrices}
\title{Merge two matrices by rowname}
\usage{
RowMergeSparseMatrices(mat1, mat2)
}
\arguments{
\item{mat1}{First matrix}

\item{mat2}{Second matrix}
}
\value{
A merged matrix

Returns a sparse matrix
}
\description{
This function is for use on sparse matrices and
should not be run on a Seurat object.
}
\details{
Shared matrix rows (with the same row name) will be merged,
and unshared rows (with different names) will be filled
with zeros in the matrix not containing the row.
}
