% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration5.R
\name{RPCAIntegration}
\alias{RPCAIntegration}
\title{Seurat-RPCA Integration}
\usage{
RPCAIntegration(
  object = NULL,
  assay = NULL,
  layers = NULL,
  orig = NULL,
  new.reduction = "integrated.dr",
  reference = NULL,
  features = NULL,
  normalization.method = c("LogNormalize", "SCT"),
  dims = 1:30,
  k.filter = NA,
  scale.layer = "scale.data",
  groups = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{Seurat} object}

\item{assay}{Name of \code{Assay} in the \code{Seurat} object}

\item{layers}{Names of layers in \code{assay}}

\item{orig}{A \link[SeuratObject:DimReduc]{dimensional reduction} to correct}

\item{new.reduction}{Name of new integrated dimensional reduction}

\item{reference}{A reference \code{Seurat} object}

\item{features}{A vector of features to use for integration}

\item{normalization.method}{Name of normalization method used: LogNormalize
or SCT}

\item{dims}{Dimensions of dimensional reduction to use for integration}

\item{k.filter}{Number of anchors to filter}

\item{scale.layer}{Name of scaled layer in \code{Assay}}

\item{groups}{A one-column data frame with grouping information}

\item{verbose}{Print progress}

\item{...}{Additional arguments passed to \code{FindIntegrationAnchors}}
}
\description{
Seurat-RPCA Integration
}
\examples{
\dontrun{
# Preprocessing
obj <- SeuratData::LoadData("pbmcsca")
obj[["RNA"]] <- split(obj[["RNA"]], f = obj$Method)
obj <- NormalizeData(obj)
obj <- FindVariableFeatures(obj)
obj <- ScaleData(obj)
obj <- RunPCA(obj)

# After preprocessing, we run integration
obj <- IntegrateLayers(object = obj, method = RPCAIntegration,
  orig.reduction = "pca", new.reduction = 'integrated.rpca',
  verbose = FALSE)

# Reference-based Integration
# Here, we use the first layer as a reference for integraion
# Thus, we only identify anchors between the reference and the rest of the datasets,
# saving computational resources
obj <- IntegrateLayers(object = obj, method = RPCAIntegration,
  orig.reduction = "pca", new.reduction = 'integrated.rpca',
  reference = 1, verbose = FALSE)

# Modifying parameters
# We can also specify parameters such as `k.anchor` to increase the strength of
# integration
obj <- IntegrateLayers(object = obj, method = RPCAIntegration,
  orig.reduction = "pca", new.reduction = 'integrated.rpca',
  k.anchor = 20, verbose = FALSE)

# Integrating SCTransformed data
obj <- SCTransform(object = obj)
obj <- IntegrateLayers(object = obj, method = RPCAIntegration,
  orig.reduction = "pca", new.reduction = 'integrated.rpca',
  assay = "SCT", verbose = FALSE)
}

}
