% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/dimensional_reduction.R
\name{RunDiffusion}
\alias{RunDiffusion}
\alias{RunDiffusion.matrix}
\alias{RunDiffusion.DimReduc}
\alias{RunDiffusion.Seurat}
\title{Run diffusion map}
\usage{
RunDiffusion(object, ...)

\method{RunDiffusion}{matrix}(object, assay = NULL, dim.embed = 2,
  q = 0.01, reduction.key = "DMC_", ...)

\method{RunDiffusion}{DimReduc}(object, cells = NULL, dims = 1:5,
  q = 0.01, dim.embed = 2, reduction.key = "DMC_", ...)

\method{RunDiffusion}{Seurat}(object, assay = NULL, cells = NULL,
  dims = 1:5, reduction = "pca", features = NULL,
  normalize.method = c("LogNormalize", "SCTransform"), q = 0.01,
  dim.embed = 2, scale.clip = 10, reduction.name = "diffusionmap",
  reduction.key = "DMC_", ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods and destiny}

\item{assay}{Name of Assay diffusion map is being run on}

\item{dim.embed}{Total Number of diffusion map components (DMCs) to compute and store (2 by default)}

\item{q}{Quantile to clip diffusion map components at. This addresses an
issue where 1-2 cells will have extreme values that obscure all other points.
0.01 by default}

\item{reduction.key}{dimensional reduction key, specifies the string before
the number for the dimension names. DMC by default}

\item{cells}{Which cells to analyze (default, all cells)}

\item{dims}{Which dimensions to use as input features.}

\item{reduction}{Which dimensional reduction (e.g. PCA, ICA) to use for
the diffusion map. Default is PCA}

\item{features}{If set, run the diffusion map on this subset of features
(instead of running on a set of reduced dimensions). Not set (NULL) by default;
\code{dims} will automaticly be NULL to run on features}

\item{normalize.method}{'LogNormalize' or 'SCTransform'. If NULL, set 'LogNormalize'
by default}

\item{scale.clip}{Max/min value for scaled data. Default is 10}

\item{reduction.name}{dimensional reduction name, specifies the position in the object$dr list. diffusionmap by default}

\item{...}{Additional arguments to the DiffusionMap call}
}
\value{
Returns Seurat object with the diffusion map calculation stored in the reductions slot
}
\description{
Run a diffusion map dimensionality reduction. For details about stored difussion map calculation
parameters, see \code{PrintDiffusionParams}.
}
