% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr.R
\name{RunEnrichr}
\alias{RunEnrichr}
\title{Run Enrichr}
\usage{
RunEnrichr(GeneList = NULL, EnrichrID = NULL,
  GeneSetLib = "GO_Biological_Process_2017", PrintList = FALSE,
  PrintID = FALSE, Download = FALSE, FileName = "test.txt", ...)
}
\arguments{
\item{GeneList}{Input gene list}

\item{EnrichrID}{Enrichr identifier for input gene list, if input genes are already uploaded}

\item{GeneSetLib}{Library to be enriched against}

\item{PrintList}{Print gene list}

\item{PrintID}{Print Enrichr identifier}

\item{Download}{Download enrichment results as a txt file}

\item{FileName}{Name for the txt file}
}
\value{
If Download=TRUE, a tab-delimited file is written in the directory specified in FileName
}
\description{
Run Enrichr in R: enrichment against available libraries in Enrichr
}
