% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{RunkUMP}
\alias{RunkUMP}
\title{Run UMAP on k nearest neighbors}
\usage{
RunkUMP(object, adjacency.use = NULL, knn.use = NULL,
  genes.for.knn = NULL, assay = "RNA", reduction = "pca",
  dims = 1:10, k.param = 30, eps = 0, min_dist = 0.3,
  metric = "correlation", ...)
}
\arguments{
\item{object}{A seurat object}

\item{adjacency.use}{A square adjacency matrix showing pairwise cell similarities}

\item{knn.use}{A list of length 2 showing the indices and distances between knn. Can be
outputs from running RANN::nn2, FNN::get.knn or similar functions}

\item{genes.for.knn}{a vector of genes for searching nearest neighbors}

\item{assay}{Name of assay to use if genes.for.knn is provided}

\item{reduction}{Dimensional reduction to use for searching nearest neighbors, used if no adjacencies, 
knn or genes.for.knn is specified. Default is pca.}

\item{dims}{Dimensions to use for reduction embeddings}

\item{k.param}{Number of k nearest neighbors to search}

\item{eps}{Error bound when performing nearest neighbor seach using RANN; 
default of 0.0 implies exact nearest neighbor search}

\item{min_dist}{min_dist: This controls how tightly the embedding is allowed compress points together. 
Larger values ensure embedded points are more evenly distributed, while smaller values allow the algorithm 
to optimise more accurately with regard to local structure. Sensible values are in the range 0.001 to 0.5.}

\item{metric}{metric: This determines the choice of metric used to measure distance in the input space. A wide 
variety of metrics are already coded, and a user defined function can be passed as long as it has been JITd by numba.}

\item{...}{Additional parameters passed to umap}
}
\description{
This function calls several functions from umap, which is available at
https://github.com/tkonopka/umap.
The function runs the Uniform Manifold Approximation and Projection (UMAP) dimensional reduction,
on an adjacency matrix or a knn object. If no input adjacencies or knn is provided, 
the function will search for knn with user-provided genes or dimensional reductions before running UMAP.
}
