% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mixscape.R
\name{RunLDA}
\alias{RunLDA}
\alias{RunLDA.Assay}
\alias{RunLDA.Seurat}
\alias{RunLDA.default}
\title{Run Linear Discriminant Analysis}
\usage{
RunLDA(object, ...)

\method{RunLDA}{Assay}(
  object,
  labels,
  assay = NULL,
  features = NULL,
  verbose = TRUE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.key = "LDA_",
  seed.use = 42,
  ...
)

\method{RunLDA}{Seurat}(
  object,
  labels,
  assay = NULL,
  features = NULL,
  verbose = TRUE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.name = "lda",
  reduction.key = "LDA_",
  seed.use = 42,
  project = TRUE,
  ...
)

\method{RunLDA}{default}(
  object,
  labels,
  assay = NULL,
  verbose = TRUE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.key = "LDA_",
  seed.use = 42,
  ...
)
}
\arguments{
\item{object}{Input values for LDA (numeric), with observations as rows}

\item{...}{Arguments passed to other methods}

\item{labels}{Observation labels for LDA}

\item{assay}{Name of Assay LDA is being run on}

\item{features}{Features to compute LDA on}

\item{verbose}{Print the top genes associated with high/low loadings for
the PCs}

\item{ndims.print}{PCs to print genes for}

\item{nfeatures.print}{Number of genes to print for each PC}

\item{reduction.key}{dimensional reduction key, specifies the string before
the number for the dimension names. LDA by default}

\item{seed.use}{Set a random seed. By default, sets the seed to 42. Setting
NULL will not set a seed.}

\item{reduction.name}{dimensional reduction name,  lda by default}

\item{project}{}
}
\description{
Run Linear Discriminant Analysis

Function to perform Linear Discriminant Analysis.
}
