% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic.R
\name{RunMagic}
\alias{RunMagic}
\title{Impute expression based on neighboring cell using method based on}
\usage{
RunMagic(object, t = 6)
}
\description{
MAGIC: A diffusion-based imputation method reveals gene-gene interactions
in single-cell RNA-sequencing data
}
\details{
David van Dijk, Juozas Nainys, Roshan Sharma, Pooja Kathail, Ambrose J Carr,
Kevin R Moon, Linas Mazutis, Guy Wolf, Smita Krishnaswamy, Dana Pe'er
doi: https://doi.org/10.1101/111591
}
