% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{RunMarkVario}
\alias{RunMarkVario}
\title{Compute a mark variogram for a marked point pattern.}
\usage{
RunMarkVario(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns
}
\description{
Wraps the functionality of markvario from the spatstat package.
}
