% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{RunMarkVario.Seurat}
\alias{RunMarkVario.Seurat}
\title{Run the mark variogram computation on a given Seurat object}
\usage{
\method{RunMarkVario}{Seurat}(object, assay = NULL, image = NULL,
  slot = "scale.data", features = NULL, r.metric = 20, ...)
}
\arguments{
\item{object}{A Seurat object or data.frame giving x/y positions}

\item{assay}{Assay to pull the features (marks) from}

\item{image}{Name of image to pull the coordinates from}

\item{slot}{Slot in the Assay to pull data from}

\item{features}{Set of features to run on}

\item{r.metric}{r value at which to report the "trans" value of the mark 
variogram}
}
\value{
Returns a Seurat object with the output from markvario stored in
Tools and the r.metric stored in the Assay's meta.features data.frame
}
\description{
Run the mark variogram computation on a given Seurat object
}
