% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixscape.R
\name{RunMixscape}
\alias{RunMixscape}
\title{Function to identify perturbed and non-perturbed gRNA expressing cells.}
\usage{
RunMixscape(
  object = NULL,
  assay = "PRTB",
  slot = "scale.data",
  labels = "gene",
  nt.class.name = "NT",
  new.class.name = "mixscape_class",
  min.de.genes = 5,
  de.assay = "RNA",
  logfc.threshold = 0.25,
  iter.num = 10,
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of class Seurat.}

\item{assay}{Assay to use for mixscape classification.}

\item{slot}{Assay data slot to use.}

\item{labels}{metadata column with target gene classifications.}

\item{nt.class.name}{Classification name of non-targeting gRNA cells.}

\item{new.class.name}{Name of mixscape classification to be stored in 
metadata.}

\item{min.de.genes}{Required number of genes that are differentially 
expressed for method to separate perturbed and non-perturbed cells.}

\item{de.assay}{Assay to use when performing differential expression analysis. 
Usually RNA.}

\item{logfc.threshold}{Limit testing to genes which show, on average, at least
X-fold difference (log-scale) between the two groups of cells. Default is 0.25
Increasing logfc.threshold speeds up the function, but can miss weaker signals.}

\item{iter.num}{Number of normalmixEM iterations to run if convergence does 
not occur.}

\item{verbose}{Display messages}
}
\value{

}
\description{
Function to identify perturbed and non-perturbed gRNA expressing cells.
}
