% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/dimensional_reduction.R
\name{RunSLSI}
\alias{RunSLSI}
\alias{RunSLSI.default}
\alias{RunSLSI.Assay}
\alias{RunSLSI.Seurat}
\title{Run Supervised Latent Semantic Indexing}
\usage{
RunSLSI(object, ...)

\method{RunSLSI}{default}(
  object,
  assay = NULL,
  n = 50,
  reduction.key = "SLSI_",
  graph = NULL,
  verbose = TRUE,
  seed.use = 42,
  ...
)

\method{RunSLSI}{Assay}(
  object,
  assay = NULL,
  features = NULL,
  n = 50,
  reduction.key = "SLSI_",
  graph = NULL,
  verbose = TRUE,
  seed.use = 42,
  ...
)

\method{RunSLSI}{Seurat}(
  object,
  assay = NULL,
  features = NULL,
  n = 50,
  reduction.name = "slsi",
  reduction.key = "SLSI_",
  graph = NULL,
  verbose = TRUE,
  seed.use = 42,
  ...
)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{assay}{Name of Assay SLSI is being run on}

\item{n}{Total Number of SLSI components to compute and store}

\item{reduction.key}{dimensional reduction key, specifies the string before
the number for the dimension names}

\item{graph}{Graph used supervised by SLSI}

\item{verbose}{Display messages}

\item{seed.use}{Set a random seed. Setting NULL will not set a seed.}

\item{features}{Features to compute SLSI on. If NULL, SLSI will be run
using the variable features for the Assay.}

\item{reduction.name}{dimensional reduction name}
}
\value{
Returns Seurat object with the SPCA calculation stored in the
reductions slot
}
\description{
Run a supervised LSI (SLSI) dimensionality reduction supervised by a
cell-cell kernel. SLSI is used to capture a linear transformation that
maximizes its dependency to the given cell-cell kernel.
}
\references{
Barshan E, Ghodsi A, Azimifar Z, Jahromi MZ.
Supervised principal component analysis: Visualization, classification and
regression on subspaces and submanifolds.
Pattern Recognition. 2011 Jul 1;44(7):1357-71. \url{https://www.sciencedirect.com/science/article/pii/S0031320310005819?casa_token=AZMFg5OtPnAAAAAA:_Udu7GJ7G2ed1-XSmr-3IGSISUwcHfMpNtCj-qacXH5SBC4nwzVid36GXI3r8XG8dK5WOQui};
}
\concept{dimensional_reduction}
