% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{RunWorkflow}
\alias{RunWorkflow}
\title{Run a command from a workflow}
\usage{
RunWorkflow(object, workflow.name = "cluster", end = "FindClusters",
  start = NULL, ...)
}
\arguments{
\item{object}{Seurat object}

\item{workflow.name}{Workflow name, should already be initialized using InitializeWorkflow}

\item{command.name}{Name of the command to run}

\item{time.stamp}{Timestamp to assign}
}
\value{
Seurat object with updated workflow
}
\description{
Run workflow
}
\examples{
\dontrun{
TouchWorkflow(object = pbmc_small,workflow.name = "cluster", command.name = "ScaleData")
}

}
