% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SaveAnnoyIndex}
\alias{SaveAnnoyIndex}
\title{Save the Annoy index}
\usage{
SaveAnnoyIndex(object, file)
}
\arguments{
\item{object}{A Neighbor object with the annoy index stored}

\item{file}{Path to file to write index to}
}
\description{
Save the Annoy index
}
\concept{utilities}
