% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{SaveAnnoyNN}
\alias{SaveAnnoyNN}
\title{Save the Annoy neighbors and C++ index to RDS}
\usage{
SaveAnnoyNN(nn, file, annoy_index_name = "annoy_index")
}
\arguments{
\item{nn}{the output from AnnoyNN with return.annoy_index = T}

\item{file}{the name of the file where the nn object is saved}

\item{annoy_index_name}{suffix name for C++ index, which is saved 
in a separate file}
}
\description{
Save the Annoy neighbors and C++ index to RDS
}
