% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R, R/preprocessing_generics.R
\name{ScaleData.Assay}
\alias{ScaleData.Assay}
\alias{ScaleData.Seurat}
\alias{ScaleData}
\title{Scale and center the data.}
\usage{
\method{ScaleData}{Assay}(object, features.use = NULL,
  vars.to.regress = NULL, model.use = "linear", use.umi = FALSE,
  do.scale = TRUE, do.center = TRUE, scale.max = 10, block.size = 1000,
  min.cells.to.block = 3000, display.progress = TRUE, ...)

\method{ScaleData}{Seurat}(features.use = NULL, assay.use = NULL,
  vars.to.regress = NULL, model.use = "linear", use.umi = FALSE,
  do.scale = TRUE, do.center = TRUE, scale.max = 10, block.size = 1000,
  min.cells.to.block = 3000, display.progress = TRUE, ...)

ScaleData(object, features.use, vars.to.regress, model.use, use.umi, do.scale,
  do.center, scale.max, block.size, min.cells.to.block, display.progress, ...)
}
\arguments{
\item{object}{An object}

\item{features.use}{Vector of features names to scale/center. Default is all features}

\item{vars.to.regress}{Variables to regress out (previously latent.vars in
RegressOut). For example, nUMI, or percent.mito.}

\item{model.use}{Use a linear model or generalized linear model
(poisson, negative binomial) for the regression. Options are 'linear'
(default), 'poisson', and 'negbinom'}

\item{use.umi}{Regress on UMI count data. Default is FALSE for linear
modeling, but automatically set to TRUE if model.use is 'negbinom' or 'poisson'}

\item{do.scale}{Whether to scale the data.}

\item{do.center}{Whether to center the data.}

\item{scale.max}{Max value to return for scaled data. The default is 10.
Setting this can help reduce the effects of feautres that are only expressed in
a very small number of cells. If regressing out latent variables and using a
non-linear model, the default is 50.}

\item{block.size}{Default size for number of feautres to scale at in a single
computation. Increasing block.size may speed up calculations but at an
additional memory cost.}

\item{min.cells.to.block}{If object contains fewer than this number of cells,
don't block for scaling calculations.}

\item{display.progress}{Displays a progress bar for scaling procedure}

\item{assay.use}{Name of Assay to scale}
}
\description{
Scales and centers features in the dataset. If variables are provided in vars.to.regress,
they are individually regressed against each feautre, and the resulting residuals are
then scaled and centered.
}
\details{
ScaleData now incorporates the functionality of the function formerly known
as RegressOut (which regressed out given the effects of provided variables
and then scaled the residuals). To make use of the regression functionality,
simply pass the variables you want to remove to the vars.to.regress parameter.

Setting center to TRUE will center the expression for each feautre by subtracting
the average expression for that feautre. Setting scale to TRUE will scale the
expression level for each feautre by dividing the centered feautre expression
levels by their standard deviations if center is TRUE and by their root mean
square otherwise.
}
\section{Methods (by class)}{
\itemize{
\item \code{Assay}: Scale an Assay object

\item \code{Seurat}: Scale a Seurat object
}}

