% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ScaleDataR}
\alias{ScaleDataR}
\title{Old R based implementation of ScaleData. Scales and centers the data}
\usage{
ScaleDataR(object, genes.use = NULL, data.use = NULL, do.scale = TRUE,
  do.center = TRUE, scale.max = 10)
}
\arguments{
\item{object}{Seurat object}

\item{genes.use}{Vector of gene names to scale/center. Default is all genes in object@data.}

\item{data.use}{Can optionally pass a matrix of data to scale, default is object@data[genes.use,]}

\item{do.scale}{Whether to scale the data.}

\item{do.center}{Whether to center the data.}

\item{scale.max}{Max value to accept for scaled data. The default is 10. Setting this can help
reduce the effects of genes that are only expressed in a very small number of cells.}
}
\value{
Returns a seurat object with object@scale.data updated with scaled and/or centered data.
}
\description{
Old R based implementation of ScaleData. Scales and centers the data
}
\examples{
\dontrun{
pbmc_small <- ScaleDataR(object = pbmc_small)
}

}
