% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{ScaleFactors}
\alias{ScaleFactors}
\alias{scalefactors}
\alias{ScaleFactors.VisiumV1}
\title{Get image scale factors}
\usage{
ScaleFactors(object, ...)

scalefactors(spot, fiducial, hires, lowres)

\method{ScaleFactors}{VisiumV1}(object, ...)

\method{ScaleFactors}{VisiumV1}(object, ...)
}
\arguments{
\item{object}{An object to get scale factors from}

\item{...}{Arguments passed to other methods}

\item{spot}{Spot full resolution scale factor}

\item{fiducial}{Fiducial full resolution scale factor}

\item{hires}{High resolutoin scale factor}

\item{lowres}{Low resolution scale factor}
}
\value{
An object of class \code{scalefactors}
}
\description{
Get image scale factors
}
\note{
\code{scalefactors} objects can be created with \code{scalefactors()}
}
\concept{objects}
\concept{spatial}
