% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackstraw.R, R/jackstraw_generics.R
\name{ScoreJackStraw.JackStrawData}
\alias{ScoreJackStraw.JackStrawData}
\alias{ScoreJackStraw.DimReduc}
\alias{ScoreJackStraw.Seurat}
\alias{ScoreJackStraw}
\title{Compute Jackstraw scores significance.}
\usage{
\method{ScoreJackStraw}{JackStrawData}(object, dims = 1:5,
  score.thresh = 1e-05, ...)

\method{ScoreJackStraw}{DimReduc}(object, dims = 1:5, score.thresh = 1e-05,
  do.plot = FALSE, ...)

\method{ScoreJackStraw}{Seurat}(object, reduction.use = "pca", dims = 1:5,
  score.thresh = 1e-05, do.plot = FALSE, ...)

ScoreJackStraw(object, slot, ...)
}
\arguments{
\item{object}{Seurat plot}

\item{score.thresh}{Threshold to use for the proportion test of PC
significance (see Details)}

\item{do.plot}{Show plot. To return ggplot object, use JackStrawPlot after
running ScoreJackStraw.}

\item{reduction.use}{Reduction associated with JackStraw to score}

\item{PCs}{Which PCs to examine}

\item{nCol}{Number of columns}
}
\value{
Returns a Seurat object
}
\description{
Significant PCs should show a p-value distribution that is
strongly skewed to the left compared to the null distribution.
The p-value for each PC is based on a proportion test comparing the number
of genes with a p-value below a particular threshold (score.thresh), compared with the
proportion of genes expected under a uniform distribution of p-values.
}
\section{Methods (by class)}{
\itemize{
\item \code{JackStrawData}: Score JackStraw results given a JackStrawData

\item \code{DimReduc}: Score JackStraw results given a DimReduc

\item \code{Seurat}: Score JackStraw results given a Seurat object
}}

\author{
Thanks to Omri Wurtzel for integrating with ggplot
}
