% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing5.R
\name{SCTransform.IterableMatrix}
\alias{SCTransform.IterableMatrix}
\title{LayerData LayerData<-}
\usage{
\method{SCTransform}{IterableMatrix}(
  object,
  cell.attr = NULL,
  reference.SCT.model = NULL,
  do.correct.umi = TRUE,
  ncells = 5000,
  residual.features = NULL,
  variable.features.n = 3000,
  variable.features.rv.th = 1.3,
  vars.to.regress = NULL,
  do.scale = FALSE,
  do.center = TRUE,
  clip.range = c(-sqrt(x = ncol(x = object)/30), sqrt(x = ncol(x = object)/30)),
  conserve.memory = FALSE,
  return.only.var.genes = TRUE,
  seed.use = 1448145,
  verbose = TRUE,
  ...
)
}
\description{
LayerData LayerData<-
}
