% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{SCTResults}
\alias{SCTResults}
\alias{SCTResults.Assay}
\alias{SCTResults.SCTAssay}
\alias{SCTResults.Seurat}
\title{Get SCT results from an Assay}
\usage{
SCTResults(object, ...)

\method{SCTResults}{Assay}(object, slot, key = "1", ...)

\method{SCTResults}{SCTAssay}(object, slot, key = "1", ...)

\method{SCTResults}{Seurat}(object, assay = "SCT", slot, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods (not used)}

\item{slot}{Which slot to pull the SCT results from}

\item{key}{Key value}

\item{assay}{Assay in the Seurat object to pull from}

\item{slot}{Which slot to pull the SCT results from}

\item{key}{Key value}
}
\description{
Get SCT results from an Assay
}
