% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{SelectIntegrationFeatures5}
\alias{SelectIntegrationFeatures5}
\title{Select integration features}
\usage{
SelectIntegrationFeatures5(
  object,
  nfeatures = 2000,
  assay = NULL,
  method = NULL,
  layers = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{nfeatures}{Number of features to return for integration}

\item{assay}{Name of assay to use for integration feature selection}

\item{method}{Which method to pull. For \code{HVFInfo} and
\code{VariableFeatures}, choose one from one of the
following:
\itemize{
 \item \dQuote{vst}
 \item \dQuote{sctransform} or \dQuote{sct}
 \item \dQuote{mean.var.plot}, \dQuote{dispersion}, \dQuote{mvp}, or
  \dQuote{disp}
}}

\item{layers}{Name of layers to use for integration feature selection}

\item{verbose}{Print messages}
}
\description{
Select integration features
}
