% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{SelectSCTIntegrationFeatures}
\alias{SelectSCTIntegrationFeatures}
\title{Select SCT integration features}
\usage{
SelectSCTIntegrationFeatures(
  object,
  nfeatures = 3000,
  assay = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{nfeatures}{Number of features to return for integration}

\item{assay}{Name of assay to use for integration feature selection}

\item{verbose}{Print messages}
}
\description{
Select SCT integration features
}
