% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{SetAllIdent}
\alias{SetAllIdent}
\title{Switch identity class definition to another variable}
\usage{
SetAllIdent(object, id = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{id}{Variable to switch identity class to (for example, 'DBclust.ident', the output
of density clustering) Default is orig.ident - the original annotation pulled from the cell name.}
}
\value{
A Seurat object where object@ident has been appropriately modified
}
\description{
Switch identity class definition to another variable
}
