% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_generics.R, R/utilities_internal.R
\docType{methods}
\name{SetCalcParams}
\alias{SetCalcParams}
\alias{SetCalcParams,seurat-method}
\alias{SetCalcParams,loom-method}
\title{Set CalcParam information}
\usage{
SetCalcParams(object, ...)

\S4method{SetCalcParams}{seurat}(object, calculation, time = TRUE, ...)

\S4method{SetCalcParams}{loom}(object, dataset.use, time = TRUE, ...)
}
\arguments{
\item{object}{Seurat object}

\item{...}{Parameters for the calculation}

\item{calculation}{The name of the calculation that was done}

\item{time}{store time of calculation as well}

\item{dataset.use}{dataset on which to store the attribute (e.g. "layers/norm_data")}
}
\value{
object with the calc.param slot modified to either append this
calculation or replace the previous instance of calculation with
a new list of parameters. In a loom representation, the calc.params are stored as
}
\description{
Set CalcParam information
}
