% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{SetDimReduction}
\alias{SetDimReduction}
\title{Dimensional Reduction Mutator Function}
\usage{
SetDimReduction(object, reduction.type, slot, new.data)
}
\arguments{
\item{object}{Seurat object}

\item{reduction.type}{Type of dimensional reduction to set}

\item{slot}{Specific information to set (i.e. rotation, x, x.full, ...)}

\item{new.data}{New data to insert}
}
\value{
Seurat object with updated slot
}
\description{
Set information for specified stored dimensional reduction analysis
}
