% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_generics.R, R/assay.R
\name{SetHVFInfo}
\alias{SetHVFInfo}
\alias{SetHVFInfo.Assay}
\title{Set highly variable feature information}
\usage{
SetHVFInfo(object, hvf.info, ...)

\method{SetHVFInfo}{Assay}(object, hvf.info, ...)
}
\arguments{
\item{object}{An object}

\item{hvf.info}{A dataframe with feature means, dispersion, and scaled dispersion}

\item{...}{Arguments passed to GetVariableFeatures}
}
\description{
Set highly variable feature information
}
\section{Methods (by class)}{
\itemize{
\item \code{Assay}: Get highly variable feature information from an Assay
}}

\seealso{
GetVariableFeatures
}
