% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackstraw.R, R/jackstraw_generics.R
\name{SetJS.JackStrawData}
\alias{SetJS.JackStrawData}
\alias{SetJS}
\title{Set JackStrawData slot}
\usage{
\method{SetJS}{JackStrawData}(object, slot, new.data)

SetJS(object, slot, new.data, ...)
}
\arguments{
\item{object}{JackStrawData object}

\item{slot}{Name of slot to set}

\item{new.data}{New data to replace slot with}
}
\value{
Returns JackStrawData object with modified slot
}
\description{
Mutator method for JackStrawData
}
\section{Methods (by class)}{
\itemize{
\item \code{JackStrawData}: Set a slot for a given JackStrawData
}}

