% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setThreads}
\alias{setThreads}
\title{Set the number of threads to use for parallel processing in Seurat}
\usage{
setThreads(n = NULL, auto = TRUE)
}
\arguments{
\item{n}{Number of threads (>= 1) to use for processing}

\item{auto}{Automatically detect the number of threads to use (number of available cores) (default: TRUE)}
}
\description{
Set the number of threads to use for parallel processing in Seurat
}
\concept{utilities}
