% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{SetWorkflowParams}
\alias{SetWorkflowParams}
\title{Set Seurat workflow parameters}
\usage{
SetWorkflowParams(object, workflow.name = NULL, ...,
  fxn.param.names = NULL, fxn.param.values = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{workflow.name}{Workflow name, should already be initialized using
InitializeWorkflow}

\item{\dots}{Global parameters to set, will be fed into workflow Seurat functions.
Parameters ending with a "." will populate all similar variable names (i.e.
 setting dims. will set both dims.compute, and dims.cluster)}

\item{fxn.param.names}{Name of the function and parameter to set (formatted as
FXNNAME_PARAM). Can take a vector to set multiple functions}

\item{fxn.param.values}{Value of the parameter to set. Should be of equal length
to fxn.param.names}
}
\value{
Object with modified workflows
}
\description{
Sets parameters for a workflow
}
\examples{
\dontrun{
pbmc_small <- SetWorkflowParams(object = pbmc_small, seed.use = 31, dims. = 20)
}

}
