% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{Seurat-deprecated}
\alias{Seurat-deprecated}
\alias{feature.plot}
\alias{find.markers}
\alias{find_all_markers}
\alias{genePlot}
\alias{jackStraw}
\alias{jackStrawPlot}
\alias{mean.var.plot}
\alias{pcHeatmap}
\alias{pca}
\alias{pca.plot}
\alias{print.pca}
\alias{project.pca}
\alias{run_tsne}
\alias{set.ident}
\alias{setup}
\alias{subsetData}
\alias{tnse.plot}
\alias{tsne.plot}
\alias{viz.pca}
\alias{vlnPlot}
\title{Deprecated function(s) in the Seurat package}
\usage{
setup(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
These functions are provided for compatibility with older version of the Seurat package.  They may eventually be completely removed.
}
\section{Details}{

\tabular{rl}{
  \code{setup} \tab now a synonym for \code{\link{Setup}}\cr
  \code{vlnPlot} \tab now a synonym for \code{\link{VlnPlot}}\cr
  \code{subsetData} \tab now a synonym for \code{\link{SubsetData}}\cr
  \code{mean.var.plot} \tab now a synonym for \code{\link{MeanVarPlot}}\cr
  \code{pca} \tab now a synonym for \code{\link{PCA}}\cr
  \code{project.pca} \tab now a synonym for \code{\link{ProjectPCA}}\cr
  \code{print.pca} \tab now a synonym for \code{\link{PrintPCA}}\cr
  \code{viz.pca} \tab now a synonym for \code{\link{VizPCA}}\cr
  \code{set.ident} \tab now a synonym for \code{\link{SetIdent}}\cr
  \code{pca.plot} \tab now a synonym for \code{\link{PCAPlot}}\cr
  \code{pcHeatmap} \tab now a synonym for \code{\link{PCHeatmap}}\cr
  \code{jackStraw} \tab now a synonym for \code{\link{JackStraw}}\cr
  \code{jackStrawPlot} \tab now a synonym for \code{\link{JackStrawPlot}}\cr
  \code{run_tsne} \tab now a synonym for \code{\link{RunTSNE}}\cr
  \code{tsne.plot} \tab now a synonym for \code{\link{TSNEPlot}}\cr
  \code{find.markers} \tab now a synonym for \code{\link{FindMarkers}}\cr
  \code{find_all_markers} \tab now a synonym for \code{\link{FindAllMarkers}}\cr
  \code{genePlot} \tab now a synonym for \code{\link{GenePlot}}\cr
  \code{feature.plot} \tab now a synonym for \code{\link{FeaturePlot}}\cr
}
}

