% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seurat.R
\name{seurat}
\alias{seurat}
\title{Constructor function}
\usage{
seurat(...)
}
\arguments{
\item{\dots}{parameters to pass into the Seurat object. Usually will pass in raw.data (data frame with rows as genes, columns as cell names)}
}
\description{
Setup a new Seurat object, pass in the raw data for downstream analysis
}

