% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{methods}
\name{[[<-,Assay-method}
\alias{[[<-,Assay-method}
\alias{[[<-,Seurat-method}
\title{Set metadata and add objects}
\usage{
\S4method{[[}{Assay}(x, i, j, ...) <- value

\S4method{[[}{Seurat}(x, i, j, ...) <- value
}
\arguments{
\item{x}{An object}

\item{i}{Name of metadata or object to add}

\item{j}{Not used}

\item{...}{Arguments passed to other methods}

\item{value}{Metadata or object to add}
}
\value{
An object with metadata or and object added
}
\description{
Set metadata and add objects
}
