% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{SeuratCommand-class}
\alias{SeuratCommand-class}
\alias{SeuratCommand}
\title{The SeuratCommand Class}
\description{
The SeuratCommand is used for logging commands that are run on a SeuratObject. It stores parameters and timestamps
}
\section{Slots}{

\describe{
\item{\code{name}}{Command name}

\item{\code{time.stamp}}{Timestamp of when command was tun}

\item{\code{assay.used}}{Optional name of assay used to generate \code{SeuratCommand} object}

\item{\code{call.string}}{String of the command call}

\item{\code{params}}{List of parameters used in the command call}
}}

