% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_themes.R
\name{DarkTheme}
\alias{DarkTheme}
\alias{NoAxes}
\alias{NoLegend}
\title{Seurat Themes}
\usage{
DarkTheme(...)

NoAxes(...)

NoLegend(...)
}
\arguments{
\item{...}{Extra parameters to be passed to \code{theme}}
}
\value{
A ggplot2 theme object
}
\description{
Various themes to be applied to ggplot2-based plots
}
\examples{
# Generate a plot with a dark theme
df <- data.frame(x = rnorm(n = 100, mean = 20, sd = 2), y = rbinom(n = 100, size = 100, prob = 0.2))
p <- ggplot(data = df, mapping = aes(x = x, y = y)) + geom_point(mapping = aes(color = 'red'))
p + DarkTheme(legend.position = 'none')

# Generate a plot with no axes
df <- data.frame(x = rnorm(n = 100, mean = 20, sd = 2), y = rbinom(n = 100, size = 100, prob = 0.2))
p <- ggplot(data = df, mapping = aes(x = x, y = y)) + geom_point(mapping = aes(color = 'red'))
p + NoAxes()

# Generate a plot with no legend
df <- data.frame(x = rnorm(n = 100, mean = 20, sd = 2), y = rbinom(n = 100, size = 100, prob = 0.2))
p <- ggplot(data = df, mapping = aes(x = x, y = y)) + geom_point(mapping = aes(color = 'red'))
p + NoLegend()

}
\seealso{
\code{\link{ggplot2::theme}}
}
