% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{SeuratWorkflow}
\alias{SeuratWorkflow}
\title{The SeuratWorkflow class}
\description{
The SeuratWorkflow class aims to create a Makefile-like approach for Seurat
analysis. A full set of parameters can be defined with a single command, and
the workflow encodes the pipeline command structure (for example,
NormalizeData -> FindVariableFeatures -> ScaleData -> RunPCA -> FindClusters).
This leads to very simple workflows being encoded in just a couple commands,
without losing the flexibility to run each step for clarity if desired.
}
\section{Slots}{

\describe{
\item{\code{name}}{Workflow name}

\item{\code{depends}}{Dependency graph encoding the relationship between commands
(for example, RunPCA depends on ScaleData)}

\item{\code{update}}{Vector specifying for each command, if it needs to be
updated/rerun}

\item{\code{params}}{List of parameters used across the workflow}

\item{\code{mostRecent}}{Vector specifying for each command, the most recent
timestamp. If the current timestamp matches this, should not need to be rerun.}
}}

