% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{ShiftDim}
\alias{ShiftDim}
\title{Shift dim scores}
\usage{
ShiftDim(object, grouping.var, reduction.type, dims.shift, ds.amt = 0.4,
  constrain.var = T)
}
\arguments{
\item{object}{Seurat object}

\item{reduction.type}{reduction to shift scores for}

\item{dims.shift}{Dims to shift, default is all}

\item{ds.amt}{percent of cells in each group for downsampling}

\item{group.var}{Name of the grouping variable for which to shift the scores}

\item{search.lower}{lower bounds on the grid search for optimal shifting parameters. Defaults to 
25% quantile for intercept parameter and 0.25 for slope}

\item{search.upper}{upper bounds on the grid search for optimal shifting parameters. Defaults to 
75% quantile for intercept parameter and 4 for slope.}

\item{search.levels}{Controls the fineness of the grid search. Larger number will evaulate more
options}
}
\value{
Returns Seurat object with the dims shifted, stored in object@dr$reduction.type.shifted
}
\description{
Shifts dim scores so that they line up across grouping variable (only implemented for case with
2 categories in grouping.var)
}
