% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{ShiftDimDTW}
\alias{ShiftDimDTW}
\title{Shift dim scores using DTW}
\usage{
ShiftDimDTW(object, reduction.type, grouping.var, dims.shift, num.genes = 20)
}
\arguments{
\item{object}{Seurat object}

\item{reduction.type}{reduction to shift scores for}

\item{dims.shift}{Dims to shift, default is all}

\item{num.genes}{Number of genes to use in construction of "metagene"}

\item{group.var}{Name of the grouping variable for which to shift the scores}
}
\value{
Returns Seurat object with the dims shifted, stored in object@dr$reduction.type.shifted
}
\description{
Shifts dim scores so that they line up across grouping variable (only implemented for case with
2 categories in grouping.var)
}
