% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,seurat-method}
\title{show method for seurat}
\usage{
\S4method{show}{seurat}(object)
}
\arguments{
\item{object}{A Seurat object}
}
\description{
show method for seurat
}
