% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFxns.R
\name{shuffleMatRow}
\alias{shuffleMatRow}
\title{Independently shuffle values within each row of a matrix}
\usage{
shuffleMatRow(x)
}
\arguments{
\item{x}{Matrix to shuffle}
}
\value{
Returns a scrambled matrix, where each row is shuffled independently
}
\description{
Creates a matrix where correlation structure has been removed, but overall values are the same
}
