% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SingleImageMap}
\alias{SingleImageMap}
\title{A single heatmap from base R using \code{\link[graphics]{image}}}
\usage{
SingleImageMap(data, order = NULL, title = NULL)
}
\arguments{
\item{data}{matrix of data to plot}

\item{order}{optional vector of cell names to specify order in plot}

\item{title}{Title for plot}
}
\value{
No return, generates a base-R heatmap using \code{\link[graphics]{image}}
}
\description{
A single heatmap from base R using \code{\link[graphics]{image}}
}
\keyword{internal}
