% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SingleImagePlot}
\alias{SingleImagePlot}
\title{Single Spatial Plot}
\usage{
SingleImagePlot(
  data,
  col.by = NA,
  cols = NULL,
  molecules = NULL,
  mols.size = 0.1,
  mols.cols = NULL,
  alpha = molecules \%iff\% 0.3 \%||\% 0.6,
  border.color = "black",
  border.size = 0.3,
  na.value = "grey50",
  ...
)
}
\arguments{
\item{data}{A data frame with the following columns:
\itemize{
 \item \dQuote{\code{x}}
 \item \dQuote{\code{y}}
 \item \dQuote{\code{cell}}
}}

\item{col.by}{...}

\item{cols}{...}

\item{molecules}{...}

\item{mols.size}{...}

\item{mols.cols}{...}

\item{border.color}{...}

\item{border.size}{...}

\item{na.value}{...}

\item{...}{...}
}
\value{
A ggplot object
}
\description{
Single Spatial Plot
}
\keyword{internal}
