% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfRenderSeurat.R
\name{situ3d}
\alias{situ3d}
\title{Draw 3D in situ predictions from Zebrafish dataset}
\usage{
situ3d(data, label = NULL, ...)
}
\arguments{
\item{data}{Predicted expression levels across Zebrafish bins}

\item{label}{Plot label}
}
\description{
From Jeff Farrell
}
