% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\docType{class}
\name{SliceImage-class}
\alias{SliceImage-class}
\alias{SliceImage}
\title{The SliceImage class}
\description{
The SliceImage class represents spatial information from the 10X Genomics Visium
platform
}
\section{Slots}{

\describe{
\item{\code{image}}{A three-dimensional array with PNG image data, see
\code{\link[png]{readPNG}} for more details}

\item{\code{scale.factors}}{An object of class \code{\link{scalefactors}}; see
\code{\link{scalefactors}} for more information}

\item{\code{coordinates}}{A data frame with tissue coordinate information}
}}

