% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\docType{class}
\name{SpatialImage-class}
\alias{SpatialImage-class}
\alias{SpatialImage}
\title{The SpatialImage class}
\description{
The SpatialImage class is a virtual class representing spatial information for
Seurat. All spatial image information must inherit from this class for use with
\code{Seurat} objects
}
\section{Slots}{

\describe{
\item{\code{assay}}{Name of assay to associate image data with; will give this image
priority for visualization when the assay is set as the active/default assay
in a \code{Seurat} object}
}}

\section{Provided methods}{

These methods are defined on the \code{SpatialImage} object and should not be
overwritten without careful thought
\itemize{
  \item \code{\link{DefaultAssay}} and \code{\link{DefaultAssay<-}}
  \item \code{\link{IsGlobal}}
}
}

\section{Required methods}{

All subclasses of the \code{SpatialImage} class must define the following methods;
simply relying on the \code{SpatialImage} method will result in errors
\describe{
  \item{\code{\link{Cells}}}{Return the cell/spot barcodes associated with each position}
  \item{\code{\link{dim}}}{...}
  \item{\code{\link{GetImage}}}{Return image data; by default, must return a grob object}
  \item{\code{\link{GetTissueCoordinates}}}{Return tissue coordinates; by default,
  must return a two-column data.frame with x-coordinates in the first column and y-coordiantes
  in the second}
  \item{\code{\link{RenameCells}}}{Rename the cell/spot barcodes for this image}
  \item{\code{\link{subset}} and \code{[}}{Subset the image data by cells/spots;
  \code{[} should only take \code{i} for subsetting by cells/spots}
}
These methods are used throughout Seurat, so defining them and setting the proper
defaults will allow subclasses of \code{SpatialImage} to work seamlessly
}

