% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{SpatialDimPlot}
\alias{SpatialDimPlot}
\alias{SpatialFeaturePlot}
\alias{SpatialPlot}
\title{Visualize spatial clustering and expression data}
\usage{
SpatialDimPlot(object, group.by = NULL, images = NULL, label = FALSE,
  label.size = 7, label.color = "white", repel = FALSE,
  ncol = NULL, combine = TRUE, pt.size.factor = 1, alpha = 1,
  stroke = 0.25, box = TRUE, do.hover = FALSE)

SpatialFeaturePlot(object, features, images = NULL, slot = "data",
  min.cutoff = NA, max.cutoff = NA, ncol = NULL, combine = TRUE,
  pt.size.factor = 1, alpha = 1, stroke = 0.25, do.hover = FALSE)

SpatialPlot(object, group.by = NULL, features = NULL, images = NULL,
  slot = "data", min.cutoff = NA, max.cutoff = NA, label = FALSE,
  label.size = 5, label.color = "white", repel = FALSE, box = TRUE,
  ncol = NULL, combine = TRUE, pt.size.factor = 1, alpha = 1,
  stroke = 0.25, do.hover = FALSE)
}
\value{
A ggplot object
}
\description{
Visualize spatial clustering and expression data
}
