% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{SplitObject}
\alias{SplitObject}
\title{Splits object into a list of subsetted objects.}
\usage{
SplitObject(object, attribute.1 = "ident")
}
\arguments{
\item{object}{Seurat object}

\item{attribute.1}{Attribute for splitting. Default is "ident". Currently only supported for class-level (i.e. non-quantitative) attributes.}

\item{\dots}{Additional parameters to pass to SubsetData}
}
\value{
A named list of Seurat objects, each containing a subset of cells from the original object.
}
\description{
Splits object based on a single attribute into a list of subsetted objects, one for each level of the attribute. For example, useful for taking an object that contains cells from many patients, and subdividing it into patient-specific objects.
}
\examples{
NEED AN EXAMPLE

}
