% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\docType{class}
\name{STARmap-class}
\alias{STARmap-class}
\alias{STARmap}
\title{The STARmap class}
\description{
The STARmap class represents spatial information from the STARmap platform
}
\section{Slots}{

\describe{
\item{\code{...}}{}
}}

