% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreduc_generics.R, R/seurat.R, R/dimreduc.R
\name{Stdev}
\alias{Stdev}
\alias{Stdev.Seurat}
\alias{Stdev.DimReduc}
\title{Get the standard deviations for an object}
\usage{
Stdev(object, ...)

\method{Stdev}{Seurat}(object, reduction.use, ...)

\method{Stdev}{DimReduc}(object)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{reduction.use}{Name of reduction to use}
}
\description{
Get the standard deviations for an object
}
\section{Methods (by class)}{
\itemize{
\item \code{Seurat}: Get the standard deviations of a dimensional reduction from a Seurat object

\item \code{DimReduc}: Get the standard deviations from a DimReduc object
}}

