% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{subc}
\alias{subc}
\title{#returns the setdiff of two vectors, i.e all elements of a that are not in b
#we should be switching to use the setdiff function instead
#' @export
setdiff <- function(x, y) {
  return(x[! x %in% y])
}
Return a subset of columns for a matrix or data frame}
\usage{
subc(data, code)
}
\arguments{
\item{data}{Matrix or data frame with column names}

\item{code}{Pattern for matching within column names}
}
\value{
Returns a subset of data, using only column names that yield a match to the pattern
}
\description{
#returns the setdiff of two vectors, i.e all elements of a that are not in b
#we should be switching to use the setdiff function instead
#' @export
setdiff <- function(x, y) {
  return(x[! x %in% y])
}
Return a subset of columns for a matrix or data frame
}
