% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFxns.R
\name{subr}
\alias{subr}
\title{Return a subset of rows for a matrix or data frame}
\usage{
subr(data, code)
}
\arguments{
\item{data}{Matrix or data frame with row names}

\item{code}{Pattern for matching within row names}
}
\value{
Returns a subset of data, using only rownames that yielded a match to the pattern
}
\description{
Return a subset of rows for a matrix or data frame
}
