% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_generics.R, R/interaction.R
\docType{methods}
\name{SubsetSeurat}
\alias{SubsetSeurat}
\alias{SubsetSeurat,loom-method}
\title{Subset a seurat/loom object}
\usage{
SubsetSeurat(object, cells, ...)

\S4method{SubsetSeurat}{loom}(object, cells,
  cell.names = "col_attrs/cell_names", gene.names = "row_attrs/gene_names",
  filename, remove.existing = FALSE, display.progress = TRUE,
  chunk.size = 1000, keep.layers = FALSE, ...)
}
\arguments{
\item{object}{An object to subset}

\item{cells}{A character vector of cells to keep}
}
\description{
Subset a seurat/loom object
}
