% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/spatial.R
\name{SVFInfo}
\alias{SVFInfo}
\alias{SVFInfo.Assay}
\title{Get spatially variable feature information}
\usage{
SVFInfo(object, ...)

\method{SVFInfo}{Assay}(object, selection.method = "markvariogram", status = FALSE, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{selection.method}{Which method to pull. Options: markvariogram}

\item{status}{Add variable status to the resulting data.frame}
}
\description{
Get spatially variable feature information
}
