% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{TCCBarPlot}
\alias{TCCBarPlot}
\title{Barplot to visualize EC and feature-level differences}
\usage{
TCCBarPlot(object, assay = "tcc", feature, group.by = NULL, group1,
  group2, ec.threshold, ambig = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Assay with the TCC info}

\item{feature}{Feature to produce barplots for}

\item{group.by}{Name of metadata column to group by.}

\item{group1}{Name of the first group to display}

\item{group2}{Name of the second group to display}

\item{ec.threshold}{Minimum counts per cell}

\item{ambig}{Include ECs mapping to more than one gene?}
}
\value{
Returns ggplot object
}
\description{
Given a feature, this function plots the mean counts for each EC that maps to
the feature and passes the filters, split by group. Also plots the aggregated
feature level mean counts for each group.
}
