% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{TCCStemPlot}
\alias{TCCStemPlot}
\title{Feature level stemplot for TCCs}
\usage{
TCCStemPlot(object, feature, assay = "tcc", log.scale = TRUE,
  ec.threshold = NULL, ec.fail.color = "grey",
  ec.pass.color = "blue", multimap.color = "red", print.info = TRUE)
}
\arguments{
\item{object}{Seurat object}

\item{feature}{Feature(s) to produce stem plots for}

\item{log.scale}{Plot UMIs on a log scale (natural log - log1p)}

\item{ec.threshold}{Minimum counts per cell}

\item{ec.fail.color}{Color for ECs that fail the threshold cutoff}

\item{ec.pass.color}{Color for ECs that pass threshold cutoff}

\item{multimap.color}{Color for ECs that map to more than 1 gene}

\item{print.info}{Print some stats for each feature plotted}
}
\value{
Returns ggplot object
}
\description{
A useful plot for visualizing information about the equivalence classes (ECs)
that map to certain features. Plots the distribution of UMI counts for each EC
with options for investigating how filtering would affect would change the ECs
that are kept.
}
