% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{TopFeatures}
\alias{TopFeatures}
\title{Find features with highest scores for a given dimensional reduction technique}
\usage{
TopFeatures(object, dim.use = 1, num.features = 20,
  projected = FALSE, do.balanced = FALSE)
}
\arguments{
\item{object}{DimReduc object}

\item{dim.use}{Dimension to use}

\item{num.features}{Number of features to return}

\item{projected}{Use the full PCA (projected PCA). Default i s FALSE}

\item{do.balanced}{Return an equal number of features with both + and - scores.}
}
\value{
Returns a vector of features
}
\description{
Return a list of features with the strongest contribution to a set of components
}
\examples{
pbmc_small
TopFeatures(object = pbmc_small, dim.use = 1, reduction.type = "pca")
# After projection:
TopFeatures(object = pbmc_small, dim.use = 1, reduction.type = "pca", use.full = TRUE)

}
