% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreduc.R
\name{TopGenes}
\alias{TopGenes}
\title{Find genes with highest scores for a given dimensional reduction technique}
\usage{
TopGenes(object, dim.use = 1, num.features = 20, projected = FALSE,
  do.balanced = FALSE)
}
\arguments{
\item{object}{DimReduc object}

\item{dim.use}{Dimension to use}

\item{num.features}{Number of genes to return}

\item{projected}{Use the full PCA (projected PCA). Default i s FALSE}

\item{do.balanced}{Return an equal number of genes with both + and - scores.}
}
\value{
Returns a vector of genes
}
\description{
Return a list of genes with the strongest contribution to a set of components
}
\examples{
pbmc_small
DimTopGenes(object = pbmc_small, dim.use = 1, reduction.type = "pca")
# After projection:
DimTopGenes(object = pbmc_small, dim.use = 1, reduction.type = "pca", use.full = TRUE)

}
