% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{TopNeighbors}
\alias{TopNeighbors}
\title{Get nearest neighbors for given cell}
\usage{
TopNeighbors(object, cell, n = 5)
}
\arguments{
\item{object}{\code{\link{Neighbor}} object}

\item{cell}{Cell of interest}

\item{n}{Number of neighbors to return}
}
\value{
Returns a vector of cell names
}
\description{
Return a vector of cell names of the nearest n cells.
}
\concept{objects}
