% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{TouchWorkflow}
\alias{TouchWorkflow}
\title{Updates workflow timestamps}
\usage{
TouchWorkflow(object, workflow.name, command.name, time.stamp = Sys.time())
}
\arguments{
\item{object}{Seurat object}

\item{workflow.name}{Workflow name, should already be initialized using InitializeWorkflow}

\item{command.name}{Name of the command to touch}

\item{time.stamp}{Timestamp to assign}
}
\value{
Seurat object with updated workflow
}
\description{
Like the touch command in linux. Updates a workflow command's timestamp, and its dependencies
}
\examples{
TouchWorkflow(object = pbmc_small,workflow.name = "cluster", command.name = "ScaleData")
#'
}
