% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seurat.R
\name{tsne.plot}
\alias{tsne.plot}
\title{Plot tSNE map}
\usage{
tsne.plot(object, do.label = FALSE, label.pt.size = 1, label.cex.text = 1,
  label.cols.use = NULL, ...)
}
\arguments{
\item{object}{Seurat object}

\item{do.label}{FALSE by default. If TRUE, plots an alternate view where the center of each
cluster is lebeled}

\item{label.pt.size}{If do.label is set, the point size}

\item{label.cex.text}{If label.cex.text is set, the size of the text labels}

\item{label.cols.use}{If do.label is set, the color palette to use for the points}

\item{\dots}{Additional parameters to dim.plot, for example, which dimensions to plot.}
}
\description{
Graphs the output of a tSNE analysis
Cells are colored by their identity class.
}
\details{
This function is a wrapper for dim.plot. See ?dim.plot for a full list of possible
arguments which can be passed in here.
}
\seealso{
dim.plot
}

