% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{TSNEPlot}
\alias{TSNEPlot}
\title{Plot tSNE map}
\usage{
TSNEPlot(object, do.label = FALSE, pt.size = 1, label.size = 4,
  cells.use = NULL, colors.use = NULL, ...)
}
\arguments{
\item{object}{Seurat object}

\item{do.label}{FALSE by default. If TRUE, plots an alternate view where the center of each
cluster is labeled}

\item{pt.size}{Set the point size}

\item{label.size}{Set the size of the text labels}

\item{cells.use}{Vector of cell names to use in the plot.}

\item{colors.use}{Manually set the color palette to use for the points}

\item{\dots}{Additional parameters to DimPlot, for example, which dimensions to plot.}
}
\description{
Graphs the output of a tSNE analysis
Cells are colored by their identity class.
}
\details{
This function is a wrapper for DimPlot. See ?DimPlot for a full list of possible
arguments which can be passed in here.
}
\seealso{
DimPlot
}
