% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{UnSketchEmbeddings}
\alias{UnSketchEmbeddings}
\title{Transfer embeddings from sketched cells to the full data}
\usage{
UnSketchEmbeddings(
  atom.data,
  atom.cells = NULL,
  orig.data,
  embeddings,
  sketch.matrix = NULL
)
}
\arguments{
\item{atom.data}{Atom data}

\item{atom.cells}{Atom cells}

\item{orig.data}{Original data}

\item{embeddings}{Embeddings of atom cells}

\item{sketch.matrix}{Sketch matrix}
}
\description{
Transfer embeddings from sketched cells to the full data
}
