% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{UpdateAssay}
\alias{UpdateAssay}
\title{Update Seurat assay}
\usage{
UpdateAssay(old.assay, assay.name)
}
\arguments{
\item{old.assay}{Seurat2 assay}

\item{assay.name}{Name to store for assay in new object}
}
\description{
Update Seurat assay
}
