% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{UpdateDimReduction}
\alias{UpdateDimReduction}
\title{Update dimension reduction}
\usage{
UpdateDimReduction(old.dr)
}
\arguments{
\item{old.dr}{Seurat2 dimension reduction slot}
}
\description{
Update dimension reduction
}
