% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{UpdateNeighbor}
\alias{UpdateNeighbor}
\title{Update old neighbor output to neighbor object}
\usage{
UpdateNeighbor(nn, cells)
}
\arguments{
\item{nn}{Neighbor list containing nn.idx and nn.dist}

\item{cells}{Names of cells for row name of nn.idx and nn.dist}
}
\value{
Returns a Neighbor object compatible with latest changes
}
\description{
Update old neighbor output to neighbor object
}
