% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{UpdateSCTAssays}
\alias{UpdateSCTAssays}
\title{Update pre-V4 Assays generated with SCTransform in the Seurat to the new
SCTAssay class}
\usage{
UpdateSCTAssays(object)
}
\arguments{
\item{object}{A Seurat object}
}
\value{
A Seurat object with updated SCTAssays
}
\description{
Update pre-V4 Assays generated with SCTransform in the Seurat to the new
SCTAssay class
}
\concept{objects}
