% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{UpdateSeuratObject}
\alias{UpdateSeuratObject}
\title{Update old Seurat object to accomodate new features}
\usage{
UpdateSeuratObject(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{
Returns a Seurat object compatible with latest changes
}
\description{
Updates Seurat objects to new structure for storing data/calculations.
For Seurat v3 objects, will validate object structure ensuring all keys and feature
names are formed properly.
}
\examples{
\dontrun{
updated_seurat_object = UpdateSeuratObject(object = old_seurat_object)
}

}
