% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seurat.R
\name{validate.clusters}
\alias{validate.clusters}
\title{Cluster Validation}
\usage{
validate.clusters(object, pc.use = NULL, top.genes = 30, SNN = NULL,
  min_connectivity = 0.01, acc.cutoff = 0.9)
}
\arguments{
\item{object}{Seurat object}

\item{pc.use}{Which PCs to use for model construction}

\item{top.genes}{Use the top X genes for model construction}

\item{SNN}{SNN matrix to use for tree building algorithm}

\item{min_connectivity}{Threshold of connectedness for comparison of two clusters}

\item{acc.cutoff}{Accuracy cutoff for classifier}
}
\value{
Returns a Seurat object, object@ident has been updated with new cluster info
}
\description{
Methods for validating the legitimacy of clusters using
classification. SVMs are used as the basis for the classification.
Merging is done recursively up the tree built by buildClusterTree using
an SNN graph to build the tree.
}

