% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cluster_validation.R
\name{ValidateSpecificClusters}
\alias{ValidateSpecificClusters}
\title{Specific Cluster Validation}
\usage{
ValidateSpecificClusters(object, cluster1 = NULL, cluster2 = 1,
  pc.use = 2, top.genes = 30, SNN.use = NULL, acc.cutoff = 0.9)
}
\arguments{
\item{object}{Seurat object}

\item{cluster1}{First cluster to check classification}

\item{cluster2}{Second cluster to check with classification}

\item{pc.use}{Which PCs to use for model construction}

\item{top.genes}{Use the top X genes for model construction}

\item{acc.cutoff}{Accuracy cutoff for classifier}

\item{SNN}{SNN matrix used for cluster comparison}
}
\value{
Returns a Seurat object, object@ident has been updated with
new cluster info
}
\description{
Methods for validating the legitimacy of two specific clusters using
classification. SVMs are used as the basis for the classification.
Merging is done based on the connectivity from an SNN graph.
}

