% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{VariableFeaturePlot}
\alias{VariableFeaturePlot}
\title{View variable features}
\usage{
VariableFeaturePlot(object, pt.size = 1, assay.use = NULL, ...)
}
\arguments{
\item{object}{Seurat object}

\item{pt.size}{Size of the points on the plot}

\item{assay.use}{Assay to pull from}

\item{...}{Ignored for now}
}
\description{
View variable features
}
\examples{
VariableFeaturePlot(object = pbmc_small)

}
