% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_generics.R, R/seurat.R, R/assay.R
\name{VariableFeatures}
\alias{VariableFeatures}
\alias{VariableFeatures<-}
\alias{VariableFeatures.Seurat}
\alias{VariableFeatures.Assay}
\alias{VariableFeatures<-.Assay}
\title{Get and set variable feature information}
\usage{
VariableFeatures(object, ...)

VariableFeatures(object, ...) <- value

\method{VariableFeatures}{Seurat}(object, assay.use = NULL, ...)

\method{VariableFeatures}{Assay}(object, ...)

\method{VariableFeatures}{Assay}(object, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{A character vector of variable features}

\item{assay.use}{Name of assay to pull variable features for}
}
\description{
Get and set variable feature information
}
\section{Methods (by class)}{
\itemize{
\item \code{Seurat}: Get the variable features of a Seurat object

\item \code{Assay}: Get the variable features of an assay object

\item \code{Assay}: Set the variable features of an assay object
}}

