% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{VisiumV1-class}
\alias{VisiumV1-class}
\alias{VisiumV1}
\title{The VisiumV1 class}
\description{
The VisiumV1 class represents spatial information from the 10X Genomics Visium
platform
}
\section{Slots}{

\describe{
\item{\code{image}}{A three-dimensional array with PNG image data, see
\code{\link[png]{readPNG}} for more details}

\item{\code{scale.factors}}{An object of class \code{\link{scalefactors}}; see
\code{\link{scalefactors}} for more information}

\item{\code{coordinates}}{A data frame with tissue coordinate information}

\item{\code{spot.radius}}{Single numeric value giving the radius of the spots}
}}

\concept{objects}
\concept{spatial}
