% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{VisiumV2-class}
\alias{VisiumV2-class}
\alias{VisiumV2}
\title{The VisiumV2 class}
\description{
The VisiumV2 class represents spatial information from the 10X Genomics 
Visium HD platform - it can also accomodate data from the standard 
Visium platform
}
\section{Slots}{

\describe{
\item{\code{image}}{A three-dimensional array with PNG image data, see
\code{\link[png]{readPNG}} for more details}

\item{\code{scale.factors}}{An object of class \code{\link{scalefactors}}; see
\code{\link{scalefactors}} for more information}
}}

\concept{objects}
\concept{spatial}
