% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{VizDimReduction}
\alias{VizDimReduction}
\title{Visualize Dimensional Reduction genes}
\usage{
VizDimReduction(object, reduction.use = "pca", dims.use = 1:5,
  num.features = 30, pt.color = "blue", projected = FALSE, nCol = NULL,
  do.balanced = FALSE, combine.plots = TRUE, ...)
}
\arguments{
\item{object}{Seurat object}

\item{reduction.use}{Reduction technique to visualize results for}

\item{dims.use}{Number of dimensions to display}

\item{num.features}{Number of genes to display}

\item{pt.color}{Color of points to use}

\item{projected}{Use reduction values for full dataset (i.e. projected dimensional reduction values)}

\item{nCol}{Number of columns to display}

\item{do.balanced}{Return an equal number of genes with + and - scores. If FALSE (default), returns
the top genes ranked by the scores absolute values}

\item{combine.plots}{Combine plots into a single gg object; note that if TRUE; themeing will not work when plotting multiple features}

\item{...}{Ignored}
}
\value{
A ggplot object
}
\description{
Visualize top genes associated with reduction components
}
\examples{
VizDimReduction(object = pbmc_small)

}
