% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{WhichCells}
\alias{WhichCells}
\alias{WhichCells.Assay}
\alias{WhichCells.Seurat}
\title{Identify cells matching certain criteria}
\usage{
WhichCells(object, cells, expression, invert, ...)

\method{WhichCells}{Assay}(object, cells = NULL, expression,
  invert = FALSE, ...)

\method{WhichCells}{Seurat}(object, cells = NULL, idents = NULL,
  expression, invert = FALSE, downsample = Inf, seed = 1, ...)
}
\arguments{
\item{object}{Seurat object}

\item{cells}{Subset of cell names}

\item{expression}{A predicate expression for feature/variable expression, can
evalue anything that can be pulled by \code{FetchData}}

\item{invert}{Invert the selection of cells}

\item{...}{Arguments passed to other methods}

\item{idents}{A vector of identity classes to keep}

\item{downsample}{Maximum number of cells per identity class, default is \code{Inf}}

\item{seed}{Random seed for downsampling}
}
\value{
A vector of cell names
}
\description{
Returns a list of cells that match a particular set of criteria such as
identity class, high/low values for particular PCs, ect..
}
\section{Methods (by class)}{
\itemize{
\item \code{Assay}: Get cells from an Assay object

\item \code{Seurat}: Get cells from a Seurat object
}}

\examples{
WhichCells(object = pbmc_small, idents = 2)
WhichCells(object = pbmc_small, expression = MS4A1 > 3)
WhichCells(object = pbmc_small, idents = c(1, 3), invert = TRUE)

}
\seealso{
\code{\link{FetchData}}
}
