% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{WorkflowStatus}
\alias{WorkflowStatus}
\title{Output status of each command in the workflow}
\usage{
WorkflowStatus(object, workflow.name, command.name)
}
\arguments{
\item{object}{Seurat object}

\item{workflow.name}{Workflow name, should already be initialized using InitializeWorkflow}

\item{command.name}{Name of the command at the end of the workflow}
}
\value{
Seurat object with updated workflow
}
\description{
For each command in the workflow, indicate whether it is up-to-date.
}
\examples{
WorkflowStatus(object = pbmc_small,workflow.name = "cluster")
#
}
