% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfRenderSeurat.R
\name{zfRenderSeurat}
\alias{zfRenderSeurat}
\alias{situ3d}
\alias{zf.cells.render}
\alias{zf.anchor.render}
\alias{zf.insitu.vec.lateral}
\alias{zf.insitu.lateral}
\alias{zf.insitu.dorsal}
\alias{zf.insitu.ventral}
\alias{zf.insitu.side}
\title{Zebrafish analysis functions}
\usage{
situ3d(data, label = NULL, ...)
}
\arguments{
\item{data}{Predicted expression levels across Zebrafish bins}

\item{label}{Plot label}

\item{...}{Extra parameters}
}
\description{
These functions are for the zebrafish analysis from Satija et al. 2015
}
\references{
Satija R*, Farrell JA*, Gennert D, Schier AF, Regev A.
Spatial reconstruction of single-cell gene expression data.
Nature Biotechnology. 2015

Draw 3D in situ predictions from Zebrafish dataset

From Jeff Farrell
}
