% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{zingeRDETest}
\alias{zingeRDETest}
\title{Differential expression using zingeR-DESeq2 pipeline}
\usage{
zingeRDETest(object, cells.1, cells.2, min.cells = 3, genes.use = NULL,
  latent.vars = NULL, print.bar = TRUE, ...)
}
\arguments{
\item{object}{Seurat object}

\item{cells.1}{Group 1 cells}

\item{cells.2}{Group 2 cells}

\item{min.cells}{Minimum number of cells expressing the gene in at least one 
of the two groups}

\item{genes.use}{Genes to use for test}
}
\value{
Returns a p-value ranked matrix of putative differentially expressed
genes.
}
\description{
Identifies differentially expressed genes between two groups of cells using
zingeR-DESeq2
}
\details{
This test does not support pre-filtering of genes based on average difference (or percent detection rate) between cell groups. 
However, genes may be pre-filtered based on their minimum detection rate (min.pct) across both cell groups. 
To use this method, please install zingeR, using the instructions at https://github.com/statOmics/zingeR
}
\examples{
pbmc_small
DESeq2DETest(pbmc_small, cells.1 = WhichCells(object = pbmc_small, ident = 1),
            cells.2 = WhichCells(object = pbmc_small, ident = 2))

}
\references{
Van Der Verge K, Soneson C, Love MI, Robinson MD,  Clement LS (2017). "zingeR: unlocking RNA-seq tools for zero-inflation and single cell applications" biorXiv.
https://github.com/statOmics/zingeR
}
