% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/utilities.R
\name{AddModuleScore}
\alias{AddModuleScore}
\alias{AddModuleScore.StdAssay}
\alias{AddModuleScore.Assay}
\title{Calculate module scores for feature expression programs in single cells}
\usage{
AddModuleScore(object, ...)

\method{AddModuleScore}{StdAssay}(
  object,
  features,
  kmeans.obj,
  pool = NULL,
  nbin = 24,
  ctrl = 100,
  k = FALSE,
  name = "Cluster",
  seed = 1,
  search = FALSE,
  slot = "data",
  ...
)

\method{AddModuleScore}{Assay}(
  object,
  features,
  kmeans.obj,
  pool = NULL,
  nbin = 24,
  ctrl = 100,
  k = FALSE,
  name = "Cluster",
  seed = 1,
  search = FALSE,
  slot = "data",
  ...
)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns a Seurat object with module scores added to object meta data;
each module is stored as \code{name#} for each module program present in
\code{features}
}
\description{
Calculate the average expression levels of each program (cluster) on single
cell level, subtracted by the aggregated expression of control feature sets.
All analyzed features are binned based on averaged expression, and the
control features are randomly selected from each bin.
}
\references{
Tirosh et al, Science (2016)

Tirosh et al, Science (2016)

Tirosh et al, Science (2016)
}
\concept{utilities}
