% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{AddSmoothedScore}
\alias{AddSmoothedScore}
\title{Calculate smoothed expression values}
\usage{
AddSmoothedScore(object, genes.fit = NULL, dim.1 = 1, dim.2 = 2,
  reduction.use = "tSNE", k = 30, do.log = FALSE, do.print = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{genes.fit}{Genes to calculate smoothed values for}

\item{dim.1}{Dimension for x-axis (default 1)}

\item{dim.2}{Dimension for y-axis (default 2)}

\item{reduction.use}{Which dimensionality reduction to use. Default is
"tsne", can also be "pca", or "ica", assuming these are precomputed.}

\item{k}{k-param for k-nearest neighbor calculation}

\item{do.log}{Whether to perform smoothing in log space. Default is false.}

\item{do.print}{Print progress (output the name of each gene after it has
been imputed).}
}
\description{
Smooths expression values across the k-nearest neighbors based on dimensional reduction
}
