% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc.R
\name{AddTCC}
\alias{AddTCC}
\title{Add in raw TCC data to Seurat object}
\usage{
AddTCC(object, tcc.counts, tx.counts, ec.map, gene.map, min.ec.filter = 0,
  display.progress = TRUE)
}
\arguments{
\item{object}{Seurat object}

\item{tcc.counts}{Path to TCC matrix}

\item{tx.counts}{Path to transcript counts file}

\item{ec.map}{Path to file mapping between equivalence classes and
transcript IDs}

\item{gene.map}{Path to file mapping transcript IDs to gene IDs}

\item{min.ec.filter}{Only keep ECs with at least this many counts across all
cells}

\item{display.progress}{prints output/progress bars}
}
\description{
Add in raw TCC data to Seurat object
}
