% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc.R
\name{AddTCC}
\alias{AddTCC}
\title{Add in raw TCC data to Seurat object}
\usage{
AddTCC(object, raw.counts, ec.map, gene.map, display.progress = TRUE)
}
\arguments{
\item{object}{Seurat object}

\item{raw.counts}{Path to TCC matrix}

\item{ec.map}{Path to file mapping between equivalence classes and
transcript IDs}

\item{gene.map}{Path to file mapping transcript IDs to gene IDs}

\item{display.progress}{prints output/progress bars}
}
\description{
Add in raw TCC data to Seurat object
}
