% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/integration.R
\name{AnnotateAnchors}
\alias{AnnotateAnchors}
\alias{AnnotateAnchors.default}
\alias{AnnotateAnchors.IntegrationAnchorSet}
\alias{AnnotateAnchors.TransferAnchorSet}
\title{Add info to anchor matrix}
\usage{
AnnotateAnchors(anchors, vars, slot, ...)

\method{AnnotateAnchors}{default}(anchors, vars = NULL, slot = NULL, object.list, ...)

\method{AnnotateAnchors}{IntegrationAnchorSet}(anchors, vars = NULL, slot = NULL, ...)

\method{AnnotateAnchors}{TransferAnchorSet}(anchors, vars = NULL, slot = NULL, ...)
}
\arguments{
\item{anchors}{An \code{\link{AnchorSet}} object}

\item{vars}{Variables to pull for each object via FetchData}

\item{slot}{Slot to pull feature data for}

\item{...}{Arguments passed to other methods}

\item{object.list}{List of Seurat objects}
}
\value{
Returns the anchor dataframe with additional columns for annotation
metadata
}
\description{
Add info to anchor matrix
}
