% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{AnnotateAnchors}
\alias{AnnotateAnchors}
\title{Add info to anchor matrix}
\usage{
AnnotateAnchors(object, vars = NULL, slot = NULL)
}
\arguments{
\item{object}{An \code{\link{AnchorSet}} object}

\item{vars}{Variables to pull for each object via FetchData}

\item{slot}{Slot to pull feature data for}
}
\value{
Returns the anchor dataframe with additional columns for annotation
metadata
}
\description{
Add info to anchor matrix
}
