% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delayedarray.R
\name{as.DelayedMatrix}
\alias{as.DelayedMatrix}
\title{Coerce to a DelayedMatrix}
\usage{
as.DelayedMatrix(x, ...)
}
\arguments{
\item{x}{A matrix-like object}

\item{...}{Ignored}
}
\value{
A \code{\link[DelayedArray]{DelayedMatrix}}
}
\description{
Wrap a matrix-like object in a \code{\link[DelayedArray]{DelayedMatrix}} so it
can be used as an out-of-memory Seurat layer. Existing \code{DelayedMatrix}
objects are returned unchanged; other inputs (\code{matrix},
\code{\link[Matrix]{dgCMatrix}}, \code{IterableMatrix}) are wrapped, preserving
dimnames.
}
\examples{
\dontrun{
mat <- Matrix::rsparsematrix(100, 20, 0.1)
rownames(mat) <- paste0("g", 1:100)
colnames(mat) <- paste0("c", 1:20)
d <- as.DelayedMatrix(mat)
}

}
\concept{utilities}
