% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.Graph}
\alias{as.Graph}
\title{Convert a matrix (or Matrix) to the Graph class.}
\usage{
as.Graph(from, ...)
}
\arguments{
\item{from}{The matrix to convert}

\item{...}{Arguments passed to other methods (ignored for now)}
}
\description{
Convert a matrix (or Matrix) to the Graph class.
}
