% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R, R/utilities.R
\name{as.sparse.H5Group}
\alias{as.sparse.H5Group}
\alias{as.data.frame.Matrix}
\title{Cast to Sparse}
\usage{
\method{as.sparse}{H5Group}(x, ...)

\method{as.data.frame}{Matrix}(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  stringsAsFactors = getOption(x = "stringsAsFactors", default = FALSE)
)
}
\arguments{
\item{x}{An object}

\item{...}{Arguments passed to other methods}

\item{row.names}{\code{NULL} or a character vector giving the row names for
the data; missing values are not allowed}

\item{optional}{logical. If \code{TRUE}, setting row names and
    converting column names (to syntactic names: see
    \code{\link[base]{make.names}}) is optional.  Note that all of \R's
    \pkg{base} package \code{as.data.frame()} methods use
    \code{optional} only for column names treatment, basically with the
    meaning of \code{\link[base]{data.frame}(*, check.names = !optional)}.
    See also the \code{make.names} argument of the \code{matrix} method.}

\item{stringsAsFactors}{logical: should the character vector be converted
    to a factor?}
}
\value{
\code{as.data.frame.Matrix}: A data frame representation of the S4 Matrix
}
\description{
Cast to Sparse
}
\seealso{
\code{\link[SeuratObject:as.sparse]{SeuratObject::as.sparse}}
}
\concept{objects}
\concept{utilities}
