% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{class}
\name{Assay-class}
\alias{Assay-class}
\title{The Assay Class}
\description{
The \code{Assay} object is the basic unit of Seurat; for more details, please
see the documentation in \code{\link[SeuratObject:Assay]{SeuratObject}}
}
\seealso{
\code{\link[SeuratObject:Assay]{SeuratObject::Assay-class}}
}
