% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Assays}
\alias{Assays}
\title{List Assay objects}
\usage{
Assays(object)
}
\arguments{
\item{object}{A Seurat object}
}
\value{
Returns a character vector
}
\description{
Lists the names of \code{\link{Assay}} objects present in 
a Seurat object.
}
\examples{
Assays(object = pbmc_small)
}
