% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Assays}
\alias{Assays}
\title{Pull Assays or assay names}
\usage{
Assays(object, slot = NULL)
}
\arguments{
\item{object}{A Seurat object}

\item{slot}{Name of Assay to return}
}
\value{
If \code{slot} is \code{NULL}, the names of all \code{Assay} objects
in this Seurat object. Otherwise, the \code{Assay} object specified
}
\description{
Lists the names of \code{\link{Assay}} objects present in
a Seurat object. If slot is provided, pulls specified Assay object.
}
\examples{
Assays(object = pbmc_small)

}
