% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{AverageExpression}
\alias{AverageExpression}
\title{Averaged gene expression by identity class}
\usage{
AverageExpression(object, genes.use = NULL, return.seurat = F,
  add.ident = NULL, ...)
}
\arguments{
\item{object}{Seurat object}

\item{genes.use}{Genes to analyze. Default is all genes.}
}
\value{
Returns a matrix with genes as rows, identity classes as columns.
}
\description{
Returns gene expression for an 'average' single cell in each identity class
}
\details{
Output is in log-space, but averaging is done in non-log space.
}

