% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{AverageExpression}
\alias{AverageExpression}
\title{Averaged gene expression by identity class}
\usage{
AverageExpression(object, genes.use = NULL, return.seurat = F,
  add.ident = NULL, use.scale = F, use.raw = F, show.progress = T, ...)
}
\arguments{
\item{object}{Seurat object}

\item{genes.use}{Genes to analyze. Default is all genes.}

\item{return.seurat}{Whether to return the data as a Seurat object. Default is false.}

\item{add.ident}{Place an additional label on each cell prior to averaging (very useful if you want to observe cluster averages, separated by replicate, for example)}

\item{use.raw}{For gene expression, use raw values. Default is FALSE}

\item{show.progress}{Show progress bar (default is T)}

\item{...}{Arguments to be passed to methods such as \code{\link{Setup}}}
}
\value{
Returns a matrix with genes as rows, identity classes as columns.
}
\description{
Returns gene expression for an 'average' single cell in each identity class
}
\details{
Output is in log-space, but averaging is done in non-log space.
}

