% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{AverageExpression}
\alias{AverageExpression}
\title{Averaged feature expression by identity class}
\usage{
AverageExpression(object, assay = NULL, features = NULL,
  return.seurat = FALSE, add.ident = NULL, use.scale = FALSE,
  use.counts = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Which assay to use. Default is all assays.}

\item{features}{Features to analyze. Default is all features in the assay.}

\item{return.seurat}{Whether to return the data as a Seurat object. Default is FALSE.}

\item{add.ident}{Place an additional label on each cell prior to averaging (very useful if you want to observe cluster averages, separated by replicate, for example).}

\item{use.scale}{Use scaled values for gene expression}

\item{use.counts}{Use count values for gene expression}

\item{verbose}{Print messages and show progress bar}

\item{...}{Arguments to be passed to methods such as \code{\link{CreateSeuratObject}}}
}
\value{
Returns a matrix with genes as rows, identity classes as columns.
If return.seurat is TRUE, returns an object of class \code{\link{Seurat}}.
}
\description{
Returns expression for an 'average' single cell in each identity class
}
\details{
Output is in log-space when \code{return.seurat = TRUE}, otherwise it's in non-log space.
Averaging is done in non-log space.
}
\examples{
head(AverageExpression(object = pbmc_small))

}
