% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R, R/utilities_generics.R
\name{BlockCov.loom}
\alias{BlockCov.loom}
\alias{BlockCov}
\title{Calculate covariance matrix}
\usage{
\method{BlockCov}{loom}(object, mat, chunk.size = 100,
  display.progress = TRUE, rows.use = NULL,
  row.names = "row_attrs/gene_names")

BlockCov(object, ...)
}
\arguments{
\item{object}{Object to calculate covariate matrix for}

\item{mat}{Path to matrix in loom file}

\item{chunk.size}{Size of chunk, defaults to 100}

\item{display.progress}{Display progress bar}

\item{rows.use}{Rows to include in covariance calculation,
should be the path to a one-dimensional boolean (logical) dataset
(eg. 'row_attrs/var_genes') OR a character vector of rownames to use.}

\item{row.names}{Location of associated rownames for labeling of final matrix}
}
\value{
Returns the covariance matrix
}
\description{
Calculate the covariance matrix for an object in blocks so that entire
matrix doesn't need to ever be loaded into memory.
}
\section{Methods (by class)}{
\itemize{
\item \code{loom}: Calculate blocked covariates in loom files
}}

