% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_generics.R, R/utilities.R
\docType{methods}
\name{BlockCov}
\alias{BlockCov}
\alias{BlockCov,loom-method}
\title{Calculate covariance matrix}
\usage{
BlockCov(object, ...)

\S4method{BlockCov}{loom}(object, mat = NULL, chunk.size = 100,
  display.progress = TRUE, rows.use = NULL,
  row.names = "row_attrs/gene_names", do.cpp = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{mat}{path to matrix in loom file}

\item{chunk.size}{size of chunk}

\item{display.progress}{display progress bar}

\item{rows.use}{rows to include in covariance calculation}

\item{row.names}{location of associated rownames for labeling of final matrix}
}
\value{
Returns the covariance matrix
}
\description{
Calculate the covariance matrix for a loom object in blocks so that entire
matrix doesn't need to ever be loaded into memory.
}
