% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{BridgeCellsRepresentation}
\alias{BridgeCellsRepresentation}
\title{Use bridge cells to represent single-modality object}
\usage{
BridgeCellsRepresentation(
  object.list,
  bridge.object,
  object.reduction.list,
  bridge.reduction.list,
  laplacian.reduction = NULL,
  laplacian.dims = NULL,
  bridge.assay.name = "Bridge",
  return.all.assays = FALSE,
  l2.norm = TRUE,
  do.center = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object.list}{A list of Seurat objects between which to
find anchors for downstream integration.}

\item{bridge.object}{A multimodal bridge seurat which connects two
single-modality objects}

\item{object.reduction.list}{A list of dimensional reductions from object.list used
to be reconstructed by bridge.obejct}

\item{bridge.reduction.list}{A list of dimensional reductions from bridge.object used
to reconstruct object.reduction.list}

\item{laplacian.reduction}{Name of bridge graph laplacian dimensional reduction}

\item{laplacian.dims}{Dimensions used for bridge graph laplacian dimensional reduction}

\item{bridge.assay.name}{Assay name used for bridge object reconstruction value}

\item{return.all.assays}{if return all assays in the object.list. 
Only bridge assay is returned by default.}

\item{verbose}{Print messages and progress}
}
\value{
Returns a object list in which each object has a bridge cell derived assay
}
\description{
Use bridge cells to represent single-modality object
}
