% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing5.R
\name{CalcDispersion}
\alias{CalcDispersion}
\title{Calculate dispersion of features}
\usage{
CalcDispersion(
  object,
  mean.function = FastExpMean,
  dispersion.function = FastLogVMR,
  num.bin = 20,
  binning.method = "equal_width",
  verbose = TRUE,
  ...
)
}
\description{
Calculate dispersion of features
}
