% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc.genes}
\alias{cc.genes}
\title{Cell cycle genes}
\format{A list of two vectors
\describe{
  \item{s.genes}{Genes associated with S-phase}
  \item{g2m.genes}{Genes associated with G2M-phase}
}}
\source{
\url{http://science.sciencemag.org/content/352/6282/189}
}
\usage{
cc.genes
}
\description{
A list of genes used in cell-cycle regression
}
\keyword{datasets}
