% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Cells.SCTModel}
\alias{Cells.SCTModel}
\alias{Cells.SlideSeq}
\alias{Cells.STARmap}
\alias{Cells.VisiumV1}
\title{Get Cell Names}
\usage{
\method{Cells}{SCTModel}(x, ...)

\method{Cells}{SlideSeq}(x, ...)

\method{Cells}{STARmap}(x, ...)

\method{Cells}{VisiumV1}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Arguments passed to other methods}
}
\description{
Get Cell Names
}
\seealso{
\code{\link[SeuratObject:Cells]{SeuratObject::Cells}}
}
\concept{objects}
\concept{spatial}
