% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{ClassifyCells}
\alias{ClassifyCells}
\title{Classify New Data}
\usage{
ClassifyCells(object, new.data = NULL, training.genes = NULL,
  training.classes = NULL, ...)
}
\arguments{
\item{object}{Seurat object on which to train the classifier}

\item{new.data}{New data to classify}

\item{training.genes}{Vector of genes to build the classifier on}

\item{training.classes}{Vector of classes to build the classifier on}

\item{...}{additional parameters passed to ranger}
}
\value{
Vector of cluster ids
}
\description{
Classify new data based on the cluster information of the provided object.
Random Forests are used as the basis of the classification.
}

