% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Command}
\alias{Command}
\alias{Command.Seurat}
\title{Get SeuratCommands}
\usage{
Command(object, ...)

\method{Command}{Seurat}(object, command = NULL, value = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{command}{Name of the command to pull, pass \code{NULL} to get the names of all commands run}

\item{value}{Name of the parameter to pull the value for}
}
\value{
Either a SeuratCommand object or the requested paramter value
}
\description{
Pull information on previously run commands in the Seurat object.
}
