% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Convert}
\alias{Convert}
\alias{Convert.anndata.base.AnnData}
\alias{Convert.seurat}
\title{Convert Seurat objects to other classes and vice versa}
\usage{
Convert(from, ...)

\method{Convert}{anndata.base.AnnData}(...)

\method{Convert}{seurat}(from, to, filename, chunk.dims = "auto",
  chunk.size = 1000, overwrite = FALSE, display.progress = TRUE,
  anndata.raw = "raw.data", anndata.X = "data", ...)
}
\arguments{
\item{from}{Object to convert from}

\item{...}{Arguments passed to other methods}

\item{to}{Class of object to convert to}

\item{filename}{Filename for writing files}

\item{chunk.dims}{Internal HDF5 chunk size}

\item{chunk.size}{Number of cells to stream to loom file at a time}

\item{overwrite}{Overwrite existing file at \code{filename}?}

\item{display.progress}{Display a progress bar}

\item{anndata.raw}{Name of matrix (raw.data, data) to put in the anndata raw slot}

\item{anndata.X}{Name of matrix (data, scale.data) to put in the anndata X slot}
}
\value{
An object of class \code{to}
}
\description{
Defunct, please use other converters to convert to and from \code{Seurat} objects
}
\seealso{
\code{\link{WriteH5AD}} \code{\link{as.Seurat}} \code{\link{as.SingleCellExperiment}}
}
