% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\docType{methods}
\name{Convert}
\alias{Convert}
\alias{Convert,seurat,character-method}
\alias{Convert,loom,character-method}
\title{Convert Seurat objects to loom and vice versa}
\usage{
Convert(from, to, ...)

\S4method{Convert}{seurat,character}(from, to, filename, chunk.dims = "auto")

\S4method{Convert}{loom,character}(from, to, raw.data = "matrix",
  gene.names = "row_attrs/gene_names", cell.names = "col_attrs/cell_names",
  norm.data = "layers/norm_data", scale.data = "layers/scale_data",
  gene.means = "row_attrs/gene.mean",
  gene.dispersion = "row_attrs/gene.dispersion",
  gene.scaled = "row_attrs/gene.dispersion.scaled",
  var.genes = "row_attrs/var_genes")
}
\arguments{
\item{from}{...}

\item{to}{...}

\item{filename}{Filename for writing loom files}

\item{raw.data}{Path to raw data matrix (required)}

\item{gene.names}{Path to gene names array (required)}

\item{cell.names}{Path to cell names array (required)}

\item{norm.data}{Path to normalized data matrix (Optional, pass NULL to skip)}

\item{scale.data}{Path to scaled data matrix (Optional, pass NULL to skip)}

\item{gene.means}{Path to gene means array (Optional, pass NULL to skip)}

\item{gene.dispersion}{Path to gene disperion array (Optional, pass NULL to skip)}

\item{gene.scaled}{Path to scaled gene dispersion array (Optional, pass NULL to skip)}

\item{var.gens}{Path to variable genes array (Optional, pass NULL to skip)}
}
\value{
An object of class \code{to}
}
\description{
Convert Seurat objects to loom and vice versa
}
