% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{as.SingleCellExperiment}
\alias{as.SingleCellExperiment}
\alias{as.seurat}
\alias{Convert}
\alias{as.SingleCellExperiment.seurat}
\alias{as.seurat.SingleCellExperiment}
\alias{Convert.anndata.base.AnnData}
\alias{Convert.SingleCellExperiment}
\alias{Convert.seurat}
\title{Convert Seurat objects to other classes and vice versa}
\usage{
as.SingleCellExperiment(from)

as.seurat(from)

Convert(from, ...)

\method{as.SingleCellExperiment}{seurat}(from, ...)

\method{as.seurat}{SingleCellExperiment}(from, ...)

\method{Convert}{anndata.base.AnnData}(from, to, X.slot = "scale.data",
  raw.slot = "data", ...)

\method{Convert}{SingleCellExperiment}(from, to,
  raw.data.slot = "counts", data.slot = "logcounts", ...)

\method{Convert}{seurat}(from, to, filename, chunk.dims = "auto",
  chunk.size = 1000, overwrite = FALSE, display.progress = TRUE,
  anndata.raw = "raw.data", anndata.X = "data", ...)
}
\arguments{
\item{from}{Object to convert from}

\item{...}{Arguments passed to other methods}

\item{to}{Class of object to convert to}

\item{X.slot}{Seurat slot to transfer anndata X into. Default is scale.data}

\item{raw.slot}{Seurat slot to transfer anndata raw into. Default is data}

\item{raw.data.slot}{name of the SingleCellExperiment assay to slot into @raw.data}

\item{data.slot}{name of the SingleCellExperiment assay to slot into @data}

\item{filename}{Filename for writing files}

\item{chunk.dims}{Internal HDF5 chunk size}

\item{chunk.size}{Number of cells to stream to loom file at a time}

\item{overwrite}{Overwrite existing file at \code{filename}?}

\item{display.progress}{Display a progress bar}

\item{anndata.raw}{Name of matrix (raw.data, data) to put in the anndata raw slot}

\item{anndata.X}{Name of matrix (data, scale.data) to put in the anndata X slot}
}
\value{
An object of class \code{to}
}
\description{
Currently, we support direct conversion to/from loom (\url{http://loompy.org/}),
SingleCellExperiment (\url{https://bioconductor.org/packages/release/bioc/html/SingleCellExperiment.html}),
and Anndata(\url{https://anndata.readthedocs.io/en/latest/}) objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{anndata.base.AnnData}: from Anndata file to a Seurat object

\item \code{SingleCellExperiment}: Convert from SingleCellExperiment to a Seurat object

\item \code{seurat}: Convert a Seurat object
}}

