% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{Convert}
\alias{Convert}
\alias{Convert.seurat}
\alias{Convert.loom}
\title{Convert Seurat objects to loom and vice versa}
\usage{
Convert(from, ...)

\method{Convert}{seurat}(from, to, filename, chunk.dims = "auto",
  chunk.size = 1000, overwrite = FALSE)

\method{Convert}{loom}(from, to, raw.data = "matrix",
  gene.names = "row_attrs/gene_names", cell.names = "col_attrs/cell_names",
  norm.data = "layers/norm_data", scale.data = "layers/scale_data",
  gene.means = "row_attrs/gene.mean",
  gene.dispersion = "row_attrs/gene.dispersion",
  gene.scaled = "row_attrs/gene.dispersion.scaled",
  var.genes = "row_attrs/var_genes")
}
\arguments{
\item{from}{Object to convert from}

\item{to}{Class of object to convert to}

\item{filename}{Filename for writing loom files}

\item{chunk.dims}{Internal HDF5 chunk size}

\item{chunk.size}{Number of cells to stream to loom file at a time}

\item{overwrite}{Overwrite existing file at \code{filename}?}

\item{raw.data}{Path to raw data matrix (required)}

\item{gene.names}{Path to gene names array (required)}

\item{cell.names}{Path to cell names array (required)}

\item{norm.data}{Path to normalized data matrix (Optional, pass NULL to skip)}

\item{scale.data}{Path to scaled data matrix (Optional, pass NULL to skip)}

\item{gene.means}{Path to gene means array (Optional, pass NULL to skip)}

\item{gene.dispersion}{Path to gene disperion array (Optional, pass NULL to skip)}

\item{gene.scaled}{Path to scaled gene dispersion array (Optional, pass NULL to skip)}

\item{var.gens}{Path to variable genes array (Optional, pass NULL to skip)}
}
\value{
An object of class \code{to}
}
\description{
Convert Seurat objects to loom and vice versa
}
\section{Methods (by class)}{
\itemize{
\item \code{seurat}: Convert a Seurat object

\item \code{loom}: Convert a loom object
}}

