% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{CreateAssayObject}
\alias{CreateAssayObject}
\title{Make an Assay object}
\usage{
CreateAssayObject(raw.data, min.cells = 0, min.genes = 0, is.expr = 0,
  names.field = 1, names.delim = "_", ...)
}
\arguments{
\item{raw.data}{Raw input data}

\item{min.cells}{Include genes with detected expression in at least this
many cells. Will subset the raw.data matrix as well. To reintroduce excluded
genes, create a new object with a lower cutoff.}

\item{min.genes}{Include cells where at least this many genes are detected.}

\item{is.expr}{Expression threshold for 'detected' gene. For most datasets, particularly UMI
datasets, will be set to 0 (default). If not, when initializing, this should be set to a level
based on pre-normalized counts (i.e. require at least 5 counts to be treated as expresesd) All
values less than this will be set to 0 (though maintained in object@raw.data).}

\item{...}{Ignored for now}
}
\description{
Make an Assay object
}
