% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CreateCategoryMatrix}
\alias{CreateCategoryMatrix}
\title{Create one hot matrix for a given label}
\usage{
CreateCategoryMatrix(
  labels,
  method = c("aggregate", "average"),
  cells.name = NULL
)
}
\description{
Create one hot matrix for a given label
}
