% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utilities.R
\name{DarkTheme}
\alias{DarkTheme}
\title{Dark Theme}
\usage{
DarkTheme(...)
}
\arguments{
\item{...}{Extra parameters to be passed to theme()}
}
\value{
A ggplot2 theme object
}
\description{
Add a dark theme to ggplot objects
}
\seealso{
\code{theme}
}
