% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{DefaultAssay}
\alias{DefaultAssay}
\alias{DefaultAssay<-}
\alias{DefaultAssay.DimReduc}
\alias{DefaultAssay.Seurat}
\alias{DefaultAssay<-.Seurat}
\title{Get the default assay}
\usage{
DefaultAssay(object, ...)

DefaultAssay(object, ...) <- value

\method{DefaultAssay}{DimReduc}(object, ...)

\method{DefaultAssay}{Seurat}(object, ...)

\method{DefaultAssay}{Seurat}(object, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{Name of assay to set as default}
}
\value{
The name of the default assay

An object with the new default assay
}
\description{
Get the default assay
}
