% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{DimHeatmap}
\alias{DimHeatmap}
\title{Dimensional reduction heatmap}
\usage{
DimHeatmap(object, assay.use = NULL, reduction.use = "pca", dims.use = 1,
  cells.use = NULL, num.features = 30, use.full = FALSE,
  disp.min = -2.5, disp.max = 2.5, slot.use = "scale.data",
  do.balanced = FALSE, check.plot = TRUE, num.col = NULL, ...)
}
\arguments{
\item{object}{Seurat object}

\item{assay.use}{A vector of assays to pull data from}

\item{reduction.use}{Which dimmensional reduction to use}

\item{dims.use}{Dimensions to plot}

\item{cells.use}{A list of cells to plot. If numeric, just plots the top cells.}

\item{disp.min}{Minimum display value (all values below are clipped)}

\item{disp.max}{Maximum display value (all values above are clipped)}

\item{slot.use}{Data slot to use, choose from 'raw.data', 'data', or 'scale.data'}

\item{do.balanced}{Plot an equal number of genes with both + and - scores.}

\item{check.plot}{Check that plotting will finish in a reasonable amount of time}

\item{num.col}{Number of columns.}

\item{...}{Extra parameters passed to superheat}

\item{num.genes}{NUmber of genes to plot}
}
\value{
Invisbly returns the final grob
}
\description{
Draws a heatmap focusing on a principal component. Both cells and genes are sorted by their
principal component scores. Allows for nice visualization of sources of heterogeneity in the dataset.
}
\examples{
DimHeatmap(object = pbmc_small)

}
\seealso{
\code{\link{superheat::superheat}}
}
