% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{DimHeatmap}
\alias{DimHeatmap}
\title{Dimensional reduction heatmap}
\usage{
DimHeatmap(object, reduction.type = "pca", dim.use = 1, cells.use = NULL,
  num.genes = 30, use.full = FALSE, disp.min = -2.5, disp.max = 2.5,
  do.return = FALSE, col.use = PurpleAndYellow(), use.scale = TRUE,
  do.balanced = FALSE, remove.key = FALSE, label.columns = NULL,
  check.plot = TRUE, ...)
}
\arguments{
\item{object}{Seurat object.}

\item{reduction.type}{Which dimmensional reduction t use}

\item{dim.use}{Dimensions to plot}

\item{cells.use}{A list of cells to plot. If numeric, just plots the top cells.}

\item{num.genes}{NUmber of genes to plot}

\item{use.full}{Use the full PCA (projected PCA). Default is FALSE}

\item{disp.min}{Minimum display value (all values below are clipped)}

\item{disp.max}{Maximum display value (all values above are clipped)}

\item{do.return}{If TRUE, returns plot object, otherwise plots plot object}

\item{col.use}{Color to plot.}

\item{use.scale}{Default is TRUE: plot scaled data. If FALSE, plot raw data on the heatmap.}

\item{do.balanced}{Plot an equal number of genes with both + and - scores.}

\item{remove.key}{Removes the color key from the plot.}

\item{label.columns}{Labels for columns}

\item{check.plot}{Check that plotting will finish in a reasonable amount of time}

\item{...}{Extra parameters for heatmap plotting.}
}
\value{
If do.return==TRUE, a matrix of scaled values which would be passed
to heatmap.2. Otherwise, no return value, only a graphical output
}
\description{
Draws a heatmap focusing on a principal component. Both cells and genes are sorted by their
principal component scores. Allows for nice visualization of sources of heterogeneity in the dataset.
}
\examples{
DimHeatmap(object = pbmc_small)

}
