% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{DimPlot}
\alias{DimPlot}
\title{Dimensional reduction plot}
\usage{
DimPlot(object, reduction.use = "pca", dims.use = c(1, 2),
  group.by = NULL, cols.use = NULL, cells.use = NULL, pt.size = 1,
  shape.by = NULL, plot.order = NULL, do.label = FALSE,
  label.size = 4, cells.highlight = NULL, cols.highlight = "red",
  sizes.highlight = 1, na.value = "grey50", do.hover = FALSE,
  data.hover = "ident", do.identify = FALSE, no.legend = FALSE,
  vector.friendly = FALSE, png.file = NULL, png.arguments = c(10, 10,
  100), ...)
}
\arguments{
\item{object}{Seurat object}

\item{reduction.use}{Which dimensionality reduction to use. Default is
"pca", can also be "tsne", or "ica", assuming these are precomputed.}

\item{dims.use}{Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions}

\item{group.by}{Group (color) cells in different ways (for example, orig.ident)}

\item{cols.use}{Vector of colors, each color corresponds to an identity class. By default, ggplot2 assigns colors.}

\item{cells.use}{Vector of cells to plot (default is all cells)}

\item{pt.size}{Adjust point size for plotting}

\item{shape.by}{If NULL, all points are circles (default). You can specify any
cell attribute (that can be pulled with FetchData) allowing for both
different colors and different shapes on cells.}

\item{plot.order}{Specify the order of plotting for the idents. This can be
useful for crowded plots if points of interest are being buried. Provide
either a full list of valid idents or a subset to be plotted last (on top).}

\item{do.label}{Whether to label the clusters}

\item{label.size}{Sets size of labels}

\item{cells.highlight}{A list of character or numeric vectors of cells to
highlight. If only one group of cells desired, can simply
pass a vector instead of a list. If set, colors selected cells to the color(s)
in \code{cols.highlight} and other cells black (white if dark.theme = TRUE);
 will also resize to the size(s) passed to \code{sizes.highlight}}

\item{cols.highlight}{A vector of colors to highlight the cells as; will
repeat to the length groups in cells.highlight}

\item{sizes.highlight}{Size of highlighted cells; will repeat to the length
groups in cells.highlight}

\item{na.value}{Color value for NA points when using custom scale.}
}
\value{
A ggplot object
}
\description{
Graphs the output of a dimensional reduction technique (PCA by default).
Cells are colored by their identity class.
}
\examples{
DimPlot(object = pbmc_small)

}
\seealso{
\code{\link{FeaturePlot}} \code{\link{FeatureMap}}
}
