% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{class}
\name{DimReduc-class}
\alias{DimReduc-class}
\title{The DimReduc Class}
\description{
The \code{DimReduc} object stores a dimensionality reduction taken out in
Seurat; for more details, please see the documentation in
\code{\link[SeuratObject:DimReduc]{SeuratObject}}
}
\seealso{
\code{\link[SeuratObject:DimReduc]{SeuratObject::DimReduc-class}}
}
