% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{DoHeatmap}
\alias{DoHeatmap}
\title{Feature expression heatmap}
\usage{
DoHeatmap(object, features = NULL, cells = NULL, group.by = "ident",
  group.bar = TRUE, disp.min = -2.5, disp.max = 2.5,
  slot = "scale.data", assay = NULL, ...)
}
\arguments{
\item{object}{Seurat object}

\item{features}{A vector of features to plot, defaults to \code{VariableFeatures(object = object)}}

\item{cells}{A vector of cells to plot}

\item{group.by}{Name of variable to group cells by}

\item{group.bar}{Add a color bar showing group status for cells}

\item{disp.min}{Minimum display value (all values below are clipped)}

\item{disp.max}{Maximum display value (all values above are clipped)}

\item{slot}{Data slot to use, choose from 'raw.data', 'data', or 'scale.data'}

\item{assay}{Assay to pull from}

\item{...}{Ignored for now}
}
\value{
A ggplot object
}
\description{
Draws a heatmap of single cell feature expression.
}
\examples{
DoHeatmap(object = pbmc_small)

}
