% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_generics.R, R/interaction.R
\docType{methods}
\name{DownsampleSeurat}
\alias{DownsampleSeurat}
\alias{DownsampleSeurat,seurat-method}
\alias{DownsampleSeurat,loom-method}
\title{Density Dependent Downsampling}
\usage{
DownsampleSeurat(object, ...)

\S4method{DownsampleSeurat}{seurat}(object, size, dims.use = 1:10,
  return.type = "seurat", filename, overwrite = FALSE, ...)

\S4method{DownsampleSeurat}{loom}(object, size, dims.use = 1:10,
  cell.names = "col_attrs/cell_names", ...)
}
\arguments{
\item{object}{An object to downsample}

\item{size}{A non-negative integer giving the number of cells to choose}

\item{dims.use}{A vector giving the PCs that are used}

\item{return.type}{Return as either a "seurat" or "loom" object}

\item{filename}{file name/path for new loom object if returning loom}

\item{overwrite}{overwrite loom file if it already exists}

\item{cell.names}{location in loom object of cell names}

\item{method}{Sampling strategy, one of 'random', 'max', 'prod', 'mean'}

\item{eps}{A small value added to the density estimate}

\item{bw.adjust}{Factor applied to default bandwidth}

\item{display.progress}{display progress of the process}
}
\description{
Density Dependent Downsampling
}
