% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Embeddings}
\alias{Embeddings}
\alias{Embeddings.DimReduc}
\title{Get cell embeddings}
\usage{
Embeddings(object, ...)

\method{Embeddings}{DimReduc}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\description{
Get cell embeddings
}
\section{Methods (by class)}{
\itemize{
\item \code{DimReduc}: Get the cell embeddings from a DimReduc object
}}

