% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ExpMean}
\alias{ExpMean}
\title{Calculate the mean of logged values}
\usage{
ExpMean(x, ...)
}
\arguments{
\item{x}{A vector of values}

\item{...}{Other arguments (not used)}
}
\value{
Returns the mean in log-space
}
\description{
Calculate mean of logged values in non-log space (return answer in log-space)
}
\examples{
ExpMean(x = c(1, 2, 3))

}
\concept{utilities}
