% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ExportToCellbrowser}
\alias{ExportToCellbrowser}
\title{Export Seurat object for UCSC cell browser}
\usage{
ExportToCellbrowser(object, dir, dataset.name, vars = NULL,
  vars.names = NULL, reductions = "tsne", markers.file = NULL,
  cluster.field = NULL, port = NULL, cb.dir = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{dir}{output directory path}

\item{dataset.name}{name of the dataset}

\item{vars}{vector of metadata fields to export}

\item{vars.names}{list that defines metadata field names after the export.
Should map metadata column name to export name}

\item{reductions}{vector of reduction names to export}

\item{markers.file}{path to file with marker genes}

\item{cluster.field}{name of the metadata field containing cell cluster}

\item{port}{\emph{experimental} on which port to run UCSC cell browser after export}

\item{cb.dir}{in which dir to create UCSC cell browser content root}
}
\value{
...
}
\description{
Export Seurat object for UCSC cell browser
}
\examples{
\dontrun{
ExportToCellbrowser(object = pbmc_small, dataset.name = "PBMC", dir = "out")
}

}
\author{
Maximilian Haeussler, Nikolay Markov
}
