% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{FeatureHeatmap}
\alias{FeatureHeatmap}
\title{Vizualization of multiple features}
\usage{
FeatureHeatmap(object, features.plot, dim.1 = 1, dim.2 = 2,
  idents.use = NULL, pt.size = 2, cols.use = rev(heat.colors(10)),
  pch.use = 16, reduction.use = "tsne")
}
\arguments{
\item{object}{Seurat object}

\item{features.plot}{Vector of features to plot}

\item{dim.1}{Dimension for x-axis (default 1)}

\item{dim.2}{Dimension for y-axis (default 2)}

\item{idents.use}{Which identity classes to display (default is all identity
classes)}

\item{pt.size}{Adjust point size for plotting}

\item{cols.use}{Ordered vector of colors to use for plotting. Default is
heat.colors(10).}

\item{pch.use}{Pch for plotting}

\item{reduction.use}{Which dimensionality reduction to use. Default is
"tsne", can also be "pca", or "ica", assuming these are precomputed.}
}
\value{
No return value, only a graphical output
}
\description{
Similar to FeaturePlot, however, also splits the plot by visualizing each
identity class separately.
}
\details{
Particularly useful for seeing if the same groups of cells co-exhibit a
common feature (i.e. co-express a gene), even within an identity class. Best
understood by example.
}

