% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{FeatureLocator}
\alias{FeatureLocator}
\title{Feature Locator}
\usage{
FeatureLocator(plot, ...)
}
\arguments{
\item{plot}{A ggplot2 plot}

\item{...}{Extra parameters, such as dark.theme, recolor, or smooth for using a dark theme,
recoloring based on selected cells, or using a smooth scatterplot, respectively}
}
\value{
The names of the points selected
}
\description{
Select points on a scatterplot and get information about them
}
\examples{
\dontrun{
plot <- DimPlot(object = pbmc_small)
# Follow instructions in the terminal to select points
cells.located <- FeatureLocator(plot = plot)
cells.located
}

}
\seealso{
\code{\link[graphics]{locator}} \code{\link[ggplot2]{ggplot_build}}
\code{\link[SDMTools]{pnt.in.poly}} \code{\link{DimPlot}} \code{\link{FeaturePlot}}
}
