% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{FeaturePlot}
\alias{FeaturePlot}
\alias{FeatureHeatmap}
\title{Visualize 'features' on a dimensional reduction plot}
\usage{
FeaturePlot(object, features, dims = c(1, 2), cells = NULL,
  cols = c("lightgrey", "blue"), pt.size = NULL, min.cutoff = NA,
  max.cutoff = NA, reduction = "tsne", group.by = NULL,
  split.by = NULL, shape.by = NULL, order = NULL, label = FALSE,
  label.size = 4, ncol = NULL, combine = TRUE, coord.fixed = FALSE,
  ...)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Vector of features to plot}

\item{dims}{Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions}

\item{cells}{Vector of cells to plot (default is all cells)}

\item{cols}{The two colors to form the gradient over. Provide as string vector with
the first color corresponding to low values, the second to high. Also accepts a Brewer
color scale or vector of colors. Note: this will bin the data into number of colors provided.}

\item{pt.size}{Adjust point size for plotting}

\item{min.cutoff}{Vector of minimum cutoff values for each feature, may specify quantile in the form of 'q##' where '##' is the quantile (eg, 1, 10)}

\item{max.cutoff}{Vector of maximum cutoff values for each feature, may specify quantile in the form of 'q##' where '##' is the quantile (eg, 1, 10)}

\item{reduction}{Which dimensionality reduction to use}

\item{group.by}{Group (color) cells in different ways (for example, orig.ident)}

\item{split.by}{A factor in object metadata to split the feature plot by, pass 'ident' to split by cell identity'; similar to the old \code{FeatureHeatmap}}

\item{shape.by}{If NULL, all points are circles (default). You can specify any
cell attribute (that can be pulled with FetchData) allowing for both
different colors and different shapes on cells}

\item{order}{Specify the order of plotting for the idents. This can be
useful for crowded plots if points of interest are being buried. Provide
either a full list of valid idents or a subset to be plotted last (on top)}

\item{label}{Whether to label the clusters}

\item{label.size}{Sets size of labels}

\item{ncol}{Number of columns to combine multiple feature plots to, ignored if \code{split.by} is not \code{NULL}}

\item{combine}{Combine plots into a single gg object; note that if TRUE; themeing will not work when plotting multiple features}

\item{coord.fixed}{Plot cartesian coordinates with fixed aspect ratio.}

\item{...}{Ignored for now}
}
\value{
A ggplot object
}
\description{
Colors single cells on a dimensional reduction plot according to a 'feature'
(i.e. gene expression, PC scores, number of genes detected, etc.)
}
\examples{
FeaturePlot(object = pbmc_small, features.plot = 'PC1')

}
\seealso{
\code{\link{DimPlot}} \code{\link{FeatureMap}}
}
