% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{FeaturePlot}
\alias{FeaturePlot}
\title{Visualize 'features' on a dimensional reduction plot}
\usage{
FeaturePlot(object, features.plot, min.cutoff = NA, max.cutoff = NA,
  reduction.use = "tsne", dims.use = c(1, 2), group.by = NULL,
  cols.use = c("yellow", "red"), cells.use = NULL, pt.size = 1,
  shape.by = NULL, plot.order = NULL, do.label = FALSE, label.size = 4,
  na.value = "grey50", combine.plots = TRUE)
}
\arguments{
\item{object}{Seurat object}

\item{features.plot}{Vector of features to plot}

\item{min.cutoff}{Vector of minimum cutoff values for each feature, may specify quantile in the form of 'q##' where '##' is the quantile (eg, 1, 10)}

\item{max.cutoff}{Vector of maximum cutoff values for each feature, may specify quantile in the form of 'q##' where '##' is the quantile (eg, 1, 10)}

\item{reduction.use}{Which dimensionality reduction to use. Default is
"tsne", can also be "pca", or "ica", assuming these are precomputed.}

\item{dims.use}{Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions}

\item{group.by}{Group (color) cells in different ways (for example, orig.ident)}

\item{cols.use}{The two colors to form the gradient over. Provide as string vector with
the first color corresponding to low values, the second to high. Also accepts a Brewer
color scale or vector of colors. Note: this will bin the data into number of colors provided.}

\item{cells.use}{Vector of cells to plot (default is all cells)}

\item{pt.size}{Adjust point size for plotting}

\item{shape.by}{If NULL, all points are circles (default). You can specify any
cell attribute (that can be pulled with FetchData) allowing for both
different colors and different shapes on cells.}

\item{plot.order}{Specify the order of plotting for the idents. This can be
useful for crowded plots if points of interest are being buried. Provide
either a full list of valid idents or a subset to be plotted last (on top).}

\item{do.label}{Whether to label the clusters}

\item{label.size}{Sets size of labels}

\item{na.value}{Color value for NA points when using custom scale.}

\item{combine.plots}{Combine plots into a single gg object; note that if TRUE; themeing will not work when plotting multiple features}
}
\value{
A ggplot object
}
\description{
Colors single cells on a dimensional reduction plot according to a 'feature'
(i.e. gene expression, PC scores, number of genes detected, etc.)
}
\examples{
FeaturePlot(object = pbmc_small, features.plot = 'PC1')

}
\seealso{
\code{\link{DimPlot}} \code{\link{FeatureMap}}
}
