% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing5.R
\name{FetchResiduals_reference}
\alias{FetchResiduals_reference}
\title{temporal function to get residuals from reference}
\usage{
FetchResiduals_reference(
  object,
  reference.SCT.model = NULL,
  features = NULL,
  nCount_UMI = NULL,
  verbose = FALSE
)
}
\description{
temporal function to get residuals from reference
}
