% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing5.R
\name{FetchResidualSCTModel}
\alias{FetchResidualSCTModel}
\title{Calculate pearson residuals of features not in the scale.data
This function is the secondary function under FetchResiduals}
\usage{
FetchResidualSCTModel(
  object,
  assay = "SCT",
  umi.assay = "RNA",
  layer = "counts",
  chunk_size = 2000,
  layer.cells = NULL,
  SCTModel = NULL,
  reference.SCT.model = NULL,
  new_features = NULL,
  clip.range = NULL,
  replace.value = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{object}{A seurat object}

\item{assay}{Name of the assay of the seurat object generated by
SCTransform. Default is "SCT"}

\item{umi.assay}{Name of the assay of the seurat object to fetch
UMIs from. Default is "RNA"}

\item{layer}{Name of the layer under `umi.assay` to fetch UMIs from.
Default is "counts"}

\item{layer.cells}{Vector of cells to calculate the residual for.
Default is NULL which uses all cells in the layer}

\item{SCTModel}{Which SCTmodel to use from the object for calculating
the residual. Will be ignored if reference.SCT.model is set}

\item{reference.SCT.model}{If a reference SCT model should be used
for calculating the residuals. When set to not NULL, ignores the `SCTModel`
paramater.}

\item{new_features}{A vector of features to calculate the residuals for}

\item{clip.range}{Numeric of length two specifying the min and max values
the Pearson residual will be clipped to. Useful if you want to change the
clip.range.}

\item{replace.value}{Whether to replace the value of residuals if it
already exists}

\item{verbose}{Whether to print messages and progress bars}
}
\value{
Returns a matrix containing centered pearson residuals of
added features
}
\description{
Calculate pearson residuals of features not in the scale.data
This function is the secondary function under FetchResiduals
}
