% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seurat.R
\name{find.clusters}
\alias{find.clusters}
\title{Cluster Determination}
\usage{
find.clusters(object, genes.use = NULL, pc.use = NULL, SNN = NULL,
  k_param = 10, plot.SNN = FALSE, prune.SNN = TRUE, save.SNN = FALSE,
  r_param = 0.7, m_param = NULL, q = 0.1, qup = 0.1, update = 0.25,
  min_cluster_size = 1)
}
\arguments{
\item{object}{Seurat object}

\item{genes.use}{Gene expression data}

\item{pc.use}{Which PCs to use for construction of the SNN graph}

\item{SNN}{Allows use of existing SNN matrix}

\item{k_param}{Defines k for the k-nearest neighbor algorithm}

\item{plot.SNN}{Plot the SNN graph}

\item{prune.SNN}{Prune the SNN graph}

\item{save.SNN}{Whether to return the SNN matrix or not. If true, returns a list with the object as the first item
and the SNN matrix as the second item.}

\item{r_param}{r defines the connectivity for the quasi-cliques. Higher r gives a more compact subgraph}

\item{m_param}{m is the threshold for merging two quasi-cliques. Higher m results in less merging}

\item{q}{Defines the percentage of quasi-cliques to examine for merging each iteration}

\item{qup}{Determines how to change q once all possible merges have been made}

\item{update}{Adjust how verbose the output is}

\item{min_cluster_size}{Smallest allowed size for a cluster}
}
\value{
Returns a Seurat object and optionally the SNN matrix, object@ident has been updated with new cluster info
}
\description{
Identify clusters of cells by a shared nearest neighbor (SNN) quasi-clique
based clustering algorithm. First calculate k-nearest neighbors and construct
the SNN graph. Then determine the quasi-cliques associated with each cell. Finally,
merge the quasi-cliques into clusters. For a full description of the algorithm, see
Xu and Su (2015) \emph{Bioinformatics}.
}

