% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{FindAllMarkers}
\alias{FindAllMarkers}
\title{Gene expression markers for all identity classes}
\usage{
FindAllMarkers(object, thresh.test = 1, test.use = "bimod",
  return.thresh = 0.01, do.print = FALSE, min.pct = 0.1,
  print.bar = TRUE, only.pos = FALSE, max.cells.per.ident = Inf)
}
\arguments{
\item{object}{Seurat object}

\item{thresh.test}{Limit testing to genes which show, on average, at least X-fold difference (log-scale) between cells in an identity class, and all other cells.}

\item{test.use}{Denotes which test to use. Seurat currently implements
"bimod" (likelihood-ratio test for single cell gene expression, McDavid et
al., Bioinformatics, 2011, default), "roc" (standard AUC classifier), "t"
(Students t-test), and "tobit" (Tobit-test for differential gene expression,
as in Trapnell et al., Nature Biotech, 2014)}

\item{return.thresh}{Only return markers that have a p-value < return.thresh, or a power > return.thresh (if the test is ROC)}

\item{do.print}{FALSE by default. If TRUE, outputs updates on progress.}

\item{min.pct}{- only test genes that are detected in a minimum fraction of min.pct cells
in either of the two populations. Meant to speed up the function by not testing genes that are very infrequently expression}

\item{print.bar}{Print a progress bar once expression testing begins (uses pbapply to do this)}

\item{only.pos}{Only return positive markers (FALSE by default)}

\item{max.cells.per.ident}{Down sample each identity class to a max number. Default is no downsampling.}
}
\value{
Matrix containing a ranked list of putative markers, and associated
statistics (p-values, ROC score, etc.)
}
\description{
Finds markers (differentially expressed genes) for each of the identity classes in a dataset
}

