% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_determination.R
\name{FindClusters}
\alias{FindClusters}
\title{Cluster Determination}
\usage{
FindClusters(object, genes.use = NULL, pc.use = NULL, k.param = 30,
  k.scale = 25, plot.SNN = FALSE, prune.SNN = 1/15, save.SNN = FALSE,
  update = 0.25, do.sparse = FALSE, modularity.fxn = 1,
  resolution = 0.8, algorithm = 1, n.start = 100, n.iter = 10,
  random.seed = 0, print.output = TRUE)
}
\arguments{
\item{object}{Seurat object}

\item{genes.use}{Gene expression data}

\item{pc.use}{Which PCs to use for construction of the SNN graph}

\item{k.param}{Defines k for the k-nearest neighbor algorithm}

\item{k.scale}{granularity option for k.param}

\item{plot.SNN}{Plot the SNN graph}

\item{prune.SNN}{Stringency of pruning for the SNN graph (0 - no pruning,
1 - prune everything)}

\item{save.SNN}{Whether to save the SNN in an object slot}

\item{update}{Adjust how verbose the output of the SNN calculation is (0 to 1)}

\item{do.sparse}{Option to store and use SNN matrix as a sparse matrix.
May be necessary datasets containing a large number of cells.}

\item{modularity.fxn}{Modularity function (1 = standard; 2 = alternative).}

\item{resolution}{Value of the resolution parameter, use a value above
(below) 1.0 if you want to obtain a larger (smaller) number of
communities.}

\item{algorithm}{Algorithm for modularity optimization (1 = original Louvain
algorithm; 2 = Louvain algorithm with multilevel refinement;
3 = SLM algorithm).}

\item{n.start}{Number of random starts.}

\item{n.iter}{Maximal number of iterations per random start.}

\item{random.seed}{Seed of the random number generator.}

\item{print.output}{Whether or not to print output to the console}
}
\value{
Returns a Seurat object and optionally the SNN matrix,
        object@ident has been updated with new cluster info
}
\description{
Identify clusters of cells by a shared nearest neighbor (SNN) modularity optimization
based clustering algorithm. First calculate k-nearest neighbors and construct
the SNN graph. Then optimize the modularity function to determine clusters.
For a full description of the algorithms, see Waltman and van Eck (2013) 
\emph{The European Physical Journal B}.
}

