% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{FindMarkers}
\alias{FindMarkers}
\title{Gene expression markers of identity classes}
\usage{
FindMarkers(object, ident.1, ident.2 = NULL, genes.use = NULL,
  logfc.threshold = 0.25, test.use = "wilcox", min.pct = 0.1,
  min.diff.pct = -Inf, print.bar = TRUE, only.pos = FALSE,
  max.cells.per.ident = Inf, random.seed = 1, latent.vars = NULL,
  min.cells.gene = 3, min.cells.group = 3, pseudocount.use = 1,
  assay.type = "RNA", ...)
}
\arguments{
\item{object}{Seurat object}

\item{ident.1}{Identity class to define markers for}

\item{ident.2}{A second identity class for comparison. If NULL (default) -
use all other cells for comparison.}

\item{genes.use}{Genes to test. Default is to use all genes}

\item{logfc.threshold}{Limit testing to genes which show, on average, at least
X-fold difference (log-scale) between the two groups of cells. Default is 0.25
Increasing logfc.threshold speeds up the function, but can miss weaker signals.}

\item{test.use}{Denotes which test to use. Available options are:
\itemize{
 \item{"wilcox"} : Wilcoxon rank sum test (default)
 \item{"bimod"} : Likelihood-ratio test for single cell gene expression,
 (McDavid et al., Bioinformatics, 2013)
 \item{"roc"} : Standard AUC classifier
 \item{"t"} : Student's t-test
 \item{"tobit"} : Tobit-test for differential gene expression (Trapnell et
 al., Nature Biotech, 2014)
 \item{"poisson"} : Likelihood ratio test assuming an underlying poisson
  distribution. Use only for UMI-based datasets
 \item{"negbinom"} :  Likelihood ratio test assuming an underlying negative
 binomial distribution. Use only for UMI-based datasets
 \item{"MAST} : GLM-framework that treates cellular detection rate as a
 covariate (Finak et al, Genome Biology, 2015)
 \item{"DESeq2} : DE based on a model using the negative binomial
 distribution (Love et al, Genome Biology, 2014)
}}

\item{min.pct}{only test genes that are detected in a minimum fraction of
min.pct cells in either of the two populations. Meant to speed up the function
by not testing genes that are very infrequently expressed. Default is 0.1}

\item{min.diff.pct}{only test genes that show a minimum difference in the
fraction of detection between the two groups. Set to -Inf by default}

\item{print.bar}{Print a progress bar once expression testing begins (uses
pbapply to do this)}

\item{only.pos}{Only return positive markers (FALSE by default)}

\item{max.cells.per.ident}{Down sample each identity class to a max number.
Default is no downsampling. Not activated by default (set to Inf)}

\item{random.seed}{Random seed for downsampling}

\item{latent.vars}{Variables to test, used only when \code{test.use} is one of
'negbinom', 'poisson', or 'MAST'}

\item{min.cells.gene}{Minimum number of cells expressing the gene in at least one
of the two groups, currently only used for poisson and negative binomial tests}

\item{min.cells.group}{Minimum number of cells in one of the groups}

\item{pseudocount.use}{Pseudocount to add to averaged expression values when
calculating logFC. 1 by default.}

\item{assay.type}{Type of assay to fetch data for (default is RNA)}

\item{\dots}{Additional parameters to pass to specific DE functions}
}
\value{
Matrix containing a ranked list of putative markers, and associated
statistics (p-values, ROC score, etc.)
}
\description{
Finds markers (differentially expressed genes) for identity classes
}
\details{
p-value adjustment is performed using bonferroni correction based on
the total number of genes in the dataset. Other correction methods are not
recommended, as Seurat pre-filters genes using the arguments above, reducing
the number of tests performed. Lastly, as Aaron Lun has pointed out, p-values
should be interpreted cautiously, as the genes used for clustering are the
same genes tested for differential expression.
}
\examples{
markers <- FindMarkers(object = pbmc_small, ident.1 = 3)
head(markers)

}
\seealso{
\code{\link{MASTDETest}}, and \code{\link{DESeq2DETest}} for more information on these methods

\code{\link{NegBinomDETest}}
}
