% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{FindModalityWeights}
\alias{FindModalityWeights}
\title{Calculate modality weights}
\usage{
FindModalityWeights(
  object,
  reduction.list,
  dims.list,
  k.nn = 20,
  snn.far.nn = TRUE,
  s.nn = k.nn,
  prune.SNN = 0,
  l2.norm = TRUE,
  sd.scale = 1,
  query = NULL,
  cross.contant.list = list(1e-04, 1e-04),
  sigma.idx = k.nn,
  smooth = FALSE,
  verbose = TRUE
)
}
\description{
This function calculates cell-specific modality weights which are used to
construct the multimodal kernel to find multimodal neighbors. It finds
neighbors from each modality and performs within- and cross- modality
prediction to calculate modality weights.
}
\details{
the kernel width
 the neighborhood overlap for the SNN construction.
dimensional reduction
objects are different. snn.far.nn does not support for query object.
used in the modality score calculation.
When snn.far.nn is TRUE, this parameter is not used
neighbors.
\code{\link{FindMultiModalNeighbors}}
}
