% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{FindMultiModalNeighbors}
\alias{FindMultiModalNeighbors}
\title{Construct multimodal neighbors, KNN and SNN Graph}
\usage{
FindMultiModalNeighbors(
  object,
  modality.weight = NULL,
  k.nn = NULL,
  prune.SNN = 1/15,
  knn.graph.name = "jknn",
  snn.graph.name = "jsnn",
  joint.nn.name = "joint.nn",
  modality.weight.name = "first.modality.weight",
  knn.range = 200,
  weighted.graph = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{A Seurat object}

\item{modality.weight}{A \code{\link{ModalityWeights}} object generated by
\code{\link{FindModalityWeights}}}

\item{k.nn}{the number of multi-modal neighbors computed}

\item{prune.SNN}{Sets the cutoff for acceptable Jaccard index when computing
the neighborhood overlap for the SNN construction}

\item{knn.graph.name}{The name of multimodal knn graph}

\item{snn.graph.name}{The name of multimodal snn graph}

\item{joint.nn.name}{The name of multimodal neighbors}

\item{modality.weight.name}{The variable name of first modality weights
stored in the meta.data.}

\item{knn.range}{The number of approximate neighbors to compute}

\item{weighted.graph}{Add consider neighbor distance as the edges to
construct KNN graph}

\item{verbose}{Print progress bars and output}
}
\value{
return an object containing multimodal KNN, SNN and neighbors
}
\description{
This function will construct multimodal neighbors, Kth Nearest Neighbors
(KNN) and Shared Nearest Neighbor (SNN) Graphs. According to the input
\code{modality.weight}, it constructs a cell-specific weighted joint kernel.
Then, for each cell, it will find \code{knn.range}s individual modal
neighbors, and get the union of those neighbors. Next, it will find
\code{k.nn} multimodal neighbors by the weighted joint kernel. Given the set
of multimodal neighbors, we construct its KNN and SNN Graph.
}
