% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/spatial.R
\name{FindSpatiallyVariableFeatures}
\alias{FindSpatiallyVariableFeatures}
\alias{FindSpatiallyVariableFeatures.default}
\alias{FindSpatiallyVariableFeatures.Assay}
\alias{FindSpatiallyVariableFeatures.Seurat}
\title{Find spatially variable features}
\usage{
FindSpatiallyVariableFeatures(object, ...)

\method{FindSpatiallyVariableFeatures}{Assay}(
  object,
  spatial.location,
  selection.method = "markvariogram",
  r.metric = 5,
  ...
)

\method{FindSpatiallyVariableFeatures}{Assay}(
  object,
  slot = "scale.data",
  features = NULL,
  spatial.location,
  selection.method = "markvariogram",
  r.metric = 5,
  nfeatures = nfeatures,
  ...
)

\method{FindSpatiallyVariableFeatures}{Seurat}(
  object,
  assay = NULL,
  slot = "scale.data",
  features = NULL,
  image = NULL,
  selection.method = "markvariogram",
  r.metric = 5,
  nfeatures = 2000,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Arguments passed to other methods}

\item{spatial.location}{}

\item{selection.method}{Method for selecting spatially variable features.
Only 'markvariogram' method is currently implemented. See 
\code{\link{RunMarkVario}} for method.}

\item{r.metric}{r value at which to report the "trans" value of the mark
variogram}

\item{slot}{Slot in the Assay to pull data from}

\item{features}{If provided, only compute on given features. Otherwise, 
compute for all features.}

\item{assay}{Assay to pull the features (marks) from}

\item{image}{Name of image to pull the coordinates from}
}
\description{
Identify features whose variability in expression can be explained to some
degree by spatial location.
}
