% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R, R/preprocessing_generics.R
\name{FindVariableFeatures.Assay}
\alias{FindVariableFeatures.Assay}
\alias{FindVariableFeatures.Seurat}
\alias{FindVariableFeatures}
\title{Find variable features}
\usage{
\method{FindVariableFeatures}{Assay}(object, mean.function = FastExpMean,
  dispersion.function = FastLogVMR, num.bin = 20,
  binning.method = "equal_width", num.features = 1000,
  mean.cutoff = c(0.1, 8), dispersion.cutoff = c(1, Inf),
  selection.method = "mean.var.plot", verbose = TRUE, ...)

\method{FindVariableFeatures}{Seurat}(object, assay.use = NULL,
  mean.function = FastExpMean, dispersion.function = FastLogVMR,
  num.bin = 20, binning.method = "equal_width", num.features = 1000,
  mean.cutoff = c(0.1, 8), dispersion.cutoff = c(1, Inf),
  selection.method = "mean.var.plot", verbose = TRUE, ...)

FindVariableFeatures(object, mean.function, dispersion.function, num.bin,
  binning.method, verbose, ...)
}
\arguments{
\item{object}{An object}

\item{mean.function}{Function to compute x-axis value (average expression). Default
is to take the mean of the detected (i.e. non-zero) values}

\item{dispersion.function}{Function to compute y-axis value (dispersion). Default is to
take the standard deviation of all values}

\item{num.bin}{Total number of bins to use in the scaled analysis (default
is 20)}

\item{binning.method}{Specifies how the bins should be computed. Available methods are:
\itemize{
  \item{equal_width:}{ each bin is of equal width along the x-axis [default]}
  \item{equal_frequency:}{ each bin contains an equal number of features (can increase
statistical power to detect overdispersed features at high expression values, at
the cost of reduced resolution along the x-axis)}
}}

\item{num.features}{Number of features to select as top variable features;
only used when \code{selection.method = 'dispersion'}}

\item{mean.cutoff}{A two-length numeric vector with low- and high-cutoffs for
feature means}

\item{dispersion.cutoff}{A two-length numeric vector with low- and high-cutoffs for
feature dispersions}

\item{selection.method}{How to choose top variable features, either 'mean.var.plot' or 'dispersion}

\item{verbose}{show progress bar for calculations}

\item{...}{Arguments passed to other methods}

\item{assay.use}{...}
}
\description{
Identifies features that are outliers on a 'mean variability plot'. First, uses
a function to calculate average expression (mean.function) and dispersion (dispersion.function)
for each feature. Next, divides features into num.bin (deafult 20) bins based on
their average expression, and calculates z-scores for dispersion within each
bin. The purpose of this is to identify variable features while controlling for
the strong relationship between variability and average expression.
}
\details{
Exact parameter settings may vary empirically from dataset to dataset, and
based on visual inspection of the plot.
Setting the y.cutoff parameter to 2 identifies features that are more than two standard
deviations away from the average dispersion within a bin. The default X-axis function
is the mean expression level, and for Y-axis it is the log(Variance/mean). All mean/variance
calculations are not performed in log-space, but the results are reported in log-space -
see relevant functions for exact details.
}
\section{Methods (by class)}{
\itemize{
\item \code{Assay}: Find variable features in an Assay object

\item \code{Seurat}: Find variable features in a Seurat object
}}

