% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_generics.R, R/seurat.R
\name{GetAssay}
\alias{GetAssay}
\alias{GetAssay.Seurat}
\title{Get an assay from an object}
\usage{
GetAssay(object, assay.use, ...)

\method{GetAssay}{Seurat}(object, assay.use = NULL)
}
\arguments{
\item{object}{An object}

\item{assay.use}{Assay to get}

\item{...}{Arguments passed to other methods}
}
\value{
Returns an Assay object
}
\description{
Get an assay from an object
}
\section{Methods (by class)}{
\itemize{
\item \code{Seurat}: Get an assay from a Seurat object
}}

