% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{GetAssay}
\alias{GetAssay}
\alias{GetAssay.Seurat}
\title{Get an Assay object from a given Seurat object.}
\usage{
GetAssay(object, ...)

\method{GetAssay}{Seurat}(object, assay = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{assay}{Assay to get}
}
\value{
Returns an Assay object
}
\description{
Get an Assay object from a given Seurat object.
}
\examples{
data("pbmc_small")
GetAssay(object = pbmc_small, assay = "RNA")

}
\concept{objects}
