% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_generics.R, R/assay.R
\name{GetAssayData}
\alias{GetAssayData}
\alias{GetAssayData.Assay}
\title{Accessor function for multimodal data}
\usage{
GetAssayData(object, slot, ...)

\method{GetAssayData}{Assay}(object, slot = "data")
}
\arguments{
\item{object}{An object}

\item{slot}{Specific information to pull (i.e. raw.data, data, scale.data,...). Default is data}

\item{...}{Arguments passed to other methods}
}
\value{
Returns assay data
}
\description{
Pull information for specified stored dimensional reduction analysis
}
\section{Methods (by class)}{
\itemize{
\item \code{Assay}: Get assay data for an Assay object
}}

