% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_modal_generics.R, R/multi_modal.R
\name{GetAssayData}
\alias{GetAssayData}
\alias{GetAssayData.seurat}
\alias{GetAssayData.loom}
\title{Accessor function for multimodal data}
\usage{
GetAssayData(object, ...)

\method{GetAssayData}{seurat}(object, assay.type = "RNA", slot = "data")

\method{GetAssayData}{loom}(object, slot = "data", cells.use = NULL,
  genes.use = NULL, cell.names = "col_attrs/cell_names",
  gene.names = "row_attrs/gene_names", chunk.size = 1000,
  display.progress = TRUE, do.sparse = TRUE, ...)
}
\arguments{
\item{object}{Seurat object}

\item{assay.type}{Type of assay to fetch data for (default is RNA)}

\item{slot}{Specific information to pull (i.e. raw.data, data, scale.data,...). Default is data}

\item{cells.use}{Return these cells, set to \code{NULL} for all cells}

\item{genes.use}{Return these genes, set to \code{NULL} for all genes}

\item{cell.names}{Path to cell names, set to \code{NULL} to skip}

\item{gene.names}{Path to gene names, set to \code{NULL} to skip}

\item{chunk.size}{What size to chunk the cells on?}

\item{display.progress}{Show progress bar?}

\item{do.sparse}{Return matrix as a sparse matrix (default is TRUE).}
}
\value{
Returns assay data
}
\description{
Pull information for specified stored dimensional reduction analysis
}
\section{Methods (by class)}{
\itemize{
\item \code{seurat}: Get assay data from a Seurat object

\item \code{loom}: Get assay data from a Seurat object
}}

\examples{
# Simulate CITE-Seq results
df <- t(x = data.frame(
  x = round(x = rnorm(n = 80, mean = 20, sd = 2)),
  y = round(x = rbinom(n = 80, size = 100, prob = 0.2))
))
pbmc_small <- SetAssayData(
  object = pbmc_small,
  assay.type = 'CITE',
  new.data = df,
  slot = 'raw.data'
)
GetAssayData(object = pbmc_small, assay.type = 'CITE', slot = 'raw.data')

}
