% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreduc.R, R/dimreduc_generics.R
\name{GetDimReduc.Seurat}
\alias{GetDimReduc.Seurat}
\alias{GetDimReduc}
\title{Accessor function for DimReduc object}
\usage{
\method{GetDimReduc}{Seurat}(object, slot)

GetDimReduc(object, slot, ...)
}
\arguments{
\item{object}{A DimReduc object}

\item{slot}{Specific information to pull (e.g. cell.embeddings, feature.loadings, ...)}

\item{...}{Arguments passed to other methods}
}
\value{
Returns slot info
}
\description{
Pull information for specified stored dimensional reduction analysis
}
\section{Methods (by class)}{
\itemize{
\item \code{Seurat}: Get a slot for a given DimReduc
}}

