% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{GetDimReduction}
\alias{GetDimReduction}
\title{Dimensional Reduction Accessor Function}
\usage{
GetDimReduction(object, reduction.type = "pca", slot = "x")
}
\arguments{
\item{object}{Seurat object}

\item{reduction.type}{Type of dimensional reduction to fetch}

\item{slot}{Specific information to pull (i.e. rotation, x, x.full, ...)}
}
\value{
Returns results from reduction technique
}
\description{
Pull information for specified stored dimensional reduction analysis
}

