% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_generics.R, R/assay.R
\name{GetIdent}
\alias{GetIdent}
\alias{GetIdent.Assay}
\title{Get the idents of multimodal data}
\usage{
GetIdent(object, ...)

\method{GetIdent}{Assay}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\value{
A factor with the cell identities
}
\description{
Get the idents of multimodal data
}
\section{Methods (by class)}{
\itemize{
\item \code{Assay}: Get cell identities from an Assay object
}}

