% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{GetIdent}
\alias{GetIdent}
\title{Get identity of cells}
\usage{
GetIdent(object, uniq = TRUE, cells.use = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{uniq}{logic, indicating whether to return unique ident values or ident
of all cells}

\item{cells.use}{A vector of cell names. If specified, only the identity of
these cells will be returned.}
}
\value{
Return the cell identites of this Seurat object
}
\description{
Get identity of cells
}
\examples{
GetIdent(pbmc_small)

}
