% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/spatial.R
\name{GetImage}
\alias{GetImage}
\alias{GetImage.Seurat}
\alias{GetImage.SliceImage}
\alias{GetImage.SpatialImage}
\title{Get image data}
\usage{
GetImage(object, mode = "grob", ...)

\method{GetImage}{Seurat}(object, mode = "grob", image = NULL, ...)

\method{GetImage}{SliceImage}(object, mode = "grob", ...)

\method{GetImage}{SpatialImage}(object, mode = "grob", ...)
}
\arguments{
\item{object}{An object}

\item{mode}{How to return the image; should accept at least 'grob' (default) and 'raster'}

\item{...}{Arguments passed to other methods}

\item{image}{Name of \code{SpatialImage} object to pull image data for; if
\code{NULL}, will attempt to select an image automatically}
}
\value{
Image data, either as a grob (if \code{mode = 'grob'}), raster (if
\code{mode = 'raster'}) or other as implemented by each method
}
\description{
Get image data
}
