% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/spatial.R
\name{GetImage}
\alias{GetImage}
\alias{GetImage.SpatialAssay}
\alias{GetImage.Seurat}
\title{Get image data}
\usage{
GetImage(object, ...)

\method{GetImage}{SpatialAssay}(object, ...)

\method{GetImage}{Seurat}(object, assay = NULL, ...)
}
\description{
Get image data
}
