% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{GetImage.SlideSeq}
\alias{GetImage.SlideSeq}
\alias{GetImage.STARmap}
\alias{GetImage.VisiumV1}
\title{Get Image Data}
\usage{
\method{GetImage}{SlideSeq}(object, mode = c("grob", "raster", "plotly", "raw"), ...)

\method{GetImage}{STARmap}(object, mode = c("grob", "raster", "plotly", "raw"), ...)

\method{GetImage}{VisiumV1}(object, mode = c("grob", "raster", "plotly", "raw"), ...)
}
\description{
Get Image Data
}
\seealso{
\code{\link[SeuratObject:GetImage]{SeuratObject::GetImage}}
}
\concept{objects}
\concept{spatial}
