% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{GetImage.SlideSeq}
\alias{GetImage.SlideSeq}
\alias{GetImage.STARmap}
\alias{GetImage.VisiumV1}
\title{Get Image Data}
\usage{
\method{GetImage}{SlideSeq}(object, mode = c("grob", "raster", "plotly", "raw"), ...)

\method{GetImage}{STARmap}(object, mode = c("grob", "raster", "plotly", "raw"), ...)

\method{GetImage}{VisiumV1}(object, mode = c("grob", "raster", "plotly", "raw"), ...)
}
\arguments{
\item{object}{An object}

\item{mode}{How to return the image; should accept one of \dQuote{grob},
\dQuote{raster}, \dQuote{plotly}, or \dQuote{raw}}

\item{...}{Arguments passed to other methods}
}
\description{
Get Image Data
}
\seealso{
\code{\link[SeuratObject:GetImage]{SeuratObject::GetImage}}
}
\concept{objects}
\concept{spatial}
