% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getThreads}
\alias{getThreads}
\title{Get the number of threads being used for parallel processing in Seurat}
\usage{
getThreads()
}
\description{
Get the number of threads being used for parallel processing in Seurat
}
\concept{utilities}
