% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/spatial.R
\name{GetTissueCoordinates}
\alias{GetTissueCoordinates}
\alias{GetTissueCoordinates.SpatialAssay}
\alias{GetTissueCoordinates.Seurat}
\title{Get tissue coordinates}
\usage{
GetTissueCoordinates(object, ...)

\method{GetTissueCoordinates}{SpatialAssay}(object, ...)

\method{GetTissueCoordinates}{Seurat}(object, assay = NULL, ...)
}
\description{
Get tissue coordinates
}
