% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{GetTissueCoordinates.SlideSeq}
\alias{GetTissueCoordinates.SlideSeq}
\alias{GetTissueCoordinates.STARmap}
\alias{GetTissueCoordinates.VisiumV1}
\title{Get Tissue Coordinates}
\usage{
\method{GetTissueCoordinates}{SlideSeq}(object, ...)

\method{GetTissueCoordinates}{STARmap}(object, qhulls = FALSE, ...)

\method{GetTissueCoordinates}{VisiumV1}(
  object,
  scale = "lowres",
  cols = c("imagerow", "imagecol"),
  ...
)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{qhulls}{return qhulls instead of centroids}

\item{scale}{A factor to scale the coordinates by; choose from: 'tissue',
'fiducial', 'hires', 'lowres', or \code{NULL} for no scaling}

\item{cols}{Columns of tissue coordinates data.frame to pull}
}
\description{
Get Tissue Coordinates
}
\seealso{
\code{\link[SeuratObject:GetTissueCoordinates]{SeuratObject::GetTissueCoordinates}}
}
\concept{objects}
\concept{spatial}
