% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_modal.R
\name{HTODemux}
\alias{HTODemux}
\title{Demultiplex samples based on data from cell 'hashing'}
\usage{
HTODemux(object, percent_cutoff = 0.995, init_centers = NULL,
  cluster_nstarts = 100, print.output = TRUE)
}
\arguments{
\item{object}{Seurat object. Assumes that the hash tag oligo (HTO) data has been added and normalized in the HTO slot.}

\item{percent_cutoff}{The quantile of inferred 'negative' distribution for each HTO - over which the cell is considered 'positive'.}

\item{init_centers}{Initial number of clusters for kmeans of the HTO oligos. Default is the # of samples + 1 (to account for negatives)}

\item{cluster_nstarts}{nstarts value for the initial k-means clustering}

\item{print.output}{Prints the output}
}
\value{
Seurat object. Demultiplexed information is stored in the object meta data.
}
\description{
Assign sample-of-origin for each cell, annotate doublets.
}
\examples{
\dontrun{
object <- HTODemux(object)
}
}
