% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{ICA}
\alias{ICA}
\title{Run Independent Component Analysis on gene expression}
\usage{
ICA(object, ic.genes = NULL, do.print = TRUE, ics.print = 1:5,
  ics.store = 50, genes.print = 50, use.imputed = FALSE, seed.use = 1,
  ...)
}
\arguments{
\item{object}{Seurat object}

\item{ic.genes}{Genes to use as input for ICA. Default is object@var.genes}

\item{do.print}{Print the top genes associated with high/low loadings for
the ICs}

\item{ics.print}{ICs to print genes for}

\item{ics.store}{Number of ICs to store}

\item{genes.print}{Number of genes to print for each IC}

\item{use.imputed}{Run ICA on imputed values (FALSE by default)}

\item{seed.use}{Random seed to use for fastICA}

\item{\dots}{Additional arguments to be passed to fastICA}
}
\value{
Returns Seurat object with an ICA embedding (object@ica.rot) and
gene projection matrix (object@ica.x). The ICA object itself is stored in
object@ica.obj[[1]]
}
\description{
Run fastICA algorithm for ICA dimensionality reduction
}

