% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction_utilities.R
\name{ICTopCells}
\alias{ICTopCells}
\title{Find cells with highest ICA scores}
\usage{
ICTopCells(object, ic.use = 1, num.cells = NULL, do.balanced = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{ic.use}{Independent component to use}

\item{num.cells}{Number of cells to return}

\item{do.balanced}{Return an equal number of cells with both + and - PC scores.}
}
\value{
Returns a vector of cells
}
\description{
Return a list of genes with the strongest contribution to a set of principal
components
}
