% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_generics.R, R/seurat.R
\name{Idents}
\alias{Idents}
\alias{Idents<-}
\alias{Idents.Seurat}
\alias{Idents<-.Seurat}
\title{Get an object's cell identities}
\usage{
Idents(object, ...)

Idents(object, ...) <- value

\method{Idents}{Seurat}(object, ...)

\method{Idents}{Seurat}(object, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{The name of the identites to pull or the identities themselves}
}
\value{
The cell identies

An object with the cell identites changed
}
\description{
Get an object's cell identities
}
\section{Methods (by class)}{
\itemize{
\item \code{Seurat}: Get the active identities of a Seurat object

\item \code{Seurat}: Set the active identities of a Seurat object
}}

