% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Indices}
\alias{Indices}
\alias{Indices.Neighbor}
\title{Get Neighbor nearest neighbor index matrices}
\usage{
Indices(object, ...)

\method{Indices}{Neighbor}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods;}
}
\value{
A matrix with the nearest neighbor indices
}
\description{
Get Neighbor nearest neighbor index matrices
}
