% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{InitialMapping}
\alias{InitialMapping}
\title{Infer spatial origins for single cells}
\usage{
InitialMapping(object, cells.use = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{cells.use}{Which cells to map}
}
\value{
Seurat object, where mapping probabilities for each bin are stored
in object@final.prob
}
\description{
Probabilistically maps single cells based on (imputed) gene expression
estimates, a set of mixture models, and an in situ spatial reference map.
}
