% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration5.R
\name{IntegrateLayers}
\alias{IntegrateLayers}
\title{Integrate Layers}
\usage{
IntegrateLayers(
  object,
  method,
  group.by = NULL,
  assay = NULL,
  features = NULL,
  layers = NULL,
  scale.layer = "scale.data",
  ...
)
}
\arguments{
\item{object}{A \code{\link[SeuratObject]{Seurat}} object}

\item{method}{Integration method function; can choose from:
\Sexpr[stage=render,results=rd]{Seurat::.rd_methods("integration")}}

\item{...}{Arguments passed on to \code{method}}
}
\value{
\code{object} with integration data added to it
}
\description{
Integrate Layers
}
\section{Integration Methods Functions}{

Integration method functions can be written by anyone to implement any
integration method in Seurat. These methods should expect to take a
\link[SeuratObject:Assay5]{v5 assay} as input and return a named list of
objects that can be added back to a \code{Seurat} object (eg. a
\link[SeuratObject:DimReduc]{dimensional reduction} or cell-level meta data)

Every integration method function should expect the following arguments:
\itemize{
 \item \dQuote{\code{object}}: an \code{\link[SeuratObject]{Assay5}} object
 \item \dQuote{\code{assay}}: name of \code{object} in the original
 \code{\link[SeuratObject]{Seurat}} object
 \item \dQuote{\code{layers}}: names of normalized layers in \code{object}
 \item \dQuote{\code{scale.layer}}: name(s) of scaled layer(s) in
 \code{object}
 \item \dQuote{\code{features}}: a vector of features for integration
 \item \dQuote{\code{groups}}: a one-column data frame with the groups for
 each cell in \code{object}; the column name will be \dQuote{group}
}
}

