% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{IntegrateSketchEmbeddings}
\alias{IntegrateSketchEmbeddings}
\title{Integrate embeddings from the integrated atoms}
\usage{
IntegrateSketchEmbeddings(
  object,
  atom.sketch.object,
  features = NULL,
  assay = "RNA",
  atom.sketch.reduction = "integrated_dr",
  reduction.name = "pca.correct",
  reduction.key = "PCcorrect_",
  dictionary.method = c("sketch", "data"),
  sketch.ratio = 0.8,
  verbose = TRUE
)
}
\arguments{
\item{object}{A Seurat object with all cells for one dataset}

\item{atom.sketch.object}{A sketched Seurat objects with integrated embeddings}

\item{features}{Features used for atomic sketch integration}

\item{assay}{Assay name for original expression}

\item{atom.sketch.reduction}{Dimensional reduction name for batch-corrected embeddings
in the sketched object}

\item{reduction.name}{dimensional reduction name, pca.correct by default}

\item{reduction.key}{dimensional reduction key, specifies the string before
the number for the dimension names. PCcorrect_ by default}

\item{dictionary.method}{Methods to construct sketch-cell representation
for all cells. sketch by default. Can be one of:
\itemize{
\item{sketch: Use random sketched data slot}
\item{data: Use data slot}
}}

\item{sketch.ratio}{Sketch ratio of data slot when dictionary.method is set to sketch}

\item{verbose}{Print progress and message}
}
\value{
Returns a Seurat object with an integrated dimensional reduction
}
\description{
First learn a atom dictionary representation to reconstruct each cell.
Then, relying on this dictionary representation, 
reconstruct the embeddings of each cell from the integrated atoms.
}
