% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{IntegrateSketchEmbeddings}
\alias{IntegrateSketchEmbeddings}
\title{Integrate embeddings from the integrated atoms}
\usage{
IntegrateSketchEmbeddings(
  object,
  atoms = "sketch",
  orig = "RNA",
  features = NULL,
  reduction = "integrated_dr",
  method = c("sketch", "data"),
  ratio = 0.8,
  reduction.name = NULL,
  reduction.key = NULL,
  layers = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{A Seurat object with all cells for one dataset}

\item{atoms}{Assay name for sketched-cell expression (default is 'sketch')}

\item{orig}{Assay name for original expression (default is 'RNA')}

\item{features}{Features used for atomic sketch integration}

\item{reduction}{Dimensional reduction name for batch-corrected embeddings
in the sketched object (default is 'integrated_dr')}

\item{method}{Methods to construct sketch-cell representation
for all cells (default is 'sketch'). Can be one of:
\itemize{
 \item \dQuote{\code{sketch}}: Use random sketched data slot
 \item \dQuote{\code{data}}: Use data slot
}}

\item{ratio}{Sketch ratio of data slot when \code{dictionary.method} is set
to \dQuote{\code{sketch}}; defaults to 0.8}

\item{reduction.name}{Name to save new reduction as; defaults to
\code{paste0(reduction, '.orig')}}

\item{reduction.key}{Key for new dimensional reduction; defaults to creating
one from \code{reduction.name}}

\item{layers}{Names of layers for correction.}

\item{verbose}{Print progress and message}
}
\value{
Returns a Seurat object with an integrated dimensional reduction
}
\description{
The main steps of this procedure are outlined below. For a more detailed
description of the methodology, please see Hao,  et al Biorxiv 2022:
\doi{10.1101/2022.02.24.481684}
}
\details{
First learn a atom dictionary representation to reconstruct each cell.
Then, using this dictionary representation,
reconstruct the embeddings of each cell from the integrated atoms.
}
