% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{IntegrateSketchEmbeddings}
\alias{IntegrateSketchEmbeddings}
\title{Integrate embeddings from the integrated atoms}
\usage{
IntegrateSketchEmbeddings(
  object,
  atom.assay = "sketch",
  orig.assay = "RNA",
  features = NULL,
  atom.sketch.reduction = "integrated_dr",
  dictionary.method = c("sketch", "data"),
  sketch.ratio = 0.8,
  layers = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{A Seurat object with all cells for one dataset}

\item{atom.assay}{Assay name for sketched-cell expression (default is 'sketch')}

\item{orig.assay}{Assay name for original expression (default is 'RNA')}

\item{features}{Features used for atomic sketch integration}

\item{atom.sketch.reduction}{Dimensional reduction name for batch-corrected embeddings
in the sketched object (default is 'integrated_dr')}

\item{dictionary.method}{Methods to construct sketch-cell representation
for all cells (default is 'sketch'). Can be one of:
\itemize{
\item{sketch: Use random sketched data slot}
\item{data: Use data slot}
}}

\item{sketch.ratio}{Sketch ratio of data slot when \code{dictionary.method} is set to 'sketch' (default is 0.8)}

\item{layers}{Names of layers for correction.}

\item{verbose}{Print progress and message (default is TRUE)}
}
\value{
Returns a Seurat object with an integrated dimensional reduction
}
\description{
The main steps of this procedure are outlined below. For a more detailed
description of the methodology, please see Hao,  et al Biorxiv 2022:
\doi{10.1101/2022.02.24.481684}
}
\details{
First learn a atom dictionary representation to reconstruct each cell.
Then, using this dictionary representation,
reconstruct the embeddings of each cell from the integrated atoms.
}
