% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{IntegrateSketchEmbeddings}
\alias{IntegrateSketchEmbeddings}
\title{Integrate embeddings from batch-corrected sketch cell embeddings
First construct a sketch-cell representation for all cells and
then use this and the batch-corrected embeddings of sketched cells to
construct the batch-corrected embeddings for all cells}
\usage{
IntegrateSketchEmbeddings(
  object.list,
  sketch.list,
  sketch.object,
  features = NULL,
  assay = "RNA",
  sketch.reduction = "integrated_dr",
  reduction.name = "pca.correct",
  reduction.key = "PCcorrect_",
  dictionary.method = c("sketch", "data", "embeddings")[1],
  sketch.ratio = 0.8,
  sketch.reduction.raw = NULL,
  merged.object = NULL,
  reference.index = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object.list}{A list of Seurat objects with all cells}

\item{sketch.list}{A list of Seurat objects with sketched cells}

\item{sketch.object}{A sketched Seurat objects with integrated embeddings}

\item{features}{Features used for sketch integration}

\item{assay}{Assay name for raw expression}

\item{sketch.reduction}{Dimensional reduction name for batch-corrected embeddings
in the sketched object}

\item{reduction.name}{dimensional reduction name, pca.correct by default}

\item{reduction.key}{dimensional reduction key, specifies the string before
the number for the dimension names. PCcorrect_ by default}

\item{dictionary.method}{Methods to construct sketch-cell representation
for all cells. sketch by default. Can be one of:
\itemize{
\item{sketch: Use random sketched data slot}
\item{data: Use data slot}
\item{embeddings: Use uncorrected dimensional reduction in the sketched object}
}}

\item{sketch.ratio}{Sketch ratio of data slot when dictionary.method is set to sketch}

\item{sketch.reduction.raw}{Uncorrected dimensional reduction name in the sketched object
when dictionary.method is set to embeddings}

\item{merged.object}{A merged seurat object containing all cells}

\item{reference.index}{Index for the integration reference}

\item{verbose}{Print progress and message}
}
\description{
Integrate embeddings from batch-corrected sketch cell embeddings
First construct a sketch-cell representation for all cells and
then use this and the batch-corrected embeddings of sketched cells to
construct the batch-corrected embeddings for all cells
}
