% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{InteractiveSpatialPlot}
\alias{InteractiveSpatialPlot}
\title{Interactive Spatial Cell Selection Tool}
\usage{
InteractiveSpatialPlot(
  object,
  image = NULL,
  image.scale = "lowres",
  group.by = NULL,
  alpha = 1
)
}
\arguments{
\item{object}{A \code{\link[SeuratObject]{Seurat}} object with spatial data}

\item{image}{Name of the spatial image stored in the object (default is \code{"anterior1"})}

\item{use.clusters}{Logical, whether to colour cells by \code{seurat_clusters} if available (default is \code{FALSE})}
}
\value{
A character vector of cell names selected via lasso (can then be used to subset the object)
}
\description{
InteractiveSpatialPlot launches an interactive gadget to lasso-select cells from a 
spatial Seurat object. Currently works only for Visium and SlideSeq data. 
It returns the cell names of the selected subset.
}
\note{
This function requires the
\href{https://cran.r-project.org/package=plotly}{\pkg{plotly}} and
\href{https://cran.r-project.org/package=magrittr}{\pkg{magrittr}} packages
to be installed.
}
\examples{
\dontrun{
selected_cells <- InteractiveSpatialPlot(object = brain, use.clusters = TRUE)
}

}
\keyword{internal}
