% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{ISpatialFeaturePlot}
\alias{ISpatialFeaturePlot}
\title{Visualize features spatially and interactively}
\usage{
ISpatialFeaturePlot(
  object,
  feature,
  image = NULL,
  slot = "data",
  alpha = c(0.1, 1)
)
}
\arguments{
\item{object}{A Seurat object}

\item{feature}{Feature to visualize}

\item{image}{Name of the image to use in the plot}

\item{slot}{If plotting a feature, which data slot to pull from (counts,
data, or scale.data)}

\item{alpha}{Controls opacity of spots. Provide as a vector specifying the
min and max for SpatialFeaturePlot. For SpatialDimPlot, provide a single
alpha value for each plot.}
}
\value{
Returns final plot as a ggplot object
}
\description{
Visualize features spatially and interactively
}
\concept{spatial}
\concept{visualization}
