% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackstraw.R
\name{JackStraw}
\alias{JackStraw}
\title{Determine statistical significance of PCA scores.}
\usage{
JackStraw(object, num.pc = 20, num.replicate = 100, prop.freq = 0.01,
  display.progress = TRUE, do.par = FALSE, num.cores = 1, maxit = 1000)
}
\arguments{
\item{object}{Seurat object}

\item{num.pc}{Number of PCs to compute significance for}

\item{num.replicate}{Number of replicate samplings to perform}

\item{prop.freq}{Proportion of the data to randomly permute for each
replicate}

\item{display.progress}{Print progress bar showing the number of replicates
that have been processed.}

\item{do.par}{use parallel processing for regressing out variables faster.
If set to TRUE, will use half of the machines available cores (FALSE by default)}

\item{num.cores}{If do.par = TRUE, specify the number of cores to use.
Note that for higher number of cores, larger free memory is needed.
If \code{num.cores = 1} and \code{do.par = TRUE}, \code{num.cores} will be set to half
of all available cores on the machine.}

\item{maxit}{maximum number of iterations to be performed by the irlba function of RunPCA}
}
\value{
Returns a Seurat object where object@dr$pca@jackstraw@emperical.p.value
represents p-values for each gene in the PCA analysis. If ProjectPCA is
subsequently run, object@dr$pca@jackstraw@emperical.p.value.full then
represents p-values for all genes.
}
\description{
Randomly permutes a subset of data, and calculates projected PCA scores for
these 'random' genes. Then compares the PCA scores for the 'random' genes
with the observed PCA scores to determine statistical signifance. End result
is a p-value for each gene's association with each principal component.
}
\examples{
\dontrun{
pbmc_small = suppressWarnings(JackStraw(pbmc_small))
head(pbmc_small@dr$pca@jackstraw@emperical.p.value)
}

}
\references{
Inspired by Chung et al, Bioinformatics (2014)
}
