% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{class}
\name{JackStrawData-class}
\alias{JackStrawData-class}
\title{The JackStrawData Class}
\description{
For more details, please see the documentation in
\code{\link[SeuratObject:JackStrawData]{SeuratObject}}
}
\seealso{
\code{\link[SeuratObject:JackStrawData]{SeuratObject::JackStrawData-class}}
}
