% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{JS}
\alias{JS}
\alias{JS<-}
\alias{JS.DimReduc}
\alias{JS.JackStrawData}
\alias{JS<-.DimReduc}
\alias{JS<-.JackStrawData}
\title{Get JackStraw information}
\usage{
JS(object, ...)

JS(object, ...) <- value

\method{JS}{DimReduc}(object, slot = NULL, ...)

\method{JS}{JackStrawData}(object, slot, ...)

\method{JS}{DimReduc}(object, slot = NULL, ...) <- value

\method{JS}{JackStrawData}(object, slot, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{JackStraw information}

\item{slot}{Name of slot to store JackStraw scores to
Can shorten to 'empirical', 'fake', 'full', or 'overall'}
}
\description{
Get JackStraw information

Set JackStraw information
}
