% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/integration.R
\name{LeverageScore}
\alias{LeverageScore}
\alias{LeverageScore.default}
\alias{LeverageScore.Assay}
\alias{LeverageScore.Seurat}
\title{Calculate Leverage score for all cells}
\usage{
LeverageScore(object, ...)

\method{LeverageScore}{default}(
  object,
  features = NULL,
  nsketch = 5000L,
  ndims = NULL,
  sampling.method = c("CountSketch", "Gaussian"),
  MARGIN = 2L,
  eps = 0.5,
  seed = 123,
  verbose = TRUE,
  ...
)

\method{LeverageScore}{Assay}(
  object,
  features = NULL,
  nsketch = 5000L,
  ndims = NULL,
  sampling.method = c("CountSketch", "Gaussian")[1],
  slot = "data",
  seed = 123,
  eps = 0.5,
  verbose = TRUE,
  ...
)

\method{LeverageScore}{Seurat}(
  object,
  features = NULL,
  assay = NULL,
  nsketch = 5000L,
  ndims = NULL,
  var.name = "leverage.score",
  sampling.method = c("CountSketch", "Gaussian")[1],
  slot = "data",
  eps = 0.5,
  seed = 123,
  over.write = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{features}{Features used to calculate leverage score}

\item{nsketch}{Number of rows in the random sketch matrix}

\item{ndims}{Number of dimensions in the JL embeddings}

\item{sampling.method}{Sampling method for generating random matrix}

\item{MARGIN}{Margin}

\item{eps}{error tolerance for JL embeddings}

\item{seed}{Set a random seed}

\item{verbose}{Print message and process}

\item{slot}{The slot used for leverage score calculation. data slot is used by default}

\item{assay}{Assay used to calculate leverage score}

\item{var.name}{Variable name stored leverage score in the meta.data}

\item{over.write}{If over write the variable with leverage score}
}
\value{
Returns a seurat object with additional column storing leverage score
}
\description{
Leverage score can be used to sample representative cells from scRNA data.
The more abundant population will be assigned less leverage score.
Leverage-score can guarantee that both abundant and rare populations will
be sampled. We used variable features in the data slot to calculate leverage
score for all cells.
}
\references{
Clarkson KL, Woodruff DP.
Low Rank Approximation and Regression in Input Sparsity Time.
Journal of the ACM (JACM). 2017 Jan 30;63(6):1-45.
\url{https://https://arxiv.org/abs/1207.6365};
}
