% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{LeverageScoreSampling}
\alias{LeverageScoreSampling}
\title{Subset objects based on Leverage score}
\usage{
LeverageScoreSampling(
  object,
  num.cells = 5000,
  assay = NULL,
  features = NULL,
  var.name = "leverage.score",
  over.write = FALSE,
  seed = 123,
  ...
)
}
\arguments{
\item{object}{A seurat object}

\item{num.cells}{Number of sampled cells}

\item{assay}{Assay used to calculate leverage score}

\item{features}{Features used to calculate leverage score}

\item{var.name}{Variable name stored leverage score in the meta.data}

\item{over.write}{If over write the variable with leverage score}

\item{seed}{Set a random seed.By default, sets the seed to 123}

\item{...}{Arguments passed to LeverageScore}
}
\value{
Returns a sub-sampled seurat object
}
\description{
Subset objects based on Leverage score
}
