% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Loadings}
\alias{Loadings}
\alias{Loadings<-}
\alias{Loadings.DimReduc}
\alias{Loadings<-.DimReduc}
\title{Get feature loadings}
\usage{
Loadings(object, ...)

Loadings(object, ...) <- value

\method{Loadings}{DimReduc}(object, projected = NULL, ...)

\method{Loadings}{DimReduc}(object, projected = TRUE, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{Feature loadings to add}

\item{projected}{Pull the projected feature loadings?}
}
\description{
Get feature loadings

Add feature loadings
}
