% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreduc_generics.R
\name{Loadings}
\alias{Loadings}
\title{Get feature loadings}
\usage{
Loadings(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\description{
Get feature loadings
}
