% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Loadings}
\alias{Loadings}
\alias{Loadings<-}
\alias{Loadings.DimReduc}
\alias{Loadings<-.DimReduc}
\title{Get feature loadings}
\usage{
Loadings(object, projected, ...)

Loadings(object, ...) <- value

\method{Loadings}{DimReduc}(object, projected = NULL, ...)

\method{Loadings}{DimReduc}(object, ...) <- value
}
\arguments{
\item{object}{An object}

\item{projected}{Pull the projected feature loadings?}

\item{...}{Arguments passed to other methods}

\item{value}{Feature loadings to add}
}
\description{
Get feature loadings

Add feature loadings
}
\section{Methods (by class)}{
\itemize{
\item \code{DimReduc}: Get the feature loadings from a DimReduc object

\item \code{DimReduc}: Add projected feature loadings to a DimReduc object
}}

