% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{LogNormalize}
\alias{LogNormalize}
\title{Normalize raw data}
\usage{
LogNormalize(data)
}
\arguments{
\item{data}{Matrix with the raw count data}
}
\value{
Returns a matrix with the normalize and log transformed data
}
\description{
Normalize count data per cell and transform to log scale
}

