% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{LogNormalize}
\alias{LogNormalize}
\title{Normalize raw data}
\usage{
LogNormalize(data, scale.factor = 10000, display.progress = TRUE)
}
\arguments{
\item{data}{Matrix with the raw count data}

\item{scale.factor}{Scale the data. Default is 1e4}

\item{display.progress}{Print progress}
}
\value{
Returns a matrix with the normalize and log transformed data
}
\description{
Normalize count data per cell and transform to log scale
}
