% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{MakeSparse}
\alias{MakeSparse}
\title{Make object sparse}
\usage{
MakeSparse(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{
Returns a seurat object with data converted to sparse matrices.
}
\description{
Converts stored data matrices to sparse matrices to save space. Converts object@raw.data and object@data to sparse matrices.
If the snn has been stored as a dense matrix, this will convert it to a sparse matrix, store it in object@snn.sparse and
remove object@snn.dense.
}
