% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{MapQueryData}
\alias{MapQueryData}
\title{Map query cells to reference dimensional reduction}
\usage{
MapQueryData(
  reference,
  query,
  proj.reduction = "pcaproject",
  anchor.reduction = NULL,
  dims,
  anchorset = NULL,
  transfer.labels = NULL,
  transfer.expression = NULL,
  verbose = TRUE,
  ident.var = "ref.ident",
  k.nn = 20,
  k.weight = 50,
  sd.weight = 1,
  eps = 0,
  nn.method = "annoy",
  reference.index = NULL
)
}
\arguments{
\item{reference}{\code{\link{Seurat}} object to use as the reference}

\item{query}{\code{\link{Seurat}} object to use as the query}

\item{proj.reduction}{Name of the projected dimensional reduction in the query 
from \code{\link{FindMappingAnchors}}}

\item{anchor.reduction}{Name of the dimensional reduction used to calculate
cell-anchor weights matrix}

\item{dims}{Number of dimensions to use in the anchor weighting procedure}

\item{anchorset}{An \code{\link{AnchorSet}} object generated by 
\code{\link{FindMappingAnchors}}}

\item{transfer.labels}{Identities to transfer to the query}

\item{transfer.expression}{Expression values to transfer to the query}

\item{verbose}{Print progress bars and output}

\item{ident.var}{The name of variable to distinguish reference and query cells}

\item{k.nn}{How many neighbors (k) to use}

\item{k.weight}{Number of neighbors to consider when weighting anchors}

\item{sd.weight}{Controls the bandwidth of the Gaussian kernel for weighting}

\item{eps}{Error bound on the neighbor finding algorithm (from 
\code{\link{RANN}})}

\item{nn.method}{Method for nearest neighbor finding. Options include: rann,
annoy}

\item{reference.index}{Optional precomputed Neighbors index object returned by NNHelper with cache.index = TRUE}
}
\description{
Map query cells to reference dimensional reduction
}
