% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{MatrixRowShuffle}
\alias{MatrixRowShuffle}
\title{Independently shuffle values within each row of a matrix}
\usage{
MatrixRowShuffle(x)
}
\arguments{
\item{x}{Matrix to shuffle}
}
\value{
Returns a scrambled matrix, where each row is shuffled independently
}
\description{
Creates a matrix where correlation structure has been removed, but overall values are the same
}
\examples{
mat <- matrix(data = rbinom(n = 25, size = 20, prob = 0.2 ), nrow = 5)
mat
MatrixRowShuffle(x = mat)

}
