% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{MergeNode}
\alias{MergeNode}
\title{Merge subchilden of a node}
\usage{
MergeNode(object, node.use = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{node.use}{Merge subchildren of this node}
}
\description{
Merge the subchilden of a node into a single identity class
}

