% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{MergeNode}
\alias{MergeNode}
\title{Merge childen of a node}
\usage{
MergeNode(object, node.use = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{node.use}{Merge children of this node}
}
\description{
Merge the childen of a node into a single identity class
}
