% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{MergeSeurat}
\alias{MergeSeurat}
\title{Merge Seurat Objects}
\usage{
MergeSeurat(object1, object2, project = NULL, min.cells = 0,
  min.genes = 0, is.expr = 0, do.logNormalize = TRUE,
  total.expr = 10000, do.scale = TRUE, do.center = TRUE,
  names.field = 1, names.delim = "_", save.raw = TRUE,
  add.cell.id1 = NULL, add.cell.id2 = NULL)
}
\arguments{
\item{object1}{First Seurat object to merge}

\item{object2}{Second Seurat object to merge}

\item{min.cells}{Include genes with detected expression in at least this
many cells}

\item{min.genes}{Include cells where at least this many genes are detected}

\item{is.expr}{Expression threshold for 'detected' gene}

\item{do.logNormalize}{whether to normalize the expression data per cell and transform to log space.}

\item{total.expr}{scale factor in the log normalization}

\item{do.scale}{In object@scale.data, perform row-scaling (gene-based
z-score)}

\item{do.center}{In object@scale.data, perform row-centering (gene-based
centering)}

\item{names.field}{For the initial identity class for each cell, choose this
field from the cell's column name}

\item{names.delim}{For the initial identity class for each cell, choose this
delimiter from the cell's column name}

\item{save.raw}{TRUE by default. If FALSE, do not save the unmodified data in object@raw.data
which will save memory downstream for large datasets}

\item{add.cell.id1}{String to be appended to the names of all cells in object1}

\item{add.cell.id2}{String to be appended to the names of all cells in object2}

\item{meta.data}{Additional metadata to add to the Seurat object. Should be
a data frame where the rows are cell names, and the columns are additional
metadata fields}
}
\value{
Merged Seurat object
}
\description{
Merge two Seurat objects
}
