% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Misc}
\alias{Misc}
\alias{Misc<-}
\alias{Misc.Assay}
\alias{Misc.Seurat}
\alias{Misc<-.Assay}
\alias{Misc<-.Seurat}
\title{Access miscellaneous data}
\usage{
Misc(object, ...)

Misc(object, ...) <- value

\method{Misc}{Assay}(object, slot = NULL, ...)

\method{Misc}{Seurat}(object, slot = NULL, ...)

\method{Misc}{Assay}(object, slot, ...) <- value

\method{Misc}{Seurat}(object, slot, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{Data to add}

\item{slot}{Name of specific bit of meta data to pull}
}
\value{
Miscellaneous data

An object with miscellaneous data added
}
\description{
Access miscellaneous data

Set miscellaneous data
}
\examples{
# Get the misc info
Misc(object = pbmc_small, slot = "example")

# Add misc info
Misc(object = pbmc_small, slot = "example") <- "testing_misc"

}
