% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R, R/unknown.R
\name{MultiModal_CCA}
\alias{MultiModal_CCA}
\title{Canonical Correlation Analysis between modalities}
\usage{
MultiModal_CCA(object, assay.1 = "RNA", assay.2 = "CITE",
  features.1 = NULL, features.2 = NULL, num.cc = 20,
  normalize.variance = TRUE)

MultiModal_CCA(object, assay.1 = "RNA", assay.2 = "CITE",
  features.1 = NULL, features.2 = NULL, num.cc = 20,
  normalize.variance = TRUE)
}
\arguments{
\item{object}{Seurat object. Assumes that scale.data exist for both assays.}

\item{assay.1}{The first assay for CCA.}

\item{assay.2}{The second assay for CCA.}

\item{features.1}{Features to use for the first assay, default is all the features (use object@var.genes if this is RNA).}

\item{features.2}{Features to use for the second assay, default is all the features.}

\item{num.cc}{Minimal number of CCs to return.}

\item{normalize.variance}{Whether to scale the embeddings. Default is TRUE.}

\item{object}{Seurat object}

\item{assay.1}{First assay for multimodal analysis. Default is RNA}

\item{assay.2}{Second assay for multimodal analysis. Default is CITE for CITE-Seq analysis.}

\item{features.1}{Features of assay 1 to consider (default is variable genes)}

\item{features.2}{Features of assay 2 to consider (default is all features, i.e. for CITE-Seq, all antibodies)}

\item{num.cc}{Number of canonical correlations to compute and store. Default is 20, but will calculate less if either assay has <20 features.}

\item{normalize.variance}{Z-score the embedding of each CC to 1, so each CC contributes equally in downstream analysis (default is T)}
}
\value{
Returns Seurat object with two CCA results stored (for example, object@dr$RNACCA and object@dr$ADTCCA).

Returns object after CCA, with results stored in dimensional reduction cca.assay1 (ie. cca.RNA) and cca.assay2. For example, results can be visualized using DimPlot(object,reduction.use="cca.RNA")
}
\description{
Runs a canonical correlation analysis between two assays (for example, RNA and ADT from CITE-seq)

CCA finds a shared correlation structure betwen two different datasets, enabling integrated downstream analysis
}
\examples{
\dontrun{
object <- MultiModal_CCA(object,assay.1 = "RNA",assay.2 = "ADT")
}

}
