% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing5.R
\name{MVP}
\alias{MVP}
\title{Find variable features based on mean.var.plot}
\usage{
MVP(
  data,
  verbose = TRUE,
  nselect = 2000L,
  mean.cutoff = c(0.1, 8),
  dispersion.cutoff = c(1, Inf),
  ...
)
}
\description{
Find variable features based on mean.var.plot
}
