% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{Neighbor-class}
\alias{Neighbor-class}
\alias{Neighbor}
\title{The Neighbor class}
\description{
The Neighbor class is used to store the results of neighbor finding
algorithms
}
\section{Slots}{

\describe{
\item{\code{nn.idx}}{Matrix containing the nearest neighbor indices}

\item{\code{nn.dist}}{Matrix containing the nearest neighbor distances}

\item{\code{alg.idx}}{The neighbor finding index (if applicable). E.g. the annoy
index}

\item{\code{alg.info}}{Any information associated with the algorithm that may be
needed downstream (e.g. distance metric used with annoy is needed when
reading in from stored file).}

\item{\code{cell.names}}{Names of the cells for which the neighbors have been
computed.}
}}

\concept{objects}
