% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{class}
\name{Neighbor-class}
\alias{Neighbor-class}
\title{The Neighbor Class}
\description{
For more details, please see the documentation in
\code{\link[SeuratObject:Neighbor]{SeuratObject}}
}
\seealso{
\code{\link[SeuratObject:Neighbor]{SeuratObject::Neighbor-class}}
}
