% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_generics.R, R/preprocessing.R
\name{NormalizeData}
\alias{NormalizeData}
\alias{NormalizeData.seurat}
\alias{NormalizeData.loom}
\title{Normalize Assay Data}
\usage{
NormalizeData(object, ...)

\method{NormalizeData}{seurat}(object, assay.type = "RNA",
  normalization.method = "LogNormalize", scale.factor = 10000,
  display.progress = TRUE)

\method{NormalizeData}{loom}(object, scale.factor = 10000,
  chunk.size = 1000, name = "norm_data", dataset.use = "matrix",
  display.progress = TRUE, overwrite = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{assay.type}{Type of assay to normalize for (default is RNA), but can be
changed for multimodal analyses.}

\item{normalization.method}{Method for normalization. Default is
log-normalization (LogNormalize). More methods to be added very shortly.}

\item{scale.factor}{Sets the scale factor for cell-level normalization}

\item{display.progress}{Display progress bar for scaling procedure}

\item{chunk.size}{Chunk size to iterate over}

\item{name}{Basename to store results matrix in 'layers'}

\item{dataset.use}{Dataset to normalize, defaults to 'matrix'}

\item{overwrite}{Overwrite existing dataset with name 'layers/\code{name}'}
}
\value{
Returns object after normalization. Normalized data is stored in data
or scale.data slot, depending on the method
}
\description{
Normalize data for a given assay
}
\section{Methods (by class)}{
\itemize{
\item \code{seurat}: Normalize data in a Seurat object

\item \code{loom}: Normalize data in a loom file
}}

\examples{
pbmc_small
pmbc_small <- NormalizeData(object = pbmc_small)

}
