% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R, R/preprocessing_generics.R
\name{NormalizeData.Assay}
\alias{NormalizeData.Assay}
\alias{NormalizeData.Seurat}
\alias{NormalizeData}
\title{Normalize Assay Data}
\usage{
\method{NormalizeData}{Assay}(object, normalization.method = "LogNormalize",
  scale.factor = 10000, display.progress = TRUE, ...)

\method{NormalizeData}{Seurat}(object, assay.use = NULL,
  normalization.method = "LogNormalize", scale.factor = 10000,
  display.progress = TRUE, ...)

NormalizeData(object, normalization.method, scale.factor, display.progress, ...)
}
\arguments{
\item{object}{An object}

\item{normalization.method}{Method for normalization. Default is
log-normalization (LogNormalize). More methods to be added very shortly.}

\item{scale.factor}{Sets the scale factor for cell-level normalization}

\item{display.progress}{display progress bar for scaling procedure.}

\item{...}{Arguments passed to other methods}

\item{assay.use}{Name of assay to use}
}
\value{
Returns object after normalization
}
\description{
Normalize Assay Data
}
\section{Methods (by class)}{
\itemize{
\item \code{Assay}: Normalize data in an Assay object

\item \code{Seurat}: Normalize data in a Seurat object
}}

\examples{
\dontrun{
pbmc_small
pmbc_small <- NormalizeData(object = pbmc_small)
}

}
