% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_modal.R
\name{NormalizeData}
\alias{NormalizeData}
\title{Normalize Assay Data}
\usage{
NormalizeData(object, assay.type = "RNA",
  normalization.method = "LogNormalize", slot = "data",
  scale.factor = 10000, display.progress = TRUE, ...)
}
\arguments{
\item{object}{Seurat object}

\item{assay.type}{Type of assay to normalize for (default is RNA), but can be changed for multimodal analyses.}

\item{normalization.method}{Method for normalization. Default is log-normalization (LogNormalize). Other options include CLR (CLR), regularized NB normalization (NBReg; RNA only)}
}
\value{
Returns object after normalization. Normalized data is stored in data or scale.data slot, depending on the method
}
\description{
Normalize data for a given assay
}
