% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/preprocessing.R
\name{NormalizeData}
\alias{NormalizeData}
\alias{NormalizeData.default}
\alias{NormalizeData.Assay}
\alias{NormalizeData.Seurat}
\title{Normalize Data}
\usage{
NormalizeData(object, ...)

\method{NormalizeData}{default}(object,
  normalization.method = "LogNormalize", scale.factor = 10000,
  verbose = TRUE, across = "features", ...)

\method{NormalizeData}{Assay}(object,
  normalization.method = "LogNormalize", scale.factor = 10000,
  verbose = TRUE, across = "features", ...)

\method{NormalizeData}{Seurat}(object, assay = NULL,
  normalization.method = "LogNormalize", scale.factor = 10000,
  verbose = TRUE, across = "features", workflow.name = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{normalization.method}{Method for normalization.
 \itemize{
  \item{LogNormalize: }{Feature counts for each cell are divided by the total
  counts for that cell and multiplied by the scale.factor. This is then
  natural-log transformed using log1p.}
  \item{CLR: }{Applies a centered log ratio transformation}
}
More methods to be added.}

\item{scale.factor}{Sets the scale factor for cell-level normalization}

\item{verbose}{display progress bar for normalization procedure}

\item{across}{If performing CLR normalization, normalize across either "features" or "cells".}

\item{assay}{Name of assay to use}

\item{workflow.name}{Name of workflow}
}
\value{
Returns object after normalization
}
\description{
Normalize the count data present in a given assay.
}
\examples{
\dontrun{
pbmc_small
pmbc_small <- NormalizeData(object = pbmc_small)
}

}
