% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seurat.R
\name{pca.plot}
\alias{pca.plot}
\title{Plot PCA map}
\usage{
pca.plot(object, ...)
}
\arguments{
\item{object}{Seurat object}

\item{do.label}{FALSE by default. If TRUE, plots an alternate view where the center of each
cluster is lebeled}

\item{label.pt.size}{If do.label is set, the point size}

\item{\dots}{Additional parameters to dim.plot, for example, which dimensions to plot.}
}
\description{
Graphs the output of a PCA analysis
Cells are colored by their identity class.
}
\details{
This function is a wrapper for dim.plot. See ?dim.plot for a full list of possible
arguments which can be passed in here.
}

