% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{PCA}
\alias{PCA}
\title{Run Principal Component Analysis on gene expression}
\usage{
PCA(object, pc.genes = NULL, do.print = TRUE, pcs.print = 5,
  pcs.store = 40, genes.print = 30, use.imputed = FALSE,
  rev.pca = FALSE, ...)
}
\arguments{
\item{object}{Seurat object}

\item{pc.genes}{Genes to use as input for PCA. Default is object@var.genes}

\item{do.print}{Print the top genes associated with high/low loadings for
the PCs}

\item{pcs.print}{Number of PCs to print genes for}

\item{pcs.store}{Number of PCs to store}

\item{genes.print}{Number of genes to print for each PC}

\item{use.imputed}{Run PCA on imputed values (FALSE by default)}

\item{rev.pca}{By default computes the PCA on the cell x gene matrix. Setting to true will 
compute it on gene x cell matrix.}

\item{\dots}{Additional arguments to be passed to prcomp}
}
\value{
Returns Seurat object with an PCA embedding (object@pca.rot) and
gene projection matrix (object@pca.x). The PCA object itself is stored in
object@pca.obj[[1]]
}
\description{
Run prcomp for PCA dimensionality reduction
}

