% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PCAPlot}
\alias{PCAPlot}
\title{Plot PCA map}
\usage{
PCAPlot(object, ...)
}
\arguments{
\item{object}{Seurat object}

\item{\dots}{Additional parameters to DimPlot, for example, which dimensions to plot.}
}
\description{
Graphs the output of a PCA analysis
Cells are colored by their identity class.
}
\details{
This function is a wrapper for DimPlot. See ?DimPlot for a full list of possible
arguments which can be passed in here.
}
