% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{PCAX}
\alias{PCAX}
\title{PCA Gene Loadings Accessor Function}
\usage{
PCAX(object, dims.use = NULL, genes.use = NULL, use.full = F)
}
\arguments{
\item{object}{Seurat object}

\item{dims.use, }{Dimensions to include (default is all stored dims)}

\item{genes.use, }{Genes to include (default is all)}
}
\value{
PCA loading matrix for given genes and PCs
}
\description{
Pull PCA Gene Loadings
}

