% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seurat.R
\name{plotClusterTree}
\alias{plotClusterTree}
\title{Plot phylogenetic tree}
\usage{
plotClusterTree(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{
Plots dendogram (must be precomputed using buildClusterTree), returns no value
}
\description{
Plots previously computed phylogenetic tree (from buildClusterTree)
}

