% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PlotClusterTree}
\alias{PlotClusterTree}
\title{Plot clusters as a tree}
\usage{
PlotClusterTree(object, ...)
}
\arguments{
\item{object}{Seurat object}

\item{\dots}{Additional arguments to ape::plot.phylo}
}
\value{
Plots dendogram (must be precomputed using BuildClusterTree), returns no value
}
\description{
Plots previously computed tree (from BuildClusterTree)
}
\examples{
pbmc_small <- BuildClusterTree(object = pbmc_small)
PlotClusterTree(object = pbmc_small)

}
