% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PlotClusterTree}
\alias{PlotClusterTree}
\title{Plot clusters as a tree}
\usage{
PlotClusterTree(object, direction = "downwards", ...)
}
\arguments{
\item{object}{Seurat object}

\item{direction}{A character string specifying the direction of the tree (default is downwards)
Possible options: "rightwards", "leftwards", "upwards", and "downwards".}

\item{\dots}{Additional arguments to
\code{\link[ape:plot.phylo]{ape::plot.phylo}}}
}
\value{
Plots dendogram (must be precomputed using BuildClusterTree), returns no value
}
\description{
Plots previously computed tree (from BuildClusterTree)
}
\examples{
if (requireNamespace("ape", quietly = TRUE)) {
  data("pbmc_small")
  pbmc_small <- BuildClusterTree(object = pbmc_small)
  PlotClusterTree(object = pbmc_small)
}
}
\concept{visualization}
