% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Print}
\alias{Print}
\alias{Print.DimReduc}
\title{Print the results of a dimensional reduction analysis}
\usage{
Print(object, ...)

\method{Print}{DimReduc}(object, dims = 1:5, nfeatures = 20,
  projected = FALSE, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{dims}{Number of dimensions to display}

\item{nfeatures}{Number of genes to display}

\item{projected}{Use projected slot}
}
\value{
Set of features defining the components
}
\description{
Prints a set of genes that most strongly define a set of components
}
