% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketching.R
\name{ProjectData}
\alias{ProjectData}
\title{Project full data to the sketch assay}
\usage{
ProjectData(
  object,
  assay = "RNA",
  sketched.assay = "sketch",
  sketched.reduction,
  full.reduction,
  dims,
  normalization.method = c("LogNormalize", "SCT"),
  refdata = NULL,
  k.weight = 50,
  umap.model = NULL,
  recompute.neighbors = FALSE,
  recompute.weights = FALSE,
  verbose = TRUE
)
}
\description{
Project full data to the sketch assay
}
