% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_generics.R, R/interaction.R
\docType{methods}
\name{ProjectSeurat}
\alias{ProjectSeurat}
\alias{ProjectSeurat,seurat-method}
\title{Add new cells to seurat object by projection}
\usage{
ProjectSeurat(object, ...)

\S4method{ProjectSeurat}{seurat}(object, umi.mat, return.type = "seurat",
  filename, overwrite = FALSE, display.progress = TRUE, ...)
}
\arguments{
\item{object}{An object to downsample}

\item{umi.mat}{A matrix of UMI counts with rows as genes and columns as cells}

\item{return.type}{Return as either a "seurat" or "loom" object}

\item{filename}{file name/path for new loom object if returning loom}

\item{overwrite}{overwrite loom file if it already exists}

\item{display.progress}{display progress of the process}
}
\description{
Add new cells to seurat object by projection
}
