% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_generics.R, R/interaction.R
\name{ProjectSeurat}
\alias{ProjectSeurat}
\alias{ProjectSeurat.Matrix}
\alias{ProjectSeurat.loom}
\title{Add new cells to seurat object by projection}
\usage{
ProjectSeurat(object, ...)

\method{ProjectSeurat}{Matrix}(object, template, is.scaled.data = FALSE,
  cluster.label = "res.0.8", tsne.proj.k = 11, class.k = 11,
  return.type = "seurat", filename, display.progress = TRUE, ...)

\method{ProjectSeurat}{loom}(object, template, do.norm.and.scale = TRUE,
  gene.names = "row_attrs/gene_names", cell.names = "col_attrs/cell_names",
  norm.data = "layers/norm_data", scale.data = "layers/scale_data",
  chunk.size = 1000, overwrite = FALSE, tsne.k = 31,
  tsne.rbf.gamma = 100, cluster.k = 11, cluster.res = "res.0.8",
  display.progress = TRUE, ...)
}
\arguments{
\item{object}{An object to project back to}

\item{...}{...}

\item{template}{...}

\item{return.type}{Return as either a "seurat" or "loom" object}

\item{filename}{file name/path for new loom object if returning loom}

\item{display.progress}{display progress of the process}

\item{gene.names}{Path to gene names dataset in loom file}

\item{cell.names}{Path to cell names dataset in loom file}

\item{norm.data}{Path to normalized dataset in loom file}

\item{scale.data}{Path to scaled dataset in loom file}

\item{chunk.size}{...}

\item{overwrite}{overwrite loom file if it already exists}

\item{tsne.k}{...}

\item{tsne.rbf.gamma}{...}

\item{cluster.k}{...}

\item{cluster.res}{...}
}
\description{
Add new cells to seurat object by projection
}
\section{Methods (by class)}{
\itemize{
\item \code{Matrix}: ...

\item \code{loom}: Project to a loom-based dataset
}}

