% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{PseudobulkExpression}
\alias{PseudobulkExpression}
\title{Pseudobulk Expression}
\usage{
PseudobulkExpression(object, ...)
}
\arguments{
\item{object}{An assay}

\item{...}{Arguments passed to other methods}
}
\value{
Returns object after normalization
}
\description{
Normalize the count data present in a given assay.
}
