% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utilities.R
\name{PurpleAndYellow}
\alias{PurpleAndYellow}
\title{A purple and yellow color palette}
\usage{
PurpleAndYellow(...)
}
\arguments{
\item{...}{Extra parameters to CustomPalette}
}
\value{
A color palette
}
\description{
A purple and yellow color palette
}
\examples{
df <- data.frame(x = rnorm(n = 100, mean = 20, sd = 2), y = rbinom(n = 100, size = 100, prob = 0.2))
plot(df, col = BlackAndWhite())

}
\seealso{
\code{CustomPalette}
}
