% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Radius}
\alias{Radius}
\alias{Radius.SlideSeq}
\alias{Radius.STARmap}
\alias{Radius.VisiumV1}
\title{Get the spot radius from an image}
\usage{
Radius(object)

\method{Radius}{SlideSeq}(object)

\method{Radius}{STARmap}(object)

\method{Radius}{VisiumV1}(object)
}
\arguments{
\item{object}{An image object}
}
\value{
The radius size
}
\description{
Get the spot radius from an image
}
