% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Radius.SlideSeq}
\alias{Radius.SlideSeq}
\alias{Radius.STARmap}
\alias{Radius.VisiumV1}
\title{Get Spot Radius}
\usage{
\method{Radius}{SlideSeq}(object)

\method{Radius}{STARmap}(object)

\method{Radius}{VisiumV1}(object)
}
\arguments{
\item{object}{An image object}
}
\description{
Get Spot Radius
}
\seealso{
\code{\link[SeuratObject:Radius]{SeuratObject::Radius}}
}
\concept{objects}
\concept{spatial}
