% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{Read10X_HD_GeoJson}
\alias{Read10X_HD_GeoJson}
\title{Load 10X Genomics GeoJson}
\usage{
Read10X_HD_GeoJson(data.dir, segmentation.type = "cell")
}
\arguments{
\item{data.dir}{Path to the directory containing matrix data}

\item{segmentation.type}{Which segmentations to load, cell or nucleus. If using nucleus the full matrix from cells is still used}
}
\value{
A FOV
}
\description{
Load 10X Genomics GeoJson
}
\concept{preprocessing}
