% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{Read10X_Image}
\alias{Read10X_Image}
\title{Load a 10X Genomics Visium Image}
\usage{
Read10X_Image(data.dir, filter.matrix = TRUE, ...)
}
\arguments{
\item{data.dir}{Directory containing the matrix.mtx, genes.tsv (or features.tsv), and barcodes.tsv
files provided by 10X. A vector or named vector can be given in order to load
several data directories. If a named vector is given, the cell barcode names
will be prefixed with the name.}

\item{...}{Ignored for now}
}
\value{
A \code{\link{SliceImage}} object
}
\description{
Load a 10X Genomics Visium Image
}
\seealso{
\code{\link{SliceImage}} \code{\link{Load10X_Spatial}}
}
