% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{Read10X_Image}
\alias{Read10X_Image}
\title{Load a 10X Genomics Visium Image}
\usage{
Read10X_Image(
  image.dir,
  image.scale = "lowres",
  assay = "Spatial",
  slice = "slice1",
  filter.matrix = TRUE
)
}
\arguments{
\item{image.dir}{Path to directory with 10X Genomics visium image data;
should include files \code{tissue_lowres_image.png},
\code{scalefactors_json.json} and \code{tissue_positions_list.csv}}

\item{image.scale}{Indicates which \code{tissue_{"hires"/"lowres"}_image.png}
file to read in}

\item{assay}{Name of associated assay}

\item{filter.matrix}{Filter spot/feature matrix to only include spots that
have been determined to be over tissue}
}
\value{
A \code{\link{VisiumV1}} object
}
\description{
Load a 10X Genomics Visium Image
}
\seealso{
\code{\link{VisiumV1}} \code{\link{Load10X_Spatial}}
}
\concept{preprocessing}
