% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ReadSlideSeq}
\alias{ReadSlideSeq}
\title{Load Slide-seq spatial data}
\usage{
ReadSlideSeq(coord.file, assay = "Spatial")
}
\arguments{
\item{coord.file}{Path to csv file containing bead coordinate positions}

\item{assay}{Name of assay to associate image to}
}
\value{
A \code{\link{SlideSeq}} object
}
\description{
Load Slide-seq spatial data
}
\seealso{
\code{\link{SlideSeq}}
}
\concept{preprocessing}
