% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{ReadSTARsolo}
\alias{ReadSTARsolo}
\title{Read output from STARsolo}
\usage{
ReadSTARsolo(data.dir, ...)
}
\arguments{
\item{data.dir}{Directory containing the data files}

\item{...}{Extra parameters passed to \code{\link{ReadMtx}}}
}
\description{
Read output from STARsolo
}
\concept{convenience}
