% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{Reductions}
\alias{Reductions}
\title{List DimReduc objects}
\usage{
Reductions(object)
}
\arguments{
\item{object}{A Seurat object}
}
\value{
Returns a character vector
}
\description{
Lists the names of \code{\link{DimReduc}} objects present in 
a Seurat object.
}
\examples{
Reductions(object = pbmc_small)
}
