% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{RegressOut}
\alias{RegressOut}
\title{Regress out technical effects and cell cycle}
\usage{
RegressOut(object, latent.vars, genes.regress = NULL, model.use = "linear",
  use.umi = F, do.scale = T, do.center = T, scale.max = 10)
}
\arguments{
\item{object}{Seurat object}

\item{latent.vars}{effects to regress out}

\item{genes.regress}{gene to run regression for (default is all genes)}

\item{model.use}{Use a linear model or generalized linear model (poisson, negative binomial) for the regression. Options are 'linear' (default), 'poisson', and 'negbinom'}

\item{use.umi}{Regress on UMI count data. Default is FALSE for linear modeling, but automatically set to TRUE if model.use is 'negbinom' or 'poisson'}

\item{do.scale}{Whether to scale the data.}

\item{do.center}{Whether to center the data.}

\item{scale.max}{Max value to accept for scaled data. The default is 10. Setting this can help
reduce the effects of genes that are only expressed in a very small number of cells.}
}
\value{
Returns Seurat object with the scale.data (object@scale.data) genes returning the residuals from the regression model
}
\description{
Remove unwanted effects from scale.data
}
