% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{RegressRegNB}
\alias{RegressRegNB}
\title{Use regularized negative binomial regression to normalize UMI count data}
\usage{
RegressRegNB(object, assay = NULL, do.correct.umi = FALSE,
  variable.features.zscore = 1, variable.features.n = NULL,
  do.scale = FALSE, do.center = FALSE,
  scale.max = .Machine$double.xmax, verbose = TRUE, ...)
}
\arguments{
\item{object}{A seurat object}

\item{assay}{Name of assay to use}

\item{do.correct.umi}{Place corrected UMI matrix in assay data slot}

\item{variable.features.zscore}{Z-score threshold for calling features highly 
variable; z-scores are based on variances of regression model pearson 
residuals of all features}

\item{variable.features.n}{Use this many features as variable features after 
ranking by variance}

\item{do.scale}{Whether to scale residuals to have unit variance}

\item{do.center}{Whether to center residuals to have mean zero}

\item{scale.max}{Max value after scaling and/or centering}

\item{verbose}{Whether to print messages and progress bars}

\item{...}{Additional parameters passed to \code{sctransform::vst}}
}
\description{
This function calls sctransform::vst. The sctransform package is available at
https://github.com/ChristophH/sctransform.
Use this function as an alternative to the NormalizeData, 
FindVariableFeatures, ScaleData workflow. Results are saved in the assay's 
data and scale.data slot, and sctransform::vst intermediate results are saved 
in misc slot of seurat object.
}
