% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{RegressRegNB}
\alias{RegressRegNB}
\title{Use regularized negative binomial regression to normalize UMI count data}
\usage{
RegressRegNB(object, assay = NULL, do.correct.umi = FALSE,
  variable.features.rv.th = 1.3, variable.features.n = NULL,
  return.dev.residuals = FALSE, clip.range = c(-10, 10),
  do.scale = FALSE, do.center = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{object}{A seurat object}

\item{assay}{Name of assay to use}

\item{do.correct.umi}{Place corrected UMI matrix in assay data slot}

\item{variable.features.rv.th}{Features with residual variance greater or equal
this value will be selected as variable features; default is 1.3}

\item{variable.features.n}{Use this many features as variable features after
ranking by residual variance}

\item{return.dev.residuals}{Place deviance residuals instead of Pearson residuals in scale.data slot; default is FALSE}

\item{clip.range}{Range to clip the residuals to; default is \code{c(-10, 10)}}

\item{do.scale}{Whether to scale residuals to have unit variance; default is FALSE}

\item{do.center}{Whether to center residuals to have mean zero; default is TRUE}

\item{verbose}{Whether to print messages and progress bars}

\item{...}{Additional parameters passed to \code{sctransform::vst}}
}
\description{
This function calls sctransform::vst. The sctransform package is available at
https://github.com/ChristophH/sctransform.
Use this function as an alternative to the NormalizeData,
FindVariableFeatures, ScaleData workflow. Results are saved in the assay's
data and scale.data slot, and sctransform::vst intermediate results are saved
in misc slot of seurat object.
}
