% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{RenameCells}
\alias{RenameCells}
\title{Rename cells}
\usage{
RenameCells(object, add.cell.id)
}
\arguments{
\item{object}{Seurat object}

\item{add.cell.id}{prefix to add cell names}
}
\value{
Seurat object with new cell names
}
\description{
Change the cell names in all the different parts of a Seurat object. Can
be useful before combining multiple object.
}
\examples{
head(pbmc_small@cell.names)
pbmc_small <- RenameCells(pbmc_small, add.cell.id = "Test")
head(pbmc_small@cell.names)

}
