% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{RenameCells}
\alias{RenameCells}
\alias{RenameCells.Assay}
\alias{RenameCells.DimReduc}
\alias{RenameCells.Seurat}
\title{Rename cells}
\usage{
RenameCells(object, ...)

\method{RenameCells}{Assay}(object, new.names = NULL, ...)

\method{RenameCells}{DimReduc}(object, new.names = NULL, ...)

\method{RenameCells}{Seurat}(object, add.cell.id = NULL,
  new.names = NULL, for.merge = FALSE, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{new.names}{vector of new cell names}

\item{add.cell.id}{prefix to add cell names}

\item{for.merge}{Only rename slots needed for merging Seurat objects.
Currently only renames the raw.data and meta.data slots.}
}
\value{
An object with new cell names
}
\description{
Change the cell names in all the different parts of an object. Can
be useful before combining multiple objects.
}
\details{
If \code{add.cell.id} is set a prefix is added to existing cell names. If
\code{new.names} is set these will be used to replace existing names.
}
\examples{
head(x = colnames(x = pbmc_small))
pbmc_small <- RenameCells(pbmc_small, add.cell.id = "Test")
head(x = colnames(x = pbmc_small))

}
