% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{RenameCells.SCTAssay}
\alias{RenameCells.SCTAssay}
\alias{RenameCells.SlideSeq}
\alias{RenameCells.STARmap}
\alias{RenameCells.VisiumV1}
\title{Rename Cells in an Object}
\usage{
\method{RenameCells}{SCTAssay}(object, new.names = NULL, ...)

\method{RenameCells}{SlideSeq}(object, new.names = NULL, ...)

\method{RenameCells}{STARmap}(object, new.names = NULL, ...)

\method{RenameCells}{VisiumV1}(object, new.names = NULL, ...)
}
\description{
Rename Cells in an Object
}
\seealso{
\code{\link[SeuratObject:RenameCells]{SeuratObject::RenameCells}}
}
\concept{objects}
