% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{RidgePlot}
\alias{RidgePlot}
\title{Single cell ridge plot}
\usage{
RidgePlot(object, features.plot, ident.include = NULL, num.col = NULL,
  do.sort = FALSE, y.max = NULL, same.y.lims = FALSE,
  cols.use = NULL, group.by = NULL, y.log = FALSE,
  combine.plots = TRUE, plot.counts = FALSE, ...)
}
\arguments{
\item{object}{Seurat object}

\item{features.plot}{Features to plot (gene expression, metrics, PC scores,
anything that can be retreived by FetchData)}

\item{ident.include}{Which classes to include in the plot (default is all)}

\item{num.col}{Number of columns if multiple plots are displayed}

\item{do.sort}{Sort identity classes (on the x-axis) by the average
expression of the attribute being potted}

\item{y.max}{Maximum y axis value}

\item{same.y.lims}{Set all the y-axis limits to the same values}

\item{cols.use}{Colors to use for plotting}

\item{group.by}{Group (color) cells in different ways (for example, orig.ident)}

\item{y.log}{plot Y axis on log scale}

\item{combine.plots}{Combine plots into a single gg object; note that if TRUE; themeing will not work when plotting multiple features}

\item{plot.counts}{Use non-normalized counts data for plotting}

\item{\dots}{additional parameters to pass to FetchData (for example, use.imputed, use.scaled, use.raw)}
}
\value{
A ggplot object
}
\description{
Draws a ridge plot of single cell data (gene expression, metrics, PC
scores, etc.)
}
\examples{
RidgePlot(object = pbmc_small, features.plot = 'PC1')

}
