% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{RidgePlot}
\alias{RidgePlot}
\title{Single cell ridge plot}
\usage{
RidgePlot(object, features, cols = NULL, idents = NULL, sort = FALSE,
  group.by = NULL, y.max = NULL, same.y.lims = FALSE, log = FALSE,
  ncol = NULL, combine = TRUE, slot = "data", ...)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Features to plot (gene expression, metrics, PC scores,
anything that can be retreived by FetchData)}

\item{cols}{Colors to use for plotting}

\item{idents}{Which classes to include in the plot (default is all)}

\item{sort}{Sort identity classes (on the x-axis) by the average
expression of the attribute being potted}

\item{group.by}{Group (color) cells in different ways (for example, orig.ident)}

\item{y.max}{Maximum y axis value}

\item{same.y.lims}{Set all the y-axis limits to the same values}

\item{log}{plot Y axis on log scale}

\item{ncol}{Number of columns if multiple plots are displayed}

\item{combine}{Combine plots into a single gg object; note that if TRUE; themeing will not work when plotting multiple features}

\item{slot}{Use non-normalized counts data for plotting}

\item{\dots}{additional parameters to pass to FetchData (for example, use.imputed, use.scaled, use.raw)}
}
\value{
A ggplot object
}
\description{
Draws a ridge plot of single cell data (gene expression, metrics, PC
scores, etc.)
}
\examples{
RidgePlot(object = pbmc_small, features = 'PC1')

}
