% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/preprocessing.R
\name{RunALRA}
\alias{RunALRA}
\alias{RunALRA.Seurat}
\title{Run Adaptively-thresholded Low Rank Approximation (ALRA)}
\usage{
RunALRA(object, k, q, ...)

\method{RunALRA}{Seurat}(object, k = NULL, q = 10, assay = NULL,
  slot = "data", setDefaultAssay = TRUE, genes.use = NULL, K = 100,
  p.val.th = 1e-10, noise.start = 80, q.k = 2, k.only = FALSE,
  do.plot = FALSE)
}
\arguments{
\item{object}{a Seurat object}

\item{k}{The rank of the rank-k approximation. Set to 0 for automated choice of k.}

\item{q}{The number of additional power iterations in randomized SVD when
computing rank k approximation. By default, q=10.}

\item{p.val.th}{: The threshold for ''significance''}

\item{noise.start}{: Index for which all smaller singular values are considered noise}

\item{q.k}{: Number of additional power iterations}

\item{assay:}{Assay to use}

\item{slot:}{slot to use}

\item{setDefaultAssay:}{set to TRUE to set imputed results as default Assay}

\item{genes.use:}{genes to impute}

\item{K:}{Number of singular values to compute. Must be less than the smallest dimension of the matrix.}

\item{k.only:}{set to TRUE if only computes optimal k WITHOUT performing ALRA}

\item{do.plot:}{set to TRUE to plot d, diff, pvals and chosen k}
}
\description{
Runs ALRA, a method for imputation of dropped out values in scRNA-seq data.
Computes the k-rank approximation to A_norm and adjusts it according to the
error distribution learned from the negative values. Described in 
Linderman, G. C., Zhao, J., Kluger, Y. (2018). "Zero-preserving imputation
of scRNA-seq data using low rank approximation." (bioRxiv:138677)
}
\section{Methods (by class)}{
\itemize{
\item \code{Seurat}: Run ALRA
}}

