% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{RunDiffusion}
\alias{RunDiffusion}
\title{Run diffusion map}
\usage{
RunDiffusion(object, cells.use = NULL, dims.use = 1:5, genes.use = NULL,
  reduction.use = "pca", q.use = 0.01, max.dim = 2, scale.clip = 10,
  ...)
}
\arguments{
\item{object}{Seurat object}

\item{cells.use}{Which cells to analyze (default, all cells)}

\item{dims.use}{Which dimensions to use as input features}

\item{genes.use}{If set, run the tSNE on this subset of genes
(instead of running on a set of reduced dimensions). Not set (NULL) by default}

\item{reduction.use}{Which dimensional reduction (PCA or ICA) to use for the tSNE. Default is PCA}

\item{q.use}{Quantile to use}

\item{max.dim}{Max dimension to keep from diffusion calculation}

\item{scale.clip}{Max/min value for scaled data. Default is 3}

\item{...}{Additional arguments to the diffuse call}
}
\value{
Returns a Seurat object with a diffusion map
}
\description{
Run diffusion map
}
