% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{RunDiffusion}
\alias{RunDiffusion}
\title{Run t-distributed Stochastic Neighbor Embedding}
\usage{
RunDiffusion(object, cells.use = NULL, dims.use = 1:5, k.seed = 1,
  do.fast = FALSE, add.iter = 0, genes.use = NULL,
  reduction.use = "pca", dim_embed = 2, q.use = 0.01, max.dim = 2,
  scale.clip = 10, ...)
}
\arguments{
\item{object}{Seurat object}

\item{cells.use}{Which cells to analyze (default, all cells)}

\item{dims.use}{Which dimensions to use as input features}

\item{k.seed}{Random seed for the t-SNE}

\item{do.fast}{If TRUE, uses the Barnes-hut implementation, which runs
faster, but is less flexible}

\item{add.iter}{If an existing tSNE has already been computed, uses the
current tSNE to seed the algorithm and then adds additional iterations on top of this}

\item{genes.use}{If set, run the tSNE on this subset of genes
(instead of running on a set of reduced dimensions). Not set (NULL) by default}

\item{reduction.use}{Which dimensional reduction (PCA or ICA) to use for the tSNE. Default is PCA}

\item{dim_embed}{The dimensional space of the resulting tSNE embedding (default is 2).
For example, set to 3 for a 3d tSNE}

\item{q.use}{Quantile to use}

\item{max.dim}{Max dimension to keep from diffusion calculation}

\item{scale.clip}{Max/min value for scaled data. Default is 3}

\item{...}{Additional arguments to the tSNE call. Most commonly used is
perplexity (expected number of neighbors default is 30)}
}
\value{
Returns a Seurat object with a tSNE embedding in object@tsne_rot
}
\description{
Run t-SNE dimensionality reduction on selected features. Has the option of running in a reduced
dimensional space (i.e. spectral tSNE, recommended), or running based on a set of genes
}
