% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{RunMarkVario}
\alias{RunMarkVario}
\title{Run the mark variogram computation on a given position matrix and expression
matrix.}
\usage{
RunMarkVario(spatial.location, data, ...)
}
\arguments{
\item{spatial.location}{A 2 column matrix giving the spatial locations of
each of the data points also in data}

\item{data}{Matrix containing the data used as "marks" (e.g. gene expression)}

\item{...}{Arguments passed to markvario}
}
\description{
Wraps the functionality of markvario from the spatstat package.
}
\concept{preprocessing}
