% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{RunMoransI}
\alias{RunMoransI}
\title{Compute Moran's I value.}
\usage{
RunMoransI(data, pos, verbose = TRUE)
}
\arguments{
\item{data}{Expression matrix}

\item{pos}{Position matrix}

\item{verbose}{Display messages/progress}
}
\description{
Wraps the functionality of the Moran.I function from the ape package.
Weights are computed as 1/distance.
}
\concept{preprocessing}
