% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{RunMultiCCA}
\alias{RunMultiCCA}
\title{Perform Canonical Correlation Analysis with more than two groups}
\usage{
RunMultiCCA(object.list, genes.use, niter = 25, num.ccs = 1,
  standardize = TRUE)
}
\arguments{
\item{object.list}{List of Seurat objects}

\item{genes.use}{Genes to use in mCCA.}

\item{niter}{Number of iterations to perform. Set by default to 25.}

\item{num.ccs}{Number of canonical vectors to calculate}

\item{standardize}{standardize scale.data matrices to be centered (mean zero)
and scaled to have a standard deviation of 1.}
}
\value{
Returns a combined Seurat object with the CCA stored in the @dr$cca slot.
}
\description{
Runs a canonical correlation analysis
}
