% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction_generics.R,
%   R/dimensional_reduction.R
\name{RunPCA}
\alias{RunPCA}
\alias{RunPCA.seurat}
\alias{RunPCA.loom}
\title{Run PCA}
\usage{
RunPCA(object, ...)

\method{RunPCA}{seurat}(object, pcs.compute = 20, pc.genes = NULL,
  use.imputed = FALSE, rev.pca = FALSE, weight.by.var = TRUE,
  do.print = TRUE, pcs.print = 1:5, genes.print = 30,
  reduction.name = "pca", reduction.key = "PC", assay.type = "RNA",
  seed.use = 42, ...)

\method{RunPCA}{loom}(object, pcs.compute = 75, pc.genes = NULL,
  online.pca = TRUE, covariance.mat = NULL, cells.initial = 50000,
  chunk.size = 1000, do.print = TRUE, pcs.print = 1:5,
  genes.print = 30, gene.names = "row_attrs/gene_names",
  gene.means = "row_attrs/gene_means",
  gene.variance = "row_attrs/gene_dispersion",
  var.genes = "row_attrs/var_genes", scale.data = "layers/scale_data",
  cell.names = "col_attrs/cell_names", ngene = NULL,
  overwrite = FALSE, display.progress = TRUE, ...)
}
\arguments{
\item{object}{An object to run PCA on}

\item{\dots}{Additional arguments to be passed to IRLBA}

\item{pcs.compute}{Total Number of PCs to compute and store (20 by default)}

\item{pc.genes}{Genes to use as input for PCA. Default is object@var.genes}

\item{use.imputed}{Run PCA on imputed values (FALSE by default)}

\item{rev.pca}{By default computes the PCA on the cell x gene matrix. Setting
to true will compute it on gene x cell matrix.}

\item{weight.by.var}{Weight the cell embeddings by the variance of each PC
(weights the gene loadings if rev.pca is TRUE)}

\item{do.print}{Print the top genes associated with high/low loadings for
the PCs}

\item{pcs.print}{PCs to print genes for}

\item{genes.print}{Number of genes to print for each PC}

\item{reduction.name}{dimensional reduction name, specifies the position in the object$dr list. pca by default}

\item{reduction.key}{dimensional reduction key, specifies the string before the number for the dimension names. PC by default}

\item{assay.type}{Data type, RNA by default. Can be changed for multimodal}

\item{seed.use}{Set a random seed. By default, sets the seed to 42. Setting
NULL will not set a seed.}

\item{online.pca}{Compute PCA using online methods}

\item{covariance.mat}{Covariance matrix to use if calculating a full PCA}

\item{cells.initial}{Number of cells to use for initial PCA calculate}

\item{chunk.size}{Number of cells to load into memory at any given time}

\item{gene.names}{Path to gene names dataset}

\item{gene.means}{Path to gene means dataset}

\item{gene.variance}{Path to gene variance dataset}

\item{scale.data}{Path to scaled data matrix}

\item{cell.names}{Path to cell names dataset}

\item{ngene}{Number of variable genes to use}

\item{overwrite}{Overwrite preexisting results}

\item{display.progress}{Show progress bars}
}
\value{
(Seurat) Returns Seurat object with the PCA calculation stored in
object@dr$pca.

(loom) Nothing, modifies loom object directly
}
\description{
Run a PCA dimensionality reduction. For details about stored PCA calculation
parameters, see \code{PrintPCAParams}.
}
\section{Methods (by class)}{
\itemize{
\item \code{seurat}: Run PCA on Seurat objects

\item \code{loom}: Run PCA on loom objects
}}

\examples{
pbmc_small
# Run PCA on variable genes (default)
pbmc_small <- RunPCA(pbmc_small)
# Run PCA on different gene set (in this case all genes)
pbmc_small=RunPCA(pbmc_small,pc.genes = rownames(pbmc_small@data))
# Run PCA but compute more than 20 dimensions
pbmc_small=RunPCA(pbmc_small,pcs.compute=30)
# Plot results
PCAPlot(pbmc_small)

}
