% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction_generics.R,
%   R/dimensional_reduction.R
\docType{methods}
\name{RunPCA}
\alias{RunPCA}
\alias{RunPCA,seurat-method}
\alias{RunPCA,loom-method}
\title{Run PCA}
\usage{
RunPCA(object, pcs.compute, ...)

\S4method{RunPCA}{seurat}(object, pc.genes = NULL, pcs.compute = 20,
  use.imputed = FALSE, rev.pca = FALSE, weight.by.var = TRUE,
  do.print = TRUE, pcs.print = 1:5, genes.print = 30,
  reduction.name = "pca", reduction.key = "PC", assay.type = "RNA", ...)

\S4method{RunPCA}{loom}(object, pcs.compute = 75, cells.initial = 50000,
  chunk.size = 1000, gene.names = "row_attrs/Gene",
  gene.means = "row_attrs/gene_means",
  gene.variance = "row_attrs/gene_dispersion",
  scale.data = "layers/scale_data", ngene = 1000, overwrite = FALSE,
  display.progress = TRUE, ...)
}
\arguments{
\item{object}{An object to run PCA on}

\item{pcs.compute}{Total Number of PCs to compute and store (20 by default)}

\item{...}{Additional arguments}

\item{pc.genes}{Genes to use as input for PCA}

\item{use.imputed}{Run PCA on imputed values (FALSE by default)}

\item{rev.pca}{By default computes the PCA on the cell x gene matrix. Setting
to true will compute it on gene x cell matrix.}

\item{weight.by.var}{Weight the cell embeddings by the variance of each PC
(weights the gene loadings if rev.pca is TRUE)}

\item{do.print}{Print the top genes associated with high/low loadings for
the PCs}

\item{pcs.print}{PCs to print genes for}

\item{genes.print}{Number of genes to print for each PC}

\item{reduction.name}{dimensional reduction name, specifies the position in the object$dr list. pca by default}

\item{reduction.key}{dimensional reduction key, specifies the string before the number for the dimension names. PC by default}

\item{assay.type}{Data type, RNA by default. Can be changed for multimodal}

\item{cells.initial}{Number of cells to use for initial PCA calculate}

\item{chunk.size}{Number of cells to load into memory at any given time}

\item{gene.names}{Path to gene names dataset}

\item{gene.means}{Path to gene means dataset}

\item{gene.variance}{Path to gene variance dataset}

\item{scale.data}{Path to scaled data matrix}

\item{ngene}{Number of variable genes to use}

\item{overwrite}{Overwrite preexisting results}

\item{display.progress}{Show progress bars}
}
\value{
Returns Seurat object with the PCA calculation stored in
object@dr$pca.
}
\description{
Run a PCA dimensionality reduction. For details about stored PCA calculation
parameters, see \code{PrintPCAParams}.
}
\examples{
pbmc_small
# Run PCA on variable genes (default)
pbmc_small <- RunPCA(pbmc_small)
# Run PCA on different gene set (in this case all genes)
pbmc_small=RunPCA(pbmc_small,pc.genes = rownames(pbmc_small@data))
# Run PCA but compute more than 20 dimensions
pbmc_small=RunPCA(pbmc_small,pcs.compute=30)
# Plot results
PCAPlot(pbmc_small)

}
