% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/dimensional_reduction.R
\name{RunPCA}
\alias{RunPCA}
\alias{RunPCA.default}
\alias{RunPCA.Assay}
\alias{RunPCA.Seurat}
\title{Run Principal Component Analysis on gene expression using IRLBA}
\usage{
RunPCA(object, ...)

\method{RunPCA}{default}(object, assay = NULL, npcs = 20,
  rev.pca = FALSE, weight.by.var = TRUE, verbose = TRUE,
  ndims.print = 1:5, nfeatures.print = 30, reduction.key = "PC",
  seed.use = 42, ...)

\method{RunPCA}{Assay}(object, assay = NULL, features = NULL,
  npcs = 20, rev.pca = FALSE, weight.by.var = TRUE, verbose = TRUE,
  ndims.print = 1:5, nfeatures.print = 30, reduction.key = "PC",
  seed.use = 42, ...)

\method{RunPCA}{Seurat}(object, assay = NULL, features = NULL,
  npcs = 20, rev.pca = FALSE, weight.by.var = TRUE, verbose = TRUE,
  ndims.print = 1:5, nfeatures.print = 30, reduction.name = "pca",
  reduction.key = "PC", seed.use = 42, workflow.name = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods and IRLBA}

\item{assay}{Name of Assay PCA is being run on}

\item{npcs}{Total Number of PCs to compute and store (20 by default)}

\item{rev.pca}{By default computes the PCA on the cell x gene matrix. Setting
to true will compute it on gene x cell matrix.}

\item{weight.by.var}{Weight the cell embeddings by the variance of each PC
(weights the gene loadings if rev.pca is TRUE)}

\item{verbose}{Print the top genes associated with high/low loadings for
the PCs}

\item{ndims.print}{PCs to print genes for}

\item{nfeatures.print}{Number of genes to print for each PC}

\item{reduction.key}{dimensional reduction key, specifies the string before
the number for the dimension names. PC by default}

\item{seed.use}{Set a random seed. By default, sets the seed to 42. Setting
NULL will not set a seed.}

\item{features}{Features to compute PCA on}

\item{reduction.name}{dimensional reduction name,  pca by default}

\item{workflow.name}{Name of workflow}
}
\value{
Returns Seurat object with the PCA calculation stored in the reductions slot
}
\description{
Run a PCA dimensionality reduction. For details about stored PCA calculation
parameters, see \code{PrintPCAParams}.
}
