% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{RunPCA}
\alias{RunPCA}
\title{Run Principal Component Analysis on gene expression using IRLBA}
\usage{
RunPCA(object, pc.genes = NULL, pcs.compute = 20, use.imputed = FALSE,
  rev.pca = FALSE, weight.by.var = TRUE, do.print = TRUE,
  pcs.print = 1:5, genes.print = 30, ...)
}
\arguments{
\item{object}{Seurat object}

\item{pc.genes}{Genes to use as input for PCA. Default is object@var.genes}

\item{pcs.compute}{Total Number of PCs to compute and store}

\item{use.imputed}{Run PCA on imputed values (FALSE by default)}

\item{rev.pca}{By default computes the PCA on the cell x gene matrix. Setting
to true will compute it on gene x cell matrix.}

\item{weight.by.var}{Weight the gene loadings by the variance of each PC
(weights the cell embeddings if rev.pca is TRUE)}

\item{do.print}{Print the top genes associated with high/low loadings for
the PCs}

\item{pcs.print}{PCs to print genes for}

\item{genes.print}{Number of genes to print for each PC}

\item{\dots}{Additional arguments to be passed to IRLBA}
}
\value{
Returns Seurat object with the PCA calculation stored in
object@dr$pca.
}
\description{
Run a PCA dimensionality reduction. For details about stored PCA calculation
parameters, see \code{\link{PrintPCAParams}}.
}
