% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R,
%   R/dimensional_reduction_generics.R
\name{RunPCA.Assay}
\alias{RunPCA.Assay}
\alias{RunPCA.Seurat}
\alias{RunPCA}
\title{Run Principal Component Analysis on gene expression using IRLBA}
\usage{
\method{RunPCA}{Assay}(object, features.use = NULL, pcs.compute = 20,
  rev.pca = FALSE, weight.by.var = TRUE, verbose = TRUE,
  pcs.print = 1:5, features.print = 30, reduction.name = "pca",
  reduction.key = "PC", seed.use = 42, ...)

\method{RunPCA}{Seurat}(object, assay.use = NULL, features.use = NULL,
  pcs.compute = 20, rev.pca = FALSE, weight.by.var = TRUE,
  verbose = TRUE, pcs.print = 1:5, features.print = 30,
  reduction.name = "pca", reduction.key = "PC", seed.use = 42, ...)

RunPCA(object, features.use, pcs.compute, rev.pca, weight.by.var, verbose,
  pcs.print, features.print, reduction.name, reduction.key, seed.use, ...)
}
\arguments{
\item{object}{Seurat object}

\item{features.use}{Features to use as input for PCA. Defaults to variable features}

\item{pcs.compute}{Total Number of PCs to compute and store (20 by default)}

\item{rev.pca}{By default computes the PCA on the cell x gene matrix. Setting
to true will compute it on gene x cell matrix.}

\item{weight.by.var}{Weight the cell embeddings by the variance of each PC
(weights the gene loadings if rev.pca is TRUE)}

\item{verbose}{Print the top genes associated with high/low loadings for
the PCs}

\item{pcs.print}{PCs to print genes for}

\item{features.print}{Number of genes to print for each PC}

\item{reduction.name}{dimensional reduction name,  pca by default}

\item{reduction.key}{dimensional reduction key, specifies the string before
the number for the dimension names. PC by default}

\item{seed.use}{Set a random seed. By default, sets the seed to 42. Setting
NULL will not set a seed.}

\item{\dots}{Additional arguments to be passed to IRLBA}

\item{assay.use}{Name of Assay to scale}
}
\value{
Returns Seurat object with the PCA calculation stored in the reductions slot
}
\description{
Run a PCA dimensionality reduction. For details about stored PCA calculation
parameters, see \code{PrintPCAParams}.
}
\section{Methods (by class)}{
\itemize{
\item \code{Assay}: Run a PCA on an Assay object

\item \code{Seurat}: Run a PCA on a Seurat object
}}

