% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_determination.R
\name{SaveClusters}
\alias{SaveClusters}
\title{Save cluster assignments to a TSV file}
\usage{
SaveClusters(object, file)
}
\arguments{
\item{object}{Seurat object with cluster assignments}

\item{file}{Path to file to write cluster assignments to}
}
\value{
No return value. Writes clusters assignments to specified file.
}
\description{
Save cluster assignments to a TSV file
}
