% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{ScaleData}
\alias{ScaleData}
\title{Scale and center the data}
\usage{
ScaleData(object, genes.use = NULL, data.use = NULL, do.scale = TRUE,
  do.center = TRUE, scale.max = 10)
}
\arguments{
\item{object}{Seurat object}

\item{genes.use}{Vector of gene names to scale/center. Default is all genes in object@data.}

\item{data.use}{Can optionally pass a matrix of data to scale, default is object@data[genes.use,]}

\item{do.scale}{Whether to scale the data.}

\item{do.center}{Whether to center the data.}

\item{scale.max}{Max value to accept for scaled data. The default is 10. Setting this can help 
reduce the effects of genes that are only expressed in a very small number of cells.}
}
\value{
Returns a seurat object with object@scale.data updated with scaled and/or centered data.
}
\description{
Scale and center the data
}
