% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaleData_fast.R
\name{ScaleData_fast}
\alias{ScaleData_fast}
\alias{ScaleData_fast.default}
\alias{ScaleData_fast.Assay}
\alias{ScaleData_fast.StdAssay}
\alias{ScaleData_fast.Seurat}
\title{Fast ScaleData implementation}
\usage{
ScaleData_fast(object, ...)

\method{ScaleData_fast}{default}(
  object,
  features = NULL,
  vars.to.regress = NULL,
  latent.data = NULL,
  split.by = NULL,
  model.use = "linear",
  use.umi = FALSE,
  do.scale = TRUE,
  do.center = TRUE,
  scale.max = 10,
  block.size = 1000,
  min.cells.to.block = 3000,
  nthreads = 1,
  verbose = TRUE,
  ...
)

\method{ScaleData_fast}{Assay}(
  object,
  features = NULL,
  vars.to.regress = NULL,
  latent.data = NULL,
  split.by = NULL,
  model.use = "linear",
  use.umi = FALSE,
  do.scale = TRUE,
  do.center = TRUE,
  scale.max = 10,
  block.size = 1000,
  min.cells.to.block = 3000,
  nthreads = 1,
  verbose = TRUE,
  ...
)

\method{ScaleData_fast}{StdAssay}(
  object,
  features = NULL,
  layer = "data",
  vars.to.regress = NULL,
  latent.data = NULL,
  by.layer = FALSE,
  split.by = NULL,
  model.use = "linear",
  use.umi = FALSE,
  do.scale = TRUE,
  do.center = TRUE,
  scale.max = 10,
  block.size = 1000,
  min.cells.to.block = 3000,
  nthreads = 1,
  save = "scale.data",
  verbose = TRUE,
  ...
)

\method{ScaleData_fast}{Seurat}(
  object,
  features = NULL,
  assay = NULL,
  vars.to.regress = NULL,
  split.by = NULL,
  model.use = "linear",
  use.umi = FALSE,
  do.scale = TRUE,
  do.center = TRUE,
  scale.max = 10,
  block.size = 1000,
  min.cells.to.block = 3000,
  nthreads = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{features}{Vector of features names to scale/center. Default is variable features.}

\item{vars.to.regress}{Variables to regress out (previously latent.vars in
RegressOut). For example, nUMI, or percent.mito.}

\item{latent.data}{Extra data to regress out, should be cells x latent data}

\item{split.by}{Name of variable in object metadata or a vector or factor defining
grouping of cells. See argument \code{f} in \code{\link[base]{split}} for more details}

\item{model.use}{Use a linear model or generalized linear model
(poisson, negative binomial) for the regression. Options are 'linear'
(default), 'poisson', and 'negbinom'}

\item{use.umi}{Regress on UMI count data. Default is FALSE for linear
modeling, but automatically set to TRUE if model.use is 'negbinom' or 'poisson'}

\item{do.scale}{Whether to scale the data.}

\item{do.center}{Whether to center the data.}

\item{scale.max}{Max value to return for scaled data. The default is 10.
Setting this can help reduce the effects of features that are only expressed in
a very small number of cells. If regressing out latent variables and using a
non-linear model, the default is 50.}

\item{block.size}{Default size for number of features to scale at in a single
computation. Increasing block.size may speed up calculations but at an
additional memory cost.}

\item{min.cells.to.block}{If object contains fewer than this number of cells,
don't block for scaling calculations.}

\item{nthreads}{Number of threads to use for the single-call fast path
(centering/scaling without regression, splitting, or multiple workers).
Values greater than 1 enable RcppParallel threading; the progress bar is
only shown when \code{nthreads = 1}.}

\item{verbose}{Displays a progress bar for scaling procedure}

\item{assay}{Name of Assay to scale}
}
\description{
Fast ScaleData implementation
}
\concept{preprocessing}
