% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{SCTAssay-class}
\alias{SCTAssay-class}
\alias{SCTAssay}
\title{The SCTAssay Class}
\description{
The SCTAssay object contains all the information found in an \code{\link{Assay}}
object, with extra information from the results of \code{\link{SCTransform}}
}
\section{Slots}{

\describe{
\item{\code{feature.attributes}}{A data.frame with the following columns:}

\item{\code{cell.attributes}}{A data.frame with the following columns:}

\item{\code{fitted.parameters}}{A matrix with the following columns:}

\item{\code{clips}}{...}

\item{\code{groups}}{...}
}}

\section{Creating an \code{SCTAssay} from an \code{Assay}}{

Conversion from an \code{Assay} object to an \code{SCTAssay} object by
is done by adding the additional slots to the object. If \code{from} has
results generated by \code{\link{SCTransform}} from Seurat v3.0.0 to v3.1.1,
the conversion will automagically fill the new slots with the data
}

\seealso{
\code{\link{Assay}}
}
