% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{SCTAssay-class}
\alias{SCTAssay-class}
\alias{SCTAssay}
\title{The SCTAssay Class}
\description{
The SCTAssay object contains all the information found in an \code{\link{Assay}}
object, with extra information from the results of \code{\link{SCTransform}}
}
\section{Slots}{

\describe{
\item{\code{...}}{}
}}

