% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{SCTAssay-class}
\alias{SCTAssay-class}
\alias{SCTModel}
\alias{SCTAssay}
\alias{levels.SCTAssay}
\alias{levels<-.SCTAssay}
\title{The SCTModel Class}
\usage{
\method{levels}{SCTAssay}(x)

\method{levels}{SCTAssay}(x) <- value
}
\arguments{
\item{x}{An \code{SCTAssay} object}

\item{value}{New levels, must be in the same order as the levels present}
}
\value{
\code{levels}: SCT grouping levels

\code{levels<-}: \code{x} with updated SCT groupings
}
\description{
The SCTAssay object contains all the information found in an \code{\link{Assay}}
object, with extra information from the results of \code{\link{SCTransform}}
}
\section{Slots}{

\describe{
\item{\code{feature.attributes}}{A data.frame with the following columns:}

\item{\code{cell.attributes}}{A data.frame with the following columns:}

\item{\code{clips}}{...}

\item{\code{umi.assay}}{...}

\item{\code{groups}}{...}

\item{\code{model}}{...}

\item{\code{arguments}}{...}

\item{\code{SCTModel.list}}{A data.frame with the following columns:}

\item{\code{groups}}{...}
}}

\section{Get and set SCT groupings}{

SCT results are grouped by initial run of \code{\link{SCTransform}} in order
to keep SCT parameters straight between runs. When working with merged
\code{SCTAssay} objects, these groups are important. \code{levels} allows the
querying of groups present. \code{levels<-} allows the changing of the groups
present, useful when merging \code{SCTAssay} objects. Note: unlike normal
\code{\link[base]{levels<-}}, \code{levels<-.SCTAssay} allows complete changing
of groups, not reordering.
}

\section{Creating an \code{SCTAssay} from an \code{Assay}}{

Conversion from an \code{Assay} object to an \code{SCTAssay} object by
is done by adding the additional slots to the object. If \code{from} has
results generated by \code{\link{SCTransform}} from Seurat v3.0.0 to v3.1.1,
the conversion will automagically fill the new slots with the data
}

\examples{
\dontrun{
# SCTAssay objects are generated from SCTransform
pbmc_small <- SCTransform(pbmc_small)
}

# SCTAssay objects are generated from SCTransform
pbmc_small <- SCTransform(pbmc_small)
pbmc_small[["SCT"]]

\dontrun{
# Query and change SCT groupings
levels(pbmc_small[['SCT']])
levels(pbmc_small[['SCT']]) <- '3'
levels(pbmc_small[['SCT']])
}

}
\seealso{
\code{\link{Assay}}

\code{\link{Assay}}
}
