% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{SCTransform}
\alias{SCTransform}
\title{Use regularized negative binomial regression to normalize UMI count data}
\usage{
SCTransform(object, assay = "RNA", do.correct.umi = TRUE,
  variable.features.n = 3000, variable.features.rv.th = 1.3,
  return.dev.residuals = FALSE, vars.to.regress = NULL,
  do.scale = FALSE, do.center = TRUE,
  clip.range = c(-sqrt(ncol(object[[assay]])/30),
  sqrt(ncol(object[[assay]])/30)), return.only.var.genes = FALSE,
  verbose = TRUE, ...)
}
\arguments{
\item{object}{A seurat object}

\item{assay}{Name of assay to pull the count data from; default is 'RNA'}

\item{do.correct.umi}{Place corrected UMI matrix in assay counts slot; default is TRUE}

\item{variable.features.n}{Use this many features as variable features after
ranking by residual variance; default is 3000}

\item{variable.features.rv.th}{Instead of setting a fixed number of variable features,
use this residual variance cutoff; this is only used when \code{variable.features.n}
is set to NULL; default is 1.3}

\item{return.dev.residuals}{Place deviance residuals instead of Pearson residuals in scale.data slot; default is FALSE}

\item{vars.to.regress}{Variables to regress out in a second non-regularized linear
regression. For example, percent.mito. Default is NULL}

\item{do.scale}{Whether to scale residuals to have unit variance; default is FALSE}

\item{do.center}{Whether to center residuals to have mean zero; default is TRUE}

\item{clip.range}{Range to clip the residuals to; default is \code{c(-sqrt(n/30), sqrt(n/30))},
where n is the number of cells}

\item{return.only.var.genes}{If set to TRUE all data matrices in output assay are
subset to contain only the variable genes; default is FALSE}

\item{verbose}{Whether to print messages and progress bars}

\item{...}{Additional parameters passed to \code{sctransform::vst}}
}
\description{
This function calls sctransform::vst. The sctransform package is available at
https://github.com/ChristophH/sctransform.
Use this function as an alternative to the NormalizeData,
FindVariableFeatures, ScaleData workflow. Results are saved in a new assay
called SCT with counts being (corrected) counts, data being log1p(counts),
scale.data being pearson residuals; sctransform::vst intermediate results are saved
in misc slot of new assay.
}
