% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_generics.R, R/seurat.R, R/assay.R
\name{SetAssayData}
\alias{SetAssayData}
\alias{SetAssayData.Seurat}
\alias{SetAssayData.Assay}
\title{Settter for multimodal data}
\usage{
SetAssayData(object, slot, new.data, ...)

\method{SetAssayData}{Seurat}(object, slot, new.data, assay.use, ...)

\method{SetAssayData}{Assay}(object, slot, new.data)
}
\arguments{
\item{object}{An object}

\item{slot}{Where to store the new data}

\item{new.data}{New data to insert}

\item{...}{Arguments passed to other methods}

\item{assay.use}{Name of assay whose data should be set}
}
\value{
object with the assay data set
}
\description{
Settter for multimodal data
}
\section{Methods (by class)}{
\itemize{
\item \code{Seurat}: Set assay data for an Assay object in a Seurat object

\item \code{Assay}: Set assay data for an Assay object
}}

