% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_modal.R
\name{SetAssayData}
\alias{SetAssayData}
\title{Assay Data Mutator Function}
\usage{
SetAssayData(object, assay.type, slot, new.data)
}
\arguments{
\item{object}{Seurat object}

\item{assay.type}{Type of assay to fetch data for (default is RNA)}

\item{slot}{Specific information to pull (i.e. raw.data, data, scale.data,...). Default is data}

\item{new.data}{New data to insert}
}
\value{
Seurat object with updated slot
}
\description{
Store information for specified assay, for multimodal analysis
}
