% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreduc_generics.R, R/dimreduc.R
\name{SetDimReduc}
\alias{SetDimReduc}
\alias{SetDimReduc.Seurat}
\title{Setter for DimReduc object}
\usage{
SetDimReduc(object, slot, new.data, ...)

\method{SetDimReduc}{Seurat}(object, slot, new.data)
}
\arguments{
\item{object}{A DimReduc object}

\item{slot}{Where to store the new data}

\item{new.data}{New data to insert}

\item{...}{Arguments passed to other methods}
}
\value{
object with the assay data set
}
\description{
Setter for DimReduc object
}
\section{Methods (by class)}{
\itemize{
\item \code{Seurat}: Set a slot for a given DimReduc
}}

