% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction_generics.R,
%   R/dimensional_reduction_utilities.R
\name{SetDimReduction}
\alias{SetDimReduction}
\alias{SetDimReduction.seurat}
\alias{SetDimReduction.loom}
\title{Dimensional Reduction Mutator Function}
\usage{
SetDimReduction(object, ...)

\method{SetDimReduction}{seurat}(object, reduction.type, slot, new.data,
  ...)

\method{SetDimReduction}{loom}(object, reduction.type, slot, new.data,
  overwrite = FALSE, ...)
}
\arguments{
\item{object}{Seurat object}

\item{reduction.type}{Type of dimensional reduction to set}

\item{slot}{Specific information to set (must be one of the following:
"cell.embeddings", "gene.loadings", "gene.loadings.full", "sdev", "key",
"misc")}

\item{new.data}{New data to set}

\item{overwrite}{Overwrite existing data?}
}
\value{
Seurat object with updated slot
}
\description{
Set information for specified stored dimensional reduction analysis
}
\section{Methods (by class)}{
\itemize{
\item \code{seurat}: Set a dimmensional reduction for a Seurat object

\item \code{loom}: Set a dimmensional reduction for a loom object
}}

\examples{
pbmc_small
# Simulate adding a new dimensional reduction
new.cell.embeddings <- GetCellEmbeddings(object = pbmc_small, reduction.type = "pca")
new.gene.loadings <- GetGeneLoadings(object = pbmc_small, reduction.type = "pca")
SetDimReduction(
  object = pbmc_small,
  reduction.type = "new.pca",
  slot = "cell.embeddings",
  new.data = new.cell.embeddings
)
SetDimReduction(
  object = pbmc_small,
  reduction.type = "new.pca",
  slot = "gene.loadings",
  new.data = new.gene.loadings
)

}
