% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{SetIntegrationData}
\alias{SetIntegrationData}
\title{Set integration data}
\usage{
SetIntegrationData(object, integration.name, slot, new.data)
}
\arguments{
\item{object}{Seurat object}

\item{integration.name}{Name of integration object}

\item{slot}{Which slot in integration object to set}

\item{new.data}{New data to insert}
}
\value{
Returns a \code{\link{Seurat}} object
}
\description{
Set integration data
}
\concept{objects}
