% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setThreads}
\alias{setThreads}
\title{Set the number of threads to use for parallel processing in Seurat}
\usage{
setThreads(n = NULL, verbose = TRUE)
}
\arguments{
\item{n}{Number of threads (>= 1) to use for processing. If NULL, will be set to (available cores / 2).}

\item{verbose}{Whether to print the number of threads set and available cores to the console}
}
\description{
Set the number of threads to use for parallel processing in Seurat
}
\details{
The default is one thread. Calling \code{setThreads()} without
specifying \code{n} selects half the available cores. This setting 
applies to supported C++ kernels in \code{NormalizeData},
\code{FindVariableFeatures}, \code{ScaleData}, \code{SCTransform},
\code{RunPCA}, \code{FindNeighbors}, \code{FindClusters}, and 
CCA/RPCA integration. \code{RunUMAP} also passes this setting to the 
\pkg{uwot} backend. Support depends on the input type, backend, and arguments; 
not every stage of these functions can be run in parallel.

Note that this thread setting is separate from \code{\link[future]{plan}}.
}
\seealso{
\code{\link{getThreads}}
}
\concept{utilities}
