% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setThreads}
\alias{setThreads}
\title{Set the number of threads to use for parallel processing in Seurat}
\usage{
setThreads(threads)
}
\arguments{
\item{threads}{Number of threads (>= 1) to use for processing}
}
\description{
Set the number of threads to use for parallel processing in Seurat
}
\concept{utilities}
