% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setThreads}
\alias{setThreads}
\title{Set the number of threads to use for parallel processing in Seurat}
\usage{
setThreads(n = NULL)
}
\arguments{
\item{n}{Number of threads (>= 1) to use for processing. If NULL, will be set to (number of cores - 1).}
}
\description{
Set the number of threads to use for parallel processing in Seurat
}
\concept{utilities}
