% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{class}
\name{Seurat-class}
\alias{Seurat-class}
\title{The Seurat Class}
\description{
The Seurat object is a representation of single-cell expression data for R;
for more details, please see the documentation in
\code{\link[SeuratObject:Seurat]{SeuratObject}}
}
\seealso{
\code{\link[SeuratObject:Seurat]{SeuratObject::Seurat-class}}
}
