% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{Seurat-deprecated}
\alias{Seurat-deprecated}
\alias{vlnPlot}
\alias{subsetData}
\alias{mean.var.plot}
\alias{pca}
\alias{PCA}
\alias{project.pca}
\alias{print.pca}
\alias{viz.pca}
\alias{set.ident}
\alias{pca.plot}
\alias{pcHeatmap}
\alias{jackStraw}
\alias{jackStrawPlot}
\alias{run_tsne}
\alias{tsne.plot}
\alias{find.markers}
\alias{find_all_markers}
\alias{genePlot}
\alias{feature.plot}
\alias{tnse.plot}
\alias{buildClusterTree}
\alias{plotClusterTree}
\alias{plotNoiseModel}
\alias{add_samples}
\alias{subsetCells}
\alias{project.samples}
\alias{run_diffusion}
\alias{ica}
\alias{ICA}
\alias{cluster.alpha}
\alias{reorder.ident}
\alias{average.pca}
\alias{average.expression}
\alias{icTopGenes}
\alias{pcTopGenes}
\alias{pcTopCells}
\alias{fetch.data}
\alias{viz.ica}
\alias{regulatorScore}
\alias{find.markers.node}
\alias{diffExp.test}
\alias{tobit.test}
\alias{batch.gene}
\alias{marker.test}
\alias{diff.t.test}
\alias{which.cells}
\alias{set.all.ident}
\alias{rename.ident}
\alias{posterior.plot}
\alias{map.cell}
\alias{get.centroids}
\alias{refined.mapping}
\alias{initial.mapping}
\alias{calc.insitu}
\alias{fit.gene.k}
\alias{fit.gene.mix}
\alias{addSmoothedScore}
\alias{addImputedScore}
\alias{getNewScore}
\alias{calcNoiseModels}
\alias{feature.plot.keynote}
\alias{feature.heatmap}
\alias{ica.plot}
\alias{dim.plot}
\alias{spatial.de}
\alias{DBclust_dimension}
\alias{Kclust_dimension}
\alias{pca.sig.genes}
\alias{doHeatMap}
\alias{icHeatmap}
\alias{doKMeans}
\alias{genes.in.cluster}
\alias{kMeansHeatmap}
\alias{cell.cor.matrix}
\alias{gene.cor.matrix}
\alias{calinskiPlot}
\alias{dot.plot}
\alias{addMetaData}
\alias{removePC}
\alias{geneScorePlot}
\alias{cellPlot}
\alias{jackStraw.permutation.test}
\alias{jackStrawMC}
\alias{jackStrawFull}
\alias{writ.table}
\alias{jackRandom}
\alias{MeanVarPlot}
\alias{HeatmapNode}
\alias{minusr}
\alias{minusc}
\alias{RegressOut}
\title{Deprecated function(s) in the Seurat package}
\usage{
vlnPlot(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
These functions are provided for compatibility with older version of the Seurat package.  They may eventually be completely removed.
}
\section{Details}{

\tabular{rl}{
  \code{vlnPlot} \tab now a synonym for \code{\link{VlnPlot}}\cr
  \code{subsetData} \tab now a synonym for \code{\link{SubsetData}}\cr
  \code{mean.var.plot} \tab now a synonym for \code{\link{MeanVarPlot}}\cr
  \code{pca} \tab now a synonym for \code{\link{RunPCA}}\cr
  \code{PCA} \tab now a synonym for \code{\link{PCA}}\cr
  \code{project.pca} \tab now a synonym for \code{\link{ProjectPCA}}\cr
  \code{print.pca} \tab now a synonym for \code{\link{PrintPCA}}\cr
  \code{viz.pca} \tab now a synonym for \code{\link{VizPCA}}\cr
  \code{set.ident} \tab now a synonym for \code{\link{SetIdent}}\cr
  \code{pca.plot} \tab now a synonym for \code{\link{PCAPlot}}\cr
  \code{pcHeatmap} \tab now a synonym for \code{\link{PCHeatmap}}\cr
  \code{jackStraw} \tab now a synonym for \code{\link{JackStraw}}\cr
  \code{jackStrawPlot} \tab now a synonym for \code{\link{JackStrawPlot}}\cr
  \code{run_tsne} \tab now a synonym for \code{\link{RunTSNE}}\cr
  \code{tsne.plot} \tab now a synonym for \code{\link{TSNEPlot}}\cr
  \code{find.markers} \tab now a synonym for \code{\link{FindMarkers}}\cr
  \code{find_all_markers} \tab now a synonym for \code{\link{FindAllMarkers}}\cr
  \code{genePlot} \tab now a synonym for \code{\link{GenePlot}}\cr
  \code{feature.plot} \tab now a synonym for \code{\link{FeaturePlot}}\cr
  \code{buildClusterTree} \tab now a synonym for \code{\link{BuildClusterTree}}\cr
  \code{plotClusterTree} \tab now a synonym for \code{\link{PlotClusterTree}}\cr
  \code{plotNoiseModel} \tab now a synonym for \code{\link{PlotNoiseModel}}\cr
  \code{add_samples} \tab now a synonym for \code{\link{AddSamples}}\cr
  \code{subsetCells} \tab now deleted\cr
  \code{project.samples} \tab now a synonym for \code{\link{ProjectSamples}}\cr
  \code{run_diffusion} \tab now a synonym for \code{\link{RunDiffusion}}\cr
  \code{ica} \tab now a synonym for \code{\link{RunICA}}\cr
  \code{ICA} \tab now a synonym for \code{\link{RunICA}}\cr
  \code{cluster.alpha} \tab now a synonym for \code{\link{AverageDetectionRate}}\cr
  \code{reorder.ident} \tab now a synonym for \code{\link{ReorderIdent}}\cr
  \code{average.pca} \tab now a synonym for \code{\link{AveragePCA}}\cr
  \code{average.expression} \tab now a synonym for \code{\link{AverageExpression}}\cr
  \code{icTopGenes} \tab now a synonym for \code{\link{ICTopGenes}}\cr
  \code{pcTopGenes} \tab now a synonym for \code{\link{PCTopGenes}}\cr
  \code{pcTopCells} \tab now a synonym for \code{\link{PCTopCells}}\cr
  \code{fetch.data} \tab now a synonym for \code{\link{FetchData}}\cr
  \code{viz.ica} \tab now a synonym for \code{\link{VizIca}}\cr
  \code{regulatorScore} \tab now deleted\cr
  \code{find.markers.node} \tab now a synonym for \code{\link{FindMarkersNode}}\cr
  \code{diffExp.test} \tab now a synonym for \code{\link{DiffExpTest}}\cr
  \code{tobit.test} \tab now a synonym for \code{\link{TobitTest}}\cr
  \code{batch.gene} \tab now a synonym for \code{\link{BatchGene}}\cr
  \code{marker.test} \tab now a synonym for \code{\link{MarkerTest}}\cr
  \code{diff.t.test} \tab now a synonym for \code{\link{DiffTTest}}\cr
  \code{which.cells} \tab now a synonym for \code{\link{WhichCells}}\cr
  \code{set.all.ident} \tab now a synonym for \code{\link{SetAllIdent}}\cr
  \code{rename.ident} \tab now a synonym for \code{\link{RenameIdent}}\cr
  \code{posterior.plot} \tab now a synonym for \code{\link{PosteriorPlot}}\cr
  \code{map.cell} \tab now a synonym for \code{\link{MapCell}}\cr
  \code{get.centroids} \tab now a synonym for \code{\link{GetCentroids}}\cr
  \code{refined.mapping} \tab now a synonym for \code{\link{RefinedMapping}}\cr
  \code{initial.mapping} \tab now a synonym for \code{\link{InitialMapping}}\cr
  \code{calc.insitu} \tab now a synonym for \code{\link{CalcInsitu}}\cr
  \code{fit.gene.k} \tab now a synonym for \code{\link{FitGeneK}}\cr
  \code{fit.gene.mix} \tab now a synonym for \code{\link{FitGeneMix}}\cr
  \code{addSmoothedScore} \tab now a synonym for \code{\link{AddSmoothedScore}}\cr
  \code{addImputedScore} \tab now a synonym for \code{\link{AddImputedScore}}\cr
  \code{getNewScore} \tab now a synonym for \code{\link{GetNewScore}}\cr
  \code{calcNoiseModels} \tab now a synonym for \code{\link{CalcNoiseModels}}\cr
  \code{feature.plot.keynote} \tab now a synonym for \code{\link{FeaturePlotKeynote}}\cr
  \code{feature.heatmap} \tab now a synonym for \code{\link{FeatureHeatmap}}\cr
  \code{ica.plot} \tab now a synonym for \code{\link{ICAPlot}}\cr
  \code{dim.plot} \tab now a synonym for \code{\link{DimPlot}}\cr
  \code{spatial.de} \tab now a synonym for \code{\link{SpatialDe}}\cr
  \code{DBclust_dimension} \tab now a synonym for \code{\link{DBClustDimension}}\cr
  \code{Kclust_dimension} \tab now a synonym for \code{\link{KClustDimension}}\cr
  \code{pca.sig.genes} \tab now a synonym for \code{\link{PCASigGenes}}\cr
  \code{doHeatMap} \tab now a synonym for \code{\link{DoHeatMap}}\cr
  \code{icHeatmap} \tab now a synonym for \code{\link{ICHeatmap}}\cr
  \code{doKMeans} \tab now a synonym for \code{\link{DoKMeans}}\cr
  \code{genes.in.cluster} \tab now a synonym for \code{\link{GenesInCluster}}\cr
  \code{kMeansHeatmap} \tab now a synonym for \code{\link{KMeansHeatmap}}\cr
  \code{cell.cor.matrix} \tab now a synonym for \code{\link{CellCorMatrix}}\cr
  \code{gene.cor.matrix} \tab now a synonym for \code{\link{GeneCorMatrix}}\cr
  \code{calinskiPlot} \tab now a synonym for \code{\link{CalinskiPlot}}\cr
  \code{dot.plot} \tab now a synonym for \code{\link{DotPlot}}\cr
  \code{addMetaData} \tab now a synonym for \code{\link{AddMetaData}}\cr
  \code{removePC} \tab now a synonym for \code{\link{RemovePC}}\cr
  \code{geneScorePlot} \tab now deleted\cr
  \code{cellPlot} \tab now a synonym for \code{\link{CellPlot}}\cr
  \code{jackStraw.permutation.test} \tab now a synonym for \code{\link{JackStrawPermutationTest}}\cr
  \code{jackStrawMC} \tab now a synonym for \code{\link{JackStrawMC}}\cr
  \code{jackStrawFull} \tab now a synonym for \code{\link{JackStrawFull}}\cr
  \code{PCAFast} \tab now a synonym for \code{\link{PCA}}\cr
  \code{writ.table} \tab is delteded without replacement\cr
  \code{jackRandom} \tab now a synonym for \code{\link{JackRandom}}\cr
  \code{MeanVarPlot} \tab now a synonym for \code{\link{FindVariableGenes}}\cr
  \code{myPalette} \tab now a synonym for \code{\link{CustomPalette}}\cr
  \code{minusr} \tab now a synonym for \code{\link{SubsetRow}}\cr
  \code{minusc} \tab now a synonym for \code{\link{SubsetColumn}}\cr
  \code{RegressOut} \tab now part of \code{\link{ScaleData}}\cr
}
}

