% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{Seurat-deprecated}
\alias{Seurat-deprecated}
\alias{vlnPlot}
\alias{subsetData}
\alias{pca}
\alias{PCA}
\alias{project.pca}
\alias{viz.pca}
\alias{set.ident}
\alias{pca.plot}
\alias{pcHeatmap}
\alias{jackStraw}
\alias{jackStrawPlot}
\alias{run_tsne}
\alias{tsne.plot}
\alias{find.markers}
\alias{find_all_markers}
\alias{genePlot}
\alias{feature.plot}
\alias{tnse.plot}
\alias{buildClusterTree}
\alias{plotClusterTree}
\alias{plotNoiseModel}
\alias{add_samples}
\alias{subsetCells}
\alias{project.samples}
\alias{run_diffusion}
\alias{ica}
\alias{ICA}
\alias{cluster.alpha}
\alias{average.pca}
\alias{average.expression}
\alias{icTopGenes}
\alias{pcTopGenes}
\alias{pcTopCells}
\alias{fetch.data}
\alias{viz.ica}
\alias{regulatorScore}
\alias{find.markers.node}
\alias{diffExp.test}
\alias{tobit.test}
\alias{batch.gene}
\alias{marker.test}
\alias{which.cells}
\alias{set.all.ident}
\alias{rename.ident}
\alias{posterior.plot}
\alias{map.cell}
\alias{get.centroids}
\alias{refined.mapping}
\alias{initial.mapping}
\alias{calc.insitu}
\alias{fit.gene.k}
\alias{fit.gene.mix}
\alias{addSmoothedScore}
\alias{addImputedScore}
\alias{getNewScore}
\alias{calcNoiseModels}
\alias{feature.plot.keynote}
\alias{feature.heatmap}
\alias{ica.plot}
\alias{spatial.de}
\alias{DBclust_dimension}
\alias{Kclust_dimension}
\alias{pca.sig.genes}
\alias{doHeatMap}
\alias{icHeatmap}
\alias{doKMeans}
\alias{genes.in.cluster}
\alias{kMeansHeatmap}
\alias{cell.cor.matrix}
\alias{gene.cor.matrix}
\alias{calinskiPlot}
\alias{dot.plot}
\alias{addMetaData}
\alias{removePC}
\alias{geneScorePlot}
\alias{cellPlot}
\alias{jackStraw.permutation.test}
\alias{jackStrawMC}
\alias{jackStrawFull}
\alias{writ.table}
\alias{jackRandom}
\alias{MeanVarPlot}
\alias{HeatmapNode}
\alias{minusr}
\alias{minusc}
\alias{RegressOut}
\alias{VizClassification}
\title{Deprecated function(s) in the Seurat package}
\usage{
vlnPlot(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
These functions are provided for compatibility with older version of the Seurat package.  They may eventually be completely removed.
}
\section{Details}{

\tabular{rl}{
  \code{vlnPlot} \tab now a synonym for \code{VlnPlot}\cr
  \code{subsetData} \tab now a synonym for \code{SubsetData}\cr
  \code{pca} \tab now a synonym for \code{RunPCA}\cr
  \code{PCA} \tab now a synonym for \code{PCA}\cr
  \code{project.pca} \tab now a synonym for \code{ProjectPCA}\cr
  \code{viz.pca} \tab now a synonym for \code{VizPCA}\cr
  \code{set.ident} \tab now a synonym for \code{SetIdent}\cr
  \code{pca.plot} \tab now a synonym for \code{PCAPlot}\cr
  \code{pcHeatmap} \tab now a synonym for \code{PCHeatmap}\cr
  \code{jackStraw} \tab now a synonym for \code{JackStraw}\cr
  \code{jackStrawPlot} \tab now a synonym for \code{JackStrawPlot}\cr
  \code{run_tsne} \tab now a synonym for \code{RunTSNE}\cr
  \code{tsne.plot} \tab now a synonym for \code{TSNEPlot}\cr
  \code{find.markers} \tab now a synonym for \code{FindMarkers}\cr
  \code{find_all_markers} \tab now a synonym for \code{FindAllMarkers}\cr
  \code{genePlot} \tab now a synonym for \code{GenePlot}\cr
  \code{feature.plot} \tab now a synonym for \code{FeaturePlot}\cr
  \code{buildClusterTree} \tab now a synonym for \code{BuildClusterTree}\cr
  \code{plotClusterTree} \tab now a synonym for \code{PlotClusterTree}\cr
  \code{plotNoiseModel} \tab has been removed and may be replaced at a later date\cr
  \code{add_samples} \tab now a synonym for \code{AddSamples}\cr
  \code{subsetCells} \tab now deleted\cr
  \code{project.samples} \tab has been removed and may be replaced at a later date\cr
  \code{run_diffusion} \tab now a synonym for \code{RunDiffusion}\cr
  \code{ica} \tab now a synonym for \code{RunICA}\cr
  \code{ICA} \tab now a synonym for \code{RunICA}\cr
  \code{cluster.alpha} \tab now a synonym for \code{AverageDetectionRate}\cr
  \code{average.pca} \tab now a synonym for \code{AveragePCA}\cr
  \code{average.expression} \tab now a synonym for \code{AverageExpression}\cr
  \code{icTopGenes} \tab now a synonym for \code{ICTopGenes}\cr
  \code{pcTopGenes} \tab now a synonym for \code{PCTopGenes}\cr
  \code{pcTopCells} \tab now a synonym for \code{PCTopCells}\cr
  \code{fetch.data} \tab now a synonym for \code{FetchData}\cr
  \code{viz.ica} \tab now a synonym for \code{VizIca}\cr
  \code{regulatorScore} \tab now deleted\cr
  \code{find.markers.node} \tab now a synonym for \code{FindMarkersNode}\cr
  \code{diffExp.test} \tab now a synonym for \code{DiffExpTest}\cr
  \code{tobit.test} \tab now a synonym for \code{TobitTest}\cr
  \code{batch.gene} \tab has been removed and may be restored at a later date\cr
  \code{marker.test} \tab now a synonym for \code{MarkerTest}\cr
  \code{which.cells} \tab now a synonym for \code{WhichCells}\cr
  \code{set.all.ident} \tab now a synonym for \code{SetAllIdent}\cr
  \code{rename.ident} \tab now a synonym for \code{RenameIdent}\cr
  \code{posterior.plot} \tab now a synonym for \code{PosteriorPlot}\cr
  \code{map.cell} \tab has been deprecated\cr
  \code{get.centroids} \tab now a synonym for \code{GetCentroids}\cr
  \code{refined.mapping} \tab now a synonym for \code{RefinedMapping}\cr
  \code{initial.mapping} \tab now a synonym for \code{InitialMapping}\cr
  \code{calc.insitu} \tab now a synonym for \code{CalcInsitu}\cr
  \code{fit.gene.k} \tab now a synonym for \code{FitGeneK}\cr
  \code{fit.gene.mix} \tab now a synonym for \code{FitGeneMix}\cr
  \code{addSmoothedScore} \tab now a synonym for \code{AddSmoothedScore}\cr
  \code{addImputedScore} \tab now a synonym for \code{AddImputedScore}\cr
  \code{getNewScore} \tab has been removed without replacement\cr
  \code{calcNoiseModels} \tab has been removed and may be replaced at a later date\cr
  \code{feature.plot.keynote} \tab has been removed without replacement\cr
  \code{feature.heatmap} \tab now a synonym for \code{FeatureHeatmap}\cr
  \code{ica.plot} \tab now a synonym for \code{ICAPlot}\cr
  \code{spatial.de} \tab has been removed and may be replaced at a later date\cr
  \code{DBclust_dimension} \tab now a synonym for \code{DBClustDimension}\cr
  \code{Kclust_dimension} \tab now a synonym for \code{KClustDimension}\cr
  \code{pca.sig.genes} \tab now a synonym for \code{PCASigGenes}\cr
  \code{doHeatMap} \tab now a synonym for \code{DoHeatMap}\cr
  \code{icHeatmap} \tab now a synonym for \code{ICHeatmap}\cr
  \code{doKMeans} \tab now a synonym for \code{DoKMeans}\cr
  \code{genes.in.cluster} \tab now a synonym for \code{GenesInCluster}\cr
  \code{kMeansHeatmap} \tab now a synonym for \code{KMeansHeatmap}\cr
  \code{cell.cor.matrix} \tab has been removed and may be replaced at a later date\cr
  \code{gene.cor.matrix} \tab has been removed and may be replaced at a later date\cr
  \code{calinskiPlot} \tab has been removed and may be replaced at a later date\cr
  \code{dot.plot} \tab now a synonym for \code{DotPlot}\cr
  \code{addMetaData} \tab now a synonym for \code{AddMetaData}\cr
  \code{removePC} \tab has been removed and may be replaced at a later date\cr
  \code{geneScorePlot} \tab now deleted\cr
  \code{cellPlot} \tab now a synonym for \code{CellPlot}\cr
  \code{jackStraw.permutation.test} \tab has been deleted\cr
  \code{jackStrawMC} \tab has been deleted\cr
  \code{jackStrawFull} \tab has been deleted\cr
  \code{PCAFast} \tab now a synonym for \code{PCA}\cr
  \code{writ.table} \tab has been removed without replacement\cr
  \code{jackRandom} \tab has been removed without replacement\cr
  \code{MeanVarPlot} \tab now a synonym for \code{FindVariableGenes}\cr
  \code{myPalette} \tab now a synonym for \code{CustomPalette}\cr
  \code{minusr} \tab now a synonym for \code{SubsetRow}\cr
  \code{minusc} \tab now a synonym for \code{SubsetColumn}\cr
  \code{RegressOut} \tab now part of \code{ScaleData}\cr
  \code{VizClassification} \tab has been removed without replacement\cr
}
}

