% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{Seurat-package}
\alias{Seurat-package}
\title{Tools for single-cell genomics}
\description{
Tools for single-cell genomics
}
\section{Package options}{


Seurat uses the following [options()] to configure behaviour:

\describe{
  \item{\code{Seurat.memsafe}}{global option to call gc() after many operations.
  This can be helpful in cleaning up the memory status of the R session and
  prevent use of swap space. However, it does add to the computational overhead
  and setting to FALSE can speed things up if you're working in an environment
  where RAM availabiliy is not a concern.}
  \item{\code{Seurat.warn.umap.uwot}}{Show warning about the default backend
  for \code{\link{RunUMAP}} changing from Python UMAP via reticulate to UWOT}
  \item{\code{Seurat.checkdots}}{For functions that have ... as a parameter,
  this controls the behavior when an item isn't used. Can be one of warn,
  stop, or silent.}
}
}

