% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{Seurat}
\alias{Seurat}
\title{The Seurat Class}
\description{
The Seurat object is ...
}
\section{Slots}{

\describe{
\item{\code{assays}}{A list of assays for this project}

\item{\code{meta.data}}{Contains meta-information about each cell, starting with number of genes detected (nGene)
and the original identity class (orig.ident); more information is added using \code{AddMetaData}}

\item{\code{active.ident}}{The active cluster identity for this Seurat object}

\item{\code{graphs}}{...}

\item{\code{neighbors}}{...}

\item{\code{dim.reduc}}{A list of dimmensional reduction objects for this Assay}

\item{\code{project.name}}{...}

\item{\code{calc.params}}{A list of calculation parameters performed on this Seurat object}

\item{\code{misc}}{A list of miscellaneous information}

\item{\code{version}}{Version of Seurat this object was built under}
}}

