% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{class}
\name{SeuratCommand-class}
\alias{SeuratCommand-class}
\title{The SeuratCommand Class}
\description{
For more details, please see the documentation in
\code{\link[SeuratObject:SeuratCommand]{SeuratObject}}
}
\seealso{
\code{\link[SeuratObject:SeuratCommand]{SeuratObject::SeuratCommand-class}}
}
