% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_themes.R
\name{SeuratTheme}
\alias{SeuratTheme}
\alias{DarkTheme}
\alias{NoAxes}
\alias{NoLegend}
\alias{NoGrid}
\alias{SeuratAxes}
\alias{BarePlot}
\title{Seurat Themes}
\usage{
SeuratTheme()

DarkTheme(...)

NoAxes(...)

NoLegend(...)

NoGrid(...)

SeuratAxes(...)

BarePlot()
}
\arguments{
\item{...}{Extra parameters to be passed to \code{theme}}
}
\value{
A ggplot2 theme object
}
\description{
Various themes to be applied to ggplot2-based plots
\describe{
  \item{\code{SeuratTheme}}{The curated Seurat theme, consists of ...}
  \item{\code{DarkTheme}}{A dark theme, axes and text turn to white, the background becomes black}
  \item{\code{NoAxes}}{Removes axis lines, text, and ticks}
  \item{\code{NoLegend}}{Removes the legend}
  \item{\code{NoGrid}}{Removes grid lines}
  \item{\code{SeuratAxes}}{Set Seurat-style axes}
  \item{\code{BarePlot}}{Remove all extraneous features from a}
}
}
\examples{
# Generate a plot with a dark theme
df <- data.frame(x = rnorm(n = 100, mean = 20, sd = 2), y = rbinom(n = 100, size = 100, prob = 0.2))
p <- ggplot(data = df, mapping = aes(x = x, y = y)) + geom_point(mapping = aes(color = 'red'))
p + DarkTheme(legend.position = 'none')

# Generate a plot with no axes
df <- data.frame(x = rnorm(n = 100, mean = 20, sd = 2), y = rbinom(n = 100, size = 100, prob = 0.2))
p <- ggplot(data = df, mapping = aes(x = x, y = y)) + geom_point(mapping = aes(color = 'red'))
p + NoAxes()

# Generate a plot with no legend
df <- data.frame(x = rnorm(n = 100, mean = 20, sd = 2), y = rbinom(n = 100, size = 100, prob = 0.2))
p <- ggplot(data = df, mapping = aes(x = x, y = y)) + geom_point(mapping = aes(color = 'red'))
p + NoLegend()

# Generate a plot with no grid lines
df <- data.frame(x = rnorm(n = 100, mean = 20, sd = 2), y = rbinom(n = 100, size = 100, prob = 0.2))
p <- ggplot(data = df, mapping = aes(x = x, y = y)) + geom_point(mapping = aes(color = 'red'))
p + NoGrid()

}
\seealso{
\code{\link{ggplot2::theme}}
}
