% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{ShiftDim}
\alias{ShiftDim}
\title{Shift dim scores}
\usage{
ShiftDim(object, grouping.var, reduction.type, dims.shift, ds.amt = 0.1)
}
\arguments{
\item{object}{Seurat object}

\item{reduction.type}{reduction to shift scores for}

\item{dims.shift}{Dims to shift, default is all}

\item{ds.amt}{percent of cells in each group for downsampling}

\item{group.var}{Name of the grouping variable for which to shift the scores}
}
\value{
Returns Seurat object with the dims shifted, stored in object@dr$reduction.type.shifted
}
\description{
Shifts dim scores so that they line up across grouping variable (only implemented for case with
2 categories in grouping.var)
}
