% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Shuffle}
\alias{Shuffle}
\title{Shuffle a vector}
\usage{
Shuffle(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector with the same values of x, just in random order
}
\description{
Shuffle a vector
}
\examples{
v <- seq(10)
v2 <- Shuffle(x = v)
v2

}
