% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\docType{class}
\name{SlideSeq-class}
\alias{SlideSeq-class}
\alias{SlideSeq}
\title{The SlideSeq class}
\description{
The SlideSeq class represents spatial information from the Slide-seq platform
}
\section{Slots}{

\describe{
\item{\code{coordinates}}{...}

\item{\code{...}}{}
}}

