% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{SlideSeq-class}
\alias{SlideSeq-class}
\alias{SlideSeq}
\title{The SlideSeq class}
\description{
The SlideSeq class represents spatial information from the Slide-seq platform
}
\section{Slots}{

\describe{
\item{\code{coordinates}}{...}
}}

\section{Slots}{


\describe{
\item{\code{assay}}{Name of assay to associate image data with; will give this image
priority for visualization when the assay is set as the active/default assay
in a \code{Seurat} object}

\item{\code{key}}{A one-length character vector with the object's key; keys must
be one or more alphanumeric characters followed by an underscore
\dQuote{\code{_}} (regex pattern
\dQuote{\code{^[a-zA-Z][a-zA-Z0-9]*_$}})}
}
}

\concept{spatial}
