% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{class}
\name{SpatialImage-class}
\alias{SpatialImage-class}
\title{The SpatialImage Class}
\description{
For more details, please see the documentation in
\code{\link[SeuratObject:SpatialImage]{SeuratObject}}
}
\seealso{
\code{\link[SeuratObject:SpatialImage]{SeuratObject::SpatialImage-class}}
}
