% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{STARmap-class}
\alias{STARmap-class}
\alias{STARmap}
\title{The STARmap class}
\description{
The STARmap class
}
\section{Slots}{


\describe{
\item{\code{assay}}{Name of assay to associate image data with; will give this image
priority for visualization when the assay is set as the active/default assay
in a \code{Seurat} object}

\item{\code{key}}{Key for the image}
}
}

\concept{objects}
\concept{spatial}
