% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{STARmap-class}
\alias{STARmap-class}
\alias{STARmap}
\title{The STARmap class}
\description{
The STARmap class
}
\section{Slots}{


\describe{
\item{\code{assay}}{Name of assay to associate image data with; will give this image
priority for visualization when the assay is set as the active/default assay
in a \code{Seurat} object}

\item{\code{key}}{A one-length character vector with the object's key; keys must
be one or more alphanumeric characters followed by an underscore
\dQuote{\code{_}} (regex pattern
\dQuote{\code{^[a-zA-Z][a-zA-Z0-9]*_$}})}
}
}

\concept{objects}
\concept{spatial}
