% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Stdev}
\alias{Stdev}
\alias{Stdev.DimReduc}
\alias{Stdev.Seurat}
\title{Get the standard deviations for an object}
\usage{
Stdev(object, ...)

\method{Stdev}{DimReduc}(object, ...)

\method{Stdev}{Seurat}(object, reduction, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{reduction}{Name of reduction to use}
}
\description{
Get the standard deviations for an object
}
