% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFxns.R
\name{subc}
\alias{subc}
\title{Return a subset of columns for a matrix or data frame}
\usage{
subc(data, code)
}
\arguments{
\item{data}{Matrix or data frame with column names}

\item{code}{Pattern for matching within column names}
}
\value{
Returns a subset of data, using only column names that yield a match to the pattern
}
\description{
Return a subset of columns for a matrix or data frame
}
