% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{FilterCells}
\alias{FilterCells}
\alias{FilterCells.Seurat}
\alias{[.Seurat}
\alias{subset.Seurat}
\alias{subset}
\title{Subset a Seurat object}
\usage{
FilterCells(object, ...)

\method{FilterCells}{Seurat}(object, vars, low = -Inf, high = -Inf,
  ...)

\method{[}{Seurat}(x, i, j, ...)

\method{subset}{Seurat}(x, subset, cells = NULL, features = NULL,
  idents = NULL, ...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{WhichCells}}}

\item{vars}{Name of variable names to filter on}

\item{low, high}{A vector of low and high threshold values; will automatically expand to the length of \code{vars}}

\item{x, object}{Seurat object to be subsetted}

\item{i, features}{A vector of features to keep}

\item{j, cells}{A vector of cells to keep}

\item{subset}{Logical expression indicating features/variables to keep}

\item{idents}{A vector of identity classes to keep}
}
\value{
A subsetted Seurat object
}
\description{
Subset a Seurat object
}
\examples{
pbmc_small[VariableFeatures(object = pbmc_small), ]
pbmc_small[, 1:10]

subset(x = pbmc_small, subset = MS4A1 > 4)
subset(x = pbmc_small, features = VariableFeatures(object = pbmc_small))

}
\seealso{
\code{\link[base]{subset}} \code{\link{WhichCells}}
}
