% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{subset.Seurat}
\alias{subset.Seurat}
\alias{subset}
\title{Subset a Seurat object}
\usage{
\method{subset}{Seurat}(x, subset, select = NULL, ...)
}
\arguments{
\item{x}{Seurat object to be subsetted}

\item{subset}{Logical expression indicating features/variables to keep}

\item{select}{A vector of cells, identity classes, or features to keep}

\item{...}{Arguments passed to \code{WhichCells}}
}
\value{
A subsetted Seurat object
}
\description{
Subset a Seurat object
}
\examples{
subset(x = pbmc_small, subset = MS4A1 > 7)
subset(x = pbmc_small, select = VariableFeatures(object = pbmc_small))

}
\seealso{
\code{\link{base::subset}} \code{\link{WhichCells}}
}
