% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{subset.Seurat}
\alias{subset.Seurat}
\alias{subset}
\title{Subset a Seurat object}
\usage{
\method{subset}{Seurat}(x, subset, ...)
}
\arguments{
\item{x}{Seurat object to be subsetted}

\item{subset}{Logical expression indicating features/variables to keep}

\item{...}{Arguments passed to other methods}
}
\value{
A subsetted Seurat object
}
\description{
Subset a Seurat object
}
\examples{
subset(x = pbmc, subset = MS4A1 > 7)

}
\seealso{
\code{\link{base::subset}}
}
