% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R, R/interaction_generics.R
\name{SubsetData.Assay}
\alias{SubsetData.Assay}
\alias{SubsetData.Seurat}
\alias{SubsetData}
\title{Return a subset of the Seurat object}
\usage{
\method{SubsetData}{Assay}(object, cells.use = NULL, subset.name = NULL,
  low.threshold = -Inf, high.threshold = Inf, accept.value = NULL,
  do.center = FALSE, do.scale = FALSE, do.clean = FALSE, ...)

\method{SubsetData}{Seurat}(object, assay.use = NULL, cells.use = NULL,
  subset.name = NULL, ident.use = NULL, ident.remove = NULL,
  low.threshold = -Inf, high.threshold = Inf, accept.value = NULL,
  do.center = FALSE, do.scale = FALSE, max.cells.per.ident = Inf,
  random.seed = 1, do.clean = FALSE, ...)

SubsetData(object, slot, ...)
}
\arguments{
\item{object}{Seurat object}

\item{cells.use}{A vector of cell names to use as a subset. If NULL
(default), then this list will be computed based on the next three
arguments. Otherwise, will return an object consissting only of these cells}

\item{subset.name}{Parameter to subset on. Eg, the name of a gene, PC1, a
column name in object@meta.data, etc. Any argument that can be retreived
using FetchData}

\item{low.threshold}{Low cutoff for the parameter (default is -Inf)}

\item{high.threshold}{High cutoff for the parameter (default is Inf)}

\item{accept.value}{Returns cells with the subset name equal to this value}

\item{do.center}{Recenter the new object@scale.data}

\item{do.scale}{Rescale the new object@scale.data. FALSE by default}

\item{do.clean}{Only keep object@raw.data and object@data. Cleans out most
other slots. Can be useful if you want to start a fresh analysis on just a
subset of the data. Also clears out stored clustering results in
object@meta.data (any columns containing "res"). Will by default subset the
raw.data slot.}

\item{\dots}{Additional arguments to be passed to FetchData (for example,
use.imputed=TRUE)}

\item{assay.use}{Assay to subset on}

\item{ident.use}{Create a cell subset based on the provided identity classes}

\item{ident.remove}{Subtract out cells from these identity classes (used for
filtration)}

\item{max.cells.per.ident}{Can be used to downsample the data to a certain
max per cell ident. Default is INF.}

\item{random.seed}{Random seed for downsampling}

\item{subset.raw}{Also subset object@raw.data}
}
\value{
Returns a Seurat object containing only the relevant subset of cells
}
\description{
Creates a Seurat object containing only a subset of the cells in the
original object. Takes either a list of cells to use as a subset, or a
parameter (for example, a gene), to subset on.
}
\section{Methods (by class)}{
\itemize{
\item \code{Assay}: Subset an Assay object

\item \code{Seurat}: Subset an Seurat object
}}

\examples{
pbmc1 <- SubsetData(object = pbmc_small, cells.use = pbmc_small@cell.names[1:40])
pbmc1

}
