% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/spatial.R
\name{SVFInfo}
\alias{SVFInfo}
\alias{SVFInfo.Assay}
\alias{SVFInfo.Seurat}
\title{Get spatially variable feature information}
\usage{
SVFInfo(object, ...)

\method{SVFInfo}{Assay}(
  object,
  selection.method = c("markvariogram", "moransi"),
  status = FALSE,
  ...
)

\method{SVFInfo}{Seurat}(
  object,
  selection.method = c("markvariogram", "moransi"),
  assay = NULL,
  status = FALSE,
  ...
)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{selection.method}{Which method to pull. Options: markvariogram, moransi}

\item{status}{Add variable status to the resulting data.frame}

\item{assay}{Name of assay to pull highly variable feature information for}
}
\description{
Get spatially variable feature information
}
