% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketching.R
\name{TransferSketchLabels}
\alias{TransferSketchLabels}
\title{Transfer data from sketch data to full data}
\usage{
TransferSketchLabels(
  object,
  atoms = "sketch",
  reduction,
  dims,
  refdata = NULL,
  k = 50,
  reduction.model = NULL,
  neighbors = NULL,
  recompute.neighbors = FALSE,
  recompute.weights = FALSE,
  verbose = TRUE
)
}
\description{
Transfer data from sketch data to full data
}
