% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{UnSketchEmbeddings}
\alias{UnSketchEmbeddings}
\title{Transfer embeddings from sketched cells to the full data}
\usage{
UnSketchEmbeddings(
  atom.data,
  atom.cells = NULL,
  orig.data,
  embeddings,
  sketch.matrix = NULL
)
}
\description{
Transfer embeddings from sketched cells to the full data
}
