% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seurat.R
\name{validate.clusters}
\alias{validate.clusters}
\title{Cluster Validation}
\usage{
validate.clusters(object, pc.use = NULL, top.genes = 30,
  min_connectivity = 0.01, acc.cutoff = 0.9, verbose = TRUE)
}
\arguments{
\item{object}{Seurat object}

\item{pc.use}{Which PCs to use for model construction}

\item{top.genes}{Use the top X genes for model construction}

\item{min_connectivity}{Threshold of connectedness for comparison of two clusters}

\item{acc.cutoff}{Accuracy cutoff for classifier}

\item{verbose}{Controls whether to display progress and merge results}
}
\value{
Returns a Seurat object, object@ident has been updated with new cluster info
}
\description{
Methods for validating the legitimacy of clusters using
classification. SVMs are used as the basis for the classification.
Merging is done based on the connectivity from an SNN graph.
}

