% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{VariableFeaturePlot}
\alias{VariableFeaturePlot}
\title{View variable features}
\usage{
VariableFeaturePlot(object, pt.size = 1, cols.use = c("black", "red"),
  assay.use = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{pt.size}{Size of the points on the plot}

\item{cols.use}{Colors to specify non-variable/variable status}

\item{assay.use}{Assay to pull variable features from}
}
\description{
View variable features
}
\examples{
VariableFeaturePlot(object = pbmc_small)

}
