% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{VariableFeaturePlot}
\alias{VariableFeaturePlot}
\alias{VariableGenePlot}
\title{View variable features}
\usage{
VariableFeaturePlot(object, cols = c("black", "red"), pt.size = 1,
  assay = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{cols}{Colors to specify non-variable/variable status}

\item{pt.size}{Size of the points on the plot}

\item{assay}{Assay to pull variable features from}
}
\description{
View variable features
}
\examples{
VariableFeaturePlot(object = pbmc_small)

}
