% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{VariableFeatures}
\alias{VariableFeatures}
\alias{VariableFeatures<-}
\alias{VariableFeatures.Assay}
\alias{VariableFeatures.Seurat}
\alias{VariableFeatures<-.Assay}
\alias{VariableFeatures<-.Seurat}
\title{Get and set variable feature information}
\usage{
VariableFeatures(object, ...)

VariableFeatures(object, ...) <- value

\method{VariableFeatures}{Assay}(object, selection.method = NULL, ...)

\method{VariableFeatures}{Seurat}(object, assay = NULL, selection.method = NULL, ...)

\method{VariableFeatures}{Assay}(object, ...) <- value

\method{VariableFeatures}{Seurat}(object, assay = NULL, ...) <- value
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{A character vector of variable features}

\item{selection.method}{Method used to set variable features}

\item{assay}{Name of assay to pull variable features for}
}
\description{
Get and set variable feature information
}
