% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{VlnPlot}
\alias{VlnPlot}
\title{Single cell violin plot}
\usage{
VlnPlot(
  object,
  features,
  cols = NULL,
  pt.size = 1,
  idents = NULL,
  sort = FALSE,
  assay = NULL,
  group.by = NULL,
  split.by = NULL,
  adjust = 1,
  y.max = NULL,
  same.y.lims = FALSE,
  log = FALSE,
  ncol = NULL,
  combine = TRUE,
  slot = "data",
  ...
)
}
\arguments{
\item{pt.size}{Point size for geom_violin}

\item{split.by}{A variable to split the violin plots by,
see \code{\link{FetchData}} for more details}

\item{adjust}{Adjust parameter for geom_violin}
}
\value{
A ggplot object
}
\description{
Draws a violin plot of single cell data (gene expression, metrics, PC
scores, etc.)
}
\examples{
VlnPlot(object = pbmc_small, features = 'PC_1')
VlnPlot(object = pbmc_small, features = 'LYZ', split.by = 'groups')

}
\seealso{
\code{\link{FetchData}}
}
