% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{VlnPlot}
\alias{VlnPlot}
\title{Single cell violin plot}
\usage{
VlnPlot(object, features.plot, ident.include = NULL, nCol = NULL,
  do.sort = FALSE, y.max = NULL, same.y.lims = FALSE, adjust.use = 1,
  point.size.use = 1, cols.use = NULL, group.by = NULL, y.log = FALSE,
  combine.plots = TRUE, ...)
}
\arguments{
\item{object}{Seurat object}

\item{features.plot}{Features to plot (gene expression, metrics, PC scores,
anything that can be retreived by FetchData)}

\item{ident.include}{Which classes to include in the plot (default is all)}

\item{nCol}{Number of columns if multiple plots are displayed}

\item{do.sort}{Sort identity classes (on the x-axis) by the average
expression of the attribute being potted}

\item{y.max}{Maximum y axis value}

\item{same.y.lims}{Set all the y-axis limits to the same values}

\item{adjust.use}{Adjust parameter for geom_violin}

\item{point.size.use}{Point size for geom_violin}

\item{cols.use}{Colors to use for plotting}

\item{group.by}{Group (color) cells in different ways (for example, orig.ident)}

\item{y.log}{plot Y axis on log scale}

\item{combine.plots}{Combine plots into a single gg object; note that if TRUE; themeing will not work when plotting multiple features}

\item{...}{additional parameters to pass to FetchData (for example, use.imputed, use.scaled, use.raw)}
}
\value{
A ggplot object
}
\description{
Draws a violin plot of single cell data (gene expression, metrics, PC
scores, etc.)
}
\examples{
VlnPlot(object = pbmc_small, features.plot = 'PC1')

}
