% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{VlnPlot}
\alias{VlnPlot}
\title{Single cell violin plot}
\usage{
VlnPlot(object, features.plot, nCol = NULL, ylab.max = 12, do.ret = TRUE,
  do.sort = FALSE, size.x.use = 16, size.y.use = 16,
  size.title.use = 20, adjust.use = 1, size.use = 1, cols.use = NULL,
  group.by = "ident", y.log = F, ...)
}
\arguments{
\item{object}{Seurat object}

\item{features.plot}{Features to plot (gene expression, metrics, PC scores,
anything that can be retreived by FetchData)}

\item{nCol}{Number of columns if multiple plots are displayed}

\item{ylab.max}{Maximum ylab value}

\item{do.ret}{FALSE by default. If TRUE, returns a list of ggplot objects.}

\item{do.sort}{Sort identity classes (on the x-axis) by the average
expression of the attribute being potted}

\item{size.x.use}{X axis font size}

\item{size.y.use}{Y axis font size}

\item{size.title.use}{Title font size}

\item{adjust.use}{Adjust parameter for geom_violin}

\item{size.use}{Point size for geom_violin}

\item{cols.use}{Colors to use for plotting}

\item{group.by}{Group (color) cells in different ways (for example, orig.ident)}

\item{y.log}{plot T axis on log scale}

\item{\dots}{additional parameters to pass to FetchData (for example, use.imputed, use.scaled, use.raw)}
}
\value{
By default, no return, only graphical output. If do.return=TRUE,
returns a list of ggplot objects.
}
\description{
Draws a violin plot of single cell data (gene expression, metrics, PC
scores, etc.)
}
