% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R, R/interaction_generics.R
\name{WhichCells.Assay}
\alias{WhichCells.Assay}
\alias{WhichCells.Seurat}
\alias{WhichCells}
\title{Identify cells matching certain criteria}
\usage{
\method{WhichCells}{Assay}(object, cells.use, subset.name = NULL,
  low.threshold = -Inf, high.threshold = Inf, accept.value = NULL, ...)

\method{WhichCells}{Seurat}(object, assay.use = NULL, cells.use = NULL,
  ident = NULL, ident.remove = NULL, max.cells.per.ident = Inf,
  random.seed = 1, subset.name = NULL, low.threshold = -Inf,
  high.threshold = Inf, accept.value = NULL, ...)

WhichCells(object, slot, ...)
}
\arguments{
\item{object}{Seurat object}

\item{cells.use}{Subset of cell names}

\item{subset.name}{Parameter to subset on. Eg, the name of a gene, PC1, a
column name in object@meta.data, etc. Any argument that can be retreived
using FetchData}

\item{accept.value}{Returns all cells with the subset name equal to this value}

\item{\dots}{Additional arguments to be passed to FetchData (for example,
use.imputed=TRUE)}

\item{assay.use}{Which assay to filter on}

\item{ident.remove}{Subtract out cells from these identity classes (used for
filtration)}

\item{max.cells.per.ident}{Can be used to downsample the data to a certain
max per cell ident. Default is INF.}

\item{random.seed}{Random seed for downsampling}

\item{ident.use}{Create a cell subset based on the provided identity classes}

\item{accept.low}{Low cutoff for the parameter (default is -Inf)}

\item{accept.high}{High cutoff for the parameter (default is Inf)}

\item{max.cells.per.ident}{Can be used to downsample the data to a certain
max per cell ident. Default is INF.}

\item{random.seed}{Random seed for downsampling}
}
\value{
A vector of cell names
}
\description{
Returns a list of cells that match a particular set of criteria such as
identity class, high/low values for particular PCs, ect..
}
\section{Methods (by class)}{
\itemize{
\item \code{Assay}: Get cells from an Assay object

\item \code{Seurat}: Get cells from a Seurat object
}}

\examples{
WhichCells(object = pbmc_small, ident = 2)

}
