% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfRenderSeurat.R
\name{zfRenderSeurat}
\alias{zfRenderSeurat}
\alias{situ3d}
\alias{zf.cells.render}
\alias{zf.anchor.render}
\alias{zf.insitu.vec.lateral}
\alias{zf.insitu.lateral}
\alias{zf.insitu.dorsal}
\alias{zf.insitu.ventral}
\alias{zf.insitu.side}
\title{Zebrafish analysis functions}
\usage{
situ3d(data, label = NULL, ...)
}
\description{
These functions are for the zebrafish analysis from Satija et al. 2015
}
\references{
Satija R*, Farrell JA*, Gennert D, Schier AF, Regev A.
Spatial reconstruction of single-cell gene expression data.
Nature Biotechnology. 2015
}
